Bryant et al. (2018)

Data from a clustered multiple baseline across participants design conducted by Bryant et al. (2018). The variables are as follows:

Study_ID. Study identifier.

school. School identifier.

group. Group identifier.

case. Student identifier.

treatment. Indicator for treatment phase.

session. Measurement occasion.

session_trt. Measurement occasion times treatment phase.

outcome. Texas Early Mathematics Inventory (TEMI-Aim Check) scores.

session_c. Measurement occasion centered at the follow-up time.

Format

A data frame with 536 rows and 8 variables

Source

Bryant, D. R., Bryant, B. R., Sorelle-Miner, D. A., Falcomata, T. S. & Nozari, M. (2018). Tier 3 intensified intervention for second grade students with severe mathematics difficulties. Archives of Psychology, 2(11), 1-24. doi:10.31296/aop.v2i11.86