Thorne (2005)

Data from an ABAB design conducted by Thorne and Kamps (2008). The variables are as follows:

case. Participant identifier.

measure. Outcome measure description (academic engagement or inappropriate verbalizations).

session. Measurement occasion.

phase_id. Categorical variable describing the phase of the study design for each case.

condition Categorical variable describing whether each phase is a baseline (A) phase or intervention (B) phase.

phase_indicator. Indicator variable equal to 1 during intervention phases.

outcome. Outcome scores

Format

A data frame with 776 rows and 7 variables

Source

Thorne, S., & Kamps, D. (2008). The effects of a group contingency intervention on academic engagement and problem behavior of at-risk students. Behavior Analysis in Practice, 1(2), 12-18.