Clean Single Case Design Data

Clean single case design data for treatment reversal and multiple baseline designs.

preprocess_SCD(
  case,
  phase,
  session,
  outcome,
  design,
  center = 0,
  round_session = TRUE,
  treatment_name = NULL,
  data = NULL
)

Arguments

case
vector of case indicators or name of a character or factor vector within data indicating unique cases.

phase
vector of treatment indicators or name of a character or factor vector within data indicating unique treatment phases.

session
vector of measurement occasions or name of numeric vector within data of measurement times.

outcome
vector of outcome data or name of numeric vector of outcome data within data.

design
Character string to specify whether data comes from a treatment reversal, "TR", or multiple baseline, "MB", design.

center
Numeric value for the centering value for session. Default is 0.

round_session
Logical indicating whether to round session to the nearest integer. Defaults to TRUE.

treatment_name
(Optional) character string corresponding to the name of the treatment phase.

data
(Optional) dataset to be cleaned. Must be a data.frame.

Value

A cleaned SCD dataset that can be used for model fitting and effect size calculation.

Note

If treatment_name is left null it will choose the second level of the phase variable to be the treatment phase.

Examples

data(Laski)
preprocess_SCD(case = case, phase = treatment,
               session = time, outcome = outcome, 
               design = "MB", center = 4, data = Laski)

#>       case treatment time outcome trt time_trt
#> 1   Case 1  baseline   -3   27.60   0        0
#> 2   Case 1  baseline   -2   23.96   0        0
#> 3   Case 1  baseline   -1   23.83   0        0
#> 4   Case 1  baseline    0   47.26   0        0
#> 5   Case 1 treatment    1   52.70   1        0
#> 6   Case 1 treatment    2   60.99   1        1
#> 7   Case 1 treatment    3   66.60   1        2
#> 8   Case 1 treatment    4   52.50   1        3
#> 9   Case 1 treatment    5   85.88   1        4
#> 10  Case 1 treatment    6   47.05   1        5
#> 11  Case 1 treatment    7   66.28   1        6
#> 12  Case 1 treatment    8   54.05   1        7
#> 13  Case 1 treatment    9   51.23   1        8
#> 14  Case 2  baseline   -3   49.67   0        0
#> 15  Case 2  baseline   -2   23.57   0        0
#> 16  Case 2  baseline   -1   26.38   0        0
#> 17  Case 2  baseline    0   28.35   0        0
#> 18  Case 2  baseline    1   45.11   0        0
#> 19  Case 2 treatment    2   70.67   1        0
#> 20  Case 2 treatment    3   79.13   1        1
#> 21  Case 2 treatment    4   84.09   1        2
#> 22  Case 2 treatment    5   88.56   1        3
#> 23  Case 2 treatment    6   80.90   1        4
#> 24  Case 2 treatment    7   91.84   1        5
#> 25  Case 2 treatment    8   63.42   1        6
#> 26  Case 2 treatment    9   70.38   1        7
#> 27  Case 4  baseline   -3    0.00   0        0
#> 28  Case 4  baseline   -2    0.00   0        0
#> 29  Case 4  baseline   -1    0.00   0        0
#> 30  Case 4  baseline    0    0.00   0        0
#> 31  Case 4  baseline    1    0.00   0        0
#> 32  Case 4  baseline    2    0.00   0        0
#> 33  Case 4  baseline    3    0.00   0        0
#> 34  Case 4 treatment    4   28.07   1        0
#> 35  Case 4 treatment    5   28.20   1        1
#> 36  Case 4 treatment    6   57.15   1        2
#> 37  Case 4 treatment    7   72.37   1        3
#> 38  Case 4 treatment    8   45.56   1        4
#> 39  Case 4 treatment    9   52.30   1        5
#> 40  Case 4 treatment   10   28.37   1        6
#> 41  Case 4 treatment   11   60.37   1        7
#> 42  Case 4 treatment   12   57.46   1        8
#> 43  Case 5  baseline   -3   32.84   0        0
#> 44  Case 5  baseline   -2   14.26   0        0
#> 45  Case 5  baseline   -1    0.00   0        0
#> 46  Case 5  baseline    0    7.99   0        0
#> 47  Case 5  baseline    1   23.55   0        0
#> 48  Case 5  baseline    2   11.51   0        0
#> 49  Case 5  baseline    3   28.22   0        0
#> 50  Case 5  baseline    4   44.94   0        0
#> 51  Case 5  baseline    5   35.02   0        0
#> 52  Case 5  baseline    6    4.34   0        0
#> 53  Case 5 treatment    7   69.26   1        0
#> 54  Case 5 treatment    8   69.96   1        1
#> 55  Case 5 treatment    9   62.98   1        2
#> 56  Case 5 treatment   10   58.13   1        3
#> 57  Case 5 treatment   11   58.17   1        4
#> 58  Case 5 treatment   12   27.67   1        5
#> 59  Case 5 treatment   13   70.85   1        6
#> 60  Case 6  baseline   -3   46.70   0        0
#> 61  Case 6  baseline   -2   56.07   0        0
#> 62  Case 6  baseline   -1   32.28   0        0
#> 63  Case 6  baseline    0   36.70   0        0
#> 64  Case 6  baseline    1   59.93   0        0
#> 65  Case 6  baseline    2   12.05   0        0
#> 66  Case 6  baseline    3   34.95   0        0
#> 67  Case 6 treatment    4   75.18   1        0
#> 68  Case 6 treatment    5   92.15   1        1
#> 69  Case 6 treatment    6   85.84   1        2
#> 70  Case 6 treatment    7   83.33   1        3
#> 71  Case 6 treatment    8   81.16   1        4
#> 72  Case 6 treatment    9   74.86   1        5
#> 73  Case 6 treatment   10   86.54   1        6
#> 74  Case 6 treatment   11   74.96   1        7
#> 75  Case 6 treatment   12   87.13   1        8
#> 76  Case 7  baseline   -3   52.32   0        0
#> 77  Case 7  baseline   -2   51.99   0        0
#> 78  Case 7  baseline   -1   65.23   0        0
#> 79  Case 7  baseline    0   57.28   0        0
#> 80  Case 7  baseline    1   67.55   0        0
#> 81  Case 7  baseline    2   57.62   0        0
#> 82  Case 7  baseline    3   55.46   0        0
#> 83  Case 7 treatment    4   78.31   1        0
#> 84  Case 7 treatment    5   89.40   1        1
#> 85  Case 7 treatment    6   72.35   1        2
#> 86  Case 7 treatment    7   86.59   1        3
#> 87  Case 7 treatment    8   81.79   1        4
#> 88  Case 7 treatment    9   89.57   1        5
#> 89  Case 7 treatment   10   84.11   1        6
#> 90  Case 8  baseline   -3   58.06   0        0
#> 91  Case 8  baseline   -2   65.79   0        0
#> 92  Case 8  baseline   -1   52.63   0        0
#> 93  Case 8  baseline    0   72.37   0        0
#> 94  Case 8  baseline    1   81.91   0        0
#> 95  Case 8  baseline    2   71.38   0        0
#> 96  Case 8  baseline    3   73.52   0        0
#> 97  Case 8  baseline    4   70.56   0        0
#> 98  Case 8 treatment    5   45.39   1        0
#> 99  Case 8 treatment    6   77.14   1        1
#> 100 Case 8 treatment    7   71.88   1        2
#> 101 Case 8 treatment    8   77.14   1        3
#> 102 Case 8 treatment    9   87.01   1        4
#> 103 Case 8 treatment   10   84.70   1        5
#> 104 Case 8 treatment   11   74.18   1        6
#> 105 Case 8 treatment   12   54.11   1        7
#> 106 Case 8 treatment   13   74.67   1        8
#> 107 Case 8 treatment   14   76.15   1        9
#> 108 Case 8 treatment   15   83.39   1       10
#> 109 Case 9  baseline   -3   41.36   0        0
#> 110 Case 9  baseline   -2   73.04   0        0
#> 111 Case 9  baseline   -1   47.25   0        0
#> 112 Case 9  baseline    0   67.13   0        0
#> 113 Case 9  baseline    1   62.10   0        0
#> 114 Case 9  baseline    2   56.91   0        0
#> 115 Case 9  baseline    3   66.66   0        0
#> 116 Case 9  baseline    4   51.00   0        0
#> 117 Case 9  baseline    5   72.87   0        0
#> 118 Case 9  baseline    6   48.41   0        0
#> 119 Case 9  baseline    7   16.96   0        0
#> 120 Case 9 treatment    8   81.37   1        0
#> 121 Case 9 treatment    9   86.48   1        1
#> 122 Case 9 treatment   10   69.98   1        2
#> 123 Case 9 treatment   11   78.41   1        3
#> 124 Case 9 treatment   12   72.21   1        4
#> 125 Case 9 treatment   13   70.51   1        5
#> 126 Case 9 treatment   14   67.47   1        6
#> 127 Case 9 treatment   15   84.53   1        7
#> 128 Case 9 treatment   16   81.00   1        8