A shiny interface for the scdhlm package

An interactive shiny interface for estimating design-comparable standardized mean difference effect sizes from single-case designs. Estimation methods for multiple baseline and treatment reversal designs are available.

shine_scd(dataset = NULL, ...)

Arguments

dataset

Optionally, a data.frame or path to a file from which to read data. If specified, the app will open with the data loaded. Default is NULL. If dataset is a data.frame, then it will be passed directly. If a file path with a .xlsx extension is specified, it will be read using read_excel. If a file path with a .csv extension is specified, it will be read using read.csv. If a file path with a different extension is specified, it will be read using read.table.

...

Further arguments passed to read_excel, read.csv, or read.table.

Examples

if (FALSE) {
shine_scd()
data(Laski)
shine_scd(dataset = Laski)
shine_scd(dataset = "SCD_data.xlsx", sheet = "Laski")
shine_scd(dataset = "Laski.csv") 
}