Simulate Model MB2 from Pustejovsky, Hedges, & Shadish (2014)

Simulates data from a linear mixed effects model, then calculates REML effect size estimator as described in Pustejovsky, Hedges, & Shadish (2014).

simulate_MB2(
  iterations,
  beta,
  rho,
  phi,
  tau1_ratio,
  tau_corr,
  design,
  m,
  n,
  MB = TRUE
)

Arguments

iterations
number of independent iterations of the simulation

beta
vector of fixed effect parameters

rho
intra-class correlation parameter

phi
autocorrelation parameter

tau1_ratio
ratio of treatment effect variance to intercept variance

tau_corr
correlation between case-specific treatment effects and intercepts

design
design matrix. If not specified, it will be calculated based on m, n, and MB.

m
number of cases. Not used if design is specified.

n
number of measurement occasions. Not used if design is specified.

MB
If true, a multiple baseline design will be used; otherwise, an AB design will be used. Not used if design is specified.

Value

A matrix reporting the mean and variance of the effect size estimates and various associated statistics.

References

Pustejovsky, J. E., Hedges, L. V., & Shadish, W. R. (2014). Design-comparable effect sizes in multiple baseline designs: A general modeling framework. Journal of Educational and Behavioral Statistics, 39(4), 211-227. doi:10.3102/1076998614547577

Examples

set.seed(8)
simulate_MB2(iterations = 10, beta = c(0,1,0,0), rho = 0.4, phi = 0.5,
             tau1_ratio = 0.5, tau_corr = -0.4, design = design_matrix(m=3, n=8))

#> Warning: nlminb problem, convergence error code = 1
#>   message = iteration limit reached without convergence (10)

#> Warning: nlminb problem, convergence error code = 1
#>   message = iteration limit reached without convergence (10)

#> Warning: nlminb problem, convergence error code = 1
#>   message = iteration limit reached without convergence (10)

#> Warning: nlminb problem, convergence error code = 1
#>   message = iteration limit reached without convergence (10)

#> Warning: nlminb problem, convergence error code = 1
#>   message = false convergence (8)

#> Warning: nlminb problem, convergence error code = 1
#>   message = iteration limit reached without convergence (10)

#> Warning: nlminb problem, convergence error code = 1
#>   message = iteration limit reached without convergence (10)

#> Warning: nlminb problem, convergence error code = 1
#>   message = iteration limit reached without convergence (10)

#>                                      mean        var
#> p_beta                          0.9437200 0.17149090
#> r_theta                         1.3408175 0.91335371
#> delta_AB                        0.9629504 0.49129233
#> nu                              4.3361853 2.38222586
#> kappa                           0.4443433 0.03362853
#> g_AB                            0.7815793 0.41146611
#> V_g_AB                          0.4953734 0.13286470
#> cnvg_warn                       0.8000000 0.17777778
#> sigma_sq                        0.6131440 0.11369113
#> phi                             0.3587512 0.08332224
#> Tau.id.var(constant)            0.7276736 0.91766909
#> Tau.id.cov(treatment,constant) -0.2336103 0.15529333
#> Tau.id.var(treatment)           0.3035845 0.12358204
#> RML_coverage1                   1.0000000 0.00000000
#> RML_coverage2                   1.0000000 0.00000000

set.seed(8)
simulate_MB2(iterations = 10, beta = c(0,1,0,0), rho = 0.4, phi = 0.5,
             tau1_ratio = 0.5, tau_corr = -0.4, m = 3, n = 8, MB = FALSE)

#> Warning: nlminb problem, convergence error code = 1
#>   message = iteration limit reached without convergence (10)

#> Warning: nlminb problem, convergence error code = 1
#>   message = iteration limit reached without convergence (10)

#> Warning: nlminb problem, convergence error code = 1
#>   message = iteration limit reached without convergence (10)

#> Warning: nlminb problem, convergence error code = 1
#>   message = iteration limit reached without convergence (10)

#> Warning: nlminb problem, convergence error code = 1
#>   message = false convergence (8)

#> Warning: nlminb problem, convergence error code = 1
#>   message = iteration limit reached without convergence (10)

#> Warning: nlminb problem, convergence error code = 1
#>   message = iteration limit reached without convergence (10)

#> Warning: nlminb problem, convergence error code = 1
#>   message = iteration limit reached without convergence (10)

#>                                      mean        var
#> p_beta                          0.9437200 0.17149090
#> r_theta                         1.3408175 0.91335371
#> delta_AB                        0.9629504 0.49129233
#> nu                              4.3361853 2.38222586
#> kappa                           0.4443433 0.03362853
#> g_AB                            0.7815793 0.41146611
#> V_g_AB                          0.4953734 0.13286470
#> cnvg_warn                       0.8000000 0.17777778
#> sigma_sq                        0.6131440 0.11369113
#> phi                             0.3587512 0.08332224
#> Tau.id.var(constant)            0.7276736 0.91766909
#> Tau.id.cov(treatment,constant) -0.2336103 0.15529333
#> Tau.id.var(treatment)           0.3035845 0.12358204
#> RML_coverage1                   1.0000000 0.00000000
#> RML_coverage2                   1.0000000 0.00000000