Simulate Model MB4 from Pustejovsky, Hedges, & Shadish (2014)

Simulates data from a linear mixed effects model, then calculates REML effect size estimator as described in Pustejovsky, Hedges, & Shadish (2014).

simulate_MB4(
  iterations,
  beta,
  rho,
  phi,
  tau2_ratio,
  tau_corr,
  p_const,
  r_const,
  design,
  m,
  n,
  MB = TRUE
)

Arguments

iterations
number of independent iterations of the simulation

beta
vector of fixed effect parameters

rho
intra-class correlation parameter

phi
autocorrelation parameter

tau2_ratio
ratio of trend variance to intercept variance

tau_corr
correlation between case-specific trends and intercepts

p_const
vector of constants for calculating numerator of effect size

r_const
vector of constants for calculating denominator of effect size

design
design matrix. If not specified, it will be calculated based on m, n, and MB.

m
number of cases. Not used if design is specified.

n
number of measurement occasions. Not used if design is specified.

MB
If true, a multiple baseline design will be used; otherwise, an AB design will be used. Not used if design is specified.

Value

A matrix reporting the mean and variance of the effect size estimates and various associated statistics.

References

Pustejovsky, J. E., Hedges, L. V., & Shadish, W. R. (2014). Design-comparable effect sizes in multiple baseline designs: A general modeling framework. Journal of Educational and Behavioral Statistics, 39(4), 211-227. doi:10.3102/1076998614547577

Examples

simulate_MB4(iterations = 10, beta = c(0,1,0,0), rho = 0.8, phi = 0.5,
             tau2_ratio = 0.5, tau_corr = 0,
             p_const = c(0,1,0,7), r_const = c(1,0,1,0,0),
             design = design_matrix(3, 16, treat_times=c(5,9,13), center = 12))

#> Warning: nlminb problem, convergence error code = 1
#>   message = iteration limit reached without convergence (10)

#> Warning: nlminb problem, convergence error code = 1
#>   message = iteration limit reached without convergence (10)

#> Warning: nlminb problem, convergence error code = 1
#>   message = iteration limit reached without convergence (10)

#> Warning: nlminb problem, convergence error code = 1
#>   message = iteration limit reached without convergence (10)

#>                                  mean        var
#> p_beta                     1.05863547 0.47188094
#> r_theta                    0.93648215 0.47415675
#> delta_AB                   1.32487438 1.52194485
#> nu                         3.83882705 6.90618944
#> kappa                      0.55672437 0.03772621
#> g_AB                       1.01480862 1.26233746
#> V_g_AB                     1.60198179 0.61646931
#> cnvg_warn                  0.40000000 0.26666667
#> sigma_sq                   0.21560229 0.01261309
#> phi                        0.41649526 0.07881559
#> Tau.id.var(constant)       0.72087986 0.56788266
#> Tau.id.cov(trend,constant) 0.09850995 0.04564587
#> Tau.id.var(trend)          0.34143681 0.04725532
#> RML_coverage1              0.80000000 0.17777778
#> RML_coverage2              0.90000000 0.10000000

simulate_MB4(iterations = 10, beta = c(0,1,0,0), rho = 0.8, phi = 0.5,
             tau2_ratio = 0.5, tau_corr = 0, m = 6, n = 8)

#> Warning: nlminb problem, convergence error code = 1
#>   message = iteration limit reached without convergence (10)

#> Warning: nlminb problem, convergence error code = 1
#>   message = iteration limit reached without convergence (10)

#> Warning: nlminb problem, convergence error code = 1
#>   message = iteration limit reached without convergence (10)

#> Warning: nlminb problem, convergence error code = 1
#>   message = iteration limit reached without convergence (10)

#> Warning: nlminb problem, convergence error code = 1
#>   message = iteration limit reached without convergence (10)

#> Warning: nlminb problem, convergence error code = 1
#>   message = iteration limit reached without convergence (10)

#>                                 mean         var
#> p_beta                     0.8414655  0.05704896
#> r_theta                    0.7704947  0.18000478
#> delta_AB                   1.0732691  0.21367460
#> nu                         8.7616597 13.56667718
#> kappa                      0.3627907  0.01433000
#> g_AB                       0.9759664  0.20441715
#> V_g_AB                     0.2355260  0.01439783
#> cnvg_warn                  0.6000000  0.26666667
#> sigma_sq                   0.3716371  0.06730303
#> phi                        0.5963326  0.05036218
#> Tau.id.var(constant)       0.3988576  0.22480726
#> Tau.id.cov(trend,constant) 0.1117234  0.05810529
#> Tau.id.var(trend)          0.4664336  0.04102929
#> RML_coverage1              1.0000000  0.00000000
#> RML_coverage2              1.0000000  0.00000000