Run simulation comparing REML and HPS estimates

Simulates data from a simple linear mixed effects model, then calculates REML and HPS effect size estimators as described in Pustejovsky, Hedges, & Shadish (2014).

compare_RML_HPS(iterations, beta, rho, phi, design, m, n, MB = TRUE)

Arguments

Value

A matrix reporting the mean and variance of the effect size estimates and various associated statistics.

References

Pustejovsky, J. E., Hedges, L. V., & Shadish, W. R. (2014). Design-comparable effect sizes in multiple baseline designs: A general modeling framework. Journal of Educational and Behavioral Statistics, 39(4), 211-227. doi:10.3102/1076998614547577

Examples

compare_RML_HPS(iterations=10, beta = c(0,1,0,0), rho = 0.3, 
                 phi = 0.5, design=design_matrix(m=3,n=8))
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#>                           mean         var
#> p_beta               1.1544955 0.196557394
#> r_theta              0.8043969 0.168801078
#> delta_AB             1.4362736 0.523099840
#> nu                   7.2824331 9.547439286
#> kappa                0.3885091 0.004864828
#> g_AB                 1.2592159 0.429506728
#> V_g_AB               0.4140726 0.043715153
#> cnvg_warn            0.0000000 0.000000000
#> sigma_sq             0.5413242 0.041940404
#> phi                  0.5185073 0.020553804
#> Tau.id.var(constant) 0.2630727 0.115882366
#> rho                  0.2456005 0.070465014
#> RML_coverage1        0.9000000 0.100000000
#> RML_coverage2        1.0000000 0.000000000
#> D_bar                1.1187100 0.162926223
#> S_sq                 0.8287006 0.167267194
#> delta_hat_unadj      1.3587520 0.416148942
#> phi_hat              0.2683946 0.014240838
#> rho_hat              0.4512576 0.029310177
#> kappa_hat            0.3491788 0.004831281
#> df                   5.9630223 4.050375118
#> delta_hat            1.1729478 0.340611721
#> V_delta_hat          0.3770404 0.036065905
#> HPS_coverage1        1.0000000 0.000000000
#> HPS_coverage2        1.0000000 0.000000000
#> cov                         NA 0.345619057