Run simulation comparing REML and HPS estimates

Simulates data from a simple linear mixed effects model, then calculates REML and HPS effect size estimators as described in Pustejovsky, Hedges, & Shadish (2014).

compare_RML_HPS(iterations, beta, rho, phi, design, m, n, MB = TRUE)

Arguments

iterations
number of independent iterations of the simulation

beta
vector of fixed effect parameters

rho
intra-class correlation parameter

phi
autocorrelation parameter

design
design matrix. If not specified, it will be calculated based on m, n, and MB.

m
number of cases. Not used if design is specified.

n
number of measurement occasions. Not used if design is specified.

MB
If true, a multiple baseline design will be used; otherwise, an AB design will be used. Not used if design is specified.

Value

A matrix reporting the mean and variance of the effect size estimates and various associated statistics.

References

Pustejovsky, J. E., Hedges, L. V., & Shadish, W. R. (2014). Design-comparable effect sizes in multiple baseline designs: A general modeling framework. Journal of Educational and Behavioral Statistics, 39(4), 211-227. doi:10.3102/1076998614547577

Examples

compare_RML_HPS(iterations=10, beta = c(0,1,0,0), rho = 0.3,
                 phi = 0.5, design=design_matrix(m=3,n=8))

#>                           mean          var
#> p_beta               1.0737095  0.135935072
#> r_theta              1.0092672  0.369981663
#> delta_AB             1.1529924  0.165081738
#> nu                   8.6143473 14.802054704
#> kappa                0.3753481  0.006263534
#> g_AB                 1.0289415  0.145255955
#> V_g_AB               0.2831985  0.005106134
#> cnvg_warn            0.0000000  0.000000000
#> sigma_sq             0.6816939  0.275395459
#> phi                  0.3306837  0.051785447
#> Tau.id.var(constant) 0.3275733  0.192661873
#> rho                  0.2733742  0.064907633
#> RML_coverage1        1.0000000  0.000000000
#> RML_coverage2        1.0000000  0.000000000
#> D_bar                1.0750218  0.155374582
#> S_sq                 1.0964351  0.543470922
#> delta_hat_unadj      1.1057702  0.151944786
#> phi_hat              0.1974794  0.014208809
#> rho_hat              0.4017487  0.043323184
#> kappa_hat            0.3495513  0.005186405
#> df                   7.2505458  9.229465331
#> delta_hat            0.9694831  0.132003621
#> V_delta_hat          0.2727806  0.005590721
#> HPS_coverage1        1.0000000  0.000000000
#> HPS_coverage2        1.0000000  0.000000000
#> cov                         NA  0.132859737