Run simulation comparing REML and HPS estimates

Simulates data from a simple linear mixed effects model, then calculates REML and HPS effect size estimators as described in Pustejovsky, Hedges, & Shadish (2014).

compare_RML_HPS(iterations, beta, rho, phi, design, m, n, MB = TRUE)

Arguments

iterations

number of independent iterations of the simulation

beta

vector of fixed effect parameters

rho

intra-class correlation parameter

phi

autocorrelation parameter

design

design matrix. If not specified, it will be calculated based on m, n, and MB.

m

number of cases. Not used if design is specified.

n

number of measurement occasions. Not used if design is specified.

MB

If true, a multiple baseline design will be used; otherwise, an AB design will be used. Not used if design is specified.

Value

A matrix reporting the mean and variance of the effect size estimates and various associated statistics.

References

Pustejovsky, J. E., Hedges, L. V., & Shadish, W. R. (2014). Design-comparable effect sizes in multiple baseline designs: A general modeling framework. Journal of Educational and Behavioral Statistics, 39(4), 211-227. doi:10.3102/1076998614547577

Examples

compare_RML_HPS(iterations=10, beta = c(0,1,0,0), rho = 0.3, 
                 phi = 0.5, design=design_matrix(m=3,n=8))
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#>                           mean          var
#> p_beta               0.9537586  0.145290273
#> r_theta              1.3323692  0.813220157
#> delta_AB             0.9327767  0.179788957
#> nu                   6.8567161 10.401360662
#> kappa                0.3516832  0.005489085
#> g_AB                 0.8114125  0.144055082
#> V_g_AB               0.2510711  0.012069121
#> cnvg_warn            0.0000000  0.000000000
#> sigma_sq             0.9501083  0.491736455
#> phi                  0.4293339  0.117861354
#> Tau.id.var(constant) 0.3822609  0.185317166
#> rho                  0.2944270  0.051317593
#> RML_coverage1        1.0000000  0.000000000
#> RML_coverage2        1.0000000  0.000000000
#> D_bar                1.0279392  0.112700858
#> S_sq                 1.2144891  0.552528093
#> delta_hat_unadj      1.0007934  0.077483952
#> phi_hat              0.1359045  0.032157728
#> rho_hat              0.4324691  0.059996290
#> kappa_hat            0.3279533  0.007359439
#> df                   7.0559244 15.057218263
#> delta_hat            0.8613239  0.063169728
#> V_delta_hat          0.2589960  0.013433859
#> HPS_coverage1        1.0000000  0.000000000
#> HPS_coverage2        1.0000000  0.000000000
#> cov                         NA  0.085953631