Clean Single Case Design Data

Clean single case design data for treatment reversal and multiple baseline designs.

preprocess_SCD(
  design,
  case,
  phase,
  session,
  outcome,
  cluster = NULL,
  series = NULL,
  center = 0,
  round_session = TRUE,
  treatment_name = NULL,
  data = NULL
)

Arguments

design

Character string to specify whether data comes from a treatment reversal ("TR"), multiple baseline across participants ("MBP"), replicated multiple baseline across behaviors ("RMBB"), or clustered multiple baseline across participants ("CMB").

case

vector of case indicators or name of a character or factor vector within data indicating unique cases.

phase

vector of treatment indicators or name of a character or factor vector within data indicating unique treatment phases.

session

vector of measurement occasions or name of numeric vector within data of measurement times.

outcome

vector of outcome data or name of numeric vector of outcome data within data.

cluster

(Optional) vector of cluster indicators or name of a character or factor vector within data indicating clusters.

series

(Optional) vector of series indicators or name of a character or factor vector within data indicating series.

center

Numeric value for the centering value for session. Default is 0.

round_session

Logical indicating whether to round session to the nearest integer. Defaults to TRUE.

treatment_name

(Optional) character string corresponding to the name of the treatment phase.

data

(Optional) dataset to use for analysis. Must be a data.frame.

Value

A cleaned SCD dataset that can be used for model fitting and effect size calculation.

Note

If treatment_name is left null it will choose the second level of the phase variable to be the treatment phase.

Examples

data(Laski)
preprocess_SCD(design = "MBP", 
               case = case, phase = treatment,
               session = time, outcome = outcome, 
               center = 4, data = Laski)
#>       case treatment time outcome trt time_trt
#> 1   Case 1  baseline   -3   27.60   0        0
#> 2   Case 1  baseline   -2   23.96   0        0
#> 3   Case 1  baseline   -1   23.83   0        0
#> 4   Case 1  baseline    0   47.26   0        0
#> 5   Case 1 treatment    1   52.70   1        1
#> 6   Case 1 treatment    2   60.99   1        2
#> 7   Case 1 treatment    3   66.60   1        3
#> 8   Case 1 treatment    4   52.50   1        4
#> 9   Case 1 treatment    5   85.88   1        5
#> 10  Case 1 treatment    6   47.05   1        6
#> 11  Case 1 treatment    7   66.28   1        7
#> 12  Case 1 treatment    8   54.05   1        8
#> 13  Case 1 treatment    9   51.23   1        9
#> 14  Case 2  baseline   -3   49.67   0        0
#> 15  Case 2  baseline   -2   23.57   0        0
#> 16  Case 2  baseline   -1   26.38   0        0
#> 17  Case 2  baseline    0   28.35   0        0
#> 18  Case 2  baseline    1   45.11   0        0
#> 19  Case 2 treatment    2   70.67   1        1
#> 20  Case 2 treatment    3   79.13   1        2
#> 21  Case 2 treatment    4   84.09   1        3
#> 22  Case 2 treatment    5   88.56   1        4
#> 23  Case 2 treatment    6   80.90   1        5
#> 24  Case 2 treatment    7   91.84   1        6
#> 25  Case 2 treatment    8   63.42   1        7
#> 26  Case 2 treatment    9   70.38   1        8
#> 27  Case 4  baseline   -3    0.00   0        0
#> 28  Case 4  baseline   -2    0.00   0        0
#> 29  Case 4  baseline   -1    0.00   0        0
#> 30  Case 4  baseline    0    0.00   0        0
#> 31  Case 4  baseline    1    0.00   0        0
#> 32  Case 4  baseline    2    0.00   0        0
#> 33  Case 4  baseline    3    0.00   0        0
#> 34  Case 4 treatment    4   28.07   1        1
#> 35  Case 4 treatment    5   28.20   1        2
#> 36  Case 4 treatment    6   57.15   1        3
#> 37  Case 4 treatment    7   72.37   1        4
#> 38  Case 4 treatment    8   45.56   1        5
#> 39  Case 4 treatment    9   52.30   1        6
#> 40  Case 4 treatment   10   28.37   1        7
#> 41  Case 4 treatment   11   60.37   1        8
#> 42  Case 4 treatment   12   57.46   1        9
#> 43  Case 5  baseline   -3   32.84   0        0
#> 44  Case 5  baseline   -2   14.26   0        0
#> 45  Case 5  baseline   -1    0.00   0        0
#> 46  Case 5  baseline    0    7.99   0        0
#> 47  Case 5  baseline    1   23.55   0        0
#> 48  Case 5  baseline    2   11.51   0        0
#> 49  Case 5  baseline    3   28.22   0        0
#> 50  Case 5  baseline    4   44.94   0        0
#> 51  Case 5  baseline    5   35.02   0        0
#> 52  Case 5  baseline    6    4.34   0        0
#> 53  Case 5 treatment    7   69.26   1        1
#> 54  Case 5 treatment    8   69.96   1        2
#> 55  Case 5 treatment    9   62.98   1        3
#> 56  Case 5 treatment   10   58.13   1        4
#> 57  Case 5 treatment   11   58.17   1        5
#> 58  Case 5 treatment   12   27.67   1        6
#> 59  Case 5 treatment   13   70.85   1        7
#> 60  Case 6  baseline   -3   46.70   0        0
#> 61  Case 6  baseline   -2   56.07   0        0
#> 62  Case 6  baseline   -1   32.28   0        0
#> 63  Case 6  baseline    0   36.70   0        0
#> 64  Case 6  baseline    1   59.93   0        0
#> 65  Case 6  baseline    2   12.05   0        0
#> 66  Case 6  baseline    3   34.95   0        0
#> 67  Case 6 treatment    4   75.18   1        1
#> 68  Case 6 treatment    5   92.15   1        2
#> 69  Case 6 treatment    6   85.84   1        3
#> 70  Case 6 treatment    7   83.33   1        4
#> 71  Case 6 treatment    8   81.16   1        5
#> 72  Case 6 treatment    9   74.86   1        6
#> 73  Case 6 treatment   10   86.54   1        7
#> 74  Case 6 treatment   11   74.96   1        8
#> 75  Case 6 treatment   12   87.13   1        9
#> 76  Case 7  baseline   -3   52.32   0        0
#> 77  Case 7  baseline   -2   51.99   0        0
#> 78  Case 7  baseline   -1   65.23   0        0
#> 79  Case 7  baseline    0   57.28   0        0
#> 80  Case 7  baseline    1   67.55   0        0
#> 81  Case 7  baseline    2   57.62   0        0
#> 82  Case 7  baseline    3   55.46   0        0
#> 83  Case 7 treatment    4   78.31   1        1
#> 84  Case 7 treatment    5   89.40   1        2
#> 85  Case 7 treatment    6   72.35   1        3
#> 86  Case 7 treatment    7   86.59   1        4
#> 87  Case 7 treatment    8   81.79   1        5
#> 88  Case 7 treatment    9   89.57   1        6
#> 89  Case 7 treatment   10   84.11   1        7
#> 90  Case 8  baseline   -3   58.06   0        0
#> 91  Case 8  baseline   -2   65.79   0        0
#> 92  Case 8  baseline   -1   52.63   0        0
#> 93  Case 8  baseline    0   72.37   0        0
#> 94  Case 8  baseline    1   81.91   0        0
#> 95  Case 8  baseline    2   71.38   0        0
#> 96  Case 8  baseline    3   73.52   0        0
#> 97  Case 8  baseline    4   70.56   0        0
#> 98  Case 8 treatment    5   45.39   1        1
#> 99  Case 8 treatment    6   77.14   1        2
#> 100 Case 8 treatment    7   71.88   1        3
#> 101 Case 8 treatment    8   77.14   1        4
#> 102 Case 8 treatment    9   87.01   1        5
#> 103 Case 8 treatment   10   84.70   1        6
#> 104 Case 8 treatment   11   74.18   1        7
#> 105 Case 8 treatment   12   54.11   1        8
#> 106 Case 8 treatment   13   74.67   1        9
#> 107 Case 8 treatment   14   76.15   1       10
#> 108 Case 8 treatment   15   83.39   1       11
#> 109 Case 9  baseline   -3   41.36   0        0
#> 110 Case 9  baseline   -2   73.04   0        0
#> 111 Case 9  baseline   -1   47.25   0        0
#> 112 Case 9  baseline    0   67.13   0        0
#> 113 Case 9  baseline    1   62.10   0        0
#> 114 Case 9  baseline    2   56.91   0        0
#> 115 Case 9  baseline    3   66.66   0        0
#> 116 Case 9  baseline    4   51.00   0        0
#> 117 Case 9  baseline    5   72.87   0        0
#> 118 Case 9  baseline    6   48.41   0        0
#> 119 Case 9  baseline    7   16.96   0        0
#> 120 Case 9 treatment    8   81.37   1        1
#> 121 Case 9 treatment    9   86.48   1        2
#> 122 Case 9 treatment   10   69.98   1        3
#> 123 Case 9 treatment   11   78.41   1        4
#> 124 Case 9 treatment   12   72.21   1        5
#> 125 Case 9 treatment   13   70.51   1        6
#> 126 Case 9 treatment   14   67.47   1        7
#> 127 Case 9 treatment   15   84.53   1        8
#> 128 Case 9 treatment   16   81.00   1        9
               
data(Anglesea) 
preprocess_SCD(design="TR",
               case=case, phase=condition,
               session=session, outcome=outcome,
               treatment_name = "treatment",
               data=Anglesea)
#>      case condition session outcome trt condition_pair
#> 1  Case 1  baseline       1  103.72   0              1
#> 2  Case 1  baseline       2  105.32   0              1
#> 3  Case 1  baseline       3   77.39   0              1
#> 4  Case 1  baseline       4  117.29   0              1
#> 5  Case 1  baseline       5  102.93   0              1
#> 6  Case 1  baseline       6   99.73   0              1
#> 7  Case 1  baseline       7   49.47   0              1
#> 8  Case 1 treatment       8  183.51   1              1
#> 9  Case 1 treatment       9  152.39   1              1
#> 10 Case 1 treatment      10  170.74   1              1
#> 11 Case 1 treatment      11  169.95   1              1
#> 12 Case 1 treatment      12  154.79   1              1
#> 13 Case 1 treatment      13  153.99   1              1
#> 14 Case 1  baseline      14   97.34   0              2
#> 15 Case 1  baseline      15   57.45   0              2
#> 16 Case 1  baseline      16   67.82   0              2
#> 17 Case 1  baseline      17  107.71   0              2
#> 18 Case 1  baseline      18  103.72   0              2
#> 19 Case 1  baseline      19   90.16   0              2
#> 20 Case 1  baseline      20   93.35   0              2
#> 21 Case 1 treatment      21  153.19   1              2
#> 22 Case 1 treatment      22  149.20   1              2
#> 23 Case 1 treatment      23  149.20   1              2
#> 24 Case 1 treatment      24  142.02   1              2
#> 25 Case 1 treatment      25  143.62   1              2
#> 26 Case 1 treatment      26  176.33   1              2
#> 27 Case 1 treatment      27  173.94   1              2
#> 28 Case 2  baseline       1   64.00   0              1
#> 29 Case 2  baseline       2   62.80   0              1
#> 30 Case 2  baseline       3   43.60   0              1
#> 31 Case 2  baseline       4   63.40   0              1
#> 32 Case 2 treatment       5  213.20   1              1
#> 33 Case 2 treatment       6  220.01   1              1
#> 34 Case 2 treatment       7  212.80   1              1
#> 35 Case 2 treatment       8  217.60   1              1
#> 36 Case 2  baseline       9   62.40   0              2
#> 37 Case 2  baseline      10   72.20   0              2
#> 38 Case 2  baseline      11   62.01   0              2
#> 39 Case 2 treatment      12  189.81   1              2
#> 40 Case 2 treatment      13  178.60   1              2
#> 41 Case 2 treatment      14  181.41   1              2
#> 42 Case 3  baseline       1   24.02   0              1
#> 43 Case 3  baseline       2   26.38   0              1
#> 44 Case 3  baseline       3   29.92   0              1
#> 45 Case 3  baseline       4   23.62   0              1
#> 46 Case 3 treatment       5   84.65   1              1
#> 47 Case 3 treatment       6   89.76   1              1
#> 48 Case 3 treatment       7   83.46   1              1
#> 49 Case 3 treatment       8   89.76   1              1
#> 50 Case 3  baseline       9   30.71   0              2
#> 51 Case 3  baseline      10   41.73   0              2
#> 52 Case 3  baseline      11   29.92   0              2
#> 53 Case 3  baseline      12   33.86   0              2
#> 54 Case 3 treatment      13   84.25   1              2
#> 55 Case 3 treatment      14   78.74   1              2
               
data(Thiemann2001)
preprocess_SCD(design = "RMBB", 
               case = case, series = series, phase = treatment, 
               session = time, outcome = outcome, 
               data = Thiemann2001)
#>      case                          series treatment time outcome trt time_trt
#> 1   Casey            contingent responses  baseline    1     1.3   0        0
#> 2   Casey            contingent responses  baseline    2     4.6   0        0
#> 3   Casey            contingent responses  baseline    3     0.0   0        0
#> 4   Casey            contingent responses  baseline    4     2.3   0        0
#> 5   Casey            contingent responses  baseline    5     0.7   0        0
#> 6   Casey            contingent responses  baseline    6     2.3   0        0
#> 7   Casey            contingent responses  baseline    7     4.6   0        0
#> 8   Casey            contingent responses  baseline    8     8.9   0        0
#> 9   Casey            contingent responses  baseline    9     4.3   0        0
#> 10  Casey            contingent responses  baseline   10     3.0   0        0
#> 11  Casey            contingent responses  baseline   11     3.0   0        0
#> 12  Casey            contingent responses  baseline   12     4.9   0        0
#> 13  Casey            contingent responses  baseline   13     0.0   0        0
#> 14  Casey            contingent responses  baseline   14     4.3   0        0
#> 15  Casey            contingent responses  baseline   15     3.0   0        0
#> 16  Casey            contingent responses  baseline   16     3.0   0        0
#> 17  Casey            contingent responses  baseline   17     3.9   0        0
#> 18  Casey            contingent responses  baseline   18     1.0   0        0
#> 19  Casey            contingent responses  baseline   19    16.1   0        0
#> 20  Casey            contingent responses  baseline   20     5.9   0        0
#> 21  Casey            contingent responses  baseline   21     5.9   0        0
#> 22  Casey            contingent responses  baseline   22     1.0   0        0
#> 23  Casey            contingent responses  baseline   23     1.0   0        0
#> 24  Casey            contingent responses  baseline   24     3.0   0        0
#> 25  Casey            contingent responses  baseline   25     5.9   0        0
#> 26  Casey            contingent responses treatment   26    30.5   1        1
#> 27  Casey            contingent responses treatment   27    20.0   1        2
#> 28  Casey            contingent responses treatment   28    22.0   1        3
#> 29  Casey            contingent responses treatment   29    31.5   1        4
#> 30  Casey            contingent responses treatment   30    25.9   1        5
#> 31  Casey initiates comments and requests  baseline    1     0.3   0        0
#> 32  Casey initiates comments and requests  baseline    2     0.7   0        0
#> 33  Casey initiates comments and requests  baseline    3     4.0   0        0
#> 34  Casey initiates comments and requests  baseline    4     0.7   0        0
#> 35  Casey initiates comments and requests  baseline    5     3.4   0        0
#> 36  Casey initiates comments and requests  baseline    6     8.6   0        0
#> 37  Casey initiates comments and requests  baseline    7     3.7   0        0
#> 38  Casey initiates comments and requests  baseline    8     4.7   0        0
#> 39  Casey initiates comments and requests  baseline    9     2.5   0        0
#> 40  Casey initiates comments and requests treatment   10    11.3   1        1
#> 41  Casey initiates comments and requests treatment   11    10.3   1        2
#> 42  Casey initiates comments and requests treatment   12     9.7   1        3
#> 43  Casey initiates comments and requests treatment   13    21.7   1        4
#> 44  Casey initiates comments and requests treatment   14    23.4   1        5
#> 45  Casey initiates comments and requests treatment   15    20.8   1        6
#> 46  Casey               secures attention  baseline    1     0.0   0        0
#> 47  Casey               secures attention  baseline    2     0.0   0        0
#> 48  Casey               secures attention  baseline    3     0.0   0        0
#> 49  Casey               secures attention  baseline    4     0.0   0        0
#> 50  Casey               secures attention  baseline    5     0.0   0        0
#> 51  Casey               secures attention  baseline    6     0.0   0        0
#> 52  Casey               secures attention  baseline    7     0.0   0        0
#> 53  Casey               secures attention  baseline    8     0.0   0        0
#> 54  Casey               secures attention  baseline    9     0.0   0        0
#> 55  Casey               secures attention  baseline   10     0.0   0        0
#> 56  Casey               secures attention  baseline   11     0.0   0        0
#> 57  Casey               secures attention  baseline   12     0.0   0        0
#> 58  Casey               secures attention  baseline   13     0.0   0        0
#> 59  Casey               secures attention  baseline   14     0.0   0        0
#> 60  Casey               secures attention  baseline   15     0.0   0        0
#> 61  Casey               secures attention treatment   16    13.7   1        1
#> 62  Casey               secures attention treatment   17    11.8   1        2
#> 63  Casey               secures attention treatment   18    12.1   1        3
#> 64  Casey               secures attention treatment   19     8.9   1        4
#> 65  Casey               secures attention treatment   20    10.8   1        5
#> 66  Casey               secures attention treatment   21    12.7   1        6
#> 67  Casey               secures attention treatment   22    10.1   1        7
#> 68  Casey               secures attention treatment   23     9.1   1        8
#> 69  Casey               secures attention treatment   24     9.4   1        9
#> 70  Casey               secures attention treatment   25     9.7   1       10
#> 71    Dan              initiates comments  baseline    1     0.1   0        0
#> 72    Dan              initiates comments  baseline    2     1.2   0        0
#> 73    Dan              initiates comments  baseline    3     4.3   0        0
#> 74    Dan              initiates comments  baseline    4    11.1   0        0
#> 75    Dan              initiates comments  baseline    5     9.1   0        0
#> 76    Dan              initiates comments  baseline    6     2.9   0        0
#> 77    Dan              initiates comments  baseline    7     8.1   0        0
#> 78    Dan              initiates comments  baseline    8     5.3   0        0
#> 79    Dan              initiates comments  baseline    9     6.3   0        0
#> 80    Dan              initiates comments  baseline   10    10.4   0        0
#> 81    Dan              initiates comments  baseline   11     7.1   0        0
#> 82    Dan              initiates comments  baseline   12     5.4   0        0
#> 83    Dan              initiates comments  baseline   13     3.7   0        0
#> 84    Dan              initiates comments  baseline   14     7.2   0        0
#> 85    Dan              initiates comments  baseline   15     5.2   0        0
#> 86    Dan              initiates comments  baseline   16    10.3   0        0
#> 87    Dan              initiates comments  baseline   17     8.3   0        0
#> 88    Dan              initiates comments  baseline   18     9.2   0        0
#> 89    Dan              initiates comments  baseline   19    10.3   0        0
#> 90    Dan              initiates comments  baseline   20     8.3   0        0
#> 91    Dan              initiates comments treatment   21    19.3   1        1
#> 92    Dan              initiates comments treatment   22    12.5   1        2
#> 93    Dan              initiates comments treatment   23    15.3   1        3
#> 94    Dan              initiates comments treatment   24    14.4   1        4
#> 95    Dan              initiates comments treatment   25    13.5   1        5
#> 96    Dan              initiates comments treatment   26    20.6   1        6
#> 97    Dan              initiates comments treatment   27    24.4   1        7
#> 98    Dan              initiates comments treatment   28    16.7   1        8
#> 99    Dan              initiates comments treatment   29    13.7   1        9
#> 100   Dan              initiates comments treatment   30    39.4   1       10
#> 101   Dan              initiates requests  baseline    1     0.0   0        0
#> 102   Dan              initiates requests  baseline    2     1.9   0        0
#> 103   Dan              initiates requests  baseline    3     0.0   0        0
#> 104   Dan              initiates requests  baseline    4     0.8   0        0
#> 105   Dan              initiates requests  baseline    5     1.9   0        0
#> 106   Dan              initiates requests  baseline    6     8.0   0        0
#> 107   Dan              initiates requests  baseline    7     3.0   0        0
#> 108   Dan              initiates requests  baseline    8     5.9   0        0
#> 109   Dan              initiates requests  baseline    9     0.8   0        0
#> 110   Dan              initiates requests  baseline   10     5.9   0        0
#> 111   Dan              initiates requests  baseline   11     4.1   0        0
#> 112   Dan              initiates requests  baseline   12     1.8   0        0
#> 113   Dan              initiates requests  baseline   13     8.1   0        0
#> 114   Dan              initiates requests  baseline   14     0.9   0        0
#> 115   Dan              initiates requests  baseline   15     9.0   0        0
#> 116   Dan              initiates requests  baseline   16     8.1   0        0
#> 117   Dan              initiates requests  baseline   17    10.1   0        0
#> 118   Dan              initiates requests  baseline   18     8.0   0        0
#> 119   Dan              initiates requests  baseline   19    22.2   0        0
#> 120   Dan              initiates requests  baseline   20     9.3   0        0
#> 121   Dan              initiates requests  baseline   21     8.0   0        0
#> 122   Dan              initiates requests  baseline   22     6.9   0        0
#> 123   Dan              initiates requests  baseline   23    14.1   0        0
#> 124   Dan              initiates requests  baseline   24     7.2   0        0
#> 125   Dan              initiates requests  baseline   25    16.2   0        0
#> 126   Dan              initiates requests  baseline   26    16.1   0        0
#> 127   Dan              initiates requests  baseline   27    13.5   0        0
#> 128   Dan              initiates requests  baseline   28     2.0   0        0
#> 129   Dan              initiates requests  baseline   29     8.1   0        0
#> 130   Dan              initiates requests  baseline   30    17.3   0        0
#> 131   Dan              initiates requests treatment   31    12.3   1        1
#> 132   Dan              initiates requests treatment   32    25.6   1        2
#> 133   Dan              initiates requests treatment   33    23.6   1        3
#> 134   Dan              initiates requests treatment   34    19.7   1        4
#> 135   Dan              initiates requests treatment   35    17.2   1        5
#> 136   Dan              initiates requests treatment   36    30.7   1        6
#> 137   Dan               secures attention  baseline    1     0.0   0        0
#> 138   Dan               secures attention  baseline    2     0.0   0        0
#> 139   Dan               secures attention  baseline    3     0.0   0        0
#> 140   Dan               secures attention  baseline    4     0.0   0        0
#> 141   Dan               secures attention  baseline    5     0.0   0        0
#> 142   Dan               secures attention  baseline    6     2.2   0        0
#> 143   Dan               secures attention  baseline    7     0.0   0        0
#> 144   Dan               secures attention  baseline    8     0.0   0        0
#> 145   Dan               secures attention  baseline    9     0.0   0        0
#> 146   Dan               secures attention  baseline   10     4.2   0        0
#> 147   Dan               secures attention  baseline   11     0.0   0        0
#> 148   Dan               secures attention  baseline   12     1.4   0        0
#> 149   Dan               secures attention  baseline   13     0.0   0        0
#> 150   Dan               secures attention  baseline   14     0.0   0        0
#> 151   Dan               secures attention treatment   15    12.1   1        1
#> 152   Dan               secures attention treatment   16    15.1   1        2
#> 153   Dan               secures attention treatment   17     9.2   1        3
#> 154   Dan               secures attention treatment   18    13.1   1        4
#> 155   Dan               secures attention treatment   19    20.1   1        5
#> 156   Dan               secures attention treatment   20    14.2   1        6
#> 157  Greg              initiates comments  baseline    1     1.9   0        0
#> 158  Greg              initiates comments  baseline    2     6.2   0        0
#> 159  Greg              initiates comments  baseline    3     3.9   0        0
#> 160  Greg              initiates comments  baseline    4     3.9   0        0
#> 161  Greg              initiates comments  baseline    5     6.9   0        0
#> 162  Greg              initiates comments  baseline    6     9.2   0        0
#> 163  Greg              initiates comments  baseline    7    11.1   0        0
#> 164  Greg              initiates comments  baseline    8     5.0   0        0
#> 165  Greg              initiates comments  baseline    9     9.2   0        0
#> 166  Greg              initiates comments  baseline   10    15.1   0        0
#> 167  Greg              initiates comments  baseline   11     5.0   0        0
#> 168  Greg              initiates comments  baseline   12     4.0   0        0
#> 169  Greg              initiates comments  baseline   13    18.7   0        0
#> 170  Greg              initiates comments  baseline   14    10.9   0        0
#> 171  Greg              initiates comments  baseline   15     0.0   0        0
#> 172  Greg              initiates comments  baseline   16    13.8   0        0
#> 173  Greg              initiates comments  baseline   17     5.7   0        0
#> 174  Greg              initiates comments  baseline   18     5.1   0        0
#> 175  Greg              initiates comments  baseline   19     7.0   0        0
#> 176  Greg              initiates comments  baseline   20     0.0   0        0
#> 177  Greg              initiates comments  baseline   21     3.5   0        0
#> 178  Greg              initiates comments  baseline   22    13.9   0        0
#> 179  Greg              initiates comments treatment   23    16.5   1        1
#> 180  Greg              initiates comments treatment   24    14.9   1        2
#> 181  Greg              initiates comments treatment   25    16.2   1        3
#> 182  Greg              initiates comments treatment   26    10.0   1        4
#> 183  Greg              initiates comments treatment   27    11.0   1        5
#> 184  Greg              initiates comments treatment   28    21.1   1        6
#> 185  Greg              initiates comments treatment   29    13.3   1        7
#> 186  Greg              initiates comments treatment   30    16.3   1        8
#> 187  Greg              initiates requests  baseline    1    11.8   0        0
#> 188  Greg              initiates requests  baseline    2     0.0   0        0
#> 189  Greg              initiates requests  baseline    3     9.5   0        0
#> 190  Greg              initiates requests  baseline    4     2.3   0        0
#> 191  Greg              initiates requests  baseline    5     1.6   0        0
#> 192  Greg              initiates requests  baseline    6     0.0   0        0
#> 193  Greg              initiates requests  baseline    7     7.1   0        0
#> 194  Greg              initiates requests  baseline    8     8.8   0        0
#> 195  Greg              initiates requests  baseline    9    13.4   0        0
#> 196  Greg              initiates requests  baseline   10     5.2   0        0
#> 197  Greg              initiates requests  baseline   11     5.1   0        0
#> 198  Greg              initiates requests  baseline   12     2.5   0        0
#> 199  Greg              initiates requests  baseline   13     4.5   0        0
#> 200  Greg              initiates requests  baseline   14    11.0   0        0
#> 201  Greg              initiates requests  baseline   15    11.7   0        0
#> 202  Greg              initiates requests  baseline   16     7.7   0        0
#> 203  Greg              initiates requests  baseline   17    10.0   0        0
#> 204  Greg              initiates requests  baseline   18     9.3   0        0
#> 205  Greg              initiates requests  baseline   19    12.9   0        0
#> 206  Greg              initiates requests  baseline   20     1.4   0        0
#> 207  Greg              initiates requests  baseline   21    14.5   0        0
#> 208  Greg              initiates requests  baseline   22    10.3   0        0
#> 209  Greg              initiates requests  baseline   23     4.0   0        0
#> 210  Greg              initiates requests  baseline   24    10.3   0        0
#> 211  Greg              initiates requests  baseline   25     2.1   0        0
#> 212  Greg              initiates requests  baseline   26     0.1   0        0
#> 213  Greg              initiates requests  baseline   27     2.7   0        0
#> 214  Greg              initiates requests  baseline   28     6.9   0        0
#> 215  Greg              initiates requests  baseline   29     4.6   0        0
#> 216  Greg              initiates requests  baseline   30     8.6   0        0
#> 217  Greg              initiates requests treatment   31    19.7   1        1
#> 218  Greg              initiates requests treatment   32    17.4   1        2
#> 219  Greg              initiates requests treatment   33    31.8   1        3
#> 220  Greg              initiates requests treatment   34    15.1   1        4
#> 221  Greg               secures attention  baseline    1     1.9   0        0
#> 222  Greg               secures attention  baseline    2     0.0   0        0
#> 223  Greg               secures attention  baseline    3     7.1   0        0
#> 224  Greg               secures attention  baseline    4     2.6   0        0
#> 225  Greg               secures attention  baseline    5     0.0   0        0
#> 226  Greg               secures attention  baseline    6     2.9   0        0
#> 227  Greg               secures attention  baseline    7     7.1   0        0
#> 228  Greg               secures attention  baseline    8     4.5   0        0
#> 229  Greg               secures attention  baseline    9     2.9   0        0
#> 230  Greg               secures attention  baseline   10    18.1   0        0
#> 231  Greg               secures attention  baseline   11     8.1   0        0
#> 232  Greg               secures attention  baseline   12     8.1   0        0
#> 233  Greg               secures attention  baseline   13     7.1   0        0
#> 234  Greg               secures attention  baseline   14     2.9   0        0
#> 235  Greg               secures attention treatment   15    26.8   1        1
#> 236  Greg               secures attention treatment   16    15.8   1        2
#> 237  Greg               secures attention treatment   17    14.5   1        3
#> 238  Greg               secures attention treatment   18    16.1   1        4
#> 239  Greg               secures attention treatment   19    17.4   1        5
#> 240  Greg               secures attention treatment   20    10.0   1        6
#> 241  Greg               secures attention treatment   21    13.2   1        7
#> 242  Greg               secures attention treatment   22    18.7   1        8
#> 243  Greg               secures attention treatment   23     9.7   1        9
#> 244  Ivan              initiates comments  baseline    1    13.1   0        0
#> 245  Ivan              initiates comments  baseline    2     8.1   0        0
#> 246  Ivan              initiates comments  baseline    3     9.0   0        0
#> 247  Ivan              initiates comments  baseline    4     4.1   0        0
#> 248  Ivan              initiates comments  baseline    5    14.2   0        0
#> 249  Ivan              initiates comments  baseline    6     4.1   0        0
#> 250  Ivan              initiates comments  baseline    7    13.2   0        0
#> 251  Ivan              initiates comments  baseline    8    18.1   0        0
#> 252  Ivan              initiates comments  baseline    9     4.1   0        0
#> 253  Ivan              initiates comments  baseline   10     5.4   0        0
#> 254  Ivan              initiates comments  baseline   11     4.2   0        0
#> 255  Ivan              initiates comments  baseline   12    10.0   0        0
#> 256  Ivan              initiates comments  baseline   13     5.1   0        0
#> 257  Ivan              initiates comments treatment   14    12.1   1        1
#> 258  Ivan              initiates comments treatment   15     8.3   1        2
#> 259  Ivan              initiates comments treatment   16     0.0   1        3
#> 260  Ivan              initiates comments treatment   17     9.0   1        4
#> 261  Ivan              initiates comments treatment   18    14.1   1        5
#> 262  Ivan              initiates comments treatment   19     9.4   1        6
#> 263  Ivan              initiates comments treatment   20     9.4   1        7
#> 264  Ivan              initiates comments treatment   21     5.2   1        8
#> 265  Ivan              initiates comments treatment   22    19.3   1        9
#> 266  Ivan              initiates comments treatment   23    12.4   1       10
#> 267  Ivan              initiates comments treatment   24    23.2   1       11
#> 268  Ivan              initiates comments treatment   25    16.2   1       12
#> 269  Ivan              initiates comments treatment   26    17.2   1       13
#> 270  Ivan               secures attention  baseline    1     2.0   0        0
#> 271  Ivan               secures attention  baseline    2     1.0   0        0
#> 272  Ivan               secures attention  baseline    3     2.0   0        0
#> 273  Ivan               secures attention  baseline    4     0.0   0        0
#> 274  Ivan               secures attention  baseline    5     0.0   0        0
#> 275  Ivan               secures attention  baseline    6     0.0   0        0
#> 276  Ivan               secures attention  baseline    7     5.0   0        0
#> 277  Ivan               secures attention  baseline    8     6.0   0        0
#> 278  Ivan               secures attention  baseline    9     0.0   0        0
#> 279  Ivan               secures attention  baseline   10     1.0   0        0
#> 280  Ivan               secures attention  baseline   11     2.0   0        0
#> 281  Ivan               secures attention  baseline   12     5.9   0        0
#> 282  Ivan               secures attention  baseline   13     0.0   0        0
#> 283  Ivan               secures attention  baseline   14     0.0   0        0
#> 284  Ivan               secures attention  baseline   15     0.9   0        0
#> 285  Ivan               secures attention  baseline   16     0.0   0        0
#> 286  Ivan               secures attention  baseline   17     0.0   0        0
#> 287  Ivan               secures attention  baseline   18     0.9   0        0
#> 288  Ivan               secures attention  baseline   19     0.0   0        0
#> 289  Ivan               secures attention  baseline   20     0.0   0        0
#> 290  Ivan               secures attention  baseline   21     0.9   0        0
#> 291  Ivan               secures attention  baseline   22     5.0   0        0
#> 292  Ivan               secures attention  baseline   23     0.0   0        0
#> 293  Ivan               secures attention  baseline   24    15.0   0        0
#> 294  Ivan               secures attention  baseline   25     4.0   0        0
#> 295  Ivan               secures attention  baseline   26     1.0   0        0
#> 296  Ivan               secures attention treatment   27     7.9   1        1
#> 297  Ivan               secures attention treatment   28    13.0   1        2
#> 298  Ivan               secures attention treatment   29    11.9   1        3
#> 299  Ivan               secures attention treatment   30    16.8   1        4
#> 300  John            contingent responses  baseline    1    24.5   0        0
#> 301  John            contingent responses  baseline    2     8.6   0        0
#> 302  John            contingent responses  baseline    3    25.5   0        0
#> 303  John            contingent responses  baseline    4    10.0   0        0
#> 304  John            contingent responses  baseline    5     3.7   0        0
#> 305  John            contingent responses  baseline    6    14.3   0        0
#> 306  John            contingent responses  baseline    7     7.5   0        0
#> 307  John            contingent responses  baseline    8    12.8   0        0
#> 308  John            contingent responses  baseline    9    33.1   0        0
#> 309  John            contingent responses  baseline   10     6.9   0        0
#> 310  John            contingent responses  baseline   11     5.4   0        0
#> 311  John            contingent responses  baseline   12     8.3   0        0
#> 312  John            contingent responses  baseline   13    20.4   0        0
#> 313  John            contingent responses  baseline   14    16.0   0        0
#> 314  John            contingent responses  baseline   15     3.4   0        0
#> 315  John            contingent responses  baseline   16    13.5   0        0
#> 316  John            contingent responses  baseline   17    13.4   0        0
#> 317  John            contingent responses  baseline   18    23.6   0        0
#> 318  John            contingent responses  baseline   19    19.7   0        0
#> 319  John            contingent responses  baseline   20     9.0   0        0
#> 320  John            contingent responses  baseline   21     9.9   0        0
#> 321  John            contingent responses treatment   22    16.2   1        1
#> 322  John            contingent responses treatment   23    17.6   1        2
#> 323  John            contingent responses treatment   24    16.6   1        3
#> 324  John            contingent responses treatment   25    14.1   1        4
#> 325  John            contingent responses treatment   26    21.4   1        5
#> 326  John            contingent responses treatment   27    15.5   1        6
#> 327  John            contingent responses treatment   28    12.1   1        7
#> 328  John initiates comments and requests  baseline    1    24.5   0        0
#> 329  John initiates comments and requests  baseline    2    19.1   0        0
#> 330  John initiates comments and requests  baseline    3    15.3   0        0
#> 331  John initiates comments and requests  baseline    4     9.5   0        0
#> 332  John initiates comments and requests  baseline    5    17.4   0        0
#> 333  John initiates comments and requests  baseline    6    15.5   0        0
#> 334  John initiates comments and requests  baseline    7     9.2   0        0
#> 335  John initiates comments and requests  baseline    8     2.5   0        0
#> 336  John initiates comments and requests  baseline    9    11.3   0        0
#> 337  John initiates comments and requests treatment   10    32.8   1        1
#> 338  John initiates comments and requests treatment   11    18.8   1        2
#> 339  John initiates comments and requests treatment   12    15.4   1        3
#> 340  John initiates comments and requests treatment   13    50.6   1        4
#> 341  John initiates comments and requests treatment   14    22.4   1        5
#> 342  John initiates comments and requests treatment   15    32.2   1        6
#> 343  John               secures attention  baseline    1    24.5   0        0
#> 344  John               secures attention  baseline    2    12.3   0        0
#> 345  John               secures attention  baseline    3     2.5   0        0
#> 346  John               secures attention  baseline    4     8.5   0        0
#> 347  John               secures attention  baseline    5     6.5   0        0
#> 348  John               secures attention  baseline    6     6.0   0        0
#> 349  John               secures attention  baseline    7     0.2   0        0
#> 350  John               secures attention  baseline    8     0.2   0        0
#> 351  John               secures attention  baseline    9     5.1   0        0
#> 352  John               secures attention  baseline   10     6.6   0        0
#> 353  John               secures attention  baseline   11     0.0   0        0
#> 354  John               secures attention  baseline   12     2.7   0        0
#> 355  John               secures attention  baseline   13     8.2   0        0
#> 356  John               secures attention  baseline   14     1.3   0        0
#> 357  John               secures attention  baseline   15     1.3   0        0
#> 358  John               secures attention treatment   16    17.6   1        1
#> 359  John               secures attention treatment   17    11.7   1        2
#> 360  John               secures attention treatment   18    14.7   1        3
#> 361  John               secures attention treatment   19    11.3   1        4
#> 362  John               secures attention treatment   20     8.8   1        5
#> 363  John               secures attention treatment   21    14.3   1        6