Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object
a g_REML object

nsim
number of models to simulate

seed
seed value. See documentation for simulate

parallel
if TRUE, run in parallel using foreach backend.

...
additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment, 
                 random = ~ 1 | case, 
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

simulate(Laski_g, nsim = 20)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#>      p_beta  r_theta delta_AB       nu      kappa     g_AB     V_g_AB cnvg_warn
#> 1  30.58691 350.6434 1.633438 43.01321 0.17294137 1.604790 0.06195337         0
#> 2  25.45917 369.2384 1.324924 19.42801 0.14412518 1.273109 0.06882501         0
#> 3  28.51362 312.4883 1.613004 33.40359 0.15817379 1.576515 0.06545153         0
#> 4  26.31315 301.5668 1.515239 22.56317 0.14329592 1.464308 0.07408421         0
#> 5  29.91532 544.7585 1.281715 11.71542 0.09919810 1.197873 0.08678418         0
#> 6  28.51201 489.7174 1.288413 17.36974 0.12868302 1.231969 0.06761377         0
#> 7  29.86602 262.7208 1.842596 31.60088 0.15327998 1.798516 0.07920355         0
#> 8  30.37861 757.7164 1.103607 12.22817 0.11712090 1.034506 0.06846706         0
#> 9  31.86545 390.6605 1.612205 30.38764 0.16532059 1.572084 0.07190390         0
#> 10 33.76747 784.9345 1.205263 10.24874 0.09758873 1.114857 0.08850075         0
#> 11 31.44918 254.1466 1.972728 70.01278 0.19958076 1.951520 0.06827691         0
#> 12 32.24431 327.5803 1.781534 32.12204 0.13620211 1.739612 0.06970074         0
#> 13 27.25406 386.8435 1.385682 15.18663 0.11869208 1.316104 0.08204524         0
#> 14 32.89969 383.7755 1.679396 33.24005 0.17586212 1.641217 0.07504063         0
#> 15 31.66597 608.3376 1.283868 14.87395 0.13909898 1.218024 0.07929311         0
#> 16 32.32817 276.1919 1.945251 39.85498 0.18857420 1.908414 0.08460376         0
#> 17 32.66023 642.3412 1.288654 13.36545 0.13056724 1.214963 0.08472296         0
#> 18 31.68119 524.6278 1.383171 15.05464 0.13433213 1.313100 0.08655112         0
#> 19 26.13145 288.9709 1.537221 23.93946 0.14937143 1.488553 0.07415114         0
#> 20 32.45690 408.7668 1.605348 18.67154 0.14764486 1.539989 0.09509968         0
#>    sigma_sq         phi Tau.case.var((Intercept))
#> 1  242.9630  0.19688892                 107.68046
#> 2  167.7876  0.23837698                 201.45081
#> 3  192.6413  0.15168589                 119.84701
#> 4  150.1146  0.16243847                 151.45222
#> 5  135.9075  0.13014632                 408.85100
#> 6  199.1112  0.15299257                 290.60627
#> 7  157.0891  0.12901905                 105.63175
#> 8  210.3096  0.30114547                 547.40685
#> 9  234.4482  0.23647424                 156.21229
#> 10 154.1102  0.28505173                 630.82429
#> 11 213.6405  0.27205780                  40.50611
#> 12 193.1267 -0.01550054                 134.45353
#> 13 137.3752  0.13443367                 249.46826
#> 14 244.0416  0.28983674                 139.73393
#> 15 222.8312  0.36260672                 385.50643
#> 16 192.8379  0.33103431                  83.35402
#> 17 206.6289  0.36564008                 435.71229
#> 18 191.3760  0.30232150                 333.25178
#> 19 150.2645  0.19100887                 138.70643
#> 20 182.3689  0.29221256                 226.39790