Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object

a g_REML object

nsim

number of models to simulate

seed

seed value. See documentation for simulate

parallel

if TRUE, run in parallel using foreach backend.

...

additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment,
                 random = ~ 1 | case,
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)

suppressWarnings(
  Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))
)

if (requireNamespace("plyr", quietly = TRUE)) {
  simulate(Laski_g, nsim = 20)
}
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#>      p_beta  r_theta  delta_AB       nu     kappa      g_AB     V_g_AB
#> 1  24.97790 491.9619 1.1261341 19.12689 0.1569867 1.0813916 0.06022382
#> 2  36.42980 512.6560 1.6089545 13.65138 0.1233379 1.5189104 0.11781471
#> 3  32.92101 440.4694 1.5686112 25.96318 0.1533276 1.5228581 0.07312228
#> 4  31.51054 392.2636 1.5909879 18.03434 0.1303781 1.5238929 0.09154635
#> 5  31.35911 423.1612 1.5244415 18.30972 0.1467371 1.4611332 0.08907058
#> 6  31.39641 532.4348 1.3606510 19.79537 0.1581472 1.3084397 0.07478837
#> 7  29.19766 586.1819 1.2059571 15.46620 0.1341003 1.1465160 0.06873732
#> 8  28.05484 544.0922 1.2027392 19.27396 0.1696019 1.1553225 0.06904138
#> 9  27.09684 412.6841 1.3338585 16.44255 0.1256555 1.2720773 0.07372797
#> 10 29.04321 335.4181 1.5858107 21.00751 0.1343464 1.5285130 0.08114168
#> 11 30.50981 487.9811 1.3811411 14.34051 0.1144008 1.3076267 0.08485405
#> 12 30.41190 367.2011 1.5870534 23.12559 0.1468014 1.5350202 0.07878207
#> 13 30.39961 355.1740 1.6130484 22.54994 0.1628333 1.5587978 0.08731925
#> 14 28.69993 241.7804 1.8457393 73.04144 0.2046373 1.8267218 0.06578059
#> 15 28.66203 532.1752 1.2424521 19.50692 0.1477058 1.1940623 0.06399717
#> 16 29.02541 278.8735 1.7381001 38.82395 0.1862902 1.7043059 0.07501059
#> 17 25.03935 698.4431 0.9474525 12.36768 0.1060982 0.8888118 0.05116806
#> 18 27.38581 203.5827 1.9193545 84.79481 0.2355625 1.9023279 0.07777842
#> 19 28.45547 232.8081 1.8649480 50.08441 0.2009637 1.8368809 0.07616305
#> 20 29.66976 528.1081 1.2910784 15.42362 0.1198886 1.2272631 0.07247922
#>    cnvg_warn sigma_sq       phi Tau.case.var((Intercept))
#> 1          0 228.8080 0.3663765              2.631539e+02
#> 2          0 164.9136 0.2646330              3.477424e+02
#> 3          0 240.5538 0.1935610              1.999156e+02
#> 4          0 164.6825 0.1489108              2.275811e+02
#> 5          0 185.8815 0.2948812              2.372797e+02
#> 6          0 253.6369 0.3573334              2.787979e+02
#> 7          0 218.8497 0.2798693              3.673323e+02
#> 8          0 267.9799 0.4917426              2.761124e+02
#> 9          0 159.5160 0.1550017              2.531681e+02
#> 10         0 157.5716 0.1129022              1.778465e+02
#> 11         0 162.2651 0.1288994              3.257160e+02
#> 12         0 186.4960 0.1828046              1.807051e+02
#> 13         0 184.2449 0.3323096              1.709292e+02
#> 14         0 207.9877 0.2951747              3.379267e+01
#> 15         0 244.3966 0.2689981              2.877786e+02
#> 16         0 191.9607 0.3217001              8.691277e+01
#> 17         0 191.2895 0.1590002              5.071537e+02
#> 18         0 203.5827 0.4358675              1.728111e-06
#> 19         0 179.7024 0.3578610              5.310570e+01
#> 20         0 190.7424 0.1366259              3.373657e+02