Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object
a g_REML object

nsim
number of models to simulate

seed
seed value. See documentation for simulate

parallel
if TRUE, run in parallel using foreach backend.

...
additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment,
                 random = ~ 1 | case,
                 correlation = corAR1(0, ~ time | case),
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

simulate(Laski_g, nsim = 20)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#>      p_beta  r_theta delta_AB        nu      kappa     g_AB     V_g_AB
#> 1  30.08566 319.8985 1.682106  38.08131 0.15827745 1.648759 0.06345377
#> 2  35.98587 506.1877 1.599470  18.13107 0.15390025 1.532382 0.09881082
#> 3  32.31619 431.4105 1.555875  18.41384 0.13322459 1.491632 0.08753525
#> 4  24.49810 264.0266 1.507678  34.05663 0.17358649 1.474230 0.06489063
#> 5  34.56217 382.5502 1.767082  27.18687 0.18093914 1.717881 0.09282090
#> 6  28.99819 320.7244 1.619217  38.68138 0.17061274 1.587617 0.06420827
#> 7  31.80785 802.9785 1.122490  12.06564 0.12681269 1.051240 0.07365497
#> 8  29.87095 250.2919 1.888103  40.24287 0.19365255 1.852695 0.08330234
#> 9  28.74957 390.3231 1.455188  20.61079 0.14745964 1.401586 0.07612451
#> 10 30.96882 385.9029 1.576470  46.20857 0.20461755 1.550743 0.06976975
#> 11 28.09908 401.1602 1.402921  20.05870 0.15193250 1.349803 0.07517207
#> 12 33.29262 561.1889 1.405379  13.64931 0.12774028 1.326716 0.09481976
#> 13 34.34023 434.3010 1.647812  18.50363 0.14816174 1.580108 0.09989115
#> 14 34.73462 350.0773 1.856438  32.76351 0.17262403 1.813615 0.08437934
#> 15 32.54514 560.4227 1.374765  17.39416 0.14176205 1.314623 0.07817842
#> 16 32.94123 926.8438 1.082023  10.15023 0.09153237 1.000054 0.07276767
#> 17 29.02143 277.8836 1.740954  55.17423 0.18956823 1.717181 0.06419513
#> 18 37.05340 244.0557 2.371830 101.73176 0.18062354 2.354301 0.06068068
#> 19 33.72578 579.6921 1.400758  13.46097 0.11643018 1.321236 0.09260653
#> 20 25.66130 534.2725 1.110190  14.77778 0.13308643 1.052876 0.06297934
#>    cnvg_warn sigma_sq        phi Tau.case.var((Intercept))
#> 1          0 207.5684 0.11737096                 112.33015
#> 2          0 225.4934 0.36964301                 280.69432
#> 3          0 184.8757 0.16497967                 246.53478
#> 4          0 168.6092 0.26488478                  95.41734
#> 5          0 228.1975 0.40502573                 154.35271
#> 6          0 213.3954 0.20656904                 107.32896
#> 7          0 228.0850 0.44055090                 574.89349
#> 8          0 177.7209 0.36586672                  72.57095
#> 9          0 185.4923 0.23923996                 204.83085
#> 10         0 294.8921 0.40606199                  91.01087
#> 11         0 189.3867 0.29296911                 211.77351
#> 12         0 182.7479 0.31334623                 378.44104
#> 13         0 192.8434 0.30221296                 241.45766
#> 14         0 219.8259 0.26942047                 130.25134
#> 15         0 234.4214 0.27798226                 326.00128
#> 16         0 174.9624 0.21481984                 751.88133
#> 17         0 214.3595 0.25657730                  63.52405
#> 18         0 219.9695 0.07917553                  24.08625
#> 19         0 180.5392 0.20063755                 399.15298
#> 20         0 190.9759 0.30373309                 343.29659