Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object

a g_REML object

nsim

number of models to simulate

seed

seed value. See documentation for simulate

parallel

if TRUE, run in parallel using foreach backend.

...

additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment,
                 random = ~ 1 | case,
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)

suppressWarnings(
  Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))
)

if (requireNamespace("plyr", quietly = TRUE)) {
  simulate(Laski_g, nsim = 20)
}
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#>      p_beta  r_theta  delta_AB        nu      kappa      g_AB     V_g_AB
#> 1  26.59413 405.0667 1.3213640 26.420105 0.15166942 1.2834955 0.05770833
#> 2  26.06392 721.7017 0.9701992 12.097631 0.11217676 0.9087819 0.05549900
#> 3  31.73751 496.0484 1.4249868 17.105464 0.14523780 1.3615806 0.08459974
#> 4  32.35655 570.8563 1.3542499 12.529825 0.11223511 1.2715380 0.09247474
#> 5  32.25010 350.5332 1.7225286 33.259683 0.16377725 1.6833917 0.07311051
#> 6  33.73461 868.3712 1.1447825  9.671419 0.08180904 1.0536511 0.08261525
#> 7  30.24085 420.7286 1.4743241 18.030629 0.13337360 1.4121360 0.08192007
#> 8  32.32279 634.1966 1.2835028 12.285746 0.11253144 1.2035222 0.08604110
#> 9  27.59991 505.3949 1.2276999 12.704263 0.11674691 1.1537674 0.07845168
#> 10 29.61110 253.3290 1.8604245 31.796324 0.17255161 1.8161937 0.08630518
#> 11 35.15072 602.2382 1.4323531 13.196538 0.12349953 1.3493761 0.09975732
#> 12 25.84104 406.1376 1.2822519 16.689012 0.13810703 1.2237515 0.07192996
#> 13 31.40883 603.0372 1.2790269 12.545109 0.11789819 1.2010065 0.08518628
#> 14 35.85654 306.9965 2.0464526 24.343550 0.14695377 1.9827492 0.11147211
#> 15 27.17689 485.7882 1.2330374 16.370585 0.13201163 1.1756711 0.06731032
#> 16 29.92589 732.3672 1.1058154 11.608506 0.11120898 1.0327987 0.07049392
#> 17 34.40778 455.1180 1.6128525 18.231633 0.13888849 1.5455816 0.09508114
#> 18 30.74615 334.4558 1.6812077 28.265768 0.16432318 1.6362007 0.07918637
#> 19 34.58222 425.8092 1.6758895 21.513515 0.16889432 1.6167781 0.09747092
#> 20 29.67765 412.0829 1.4619659 19.144457 0.15212127 1.4039344 0.08249891
#>    cnvg_warn sigma_sq       phi Tau.case.var((Intercept))
#> 1          0 222.5003 0.1735588                 182.56648
#> 2          0 194.6428 0.2536869                 527.05888
#> 3          0 206.8314 0.3223019                 289.21702
#> 4          0 161.3391 0.2182777                 409.51713
#> 5          0 217.7053 0.1965525                 132.82797
#> 6          0 142.4772 0.1533712                 725.89409
#> 7          0 177.5646 0.1775234                 243.16402
#> 8          0 174.8015 0.2415697                 459.39511
#> 9          0 147.0492 0.2569373                 358.34573
#> 10         0 157.1927 0.2784093                  96.13631
#> 11         0 186.7380 0.2972320                 415.50026
#> 12         0 163.1357 0.2683223                 243.00181
#> 13         0 173.2954 0.2829998                 429.74182
#> 14         0 160.4087 0.1639514                 146.58781
#> 15         0 189.4546 0.2206334                 296.33361
#> 16         0 185.8369 0.2892412                 546.53031
#> 17         0 195.7651 0.2235387                 259.35282
#> 18         0 194.1335 0.2529748                 140.32238
#> 19         0 219.8559 0.4117547                 205.95327
#> 20         0 189.0539 0.3199004                 223.02900