Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object
a g_REML object

nsim
number of models to simulate

seed
seed value. See documentation for simulate

parallel
if TRUE, run in parallel using foreach backend.

...
additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment, 
                 random = ~ 1 | case, 
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

simulate(Laski_g, nsim = 20)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#>      p_beta  r_theta delta_AB        nu      kappa      g_AB     V_g_AB
#> 1  30.78368 281.4761 1.834846 42.153002 0.18135868 1.8020055 0.07410825
#> 2  27.66366 306.7513 1.579488 39.711716 0.19329829 1.5494691 0.07000434
#> 3  26.45511 587.5959 1.091365 18.239912 0.16367171 1.0458663 0.06203979
#> 4  30.21580 785.8715 1.077850  9.911992 0.08716922 0.9941833 0.07317338
#> 5  26.11061 603.5308 1.062839 14.720804 0.11368189 1.0077534 0.05445100
#> 6  29.46074 446.1050 1.394843 18.834197 0.13596964 1.3385511 0.07338219
#> 7  26.17675 501.4267 1.168993 13.176087 0.11672571 1.1011656 0.07016649
#> 8  33.79532 687.7421 1.288676 12.648799 0.12022964 1.2107244 0.08619043
#> 9  30.36690 477.7864 1.389260 13.288531 0.12928045 1.3093476 0.09572091
#> 10 26.29463 424.6095 1.276063 18.760547 0.14262744 1.2243603 0.06662476
#> 11 29.82283 235.7457 1.942348 59.948340 0.19813724 1.9179456 0.07154074
#> 12 30.60556 420.9324 1.491743 23.290021 0.15992062 1.4431840 0.07592184
#> 13 31.05227 262.6208 1.916147 41.785522 0.17839548 1.8815476 0.07712885
#> 14 36.62906 495.2315 1.645969 18.319706 0.14704725 1.5776513 0.10020494
#> 15 33.83958 439.8941 1.613433 20.449342 0.15616790 1.5535263 0.09172879
#> 16 28.14355 347.7636 1.509166 36.150003 0.17345536 1.4776370 0.06284082
#> 17 30.94687 548.2043 1.321738 13.286494 0.12180017 1.2456978 0.08634795
#> 18 30.87299 387.2580 1.568839 27.014215 0.15848604 1.5248764 0.07277513
#> 19 32.65272 300.5769 1.883396 34.182975 0.16818609 1.8417684 0.08202665
#> 20 30.52968 457.2770 1.427686 22.273893 0.12743019 1.3790674 0.06437514
#>    cnvg_warn sigma_sq        phi Tau.case.var((Intercept))
#> 1          0 196.8319 0.26307048                  84.64420
#> 2          0 216.7074 0.36721917                  90.04392
#> 3          0 273.2748 0.46714823                 314.32115
#> 4          0 138.9422 0.19090387                 646.92927
#> 5          0 205.9580 0.10268354                 397.57280
#> 6          0 195.2777 0.17906303                 250.82729
#> 7          0 152.4416 0.22252852                 348.98507
#> 8          0 201.0313 0.30173422                 486.71086
#> 9          0 152.0761 0.35652735                 325.71026
#> 10         0 187.8461 0.24292151                 236.76339
#> 11         0 189.5929 0.29578524                  46.15285
#> 12         0 220.4316 0.29230195                 200.50089
#> 13         0 182.0345 0.24381917                  80.58625
#> 14         0 217.7924 0.29749187                 277.43912
#> 15         0 212.3768 0.32138882                 227.51728
#> 16         0 227.0032 0.24637704                 120.76034
#> 17         0 170.4640 0.27177285                 377.74033
#> 18         0 217.6466 0.22171160                 169.61139
#> 19         0 190.1953 0.22249600                 110.38153
#> 20         0 220.6976 0.02588465                 236.57945