Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object
a g_REML object

nsim
number of models to simulate

seed
seed value. See documentation for simulate

parallel
if TRUE, run in parallel using foreach backend.

...
additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment,
                 random = ~ 1 | case,
                 correlation = corAR1(0, ~ time | case),
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))
simulate(Laski_g, nsim = 20)

#>      p_beta  r_theta delta_AB        nu      kappa      g_AB     V_g_AB
#> 1  31.18724 929.5305 1.022928  9.834648 0.08387291 0.9428838 0.06663316
#> 2  24.96948 435.7257 1.196198 15.904382 0.14427079 1.1388881 0.06939451
#> 3  29.86110 250.3779 1.887156 40.770553 0.16691269 1.8522264 0.07288374
#> 4  32.94596 375.0714 1.701160 31.845140 0.16564853 1.6607783 0.07467445
#> 5  32.34954 388.0592 1.642174 30.875940 0.18280018 1.6019585 0.07899535
#> 6  32.88510 495.3139 1.477607 15.790471 0.14363547 1.4062959 0.09491601
#> 7  27.99305 378.1543 1.439513 20.929018 0.16932344 1.3873040 0.08123493
#> 8  31.65345 391.6620 1.599430 32.176677 0.18249187 1.5618576 0.07476624
#> 9  34.83347 391.0881 1.761406 26.097539 0.15472592 1.7102965 0.08604357
#> 10 25.35501 254.5464 1.589206 59.514703 0.19936276 1.5690948 0.06163413
#> 11 27.73393 375.5571 1.431112 20.281188 0.14157981 1.3775285 0.07351164
#> 12 29.03901 739.1616 1.068100 12.108491 0.09951837 1.0005476 0.06159274
#> 13 32.68723 469.5571 1.508460 16.802002 0.12742515 1.4401094 0.08851294
#> 14 30.38417 298.6825 1.758096 24.134142 0.14065736 1.7028887 0.08680525
#> 15 29.74626 462.9280 1.382534 13.049561 0.10619890 1.3015229 0.09083465
#> 16 33.22494 314.5300 1.873411 29.181958 0.17352315 1.8248471 0.09274115
#> 17 32.23083 555.4004 1.367629 13.903272 0.13867745 1.2925027 0.09226832
#> 18 32.50267 569.8749 1.361536 14.130728 0.12698227 1.2879701 0.08712963
#> 19 31.17475 612.3176 1.259838 14.668516 0.11768521 1.1943053 0.07224737
#> 20 29.60238 190.0819 2.147119 79.115010 0.19727952 2.1267003 0.06863891
#>    cnvg_warn sigma_sq       phi Tau.case.var((Intercept))
#> 1          0 159.8617 0.1553372                 769.66882
#> 2          0 171.4993 0.3638856                 264.22634
#> 3          0 168.6669 0.1656008                  81.71091
#> 4          0 229.6581 0.2241855                 145.41327
#> 5          0 244.3515 0.3694748                 143.70770
#> 6          0 193.4572 0.3627511                 301.85666
#> 7          0 193.0099 0.4320538                 185.14436
#> 8          0 250.1583 0.3522213                 141.50372
#> 9          0 214.6691 0.2032650                 176.41902
#> 10         0 204.9194 0.3055239                  49.62704
#> 11         0 174.6256 0.1941514                 200.93146
#> 12         0 193.8646 0.1005236                 545.29698
#> 13         0 185.4263 0.1596795                 284.13084
#> 14         0 153.7989 0.1131526                 144.88360
#> 15         0 136.0195 0.1116532                 326.90855
#> 16         0 188.9445 0.3154443                 125.58544
#> 17         0 191.6271 0.4209966                 363.77337
#> 18         0 191.8742 0.2749731                 378.00075
#> 19         0 209.7194 0.1477270                 402.59820
#> 20         0 164.1920 0.2329052                  25.88992