Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object

a g_REML object

nsim

number of models to simulate

seed

seed value. See documentation for simulate

parallel

if TRUE, run in parallel using foreach backend.

...

additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment,
                 random = ~ 1 | case,
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)

suppressWarnings(
  Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))
)

if (requireNamespace("plyr", quietly = TRUE)) {
  simulate(Laski_g, nsim = 20)
}
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#>      p_beta  r_theta  delta_AB       nu     kappa      g_AB     V_g_AB
#> 1  22.30203 628.3714 0.8896850 14.51188 0.1453929 0.8428985 0.05118517
#> 2  23.71540 630.3440 0.9445863 13.92978 0.1200416 0.8927989 0.04940497
#> 3  30.00090 366.2114 1.5677194 22.85992 0.1708396 1.5157162 0.08590859
#> 4  31.66116 329.5493 1.7440811 39.02430 0.1835888 1.7103458 0.07405887
#> 5  37.13939 247.6462 2.3600379 37.87997 0.1648523 2.3130002 0.10287032
#> 6  32.41721 276.3604 1.9500141 31.63671 0.1643217 1.9034176 0.08933396
#> 7  33.37595 550.1415 1.4229720 13.77948 0.1170856 1.3440903 0.09323455
#> 8  30.10848 340.0939 1.6326360 26.87703 0.1705640 1.5866497 0.08101595
#> 9  28.94725 353.1755 1.5403238 20.06661 0.1325678 1.4820271 0.08005026
#> 10 28.98658 246.1914 1.8473982 56.16984 0.1877688 1.8226207 0.06645545
#> 11 31.10194 368.9489 1.6192146 22.02548 0.1524251 1.5634450 0.08594233
#> 12 28.29430 848.2259 0.9715009 13.31080 0.1363156 0.9157136 0.05762144
#> 13 30.70048 325.6848 1.7011650 26.96616 0.1562478 1.6534084 0.08044870
#> 14 31.08522 322.8970 1.7299036 30.07598 0.1653358 1.6864036 0.07912933
#> 15 29.73734 474.4424 1.3652445 15.94937 0.1458466 1.3000231 0.08414727
#> 16 33.81079 817.4655 1.1825530 11.47393 0.1024151 1.1035330 0.07762325
#> 17 27.11895 654.3245 1.0601721 12.53870 0.1053656 0.9954680 0.06018307
#> 18 30.32690 513.0796 1.3388614 13.22946 0.1100224 1.2614972 0.08568299
#> 19 30.65428 323.4339 1.7045054 32.61583 0.1896493 1.6650076 0.08237271
#> 20 26.99570 360.7714 1.4212766 20.56411 0.1499817 1.3688028 0.07454146
#>    cnvg_warn sigma_sq       phi Tau.case.var((Intercept))
#> 1          0 231.7815 0.4584052                 396.58991
#> 2          0 205.5689 0.2118966                 424.77509
#> 3          0 195.9809 0.3972331                 170.23050
#> 4          0 225.7657 0.3002249                 103.78353
#> 5          0 161.9574 0.1688195                  85.68878
#> 6          0 168.3476 0.2158568                 108.01278
#> 7          0 176.1531 0.1885248                 373.98841
#> 8          0 195.8050 0.3223938                 144.28888
#> 9          0 160.5676 0.1169881                 192.60787
#> 10         0 190.1842 0.2404268                  56.00725
#> 11         0 184.2915 0.2522753                 184.65739
#> 12         0 278.5825 0.4399595                 569.64342
#> 13         0 182.1305 0.2040418                 143.55434
#> 14         0 192.9428 0.2399642                 129.95419
#> 15         0 188.1585 0.3782726                 286.28390
#> 16         0 198.8800 0.1933775                 618.58546
#> 17         0 182.2805 0.1373652                 472.04394
#> 18         0 154.4562 0.1436671                 358.62339
#> 19         0 212.1631 0.4047449                 111.27080
#> 20         0 172.1458 0.2628761                 188.62559