Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object
a g_REML object

nsim
number of models to simulate

seed
seed value. See documentation for simulate

parallel
if TRUE, run in parallel using foreach backend.

...
additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment, 
                 random = ~ 1 | case, 
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

simulate(Laski_g, nsim = 20)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#>      p_beta  r_theta delta_AB       nu     kappa     g_AB     V_g_AB cnvg_warn
#> 1  38.28594 490.2770 1.729094 14.38443 0.1292604 1.637345 0.12866228         0
#> 2  31.62513 397.4530 1.586315 30.03206 0.1638998 1.546367 0.07055111         0
#> 3  33.30318 473.8317 1.529937 20.32551 0.1537212 1.472780 0.08461628         0
#> 4  39.20358 472.7048 1.803145 21.42936 0.1546246 1.739292 0.10381218         0
#> 5  31.20921 515.1326 1.375065 13.66015 0.1141253 1.298161 0.08799751         0
#> 6  29.65248 305.8400 1.695563 28.23304 0.1782933 1.650119 0.08501499         0
#> 7  32.11406 320.6265 1.793476 26.82907 0.1791640 1.742868 0.09481296         0
#> 8  30.60007 758.4502 1.111115 12.71602 0.1281738 1.044266 0.06972358         0
#> 9  33.12481 543.2881 1.421145 14.63149 0.1173697 1.347032 0.08813059         0
#> 10 32.77482 200.7092 2.313432 98.11530 0.1914600 2.295702 0.06437051         0
#> 11 28.31224 584.0086 1.171560 14.57215 0.1356614 1.110210 0.06952651         0
#> 12 29.86650 594.7103 1.224705 11.30211 0.1095528 1.141596 0.08525934         0
#> 13 27.36098 214.2967 1.869063 54.74016 0.1750911 1.843337 0.06338782         0
#> 14 33.77587 469.2158 1.559266 16.02759 0.1369640 1.485145 0.10002035         0
#> 15 34.01914 353.7860 1.808644 20.33801 0.1329552 1.741117 0.10242487         0
#> 16 33.06384 347.7243 1.773111 24.10149 0.1572490 1.717356 0.09309973         0
#> 17 29.86055 344.4568 1.608905 36.25758 0.1786123 1.575393 0.06898159         0
#> 18 28.17235 436.4292 1.348547 16.66056 0.1385342 1.286915 0.07769480         0
#> 19 33.66364 345.5591 1.810923 54.91779 0.2067870 1.786078 0.07351913         0
#> 20 30.74986 450.1836 1.449267 16.85393 0.1326042 1.383803 0.08416936         0
#>    sigma_sq       phi Tau.case.var((Intercept))
#> 1  169.0393 0.2847898                 321.23775
#> 2  236.6868 0.2290820                 160.76617
#> 3  226.4996 0.3023556                 247.33210
#> 4  233.5925 0.2842480                 239.11228
#> 5  162.3346 0.1630742                 352.79808
#> 6  183.6155 0.3668477                 122.22449
#> 7  189.1205 0.3954581                 131.50601
#> 8  229.4392 0.3909265                 529.01101
#> 9  185.4887 0.1461794                 357.79943
#> 10 182.0401 0.1547845                  18.66915
#> 11 207.7089 0.3426996                 376.29974
#> 12 144.2325 0.3021511                 450.47779
#> 13 160.7095 0.1583258                  53.58726
#> 14 181.8999 0.2837735                 287.31591
#> 15 162.3899 0.1144710                 191.39613
#> 16 184.2596 0.2550496                 163.46466
#> 17 227.5158 0.2843089                 116.94095
#> 18 175.2498 0.2736057                 261.17943
#> 19 277.9987 0.3744585                  67.56047
#> 20 180.0299 0.2085185                 270.15367