Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object

a g_REML object

nsim

number of models to simulate

seed

seed value. See documentation for simulate

parallel

if TRUE, run in parallel using foreach backend.

...

additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment,
                 random = ~ 1 | case,
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
simulate(Laski_g, nsim = 20)
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#>      p_beta   r_theta delta_AB        nu      kappa      g_AB     V_g_AB
#> 1  33.07967 1010.8290 1.040453  9.849588 0.09699491 0.9591637 0.07110208
#> 2  30.21901  615.2527 1.218298 16.613789 0.14786020 1.1624597 0.06997082
#> 3  31.19831  350.0439 1.667515 21.296167 0.14828471 1.6080919 0.09081218
#> 4  33.06264  479.5159 1.509858 18.592158 0.13237475 1.4481206 0.08260129
#> 5  32.00372  540.2130 1.376949 16.691917 0.13615785 1.3141396 0.07935908
#> 6  32.51574  424.9982 1.577249 29.449186 0.18586801 1.5367360 0.07876804
#> 7  30.28818  294.3773 1.765310 28.763708 0.16047813 1.7188767 0.08218257
#> 8  30.58161  394.5223 1.539659 26.577373 0.17581720 1.4957982 0.07773430
#> 9  27.33111  653.2643 1.069332 15.143358 0.12786916 1.0154830 0.05725563
#> 10 34.37425  639.2494 1.359559 13.656256 0.13423683 1.2835000 0.09156725
#> 11 29.46548  280.8361 1.758276 63.833745 0.17720132 1.7375364 0.05621391
#> 12 28.20787  314.5234 1.590537 31.796013 0.15388456 1.5527226 0.06503259
#> 13 29.33967  398.0258 1.470617 20.060014 0.14452105 1.4149400 0.07799274
#> 14 30.53010  378.5874 1.569080 24.434377 0.17468035 1.5204205 0.08354870
#> 15 29.01896  388.5353 1.472199 22.073135 0.16185106 1.4216037 0.07809446
#> 16 27.36116  650.0128 1.073182 14.892988 0.12739062 1.0182150 0.05816348
#> 17 28.57343  645.9458 1.124254 11.819724 0.11356630 1.0513747 0.07180332
#> 18 31.68405  507.3790 1.406612 13.200504 0.11198559 1.3251514 0.09390961
#> 19 31.02921  512.1200 1.371149 14.174372 0.12236303 1.2972960 0.08668649
#> 20 28.55170  651.3938 1.118691 13.212083 0.13107460 1.0539623 0.06902371
#>    cnvg_warn sigma_sq       phi Tau.case.var((Intercept))
#> 1          0 182.4326 0.3489345                 828.39643
#> 2          0 253.1890 0.3686144                 362.06367
#> 3          0 169.9771 0.2312449                 180.06679
#> 4          0 206.6279 0.1520082                 272.88798
#> 5          0 216.0269 0.2491310                 324.18613
#> 6          0 265.6917 0.4133516                 159.30644
#> 7          0 170.9233 0.2161011                 123.45400
#> 8          0 229.3602 0.3708670                 165.16210
#> 9          0 235.7185 0.2311778                 417.54577
#> 10         0 212.9801 0.3866770                 426.26932
#> 11         0 220.8608 0.1423705                  59.97534
#> 12         0 188.7273 0.1320387                 125.79604
#> 13         0 184.8323 0.2259523                 213.19348
#> 14         0 211.5481 0.3988845                 167.03930
#> 15         0 198.7698 0.3339021                 189.76553
#> 16         0 230.6465 0.2383157                 419.36630
#> 17         0 169.4144 0.2968097                 476.53140
#> 18         0 152.9528 0.1676751                 354.42621
#> 19         0 171.0114 0.2230033                 341.10865
#> 20         0 207.5025 0.3836223                 443.89129