Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment,
                 random = ~ 1 | case,
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
simulate(Laski_g, nsim = 20)
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#>      p_beta  r_theta  delta_AB       nu     kappa      g_AB     V_g_AB
#> 1  32.50048 272.7090 1.9680679 26.92034 0.1661213 1.9127236 0.10262324
#> 2  29.80425 287.3375 1.7582558 45.30590 0.2105448 1.7289879 0.07965222
#> 3  26.91910 515.1144 1.1860655 17.54633 0.1295002 1.1346356 0.05966369
#> 4  28.53232 536.7605 1.2315353 14.53719 0.1191074 1.1668864 0.07057151
#> 5  35.61114 404.4452 1.7707452 32.43707 0.1974146 1.7294845 0.08934225
#> 6  33.24888 233.0684 2.1778872 43.82167 0.1686462 2.1403991 0.08403602
#> 7  36.46327 475.0106 1.6730310 17.75078 0.1520118 1.6013329 0.10707777
#> 8  27.78034 475.4086 1.2741014 23.42702 0.1748376 1.2328720 0.06735446
#> 9  32.19121 570.9776 1.3471866 14.04426 0.1299984 1.2739394 0.08692563
#> 10 31.45656 360.1320 1.6576024 37.49467 0.1700961 1.6242231 0.06688808
#> 11 25.57143 559.5026 1.0810703 14.45154 0.1384206 1.0239777 0.06322600
#> 12 30.69367 377.3059 1.5801635 18.85285 0.1501093 1.5164570 0.09287319
#> 13 35.86533 449.0384 1.6925172 18.70892 0.1394816 1.6237489 0.10061437
#> 14 33.99898 475.3922 1.5593366 19.87928 0.1488016 1.4997571 0.08692263
#> 15 30.35940 636.8062 1.2030663 13.87455 0.1333783 1.1368403 0.07455999
#> 16 29.36047 600.6825 1.1979551 14.22829 0.1261089 1.1336793 0.07060900
#> 17 29.81785 332.2910 1.6357506 30.26194 0.1540421 1.5948731 0.06973129
#> 18 30.38835 654.3629 1.1879491 12.74878 0.1204683 1.1166652 0.07505773
#> 19 25.59138 742.2099 0.9393562 11.20869 0.1061391 0.8750677 0.05499692
#> 20 32.41837 328.0795 1.7897883 22.38693 0.1551448 1.7291502 0.09928519
#>    cnvg_warn sigma_sq       phi Tau.case.var((Intercept))
#> 1          0 155.4569 0.2853009                 117.25211
#> 2          0 223.5343 0.4566088                  63.80317
#> 3          0 211.4195 0.1552939                 303.69489
#> 4          0 182.6759 0.1692460                 354.08466
#> 5          0 273.2126 0.4731024                 131.23253
#> 6          0 161.3661 0.1610165                  71.70227
#> 7          0 207.4290 0.3648020                 267.58161
#> 8          0 260.9926 0.4214715                 214.41601
#> 9          0 192.6933 0.3128242                 378.28429
#> 10         0 236.7302 0.2110129                 123.40172
#> 11         0 199.4255 0.3803601                 360.07714
#> 12         0 170.4834 0.3097874                 206.82246
#> 13         0 196.9719 0.2151876                 252.06648
#> 14         0 221.5288 0.2691919                 253.86347
#> 15         0 214.7130 0.3613189                 422.09327
#> 16         0 203.1956 0.2599941                 397.48692
#> 17         0 195.2045 0.1481504                 137.08646
#> 18         0 193.2283 0.2964410                 461.13465
#> 19         0 175.7634 0.2691556                 566.44647
#> 20         0 166.3631 0.2686233                 161.71632