Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object
a g_REML object

nsim
number of models to simulate

seed
seed value. See documentation for simulate

parallel
if TRUE, run in parallel using foreach backend.

...
additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment,
                 random = ~ 1 | case,
                 correlation = corAR1(0, ~ time | case),
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

simulate(Laski_g, nsim = 20)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#>      p_beta  r_theta delta_AB       nu     kappa      g_AB     V_g_AB cnvg_warn
#> 1  36.18191 421.4381 1.762482 25.61474 0.1796226 1.7103676 0.09583985         0
#> 2  30.52067 431.0632 1.470021 22.85246 0.1491808 1.4212424 0.07206179         0
#> 3  28.50999 478.6313 1.303157 16.27419 0.1195516 1.2421632 0.07017705         0
#> 4  30.26784 266.7822 1.853118 26.18797 0.1547822 1.7995346 0.09239675         0
#> 5  27.09972 628.7620 1.080741 15.84209 0.1502666 1.0287560 0.06260894         0
#> 6  28.87735 415.5545 1.416587 17.46193 0.1346772 1.3548605 0.07944767         0
#> 7  26.24668 439.9391 1.251348 18.85256 0.1440957 1.2008974 0.06507721         0
#> 8  32.02866 638.3015 1.267727 15.34815 0.1344681 1.2047530 0.07456230         0
#> 9  31.66457 394.2488 1.594735 19.26853 0.1290093 1.5318460 0.08647073         0
#> 10 30.09088 369.5835 1.565231 33.85126 0.1823661 1.5302946 0.07096637         0
#> 11 27.15917 294.1642 1.583513 48.61776 0.1926280 1.5589586 0.06377869         0
#> 12 32.10400 419.8607 1.566775 20.15737 0.1549004 1.5077472 0.08849140         0
#> 13 34.01656 506.3589 1.511684 15.93698 0.1477474 1.4394103 0.09882500         0
#> 14 28.33014 343.9677 1.527530 21.01414 0.1534154 1.4723561 0.08223928         0
#> 15 31.25723 539.9590 1.345148 13.68992 0.1272234 1.2700837 0.08792693         0
#> 16 25.98564 424.5837 1.261106 22.20514 0.1580388 1.2180264 0.06296689         0
#> 17 30.74471 355.8128 1.629895 25.92566 0.1586256 1.5822850 0.07880325         0
#> 18 28.49965 404.5152 1.417007 17.31052 0.1433047 1.3547141 0.08253188         0
#> 19 27.83792 743.8938 1.020660 12.08173 0.1134923 0.9559619 0.06039109         0
#> 20 30.46350 452.4112 1.432231 14.42245 0.1299283 1.3564383 0.09369943         0
#>    sigma_sq        phi Tau.case.var((Intercept))
#> 1  244.4380 0.42065854                 177.00012
#> 2  218.4237 0.20821052                 212.63944
#> 3  181.4301 0.09985789                 297.20122
#> 4  146.7015 0.20247279                 120.08068
#> 5  252.5504 0.43207292                 376.21162
#> 6  171.6066 0.20775372                 243.94783
#> 7  195.9120 0.25389650                 244.02712
#> 8  236.9440 0.28921663                 401.35750
#> 9  173.3328 0.10275805                 220.91597
#> 10 240.3418 0.33413813                 129.24175
#> 11 220.3422 0.30718671                  73.82199
#> 12 200.3128 0.31740281                 219.54794
#> 13 202.1149 0.39919759                 304.24394
#> 14 167.5804 0.28294514                 176.38730
#> 15 176.1064 0.30492253                 363.85257
#> 16 215.7821 0.29752731                 208.80155
#> 17 196.1238 0.23792700                 159.68902
#> 18 169.2605 0.29584336                 235.25470
#> 19 201.0131 0.27089817                 542.88066
#> 20 156.6679 0.28976670                 295.74331