Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object

a g_REML object

nsim

number of models to simulate

seed

seed value. See documentation for simulate

parallel

if TRUE, run in parallel using foreach backend.

...

additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment,
                 random = ~ 1 | case,
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)

suppressWarnings(
  Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))
)

if (requireNamespace("plyr", quietly = TRUE)) {
  simulate(Laski_g, nsim = 20)
}
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#>      p_beta  r_theta delta_AB        nu     kappa     g_AB     V_g_AB cnvg_warn
#> 1  32.07350 450.0966 1.511797  18.22369 0.1386987 1.448713 0.08593123         0
#> 2  28.35658 463.7990 1.316706  19.98790 0.1283495 1.266674 0.06241846         0
#> 3  31.89609 504.7637 1.419690  13.13861 0.1021422 1.337077 0.09363779         0
#> 4  33.27809 185.2400 2.445068 107.66618 0.2156683 2.427996 0.07472700         0
#> 5  25.51965 527.0286 1.111623  13.19474 0.1098840 1.047217 0.06311270         0
#> 6  29.92250 608.9476 1.212573  12.84678 0.1245000 1.140378 0.07807504         0
#> 7  37.29864 288.3523 2.196500  35.46439 0.1596122 2.149719 0.09565968         0
#> 8  35.24680 194.5859 2.526760 107.58186 0.2098924 2.509104 0.07418292         0
#> 9  38.26358 811.8819 1.342886  11.68745 0.1094309 1.254828 0.09672381         0
#> 10 32.98171 305.3409 1.887473  33.23204 0.1759581 1.844552 0.08656535         0
#> 11 35.79442 248.4071 2.271085  53.10768 0.1736698 2.238860 0.07984981         0
#> 12 29.12035 453.8515 1.366909  16.94299 0.1209320 1.305495 0.07349139         0
#> 13 28.82146 524.8212 1.258086  18.07738 0.1485282 1.205158 0.06892108         0
#> 14 26.41577 483.0065 1.201951  19.58197 0.1512922 1.155320 0.06227660         0
#> 15 33.37669 364.2394 1.748839  19.48526 0.1525788 1.680650 0.10636602         0
#> 16 25.12826 424.7380 1.219276  20.13494 0.1416342 1.173288 0.05932142         0
#> 17 31.63750 316.3836 1.778669  28.38730 0.1746851 1.731258 0.08867164         0
#> 18 33.24673 575.2540 1.386178  16.04199 0.1317077 1.320346 0.08159897         0
#> 19 30.36435 578.4110 1.262542  13.29670 0.1222195 1.189963 0.08019760         0
#> 20 29.01325 448.2168 1.370416  23.34203 0.1653684 1.325906 0.06987568         0
#>    sigma_sq        phi Tau.case.var((Intercept))
#> 1  193.4681 0.22199520              2.566285e+02
#> 2  208.9615 0.07962729              2.548375e+02
#> 3  148.4979 0.05909912              3.562658e+02
#> 4  185.2400 0.26878502              1.131778e-06
#> 5  158.1079 0.14432073              3.689208e+02
#> 6  183.7652 0.33553294              4.251823e+02
#> 7  182.4460 0.14626832              1.059063e+02
#> 8  190.3875 0.24020731              4.198355e+00
#> 9  206.9561 0.25908714              6.049258e+02
#> 10 194.1895 0.29065202              1.111513e+02
#> 11 184.2077 0.15461385              6.419939e+01
#> 12 178.4850 0.09064459              2.753665e+02
#> 13 229.6162 0.31945593              2.952050e+02
#> 14 224.2886 0.29980443              2.587179e+02
#> 15 169.1751 0.31433901              1.950643e+02
#> 16 196.5933 0.19807674              2.281447e+02
#> 17 188.4595 0.33507263              1.279241e+02
#> 18 220.3954 0.23065512              3.548585e+02
#> 19 180.2182 0.27576058              3.981928e+02
#> 20 238.2697 0.33835255              2.099471e+02