Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object
a g_REML object

nsim
number of models to simulate

seed
seed value. See documentation for simulate

parallel
if TRUE, run in parallel using foreach backend.

...
additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment,
                 random = ~ 1 | case,
                 correlation = corAR1(0, ~ time | case),
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))
simulate(Laski_g, nsim = 20)

#>      p_beta  r_theta delta_AB       nu     kappa     g_AB     V_g_AB cnvg_warn
#> 1  27.91413 231.3497 1.835226 62.84880 0.1895701 1.813238 0.06341412         0
#> 2  28.15247 296.5954 1.634686 25.33358 0.1643169 1.585809 0.08230502         0
#> 3  30.64176 409.4735 1.514261 21.74377 0.1467005 1.461423 0.07713932         0
#> 4  29.75714 245.7507 1.898208 55.00905 0.1918268 1.872210 0.07043851         0
#> 5  31.84569 335.0387 1.739815 45.06316 0.2156754 1.710697 0.08132863         0
#> 6  33.35475 582.5499 1.381946 12.37202 0.1070272 1.296444 0.09571079         0
#> 7  30.63356 447.5513 1.448025 19.51455 0.1448515 1.391651 0.07799927         0
#> 8  35.04613 372.0964 1.816822 18.14557 0.1427747 1.740679 0.11691106         0
#> 9  30.25815 311.5365 1.714304 26.03515 0.1687058 1.664441 0.08755666         0
#> 10 34.90246 444.8713 1.654775 20.99774 0.1613597 1.594957 0.09494129         0
#> 11 34.27794 466.9011 1.586362 17.72346 0.1415603 1.518272 0.09563437         0
#> 12 26.77094 311.4028 1.517058 31.47264 0.1743650 1.480617 0.06857358         0
#> 13 29.33925 344.0363 1.581783 25.81397 0.1586417 1.535376 0.07594670         0
#> 14 32.64121 302.3008 1.877356 25.43998 0.1449966 1.821460 0.09333472         0
#> 15 31.20308 669.5426 1.205892 12.48450 0.1040672 1.131968 0.07445804         0
#> 16 32.52040 471.8468 1.497116 16.37316 0.1224953 1.427474 0.08815018         0
#> 17 32.49222 379.6707 1.667538 19.84503 0.1389390 1.603713 0.09334951         0
#> 18 32.23549 381.3645 1.650685 18.38926 0.1442646 1.582434 0.09950001         0
#> 19 29.52360 362.1036 1.551504 19.20266 0.1423170 1.490107 0.08671952         0
#> 20 30.16129 425.2965 1.462527 20.60758 0.1783740 1.408646 0.08702612         0
#>    sigma_sq       phi Tau.case.var((Intercept))
#> 1  185.0289 0.2289391                  46.32084
#> 2  163.6675 0.2941168                 132.92787
#> 3  200.6828 0.2079621                 208.79073
#> 4  190.3177 0.2725999                  55.43304
#> 5  266.9007 0.4993066                  68.13801
#> 6  159.9492 0.1696390                 422.60064
#> 7  204.2868 0.2427341                 243.26454
#> 8  160.8784 0.2627152                 211.21800
#> 9  175.9915 0.3196734                 135.54493
#> 10 221.2342 0.3547933                 223.63709
#> 11 197.9187 0.2648256                 268.98239
#> 12 193.2323 0.2959159                 118.17044
#> 13 189.2403 0.2396875                 154.79602
#> 14 161.1661 0.1310832                 141.13470
#> 15 184.9494 0.1260723                 484.59323
#> 16 180.7021 0.1259603                 291.14468
#> 17 173.1997 0.1805498                 206.47101
#> 18 167.0143 0.2691638                 214.35029
#> 19 162.6503 0.2278041                 199.45333
#> 20 227.3609 0.5455989                 197.93561