Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object
a g_REML object

nsim
number of models to simulate

seed
seed value. See documentation for simulate

parallel
if TRUE, run in parallel using foreach backend.

...
additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment, 
                 random = ~ 1 | case, 
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

simulate(Laski_g, nsim = 20)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#>      p_beta  r_theta delta_AB       nu     kappa      g_AB     V_g_AB cnvg_warn
#> 1  26.84772 287.8232 1.582503 40.26644 0.1776322 1.5528431 0.06377721         0
#> 2  30.84187 876.6153 1.041685 10.94056 0.1115279 0.9686048 0.06759830         0
#> 3  29.18071 404.3264 1.451208 22.78713 0.1659375 1.4029146 0.07636393         0
#> 4  35.81791 780.6245 1.281974 11.35153 0.0948556 1.1953658 0.08865027         0
#> 5  24.23586 392.3211 1.223595 22.53827 0.1439777 1.1824209 0.05587189         0
#> 6  27.60922 501.1407 1.233316 12.79903 0.1211426 1.1596061 0.07960779         0
#> 7  32.79933 273.1586 1.984529 50.04481 0.2054276 1.9546385 0.08270790         0
#> 8  31.85400 469.6155 1.469917 14.93740 0.1179955 1.3948566 0.09168720         0
#> 9  28.91980 393.3553 1.458152 19.47369 0.1420001 1.4012630 0.07807705         0
#> 10 27.75512 412.9098 1.365890 23.23146 0.1668710 1.3213138 0.07032071         0
#> 11 33.13901 348.6726 1.774723 19.70899 0.1418980 1.7063207 0.10456504         0
#> 12 32.19070 335.1717 1.758316 27.23392 0.1735042 1.7094443 0.08943798         0
#> 13 29.85927 387.2790 1.517285 16.53506 0.1326428 1.4474074 0.09201657         0
#> 14 27.07698 420.7149 1.320098 16.53501 0.1426853 1.2593015 0.07694279         0
#> 15 36.01501 294.6304 2.098190 30.23760 0.1889350 2.0457134 0.11138877         0
#> 16 31.49243 491.6556 1.420286 14.21529 0.1293613 1.3440098 0.09346932         0
#> 17 28.11957 625.1046 1.124689 15.37052 0.1397785 1.0689025 0.06412646         0
#> 18 29.74537 554.2329 1.263494 11.06092 0.1045294 1.1758402 0.09067643         0
#> 19 29.26200 447.4214 1.383393 15.93200 0.1334217 1.3172317 0.08228675         0
#> 20 30.77069 467.8932 1.422538 15.47655 0.1228827 1.3524692 0.08529362         0
#>    sigma_sq       phi Tau.case.var((Intercept))
#> 1  196.7509 0.2478737                  91.07227
#> 2  203.7796 0.3749632                 672.83565
#> 3  212.7136 0.3547023                 191.61276
#> 4  183.4677 0.1097198                 597.15681
#> 5  195.3926 0.1688445                 196.92856
#> 6  148.9771 0.3002684                 352.16368
#> 7  213.6730 0.3894139                  59.48568
#> 8  163.8880 0.1383366                 305.72746
#> 9  178.2421 0.2177620                 215.11320
#> 10 219.9162 0.3537195                 192.99358
#> 11 159.2267 0.2108458                 189.44584
#> 12 195.5824 0.3414801                 139.58926
#> 13 152.4977 0.2205824                 234.78126
#> 14 169.7949 0.3196636                 250.92007
#> 15 187.9808 0.4286188                 106.64966
#> 16 167.6095 0.2956067                 324.04612
#> 17 235.9286 0.3431962                 389.17598
#> 18 127.8188 0.2658242                 426.41408
#> 19 171.0862 0.2523497                 276.33520
#> 20 170.4879 0.1651467                 297.40527