Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object
a g_REML object

nsim
number of models to simulate

seed
seed value. See documentation for simulate

parallel
if TRUE, run in parallel using foreach backend.

...
additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment, 
                 random = ~ 1 | case, 
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

simulate(Laski_g, nsim = 20)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#>      p_beta  r_theta delta_AB       nu      kappa      g_AB     V_g_AB
#> 1  29.40286 283.2349 1.747093 38.29466 0.18556160 1.7126518 0.07572619
#> 2  30.92616 428.2107 1.494505 17.60440 0.15211126 1.4299171 0.09088838
#> 3  31.35137 398.6435 1.570233 17.31420 0.14195783 1.5012191 0.09609384
#> 4  33.70731 240.4697 2.173671 71.13894 0.17930200 2.1506740 0.06601179
#> 5  31.40807 406.4341 1.557924 16.84743 0.14645858 1.4875245 0.09846477
#> 6  33.25683 284.8072 1.970631 34.86671 0.17950432 1.9279356 0.08988884
#> 7  30.98830 339.3332 1.682227 30.10361 0.18615465 1.6399652 0.08374364
#> 8  29.78857 432.7711 1.431926 18.53478 0.15994814 1.3731914 0.08461179
#> 9  32.21287 482.7576 1.466104 17.13519 0.14547955 1.4009833 0.08822445
#> 10 29.76484 669.5641 1.150290 14.01803 0.12603003 1.0876295 0.06719202
#> 11 33.82014 562.2966 1.426240 14.13907 0.13116311 1.3492240 0.09504760
#> 12 32.74077 326.3403 1.812398 28.45775 0.16801737 1.7642090 0.08839228
#> 13 30.95473 374.5223 1.599515 28.07412 0.18094038 1.5563998 0.08047067
#> 14 24.39924 223.9073 1.630581 79.87392 0.19606603 1.6152216 0.05552057
#> 15 31.61526 224.1190 2.111822 59.32922 0.18120693 2.0850132 0.07128293
#> 16 29.94580 963.5467 0.964716 10.41358 0.09716452 0.8935268 0.05935054
#> 17 33.30606 630.9670 1.325928 12.03398 0.10773560 1.2415384 0.09161469
#> 18 33.04712 432.4110 1.589224 19.09929 0.14623351 1.5259902 0.09152293
#> 19 29.01794 381.7605 1.485153 19.11669 0.13686009 1.4261139 0.07992679
#> 20 34.26099 426.7457 1.658499 14.58606 0.12486008 1.5717339 0.11706603
#>    cnvg_warn sigma_sq       phi Tau.case.var((Intercept))
#> 1          0 193.7339 0.3202835                  89.50106
#> 2          0 186.0302 0.3712947                 242.18047
#> 3          0 166.2612 0.2826931                 232.38229
#> 4          0 195.5028 0.1362065                  44.96693
#> 5          0 168.1994 0.3445801                 238.23466
#> 6          0 185.8762 0.3028198                  98.93102
#> 7          0 214.0052 0.4068189                 125.32795
#> 8          0 199.6605 0.4156505                 233.11058
#> 9          0 201.6823 0.3235437                 281.07528
#> 10         0 223.0476 0.2713667                 446.51646
#> 11         0 191.6901 0.3196915                 370.60652
#> 12         0 191.4878 0.2802718                 134.85245
#> 13         0 226.2021 0.3913293                 148.32018
#> 14         0 193.4173 0.2234184                  30.48999
#> 15         0 173.7530 0.1842451                  50.36604
#> 16         0 194.9490 0.2535545                 768.59774
#> 17         0 166.9946 0.2060780                 463.97236
#> 18         0 195.1917 0.2666641                 237.21927
#> 19         0 169.1418 0.1797966                 212.61873
#> 20         0 147.5340 0.2274025                 279.21173