Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object
a g_REML object

nsim
number of models to simulate

seed
seed value. See documentation for simulate

parallel
if TRUE, run in parallel using foreach backend.

...
additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment, 
                 random = ~ 1 | case, 
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

simulate(Laski_g, nsim = 20)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#>      p_beta  r_theta  delta_AB       nu     kappa      g_AB     V_g_AB
#> 1  30.84974 275.7829 1.8576665 42.76981 0.1719832 1.8248994 0.07115298
#> 2  26.06232 730.3412 0.9643845 12.72659 0.1245748 0.9064129 0.05578219
#> 3  34.11348 374.2480 1.7633814 22.08079 0.1620010 1.7028002 0.10038876
#> 4  33.74386 310.0634 1.9163267 31.25718 0.1605058 1.8699747 0.08670782
#> 5  31.46155 329.3827 1.7335235 39.76940 0.1793124 1.7006246 0.07123951
#> 6  31.03941 744.5310 1.1375542 11.65816 0.1094103 1.0627686 0.07313318
#> 7  29.02339 348.9961 1.5535949 24.63817 0.1669101 1.5058178 0.07935887
#> 8  30.29287 491.7931 1.3659954 16.31309 0.1444075 1.3022158 0.08228561
#> 9  33.89878 304.6572 1.9421301 33.36807 0.1613486 1.8981482 0.08454863
#> 10 25.64097 296.8515 1.4882123 45.49299 0.1793877 1.4635420 0.05738465
#> 11 35.48903 385.1817 1.8082621 20.47449 0.1401653 1.7412050 0.10381962
#> 12 23.32608 355.9471 1.2363714 20.48895 0.1247653 1.1905549 0.05506356
#> 13 30.40367 425.7951 1.4734173 21.48941 0.1663480 1.4213884 0.08117384
#> 14 35.77969 455.9863 1.6755626 23.56030 0.1659499 1.6216520 0.09018733
#> 15 30.92415 359.8156 1.6302632 19.61045 0.1326704 1.5671088 0.08924064
#> 16 29.00305 273.1404 1.7548936 32.54754 0.1620991 1.7141422 0.07537629
#> 17 30.32310 343.5354 1.6360173 36.33075 0.1814596 1.6020099 0.07118690
#> 18 32.75703 314.8695 1.8460320 21.33342 0.1394684 1.7803632 0.10344812
#> 19 32.55414 468.7531 1.5036061 21.12867 0.1470455 1.4495938 0.07814604
#> 20 29.74718 612.1304 1.2023303 12.88339 0.1083200 1.1309522 0.07289868
#>    cnvg_warn sigma_sq        phi Tau.case.var((Intercept))
#> 1          0 190.4194 0.19127166                  85.36355
#> 2          0 218.1111 0.34641058                 512.23011
#> 3          0 191.5745 0.33506373                 182.67358
#> 4          0 186.6645 0.18962185                 123.39892
#> 5          0 224.9781 0.26351744                 104.40457
#> 6          0 189.0800 0.26177808                 555.45102
#> 7          0 191.2635 0.32755576                 157.73263
#> 8          0 197.4504 0.34648851                 294.34267
#> 9          0 188.6855 0.17675725                 115.97177
#> 10         0 211.9034 0.22988470                  84.94802
#> 11         0 179.7287 0.17697036                 205.45300
#> 12         0 162.2709 0.03517618                 193.67619
#> 13         0 217.7570 0.38839612                 208.03809
#> 14         0 243.9090 0.33913838                 212.07728
#> 15         0 161.0790 0.12860884                 198.73664
#> 16         0 167.6770 0.18989521                 105.46339
#> 17         0 228.5434 0.30513827                 114.99194
#> 18         0 150.5904 0.15232308                 164.27913
#> 19         0 225.9775 0.22391707                 242.77559
#> 20         0 177.7214 0.14714721                 434.40901