Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object

a g_REML object

nsim

number of models to simulate

seed

seed value. See documentation for simulate

parallel

if TRUE, run in parallel using foreach backend.

...

additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment,
                 random = ~ 1 | case,
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)

suppressWarnings(
  Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))
)

if (requireNamespace("plyr", quietly = TRUE)) {
  simulate(Laski_g, nsim = 20)
}
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#>      p_beta  r_theta delta_AB       nu     kappa     g_AB     V_g_AB cnvg_warn
#> 1  28.29547 372.8873 1.465305 26.93564 0.1573123 1.424123 0.06650530         0
#> 2  28.28812 333.3961 1.549258 28.87051 0.1983437 1.508660 0.08300388         0
#> 3  30.68829 345.6928 1.650546 26.87320 0.1694908 1.604048 0.08178376         0
#> 4  30.14196 287.2174 1.778550 41.58655 0.1856515 1.746281 0.07377844         0
#> 5  32.29677 412.4187 1.590340 29.12111 0.1777581 1.549027 0.07701423         0
#> 6  31.82269 497.6913 1.426451 18.67785 0.1348398 1.368396 0.07606481         0
#> 7  33.65877 451.5483 1.583967 20.15016 0.1332473 1.524270 0.08351004         0
#> 8  30.97809 279.5447 1.852802 48.79721 0.1985691 1.824178 0.07568683         0
#> 9  28.19128 227.2686 1.870015 64.01629 0.1933314 1.848020 0.06537854         0
#> 10 32.54108 280.5847 1.942674 31.29478 0.1614992 1.895742 0.08863070         0
#> 11 36.08948 303.0348 2.073167 34.60236 0.1565064 2.027905 0.08864664         0
#> 12 29.96143 394.3318 1.508800 21.79747 0.1511742 1.456283 0.07796724         0
#> 13 32.31509 438.1682 1.543778 19.05875 0.1491663 1.482220 0.08864577         0
#> 14 33.82532 365.8013 1.768558 19.15959 0.1517727 1.698413 0.10950527         0
#> 15 30.51486 564.7188 1.284090 15.06538 0.1364854 1.219085 0.07774684         0
#> 16 33.32648 590.5977 1.371335 16.86928 0.1459914 1.309449 0.08106447         0
#> 17 31.54675 269.7593 1.920730 48.10571 0.2078968 1.890628 0.08276594         0
#> 18 24.87641 279.5285 1.487903 31.54262 0.1591049 1.452242 0.06188039         0
#> 19 30.23608 394.4239 1.522453 22.05653 0.1619984 1.470091 0.08174460         0
#> 20 27.60374 392.1717 1.393894 17.39219 0.1457149 1.332909 0.08096869         0
#>    sigma_sq       phi Tau.case.var((Intercept))
#> 1  208.8213 0.2131566                 164.06601
#> 2  220.7818 0.5424488                 112.61431
#> 3  198.4810 0.3136055                 147.21173
#> 4  201.8472 0.2977803                  85.37018
#> 5  250.2937 0.3504352                 162.12504
#> 6  216.1211 0.1727866                 281.57020
#> 7  206.0847 0.1213483                 245.46359
#> 8  212.8158 0.3478408                  66.72885
#> 9  184.0540 0.2500366                  43.21462
#> 10 169.2935 0.1970749                 111.29118
#> 11 189.0123 0.1288163                 114.02250
#> 12 195.2810 0.2460411                 199.05084
#> 13 198.9088 0.2949253                 239.25933
#> 14 167.7480 0.3162069                 198.05323
#> 15 207.4147 0.3243099                 357.30417
#> 16 244.2995 0.3390249                 346.29820
#> 17 210.6770 0.4182432                  59.08230
#> 18 168.5312 0.1758059                 110.99725
#> 19 201.7832 0.3355760                 192.64064
#> 20 165.7198 0.3162658                 226.45186