Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object
a g_REML object

nsim
number of models to simulate

seed
seed value. See documentation for simulate

parallel
if TRUE, run in parallel using foreach backend.

...
additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment, 
                 random = ~ 1 | case, 
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

simulate(Laski_g, nsim = 20)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#>      p_beta  r_theta delta_AB       nu     kappa      g_AB     V_g_AB cnvg_warn
#> 1  28.39157 445.5061 1.345125 20.35686 0.1463411 1.2949508 0.06856889         0
#> 2  28.87631 505.2631 1.284644 13.85903 0.1233427 1.2138471 0.07980212         0
#> 3  28.40121 291.2184 1.664284 34.55870 0.1539956 1.6279023 0.06523420         0
#> 4  34.55209 400.2015 1.727170 23.78983 0.1619025 1.6721406 0.09203805         0
#> 5  31.82437 390.1137 1.611255 22.66800 0.1597139 1.5573496 0.08582434         0
#> 6  26.92899 657.0718 1.050543 13.05481 0.1238904 0.9890106 0.06167666         0
#> 7  35.44410 507.6163 1.573172 17.41410 0.1269754 1.5044308 0.09174868         0
#> 8  34.00336 596.3808 1.392387 13.57117 0.1220991 1.3139937 0.09239697         0
#> 9  26.51449 263.0523 1.634791 30.12231 0.1637168 1.5937465 0.07302738         0
#> 10 26.86284 536.2972 1.159976 13.02581 0.1312277 1.0918802 0.07377013         0
#> 11 30.22946 390.7075 1.529342 22.79901 0.1434607 1.4784744 0.07453504         0
#> 12 34.27226 391.1182 1.732961 21.30235 0.1482493 1.6712233 0.09624068         0
#> 13 28.72956 366.8029 1.500074 25.94806 0.1653610 1.4562942 0.07286775         0
#> 14 32.48402 437.1109 1.553724 17.66317 0.1433799 1.4868039 0.09338026         0
#> 15 32.56565 533.6831 1.409672 13.89104 0.1204847 1.3321664 0.09193090         0
#> 16 30.10715 811.6678 1.056770 12.31962 0.1287083 0.9911029 0.06655861         0
#> 17 28.69706 340.8675 1.554335 22.04128 0.1292672 1.5008388 0.07433162         0
#> 18 30.73412 634.8578 1.219783 14.21731 0.1252128 1.1542845 0.07240872         0
#> 19 27.13132 324.5186 1.506091 28.90292 0.1762323 1.4666683 0.07215894         0
#> 20 30.77015 358.2103 1.625776 24.40129 0.1540519 1.5752884 0.08054508         0
#>    sigma_sq        phi Tau.case.var((Intercept))
#> 1  209.6006 0.23519108                  235.9055
#> 2  165.1801 0.25181461                  340.0830
#> 3  180.9037 0.10946127                  110.3148
#> 4  212.8811 0.30001384                  187.3203
#> 5  201.3012 0.30262948                  188.8125
#> 6  201.5242 0.31214936                  455.5476
#> 7  206.1881 0.13499932                  301.4281
#> 8  190.0101 0.25614405                  406.3706
#> 9  156.7967 0.22665593                  106.2556
#> 10 168.5133 0.40105477                  367.7839
#> 11 195.7118 0.15962794                  194.9957
#> 12 189.9415 0.23079905                  201.1767
#> 13 205.2180 0.29323734                  161.5849
#> 14 185.6452 0.28421303                  251.4657
#> 15 173.7021 0.21894280                  359.9810
#> 16 238.0488 0.43751411                  573.6190
#> 17 163.8418 0.04680896                  177.0257
#> 18 214.1077 0.25104405                  420.7501
#> 19 195.5383 0.34086147                  128.9803
#> 20 189.8078 0.22311304                  168.4025