Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object
a g_REML object

nsim
number of models to simulate

seed
seed value. See documentation for simulate

parallel
if TRUE, run in parallel using foreach backend.

...
additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment,
                 random = ~ 1 | case,
                 correlation = corAR1(0, ~ time | case),
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))
simulate(Laski_g, nsim = 20)

#>      p_beta  r_theta  delta_AB       nu     kappa      g_AB     V_g_AB
#> 1  28.41318 370.4676 1.4761984 22.10621 0.1436757 1.4255424 0.07259791
#> 2  31.70145 390.6088 1.6040140 21.16523 0.1537637 1.5464956 0.08782589
#> 3  33.90855 578.3429 1.4099915 12.85299 0.1215125 1.3260833 0.09903538
#> 4  29.65691 482.7839 1.3497382 17.46578 0.1459680 1.2909373 0.07711346
#> 5  30.12869 250.3076 1.9043348 33.49480 0.1733385 1.8613732 0.08616768
#> 6  27.48032 397.0350 1.3791367 21.79011 0.1601246 1.3311169 0.07187246
#> 7  36.83389 350.6830 1.9669361 30.45131 0.1954708 1.9180905 0.10436719
#> 8  27.41218 359.3895 1.4459753 20.36756 0.1585261 1.3920681 0.07959990
#> 9  26.98613 751.2267 0.9845893 11.25513 0.1019122 0.9174895 0.05809963
#> 10 33.61045 509.4599 1.4890842 14.72700 0.1203795 1.4119402 0.09553060
#> 11 34.26416 283.6065 2.0346139 30.39098 0.1884197 1.9839865 0.10634048
#> 12 33.77373 288.2798 1.9891701 24.02769 0.1490031 1.9264273 0.10832218
#> 13 32.19660 429.8932 1.5528507 28.50462 0.1599076 1.5116313 0.06975293
#> 14 34.62056 478.0007 1.5835064 15.14505 0.1258694 1.5037732 0.10475086
#> 15 24.88858 361.0073 1.3099120 37.24924 0.1910977 1.2833592 0.06064565
#> 16 32.23778 374.3450 1.6662077 27.82251 0.1859374 1.6208852 0.08683386
#> 17 27.04428 395.3893 1.3600752 20.73103 0.1410369 1.3102703 0.06713019
#> 18 26.43078 364.0645 1.3852266 18.89862 0.1147249 1.3295163 0.06695391
#> 19 26.22477 413.4603 1.2897179 22.22402 0.1723484 1.2456982 0.06949274
#> 20 28.83247 369.8912 1.4991493 18.16867 0.1335920 1.4364012 0.08360581
#>    cnvg_warn sigma_sq         phi Tau.case.var((Intercept))
#> 1          0 182.3380  0.17429616                 188.12954
#> 2          0 191.2587  0.28257073                 199.35008
#> 3          0 172.9727  0.30030694                 405.37022
#> 4          0 204.7702  0.31605305                 278.01374
#> 5          0 158.7585  0.26806322                  91.54904
#> 6          0 200.7786  0.32486999                 196.25638
#> 7          0 230.7370  0.48404626                 119.94597
#> 8          0 174.2451  0.34507476                 185.14442
#> 9          0 176.9742  0.20993458                 574.25253
#> 10         0 176.1155  0.17329755                 333.34447
#> 11         0 180.9046  0.42211905                 102.70198
#> 12         0 150.1039  0.18643151                 138.17589
#> 13         0 248.3325  0.21456685                 181.56071
#> 14         0 171.7316  0.21052504                 306.26916
#> 15         0 248.3918  0.37043962                 112.61552
#> 16         0 229.4551  0.43872045                 144.88993
#> 17         0 186.2508  0.17909543                 209.13853
#> 18         0 154.7717 -0.02612698                 209.29284
#> 19         0 219.6335  0.42651510                 193.82686
#> 20         0 157.0614  0.17550666                 212.82983