Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object

a g_REML object

nsim

number of models to simulate

seed

seed value. See documentation for simulate

parallel

if TRUE, run in parallel using foreach backend.

...

additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment,
                 random = ~ 1 | case,
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)

suppressWarnings(
  Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))
)

if (requireNamespace("plyr", quietly = TRUE)) {
  simulate(Laski_g, nsim = 20)
}
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#>      p_beta   r_theta  delta_AB        nu      kappa      g_AB     V_g_AB
#> 1  34.80620  274.6756 2.1001320 35.326635 0.16513834 2.0552275 0.09174841
#> 2  30.11340  338.5082 1.6367228 28.567425 0.17210809 1.5933734 0.07861681
#> 3  28.12932 1010.8539 0.8847387  9.627158 0.08471669 0.8139759 0.05297766
#> 4  28.67075  443.5802 1.3612972 20.186212 0.16194746 1.3100852 0.07507059
#> 5  35.66196  340.1437 1.9336328 24.625289 0.14751424 1.8741371 0.10110075
#> 6  29.22269  428.6905 1.4113939 19.742603 0.15016801 1.3570889 0.07613572
#> 7  29.07841  782.2880 1.0396505 11.640267 0.10016819 0.9711940 0.06120972
#> 8  28.76086  918.4727 0.9490056 14.219438 0.14468191 0.8980548 0.05574787
#> 9  31.61002  721.9534 1.1764409 11.686005 0.10849072 1.0992872 0.07695906
#> 10 29.71899  376.4782 1.5316663 18.384673 0.13891721 1.4683208 0.08710849
#> 11 27.17406  467.0197 1.2574391 15.573934 0.13254215 1.1958962 0.07226407
#> 12 30.22141  360.8123 1.5910142 32.726208 0.17011796 1.5542716 0.06919209
#> 13 30.30466  302.7541 1.7416642 48.944753 0.19388451 1.7148389 0.06956019
#> 14 32.07929  522.8830 1.4028862 16.088056 0.12244802 1.3364535 0.08050170
#> 15 30.98126  566.6677 1.3014721 13.869718 0.11493123 1.2298037 0.07934157
#> 16 26.35843  518.4266 1.1576464 15.754564 0.12217609 1.1016477 0.06071931
#> 17 27.85821  381.3572 1.4265508 24.236639 0.16894543 1.3819463 0.07282912
#> 18 32.03002  600.9866 1.3065464 11.821123 0.11669589 1.2218606 0.09296482
#> 19 31.34372  352.2883 1.6699423 25.711267 0.15391164 1.6207516 0.08042790
#> 20 27.28661  473.1831 1.2543968 14.737886 0.12507789 1.1894600 0.07332288
#>    cnvg_warn sigma_sq       phi Tau.case.var((Intercept))
#> 1          0 175.0747 0.1898009                  99.60084
#> 2          0 200.8551 0.3118139                 137.65316
#> 3          0 165.3130 0.2063156                 845.54098
#> 4          0 216.2667 0.3819991                 227.31344
#> 5          0 178.9890 0.1643624                 161.15475
#> 6          0 199.5385 0.2851892                 229.15202
#> 7          0 193.6391 0.1491490                 588.64891
#> 8          0 332.6169 0.4717992                 585.85579
#> 9          0 183.5120 0.2476399                 538.44143
#> 10         0 162.9237 0.2192579                 213.55452
#> 11         0 174.7413 0.2590857                 292.27846
#> 12         0 225.2999 0.2504268                 135.51239
#> 13         0 227.9042 0.3143002                  74.84982
#> 14         0 197.1732 0.1360884                 325.70978
#> 15         0 181.9397 0.1598289                 384.72801
#> 16         0 191.7586 0.1464048                 326.66793
#> 17         0 208.5355 0.3522574                 172.82172
#> 18         0 159.3131 0.3355262                 441.67348
#> 19         0 191.6636 0.2019837                 160.62476
#> 20         0 165.3392 0.2219530                 307.84382