Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object

a g_REML object

nsim

number of models to simulate

seed

seed value. See documentation for simulate

parallel

if TRUE, run in parallel using foreach backend.

...

additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment,
                 random = ~ 1 | case,
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
simulate(Laski_g, nsim = 20)
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#>      p_beta  r_theta delta_AB       nu     kappa     g_AB     V_g_AB cnvg_warn
#> 1  25.26386 304.0049 1.448970 27.46801 0.1361662 1.409043 0.05844060         0
#> 2  32.71847 430.1804 1.577494 18.52896 0.1575772 1.512768 0.09630222         0
#> 3  32.30792 307.4762 1.842482 27.83402 0.1621989 1.792385 0.08987166         0
#> 4  30.16953 290.8094 1.769149 37.62083 0.1980469 1.733644 0.08239827         0
#> 5  31.85708 473.3805 1.464201 17.09214 0.1283453 1.398999 0.08338055         0
#> 6  32.47912 454.1603 1.524052 22.46647 0.1587520 1.472602 0.07973133         0
#> 7  36.04372 582.1755 1.493835 13.63966 0.1188578 1.410160 0.10269979         0
#> 8  35.01185 480.6845 1.596927 15.93786 0.1453863 1.520581 0.10693751         0
#> 9  31.73535 358.6155 1.675825 16.42779 0.1188708 1.598134 0.10528955         0
#> 10 27.44115 414.4800 1.347878 17.81360 0.1423914 1.290321 0.07474489         0
#> 11 35.50372 419.2123 1.734031 21.03650 0.1373325 1.671465 0.09411642         0
#> 12 33.71603 395.8980 1.694512 25.73579 0.1621553 1.644646 0.08469821         0
#> 13 31.97810 321.5434 1.783334 32.04618 0.1685025 1.741270 0.07993638         0
#> 14 29.24824 306.4204 1.670863 49.78271 0.1817715 1.645564 0.06194241         0
#> 15 31.15160 518.4754 1.368095 14.38468 0.1145557 1.295502 0.08331882         0
#> 16 33.78619 289.2665 1.986507 30.50766 0.1656974 1.937268 0.09460383         0
#> 17 26.14673 498.6561 1.170892 17.23474 0.1250837 1.119188 0.05822840         0
#> 18 38.92396 367.2069 2.031242 18.25952 0.1425651 1.946651 0.14002108         0
#> 19 29.96371 638.1185 1.186165 13.94083 0.1221716 1.121185 0.06972656         0
#> 20 31.20978 293.8643 1.820611 29.07872 0.1547260 1.773247 0.08322340         0
#>    sigma_sq        phi Tau.case.var((Intercept))
#> 1  166.5257 0.03000197                 137.47920
#> 2  196.6792 0.39196050                 233.50122
#> 3  176.4841 0.24126341                 130.99205
#> 4  205.3049 0.42114768                  85.50449
#> 5  189.8887 0.15879518                 283.49182
#> 6  232.6852 0.29835045                 221.47509
#> 7  185.0482 0.21627306                 397.12734
#> 8  190.2227 0.37397547                 290.46185
#> 9  136.9109 0.08745748                 221.70457
#> 10 176.6919 0.26975426                 237.78808
#> 11 198.0446 0.13928836                 221.16767
#> 12 219.0479 0.26991122                 176.85003
#> 13 198.4706 0.24443868                 123.07281
#> 14 225.9129 0.22572138                  80.50754
#> 15 173.0147 0.12833612                 345.46072
#> 16 174.0143 0.23816433                 115.25222
#> 17 200.2293 0.12223236                 298.42678
#> 18 159.3925 0.25720387                 207.81446
#> 19 209.3154 0.23427105                 428.80318
#> 20 169.7616 0.16622414                 124.10269