Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object
a g_REML object

nsim
number of models to simulate

seed
seed value. See documentation for simulate

parallel
if TRUE, run in parallel using foreach backend.

...
additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment, 
                 random = ~ 1 | case, 
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

simulate(Laski_g, nsim = 20)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#>      p_beta   r_theta  delta_AB        nu      kappa      g_AB     V_g_AB
#> 1  35.69355  495.9828 1.6027157 16.851659 0.15145753 1.5303111 0.10443266
#> 2  26.62190  324.9177 1.4769042 25.077848 0.15726840 1.4322899 0.07041555
#> 3  32.23810  335.0320 1.7612714 47.961192 0.16886468 1.7335850 0.06178305
#> 4  31.13745  514.4650 1.3727931 14.298830 0.12186223 1.2995062 0.08605502
#> 5  32.60655  359.0760 1.7207259 32.028214 0.17870451 1.6801149 0.08004345
#> 6  30.93093  545.8589 1.3238928 16.502241 0.14211651 1.2627984 0.07721716
#> 7  28.23395  270.9565 1.7152280 72.566572 0.22271978 1.6974393 0.07047985
#> 8  27.11010  364.7056 1.4195802 19.605042 0.13603615 1.3645720 0.07299051
#> 9  28.40298  366.3596 1.4839190 31.181148 0.15404896 1.4479378 0.06051140
#> 10 25.96256  409.5063 1.2829719 18.772291 0.16025308 1.2310221 0.07259058
#> 11 32.38326  625.9691 1.2943272 12.877023 0.12038175 1.2174487 0.08545612
#> 12 35.15968  408.6386 1.7393031 16.140816 0.11727511 1.6572131 0.11375891
#> 13 32.36874  803.7100 1.1417633 11.611223 0.11062852 1.0663909 0.07413881
#> 14 37.80562  403.3598 1.8823920 30.962503 0.18214046 1.8364240 0.09270430
#> 15 29.44123  392.7755 1.4855379 17.945324 0.11866719 1.4225747 0.07939309
#> 16 29.07255 1222.8985 0.8313578  9.385013 0.08401699 0.7631019 0.04873545
#> 17 25.88346  426.5294 1.2532791 23.325667 0.14306545 1.2125454 0.05600582
#> 18 31.88410  312.0714 1.8048751 26.868462 0.18893693 1.7540211 0.09917854
#> 19 30.41429  315.6334 1.7119307 17.330471 0.13089761 1.6367602 0.10708283
#> 20 28.17450  317.0733 1.5822544 50.169221 0.20198637 1.5584822 0.06662927
#>    cnvg_warn sigma_sq        phi Tau.case.var((Intercept))
#> 1          0 208.7770 0.39543481                 287.20582
#> 2          0 175.6995 0.23937005                 149.21820
#> 3          0 238.9825 0.14207892                  96.04950
#> 4          0 173.1547 0.21085628                 341.31031
#> 5          0 226.6755 0.32401359                 132.40048
#> 6          0 219.6135 0.31533952                 326.24542
#> 7          0 246.1444 0.40787596                  24.81204
#> 8          0 164.1662 0.15981312                 200.53943
#> 9          0 218.0508 0.13912762                 148.30874
#> 10         0 190.5312 0.41021282                 218.97517
#> 11         0 187.0048 0.28544152                 438.96433
#> 12         0 153.2060 0.08152778                 255.43256
#> 13         0 203.6431 0.28171616                 600.06690
#> 14         0 253.7584 0.36312694                 149.60132
#> 15         0 161.2935 0.03775281                 231.48196
#> 16         0 186.6292 0.24647229                1036.26933
#> 17         0 216.3078 0.14706787                 210.22155
#> 18         0 191.7252 0.48414484                 120.34620
#> 19         0 128.6245 0.17569543                 187.00882
#> 20         0 245.5812 0.36456879                  71.49209