Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object
a g_REML object

nsim
number of models to simulate

seed
seed value. See documentation for simulate

parallel
if TRUE, run in parallel using foreach backend.

...
additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment, 
                 random = ~ 1 | case, 
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

simulate(Laski_g, nsim = 20)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#>      p_beta  r_theta  delta_AB       nu      kappa      g_AB     V_g_AB
#> 1  28.96286 513.9747 1.2775283 14.89731 0.10490377 1.2121138 0.06993422
#> 2  25.68773 403.0353 1.2795408 20.79148 0.15764276 1.2328228 0.06673014
#> 3  30.49824 370.9662 1.5834617 24.65735 0.15434508 1.5348044 0.07716334
#> 4  32.41753 535.8250 1.4004526 14.87054 0.13527848 1.3286126 0.08945516
#> 5  30.95171 388.7506 1.5698171 18.22732 0.13893291 1.5043256 0.09115317
#> 6  33.22513 282.2684 1.9775870 28.50774 0.18224064 1.9250991 0.10470317
#> 7  34.07688 315.7091 1.9178575 23.45132 0.16159223 1.8558613 0.10835027
#> 8  24.36096 332.3450 1.3362875 30.12383 0.17622420 1.3027392 0.06217489
#> 9  31.05649 321.9689 1.7307942 33.90299 0.18055812 1.6922213 0.07850925
#> 10 34.41097 620.2794 1.3816665 13.94778 0.11715880 1.3060156 0.08775881
#> 11 32.23805 645.7718 1.2686133 11.91749 0.09737634 1.1870653 0.08344907
#> 12 32.02263 348.4382 1.7155136 35.66703 0.18443517 1.6791854 0.07685550
#> 13 27.34339 762.1612 0.9904419 12.19873 0.11708188 0.9282737 0.05805675
#> 14 27.79503 827.5199 0.9662243 11.13842 0.11637303 0.8996703 0.06022715
#> 15 32.16567 334.4927 1.7587307 25.33736 0.14927613 1.7061525 0.08609513
#> 16 34.96114 306.6923 1.9963381 42.85929 0.18998676 1.9611990 0.08402981
#> 17 33.71415 699.3412 1.2748752 11.86457 0.12731693 1.1925514 0.09162231
#> 18 31.84339 474.1019 1.4624582 13.75139 0.11968260 1.3812189 0.09850294
#> 19 32.26445 390.7534 1.6321979 20.27354 0.13828756 1.5710624 0.08850488
#> 20 32.56339 457.9669 1.5216424 18.66671 0.14093027 1.4596752 0.08574373
#>    cnvg_warn sigma_sq           phi Tau.case.var((Intercept))
#> 1          0 175.1374 -0.0002992244                 338.83734
#> 2          0 197.2375  0.3261529871                 205.79774
#> 3          0 197.7439  0.2214846033                 173.22229
#> 4          0 193.8492  0.3219057444                 341.97586
#> 5          0 167.2047  0.2240857828                 221.54590
#> 6          0 172.2183  0.3959890053                 110.05003
#> 7          0 166.5886  0.3036160821                 149.12048
#> 8          0 203.5991  0.3256564611                 128.74586
#> 9          0 208.5369  0.3196946375                 113.43201
#> 10         0 201.3005  0.1796638828                 418.97891
#> 11         0 164.7526  0.0898840142                 481.01923
#> 12         0 232.0210  0.3333144234                 116.41722
#> 13         0 210.8598  0.3033544870                 551.30138
#> 14         0 202.0046  0.4147722342                 625.51532
#> 15         0 179.1811  0.1686501955                 155.31163
#> 16         0 219.8401  0.3210482318                  86.85224
#> 17         0 195.8411  0.4734492977                 503.50008
#> 18         0 152.3470  0.2185049548                 321.75492
#> 19         0 180.5680  0.1645546333                 210.18542
#> 20         0 201.1706  0.2298504918                 256.79635