Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object
a g_REML object

nsim
number of models to simulate

seed
seed value. See documentation for simulate

parallel
if TRUE, run in parallel using foreach backend.

...
additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment,
                 random = ~ 1 | case,
                 correlation = corAR1(0, ~ time | case),
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

simulate(Laski_g, nsim = 20)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#>      p_beta  r_theta delta_AB        nu     kappa     g_AB     V_g_AB cnvg_warn
#> 1  33.36864 441.6617 1.587792  15.95256 0.1434818 1.511954 0.10531371         0
#> 2  26.50308 420.9240 1.291798  17.21299 0.1342450 1.234682 0.06988757         0
#> 3  26.88072 319.3223 1.504272  20.55738 0.1545083 1.448715 0.08215943         0
#> 4  32.79179 588.8968 1.351280  15.02601 0.1312064 1.282692 0.08273204         0
#> 5  34.10034 286.3235 2.015256  29.24011 0.1571609 1.963119 0.09683592         0
#> 6  31.97919 227.1899 2.121646 101.15925 0.2152791 2.105877 0.06902499         0
#> 7  32.98463 476.7127 1.510717  16.07787 0.1418482 1.439132 0.09623741         0
#> 8  31.12981 339.4384 1.689647  23.12126 0.1303135 1.634240 0.08168500         0
#> 9  26.58612 262.0859 1.642226  28.66291 0.1762119 1.598878 0.08022966         0
#> 10 27.93117 403.0248 1.391308  23.90579 0.1786138 1.347197 0.07474652         0
#> 11 28.30483 372.4947 1.466562  21.95602 0.1595229 1.415889 0.07722327         0
#> 12 27.25688 282.4018 1.621968  48.18767 0.1976684 1.596592 0.06731346         0
#> 13 30.48443 492.3766 1.373818  20.63651 0.1608860 1.323277 0.07446729         0
#> 14 27.16876 563.2546 1.144768  11.78348 0.1090104 1.070326 0.07308657         0
#> 15 32.45888 376.6291 1.672540  23.48230 0.1578312 1.618546 0.08748937         0
#> 16 28.48125 562.9504 1.200395  12.75090 0.1193511 1.128376 0.07602629         0
#> 17 30.96545 311.8422 1.753517  36.74109 0.1768900 1.717477 0.07464334         0
#> 18 32.97416 384.6935 1.681188  16.45995 0.1313344 1.603403 0.10890859         0
#> 19 30.15666 432.9348 1.449345  17.84308 0.1407044 1.387559 0.08254901         0
#> 20 40.25750 362.3607 2.114834  32.40080 0.1796133 2.065500 0.10380276         0
#>    sigma_sq        phi Tau.case.var((Intercept))
#> 1  173.7922 0.35338389                267.869506
#> 2  171.7483 0.21191691                249.175756
#> 3  153.9810 0.30365748                165.341251
#> 4  212.6730 0.27128542                376.223746
#> 5  166.4919 0.18405179                119.831552
#> 6  220.8865 0.28480930                  6.303391
#> 7  187.7876 0.33086839                288.925133
#> 8  168.1539 0.03783893                171.284493
#> 9  157.3778 0.34385020                104.708094
#> 10 226.5455 0.44643935                176.479296
#> 11 188.8491 0.31587674                183.645615
#> 12 213.7139 0.34477217                 68.687819
#> 13 242.1841 0.35973262                250.192542
#> 14 145.1284 0.24450158                418.126143
#> 15 197.1357 0.27086806                179.493411
#> 16 165.7238 0.28334589                397.226602
#> 17 206.1835 0.26754187                105.658624
#> 18 150.4938 0.21054000                234.199680
#> 19 184.0600 0.25277861                248.874819
#> 20 230.2726 0.32723156                132.088074