Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object
a g_REML object

nsim
number of models to simulate

seed
seed value. See documentation for simulate

parallel
if TRUE, run in parallel using foreach backend.

...
additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment,
                 random = ~ 1 | case,
                 correlation = corAR1(0, ~ time | case),
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

simulate(Laski_g, nsim = 20)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#>      p_beta  r_theta  delta_AB       nu     kappa      g_AB     V_g_AB
#> 1  30.10466 560.8861 1.2711486 16.13924 0.1419436 1.2111481 0.07403278
#> 2  34.51208 405.0252 1.7148658 22.96885 0.1432792 1.6582542 0.08771365
#> 3  23.26256 426.4743 1.1264476 23.70745 0.1616144 1.0904319 0.05457047
#> 4  32.08192 388.3042 1.6280748 23.04801 0.1622138 1.5745150 0.08684109
#> 5  25.92576 199.4509 1.8357498 91.90385 0.2223510 1.8207279 0.06822877
#> 6  27.40297 460.0600 1.2775868 17.58099 0.1367233 1.2222991 0.06833248
#> 7  30.71095 260.0416 1.9044596 56.75734 0.2049586 1.8791824 0.07485626
#> 8  30.74676 278.8470 1.8412654 34.19651 0.1573294 1.8005853 0.07605851
#> 9  33.30434 386.2851 1.6945207 16.75146 0.1366613 1.6175038 0.11011125
#> 10 27.52027 770.0078 0.9917567 11.15761 0.1077266 0.9235642 0.06054481
#> 11 23.64306 284.9651 1.4005794 50.59203 0.1940640 1.3797135 0.05778704
#> 12 30.92916 516.2233 1.3612859 20.20236 0.1364773 1.3101157 0.06721493
#> 13 28.01441 171.1864 2.1411488 87.98262 0.1909614 2.1228448 0.06299791
#> 14 34.12242 671.9210 1.3163785 12.60783 0.1037085 1.2364871 0.08566300
#> 15 29.28871 636.2312 1.1611620 15.81506 0.1370060 1.1052116 0.06479230
#> 16 31.27389 680.2184 1.1991063 11.45350 0.1056006 1.1188341 0.08033455
#> 17 29.98624 398.6186 1.5019076 32.22614 0.1841798 1.4666804 0.07050146
#> 18 30.87921 442.2397 1.4683760 14.67557 0.1159917 1.3920336 0.09252630
#> 19 30.91273 498.9013 1.3839807 18.90388 0.1547194 1.3283362 0.07794384
#> 20 26.03049 446.2731 1.2322022 20.62289 0.1453387 1.1868404 0.06024181
#>    cnvg_warn sigma_sq       phi Tau.case.var((Intercept))
#> 1          0 221.6718 0.3291074                339.214314
#> 2          0 203.6859 0.1550518                201.339238
#> 3          0 226.3000 0.2990565                200.174328
#> 4          0 203.3907 0.3167490                184.913511
#> 5          0 192.5671 0.3466279                  6.883823
#> 6          0 191.7815 0.2234260                268.278494
#> 7          0 209.8320 0.3539027                 50.209577
#> 8          0 173.3165 0.1384959                105.530489
#> 9          0 155.0902 0.2518165                231.194836
#> 10         0 181.9385 0.2956818                588.069281
#> 11         0 216.6412 0.3074265                 68.323959
#> 12         0 237.2593 0.1497224                278.963996
#> 13         0 150.3339 0.1727468                 20.852489
#> 14         0 187.9417 0.1128215                483.979246
#> 15         0 243.9520 0.2934940                392.279182
#> 16         0 166.3997 0.2353924                513.818621
#> 17         0 255.9655 0.3650699                142.653107
#> 18         0 151.0632 0.1294306                291.176534
#> 19         0 228.7642 0.3517153                270.137046
#> 20         0 211.2423 0.2200148                235.030794