Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object
a g_REML object

nsim
number of models to simulate

seed
seed value. See documentation for simulate

parallel
if TRUE, run in parallel using foreach backend.

...
additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment, 
                 random = ~ 1 | case, 
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

simulate(Laski_g, nsim = 20)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#>      p_beta  r_theta delta_AB       nu      kappa      g_AB     V_g_AB
#> 1  30.36571 820.1854 1.060297 10.85978 0.09428265 0.9853448 0.06627620
#> 2  35.04701 405.4950 1.740437 25.69548 0.15940023 1.6891377 0.08707143
#> 3  27.35778 480.6608 1.247847 15.79992 0.12981800 1.1876616 0.06987749
#> 4  31.46258 458.6354 1.469131 15.30016 0.12453471 1.3959196 0.09127786
#> 5  22.61727 453.2182 1.062396 21.02485 0.16261071 1.0240424 0.05522597
#> 6  30.98160 611.1671 1.253210 12.69433 0.11850593 1.1776811 0.08163735
#> 7  29.08693 524.0937 1.270555 15.99300 0.13806182 1.2100251 0.07338237
#> 8  28.68792 389.6629 1.453298 20.55720 0.15495081 1.3996235 0.07846102
#> 9  31.50224 270.9901 1.913660 41.19830 0.17522467 1.8786097 0.07653572
#> 10 31.77656 509.9707 1.407130 23.47320 0.16019608 1.3616863 0.07016450
#> 11 28.60544 250.7004 1.806638 62.56276 0.20006549 1.7848932 0.06682086
#> 12 27.81232 651.2282 1.089860 13.11460 0.12176998 1.0263214 0.06436255
#> 13 24.66117 468.1613 1.139767 14.52130 0.12034495 1.0798686 0.06293042
#> 14 29.47070 385.2787 1.501423 20.44529 0.15482269 1.4456637 0.08237587
#> 15 34.02681 333.7823 1.862471 30.31229 0.17957981 1.8160059 0.09181143
#> 16 29.34325 318.9199 1.643113 31.35525 0.15283827 1.6034946 0.06809750
#> 17 26.96534 314.0685 1.521576 35.83066 0.15692300 1.4895030 0.05813563
#> 18 31.83479 477.6133 1.456679 13.68939 0.11776882 1.3753872 0.09777879
#> 19 30.70522 371.8520 1.592308 35.67464 0.17519551 1.5585963 0.06761306
#> 20 30.35196 569.4510 1.271916 14.12752 0.11858781 1.2031765 0.07604313
#>    cnvg_warn sigma_sq       phi Tau.case.var((Intercept))
#> 1          0 178.2170 0.1555464                 641.96840
#> 2          0 222.8879 0.2477107                 182.60706
#> 3          0 180.9938 0.2217641                 299.66700
#> 4          0 165.8863 0.1897604                 292.74913
#> 5          0 226.4007 0.3655872                 226.81754
#> 6          0 178.5188 0.2780308                 432.64833
#> 7          0 203.3963 0.2963006                 320.69738
#> 8          0 188.1099 0.3076523                 201.55303
#> 9          0 185.8272 0.2243037                  85.16289
#> 10         0 268.3220 0.2912409                 241.64869
#> 11         0 204.8524 0.2990855                  45.84799
#> 12         0 199.5145 0.2835853                 451.71365
#> 13         0 159.5541 0.1831444                 308.60717
#> 14         0 185.3338 0.3092959                 199.94489
#> 15         0 206.8084 0.3502775                 126.97389
#> 16         0 189.9298 0.1277849                 128.99017
#> 17         0 198.7259 0.1226758                 115.34265
#> 18         0 152.0505 0.2013301                 325.56277
#> 19         0 242.4060 0.2633111                 129.44606
#> 20         0 187.9261 0.1851276                 381.52491