Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object
a g_REML object

nsim
number of models to simulate

seed
seed value. See documentation for simulate

parallel
if TRUE, run in parallel using foreach backend.

...
additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment, 
                 random = ~ 1 | case, 
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

simulate(Laski_g, nsim = 20)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#>      p_beta  r_theta delta_AB       nu     kappa     g_AB     V_g_AB cnvg_warn
#> 1  28.66545 261.2701 1.773430 66.42108 0.1944677 1.753329 0.06210956         0
#> 2  26.87652 287.6514 1.584673 26.44943 0.1657599 1.539310 0.07721658         0
#> 3  34.60893 339.0780 1.879483 37.34775 0.1760687 1.841486 0.07990514         0
#> 4  31.70074 428.2449 1.531875 20.42582 0.1408194 1.474930 0.08054837         0
#> 5  32.88504 596.7808 1.346142 12.32451 0.1241779 1.262527 0.09593162         0
#> 6  28.13864 435.8686 1.347799 18.62280 0.1300493 1.292781 0.06877106         0
#> 7  27.11211 372.0249 1.405650 16.27130 0.1312480 1.339848 0.08228060         0
#> 8  26.94683 337.7724 1.466208 40.31856 0.2036625 1.438764 0.06929330         0
#> 9  24.55061 461.3650 1.142984 18.05218 0.1548674 1.094830 0.06290314         0
#> 10 29.93983 423.8432 1.454276 21.21543 0.1589551 1.402252 0.07804913         0
#> 11 31.67717 343.7763 1.708474 32.54269 0.1854701 1.668795 0.08110057         0
#> 12 31.61679 475.1949 1.450380 14.64409 0.1179553 1.374809 0.09126979         0
#> 13 29.03188 338.2229 1.578605 20.18377 0.1424246 1.519211 0.08555547         0
#> 14 29.75165 442.6906 1.414037 30.26935 0.1719697 1.378709 0.06415395         0
#> 15 31.11748 374.8453 1.607231 18.24011 0.1371966 1.540226 0.09403914         0
#> 16 32.86873 709.1305 1.234297 11.55499 0.1072423 1.152411 0.08407743         0
#> 17 28.97787 461.3315 1.349149 21.22336 0.1546417 1.300904 0.06938083         0
#> 18 35.07226 746.6068 1.283565 12.23758 0.1260429 1.203259 0.08975057         0
#> 19 33.27008 558.5225 1.407776 14.92929 0.1225714 1.335849 0.08648308         0
#> 20 31.83815 549.3908 1.358336 17.20660 0.1481857 1.298256 0.07939651         0
#>    sigma_sq       phi Tau.case.var((Intercept))
#> 1  214.1953 0.2498894                  47.07481
#> 2  162.5211 0.2890706                 125.13024
#> 3  225.1193 0.2567611                 113.95878
#> 4  199.7539 0.1839678                 228.49098
#> 5  171.4852 0.3782090                 425.29560
#> 6  187.3033 0.1292269                 248.56532
#> 7  144.1142 0.2169324                 227.91068
#> 8  248.2513 0.4418534                  89.52109
#> 9  205.6231 0.3820686                 255.74190
#> 10 210.5434 0.3276892                 213.29978
#> 11 222.3586 0.3718804                 121.41768
#> 12 162.5641 0.1517697                 312.63083
#> 13 157.0310 0.2040589                 181.19186
#> 14 268.9668 0.2902147                 173.72376
#> 15 160.7353 0.2073875                 214.10999
#> 16 176.6306 0.2453857                 532.49986
#> 17 226.7179 0.2890351                 234.61365
#> 18 214.5437 0.4109054                 532.06310
#> 19 196.5582 0.1865106                 361.96432
#> 20 232.0473 0.3475185                 317.34350