Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object

a g_REML object

nsim

number of models to simulate

seed

seed value. See documentation for simulate

parallel

if TRUE, run in parallel using foreach backend.

...

additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment,
                 random = ~ 1 | case,
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)

suppressWarnings(
  Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))
)

if (requireNamespace("plyr", quietly = TRUE)) {
  simulate(Laski_g, nsim = 20)
}
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#>      p_beta  r_theta delta_AB       nu      kappa     g_AB     V_g_AB cnvg_warn
#> 1  32.21415 384.7127 1.642398 17.36299 0.13976555 1.570418 0.10228537         0
#> 2  28.56578 545.8105 1.222715 15.94663 0.14392867 1.164292 0.07129196         0
#> 3  30.54261 313.4605 1.725102 25.58506 0.16602040 1.674033 0.08847162         0
#> 4  31.11201 487.1900 1.409545 19.85702 0.15138465 1.355628 0.07605023         0
#> 5  33.04732 586.7344 1.364317 15.02863 0.13910082 1.295080 0.08619470         0
#> 6  32.44618 241.6719 2.087134 92.11061 0.21750374 2.070094 0.07144725         0
#> 7  31.69471 334.3780 1.733277 24.92753 0.13804839 1.680599 0.08203828         0
#> 8  32.62499 541.0219 1.402629 14.20440 0.13596583 1.327243 0.09347884         0
#> 9  29.21614 339.9077 1.584683 35.76703 0.19006448 1.551220 0.07267997         0
#> 10 26.23378 483.0111 1.193664 14.08611 0.11295426 1.128961 0.06753611         0
#> 11 30.35432 221.4129 2.039948 52.26690 0.18625830 2.010535 0.07554169         0
#> 12 39.72175 497.8401 1.780260 18.11577 0.14340210 1.705525 0.11343744         0
#> 13 32.86096 359.0347 1.734251 23.22152 0.14452563 1.677630 0.08881695         0
#> 14 31.66776 571.5076 1.324666 13.20480 0.12313485 1.247976 0.08748671         0
#> 15 25.75256 494.6215 1.157934 17.22073 0.14026178 1.106761 0.06150153         0
#> 16 35.11309 322.6955 1.954667 41.31724 0.17002710 1.918969 0.07654556         0
#> 17 34.09076 969.5160 1.094861  9.91602 0.08047967 1.009910 0.07401746         0
#> 18 34.29823 519.7573 1.504428 22.19871 0.16867120 1.453021 0.08235259         0
#> 19 31.04103 543.6192 1.331339 13.88645 0.12127592 1.258116 0.08385856         0
#> 20 27.01235 541.5531 1.160758 15.02354 0.12479488 1.101831 0.06403989         0
#>    sigma_sq        phi Tau.case.var((Intercept))
#> 1  159.9434 0.25987527                 224.76928
#> 2  215.0245 0.35830213                 330.78593
#> 3  174.4923 0.30433987                 138.96820
#> 4  228.2631 0.29325220                 258.92692
#> 5  216.8222 0.35401529                 369.91218
#> 6  229.5266 0.32010237                  12.14535
#> 7  174.5907 0.07877110                 159.78737
#> 8  188.1968 0.36859280                 352.82509
#> 9  230.1319 0.37566062                 109.77577
#> 10 157.1347 0.12642742                 325.87632
#> 11 167.0469 0.24594358                  54.36593
#> 12 215.3094 0.26956816                 282.53068
#> 13 182.0583 0.16148526                 176.97637
#> 14 177.1450 0.29251655                 394.36266
#> 15 204.6062 0.26964905                 290.01530
#> 16 219.4937 0.18538730                 103.20173
#> 17 168.6481 0.09092553                 800.86797
#> 18 272.2620 0.39257230                 247.49522
#> 19 177.2001 0.22777941                 366.41912
#> 20 192.6408 0.20512142                 348.91230