Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object

a g_REML object

nsim

number of models to simulate

seed

seed value. See documentation for simulate

parallel

if TRUE, run in parallel using foreach backend.

...

additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment,
                 random = ~ 1 | case,
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)

suppressWarnings(
  Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))
)

if (requireNamespace("plyr", quietly = TRUE)) {
  simulate(Laski_g, nsim = 20)
}
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#>      p_beta  r_theta delta_AB       nu     kappa     g_AB     V_g_AB cnvg_warn
#> 1  28.46898 509.7045 1.260993 16.07008 0.1417696 1.201212 0.07337945         0
#> 2  32.48180 635.3926 1.288602 17.94010 0.1523353 1.233970 0.07276808         0
#> 3  34.87304 335.9218 1.902701 18.12447 0.1399276 1.822865 0.12548968         0
#> 4  30.06771 369.9009 1.563355 27.95210 0.1765187 1.521029 0.07696031         0
#> 5  34.87558 474.4473 1.601133 14.48983 0.1365449 1.516803 0.11406736         0
#> 6  31.98073 268.0792 1.953245 62.97140 0.2117246 1.929889 0.07592470         0
#> 7  30.14556 724.6757 1.119828 12.84066 0.1127388 1.053122 0.06608218         0
#> 8  28.47576 593.5703 1.168797 13.32614 0.1226934 1.101759 0.07095058         0
#> 9  30.65568 644.3221 1.207701 11.84377 0.1042837 1.129575 0.07848791         0
#> 10 28.16848 624.6623 1.127044 17.37153 0.1553671 1.077674 0.06358274         0
#> 11 30.57510 320.6455 1.707479 33.71534 0.1547312 1.669212 0.06875297         0
#> 12 30.00342 505.2903 1.334751 18.84328 0.1510944 1.280911 0.07324925         0
#> 13 31.90523 326.7607 1.765009 34.80173 0.1935732 1.726697 0.08399908         0
#> 14 27.10030 471.9046 1.247518 17.10286 0.1323867 1.192000 0.06625283         0
#> 15 30.24291 484.0255 1.374641 17.10605 0.1435829 1.313477 0.07974492         0
#> 16 29.75210 331.1697 1.634904 27.88123 0.1623936 1.590527 0.07644238         0
#> 17 27.85459 395.9607 1.399815 34.38477 0.1817201 1.369059 0.06280100         0
#> 18 30.34323 389.5384 1.537399 25.56466 0.1584373 1.491850 0.07358705         0
#> 19 32.39436 431.5027 1.559472 14.17978 0.1001536 1.475508 0.10236430         0
#> 20 32.71207 730.8416 1.210031 12.86942 0.1342906 1.138116 0.08035047         0
#>    sigma_sq         phi Tau.case.var((Intercept))
#> 1  200.6601  0.33041638                 309.04442
#> 2  279.7014  0.36108223                 355.69121
#> 3  144.1978  0.23656746                 191.72406
#> 4  220.0432  0.35610996                 149.85777
#> 5  168.3762  0.35720563                 306.07108
#> 6  226.5877  0.37354779                  41.49155
#> 7  211.5445  0.20079844                 513.13124
#> 8  185.6413  0.27902940                 407.92897
#> 9  165.2745  0.18118721                 479.04761
#> 10 272.2541  0.41397643                 352.40820
#> 11 197.4048  0.12203307                 123.24067
#> 12 228.8377  0.31931115                 276.45261
#> 13 221.8281  0.41280918                 104.93261
#> 14 190.8578  0.19771078                 281.04676
#> 15 200.9075  0.30606826                 283.11803
#> 16 190.3036  0.24223900                 140.86608
#> 17 258.5126  0.32411661                 137.44805
#> 18 213.1654  0.24168330                 176.37304
#> 19 138.5198 -0.02075254                 292.98292
#> 20 230.4323  0.45508503                 500.40932