Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object

a g_REML object

nsim

number of models to simulate

seed

seed value. See documentation for simulate

parallel

if TRUE, run in parallel using foreach backend.

...

additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment,
                 random = ~ 1 | case,
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)

suppressWarnings(
  Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))
)

if (requireNamespace("plyr", quietly = TRUE)) {
  simulate(Laski_g, nsim = 20)
}
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#>      p_beta  r_theta delta_AB       nu     kappa     g_AB     V_g_AB cnvg_warn
#> 1  32.53944 394.4407 1.638397 20.24421 0.1447818 1.576940 0.09102503         0
#> 2  32.07189 609.4973 1.299088 13.58822 0.1140910 1.226041 0.08037908         0
#> 3  29.21846 365.9190 1.527442 24.47704 0.1519433 1.480157 0.07312186         0
#> 4  33.05894 318.1234 1.853494 33.67381 0.1728561 1.811903 0.08277584         0
#> 5  30.25781 309.6420 1.719521 25.59396 0.1810718 1.668636 0.09339675         0
#> 6  25.10086 518.4136 1.102428 15.78233 0.1211694 1.049196 0.05617450         0
#> 7  33.77577 395.5324 1.698300 24.97436 0.1710009 1.646783 0.08982993         0
#> 8  28.01394 626.7569 1.118986 13.76882 0.1254505 1.056907 0.06527362         0
#> 9  33.07828 369.7231 1.720302 18.05033 0.1255277 1.647818 0.10270313         0
#> 10 30.30124 310.4753 1.719677 24.45215 0.1772154 1.666386 0.09494286         0
#> 11 31.40423 320.4116 1.754422 19.74430 0.1493825 1.686924 0.10474504         0
#> 12 28.62284 310.0199 1.625616 27.82136 0.1655443 1.581396 0.07704482         0
#> 13 30.25678 559.6359 1.278998 14.93236 0.1296702 1.213665 0.07596023         0
#> 14 30.86243 415.0730 1.514844 21.56199 0.1561678 1.461534 0.08061539         0
#> 15 26.26513 586.1273 1.084885 16.33867 0.1371866 1.034311 0.05770720         0
#> 16 35.52126 438.1828 1.696917 16.96794 0.1357708 1.620790 0.10887354         0
#> 17 34.68594 473.9150 1.593321 16.61977 0.1259290 1.520321 0.09735697         0
#> 18 33.79725 517.7572 1.485314 14.44196 0.1157351 1.406820 0.09569552         0
#> 19 27.82917 607.0664 1.129489 16.65152 0.1292302 1.077841 0.05787764         0
#> 20 31.72035 431.7867 1.526523 17.74766 0.1237667 1.461092 0.08511151         0
#>    sigma_sq        phi Tau.case.var((Intercept))
#> 1  184.3368 0.22386248                  210.1038
#> 2  190.9087 0.16696996                  418.5886
#> 3  193.4357 0.20412557                  172.4833
#> 4  201.9785 0.26273125                  116.1449
#> 5  180.5473 0.43424865                  129.0947
#> 6  191.7112 0.13503238                  326.7024
#> 7  220.6118 0.35660866                  174.9206
#> 8  204.6175 0.28034328                  422.1393
#> 9  154.0761 0.10187674                  215.6470
#> 10 175.1339 0.42151041                  135.3414
#> 11 148.8684 0.27798620                  171.5432
#> 12 179.1405 0.26845866                  130.8794
#> 13 200.1666 0.26002310                  359.4693
#> 14 206.5943 0.29493185                  208.4787
#> 15 230.9537 0.27304287                  355.1737
#> 16 177.3652 0.23522664                  260.8175
#> 17 184.9470 0.15132722                  288.9680
#> 18 173.8769 0.13811535                  343.8803
#> 19 238.7643 0.18275789                  368.3022
#> 20 177.1026 0.09355983                  254.6841