Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object

a g_REML object

nsim

number of models to simulate

seed

seed value. See documentation for simulate

parallel

if TRUE, run in parallel using foreach backend.

...

additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment,
                 random = ~ 1 | case,
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)

suppressWarnings(
  Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))
)

if (requireNamespace("plyr", quietly = TRUE)) {
  simulate(Laski_g, nsim = 20)
}
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#>      p_beta  r_theta delta_AB       nu     kappa     g_AB     V_g_AB cnvg_warn
#> 1  31.51554 232.0769 2.068754 28.96078 0.1485025 2.014712 0.09875357         0
#> 2  32.51652 581.4900 1.348444 15.79285 0.1438531 1.283377 0.08267978         0
#> 3  34.31373 245.1825 2.191408 32.24844 0.1553471 2.140045 0.10114732         0
#> 4  33.30675 594.5187 1.365996 14.77463 0.1366183 1.295461 0.08688404         0
#> 5  29.87792 647.3501 1.174305 11.98690 0.1058891 1.099265 0.07434791         0
#> 6  30.17788 377.4215 1.553372 21.31337 0.1482383 1.498061 0.08172449         0
#> 7  35.42837 467.4627 1.638618 14.37022 0.1261552 1.551582 0.11660446         0
#> 8  28.29825 272.9576 1.712821 26.19573 0.1394953 1.663309 0.07788843         0
#> 9  24.70640 362.0446 1.298460 26.45446 0.1747486 1.261296 0.06417659         0
#> 10 33.93319 508.9047 1.504203 15.12726 0.1291838 1.428372 0.09710880         0
#> 11 28.34037 325.8630 1.569958 40.84496 0.1777432 1.540952 0.06285491         0
#> 12 36.06815 330.9542 1.982622 35.57848 0.1866789 1.940532 0.09205399         0
#> 13 33.26534 344.0433 1.793434 35.56615 0.1807882 1.755347 0.07957231         0
#> 14 38.11470 394.5398 1.918877 24.78586 0.1818456 1.860221 0.11093447         0
#> 15 33.15600 578.3965 1.378635 13.97243 0.1166450 1.303286 0.08717118         0
#> 16 32.37918 490.1503 1.462518 17.02521 0.1217548 1.397130 0.08180264         0
#> 17 32.91170 304.9813 1.884576 24.47357 0.1539017 1.826226 0.09960592         0
#> 18 34.95534 552.9189 1.486561 14.68935 0.1256272 1.409347 0.09682405         0
#> 19 28.63544 602.3831 1.166722 12.38215 0.1209740 1.094596 0.07495841         0
#> 20 31.00577 342.9819 1.674199 24.39016 0.1638638 1.622184 0.08716598         0
#>    sigma_sq        phi Tau.case.var((Intercept))
#> 1  132.5448 0.11685747                  99.53211
#> 2  227.3084 0.36491970                 354.18167
#> 3  148.3256 0.13954086                  96.85689
#> 4  214.7329 0.34144968                 379.78581
#> 5  169.6220 0.18795459                 477.72812
#> 6  183.3425 0.23046593                 194.07900
#> 7  159.7303 0.25257183                 307.73236
#> 8  146.5835 0.07406214                 126.37405
#> 9  209.3972 0.36376655                 152.64735
#> 10 184.0090 0.24546832                 324.89570
#> 11 223.8770 0.24494650                 101.98602
#> 12 221.5111 0.35125898                 109.44303
#> 13 226.8177 0.30648719                 117.22560
#> 14 228.0581 0.45806334                 166.48171
#> 15 187.8673 0.17271230                 390.52917
#> 16 193.7921 0.09614773                 296.35819
#> 17 161.8048 0.22068369                 143.17657
#> 18 192.8220 0.23014805                 360.09691
#> 19 171.9982 0.33350576                 430.38484
#> 20 185.6109 0.30601619                 157.37107