Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object

a g_REML object

nsim

number of models to simulate

seed

seed value. See documentation for simulate

parallel

if TRUE, run in parallel using foreach backend.

...

additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment,
                 random = ~ 1 | case,
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
simulate(Laski_g, nsim = 20)
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#>      p_beta  r_theta delta_AB       nu      kappa      g_AB     V_g_AB
#> 1  29.43336 366.1402 1.538212 21.98934 0.15470253 1.4851439 0.08069743
#> 2  33.45850 295.3615 1.946837 39.18727 0.17778836 1.9093373 0.08153539
#> 3  31.89695 492.1046 1.437873 18.07960 0.13858667 1.3773887 0.08010355
#> 4  31.15388 245.5634 1.988064 86.94799 0.20026556 1.9708656 0.06330601
#> 5  28.85532 400.7093 1.441489 26.64157 0.14222946 1.4005241 0.06102726
#> 6  28.07529 366.1955 1.467127 20.01673 0.12374305 1.4114604 0.07212242
#> 7  29.86738 642.4049 1.178400 12.54120 0.12017774 1.1064950 0.07510346
#> 8  28.16446 535.3392 1.217270 13.14231 0.11174900 1.1464563 0.07386154
#> 9  26.69706 492.2889 1.203243 18.16422 0.15354716 1.1528680 0.06633111
#> 10 29.67248 712.3430 1.111756 12.72111 0.11666402 1.0448957 0.06680720
#> 11 29.84297 542.2145 1.281612 11.77877 0.10116665 1.1982369 0.08673208
#> 12 27.18255 404.3624 1.351776 21.29369 0.15644333 1.3035989 0.06997278
#> 13 31.05293 488.3154 1.405246 17.47163 0.12433550 1.3440478 0.07567676
#> 14 31.21962 599.5958 1.274965 14.32598 0.13025091 1.2070321 0.07843103
#> 15 35.05169 242.1950 2.252301 80.48884 0.20250019 2.2312482 0.07313289
#> 16 35.19204 829.5589 1.221858 10.71162 0.09668348 1.1342624 0.08658252
#> 17 29.79402 220.5958 2.005998 36.92853 0.16646059 1.9649798 0.08395866
#> 18 30.30316 252.2056 1.908141 61.63928 0.21059400 1.8848293 0.07469201
#> 19 26.93892 632.8718 1.070835 11.47145 0.09923295 0.9992639 0.06497740
#> 20 31.28577 276.6876 1.880841 28.20064 0.15210816 1.8303720 0.08840004
#>    cnvg_warn sigma_sq        phi Tau.case.var((Intercept))
#> 1          0 183.6349 0.27288648                 182.50528
#> 2          0 200.1062 0.25628023                  95.25524
#> 3          0 210.2614 0.22531601                 281.84321
#> 4          0 218.9903 0.23264191                  26.57307
#> 5          0 217.8574 0.09166983                 182.85195
#> 6          0 164.0950 0.03562875                 202.10055
#> 7          0 185.8155 0.31008827                 456.58939
#> 8          0 160.4439 0.16881267                 374.89535
#> 9          0 219.3076 0.36500369                 272.98129
#> 10         0 207.5559 0.25468152                 504.78714
#> 11         0 136.9720 0.14880794                 405.24247
#> 12         0 199.9757 0.30342569                 204.38663
#> 13         0 197.9961 0.10745006                 290.31921
#> 14         0 206.4781 0.29871511                 393.11774
#> 15         0 212.6080 0.26185926                  29.58706
#> 16         0 176.8333 0.20587215                 652.72561
#> 17         0 143.2949 0.18772275                  77.30087
#> 18         0 211.3046 0.37133572                  40.90097
#> 19         0 152.7644 0.15362293                 480.10742
#> 20         0 156.9335 0.15493674                 119.75410