Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object
a g_REML object

nsim
number of models to simulate

seed
seed value. See documentation for simulate

parallel
if TRUE, run in parallel using foreach backend.

...
additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment, 
                 random = ~ 1 | case, 
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

simulate(Laski_g, nsim = 20)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#>      p_beta  r_theta delta_AB       nu     kappa      g_AB     V_g_AB cnvg_warn
#> 1  27.58816 402.0859 1.375825 26.49539 0.1574190 1.3365090 0.06229090         0
#> 2  36.26670 634.4187 1.439859 14.01437 0.1334027 1.3614035 0.09790755         0
#> 3  33.76129 464.7482 1.566067 18.78872 0.1547185 1.5027100 0.09333274         0
#> 4  29.95856 480.9123 1.366117 16.56397 0.1307046 1.3033129 0.07740237         0
#> 5  30.44950 373.9008 1.574715 22.03570 0.1608825 1.5205033 0.08526319         0
#> 6  26.59863 349.4853 1.422803 21.07356 0.1630818 1.3715582 0.07753953         0
#> 7  37.13415 256.6069 2.318138 32.66398 0.1420959 2.2645007 0.10512953         0
#> 8  26.94015 645.3276 1.060498 12.10323 0.1054044 0.9933962 0.06215890         0
#> 9  32.93018 450.1235 1.552131 17.45810 0.1383729 1.4844822 0.09268088         0
#> 10 33.10291 418.6961 1.617770 29.34568 0.1725582 1.5760688 0.07634681         0
#> 11 33.10902 420.0525 1.615454 21.22463 0.1507988 1.5576891 0.08761769         0
#> 12 34.61913 482.9330 1.575334 14.38762 0.1222848 1.4917629 0.10790471         0
#> 13 32.30312 419.8704 1.576474 17.95016 0.1515841 1.5096751 0.09673674         0
#> 14 29.99140 542.4514 1.287705 18.57377 0.1490598 1.2349984 0.06989670         0
#> 15 32.82584 416.8640 1.607751 19.41454 0.1397355 1.5448317 0.09001171         0
#> 16 28.74777 334.2585 1.572400 23.40200 0.1437052 1.5214627 0.07612624         0
#> 17 27.70585 321.5294 1.545116 27.35090 0.1702269 1.5023562 0.07470652         0
#> 18 24.61548 380.3402 1.262183 31.13817 0.1522452 1.2315353 0.04993057         0
#> 19 34.93255 370.7865 1.814130 22.73191 0.1503494 1.7536100 0.09860246         0
#> 20 29.64976 821.5753 1.034421 10.71669 0.1031214 0.9602991 0.06619543         0
#>    sigma_sq        phi Tau.case.var((Intercept))
#> 1  223.4493 0.21995757                  178.6367
#> 2  215.9862 0.35226697                  418.4325
#> 3  212.2942 0.35537089                  252.4540
#> 4  188.8758 0.20045198                  292.0365
#> 5  190.6110 0.32614550                  183.2898
#> 6  175.0553 0.36868724                  174.4300
#> 7  153.3759 0.02901559                  103.2310
#> 8  171.2475 0.17214567                  474.0801
#> 9  187.3153 0.24327142                  262.8083
#> 10 251.5121 0.30570343                  167.1840
#> 11 204.6416 0.25499901                  215.4109
#> 12 163.7278 0.21060503                  319.2052
#> 13 184.4616 0.35336228                  235.4088
#> 14 242.0405 0.30848094                  300.4108
#> 15 187.6729 0.19855405                  229.1911
#> 16 170.0025 0.15132355                  164.2559
#> 17 186.3866 0.31284369                  135.1428
#> 18 225.6313 0.12508622                  154.7089
#> 19 187.7448 0.22051419                  183.0417
#> 20 178.7220 0.29162126                  642.8533