Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object
a g_REML object

nsim
number of models to simulate

seed
seed value. See documentation for simulate

parallel
if TRUE, run in parallel using foreach backend.

...
additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment, 
                 random = ~ 1 | case, 
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

simulate(Laski_g, nsim = 20)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#>      p_beta  r_theta delta_AB       nu     kappa     g_AB     V_g_AB cnvg_warn
#> 1  34.06913 437.8778 1.628113 17.23813 0.1434402 1.556234 0.10256468         0
#> 2  30.51021 655.5469 1.191635 12.69172 0.1265089 1.119802 0.07729239         0
#> 3  34.42911 378.2152 1.770339 22.22728 0.1355345 1.709924 0.09238166         0
#> 4  30.63653 417.3605 1.499629 24.31969 0.1556907 1.452902 0.07283939         0
#> 5  28.22870 364.4688 1.478634 24.05340 0.1709524 1.432045 0.07714879         0
#> 6  27.70462 440.6559 1.319783 19.96291 0.1455187 1.269570 0.06752760         0
#> 7  26.54510 467.3953 1.227841 15.91951 0.1413852 1.169072 0.07105100         0
#> 8  30.00971 367.4666 1.565499 21.32592 0.1607011 1.509790 0.08656624         0
#> 9  29.08553 577.2528 1.210581 13.44663 0.1051950 1.141781 0.07021928         0
#> 10 33.35858 381.3747 1.708172 28.84692 0.1873509 1.663372 0.08798566         0
#> 11 31.30628 276.5015 1.882707 26.67670 0.1510135 1.829275 0.09206507         0
#> 12 23.10154 277.5919 1.386556 33.35789 0.1713774 1.355146 0.05946477         0
#> 13 29.12454 240.4094 1.878380 42.93201 0.1917845 1.845373 0.07920477         0
#> 14 31.47600 414.2642 1.546467 17.13743 0.1408912 1.477786 0.09432831         0
#> 15 33.31852 325.4499 1.846901 31.93341 0.1771451 1.803182 0.08688058         0
#> 16 35.19925 623.8558 1.409261 14.48525 0.1381689 1.335012 0.09323020         0
#> 17 31.28705 298.0288 1.812322 28.43294 0.1560017 1.764093 0.08447368         0
#> 18 34.03272 296.4407 1.976641 40.08675 0.1659575 1.939427 0.07775957         0
#> 19 27.60864 582.7618 1.143667 15.04088 0.1377420 1.085675 0.06611569         0
#> 20 27.29978 474.9982 1.252602 15.38077 0.1263710 1.190513 0.07092069         0
#>    sigma_sq       phi Tau.case.var((Intercept))
#> 1  182.7201 0.2997870                 255.15770
#> 2  196.5727 0.3726910                 458.97421
#> 3  184.3320 0.1000280                 193.88316
#> 4  221.4810 0.2382520                 195.87953
#> 5  199.7872 0.3732649                 164.68163
#> 6  204.4140 0.2373548                 236.24191
#> 7  182.4238 0.3333243                 284.97147
#> 8  183.9472 0.3407653                 183.51949
#> 9  175.6294 0.0756748                 401.62347
#> 10 238.0312 0.4350180                 143.34342
#> 11 152.4377 0.1647353                 124.06382
#> 12 175.0766 0.2542605                 102.51531
#> 13 173.2028 0.3335966                  67.20663
#> 14 171.0004 0.2787318                 243.26381
#> 15 204.4285 0.3128092                 121.02146
#> 16 222.6091 0.3754415                 401.24667
#> 17 170.7747 0.1838904                 127.25413
#> 18 198.2143 0.1626764                  98.22647
#> 19 214.5627 0.3388720                 368.19915
#> 20 173.3054 0.2049654                 301.69278