Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object

a g_REML object

nsim

number of models to simulate

seed

seed value. See documentation for simulate

parallel

if TRUE, run in parallel using foreach backend.

...

additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment,
                 random = ~ 1 | case,
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
simulate(Laski_g, nsim = 20)
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#>      p_beta  r_theta delta_AB       nu     kappa     g_AB     V_g_AB cnvg_warn
#> 1  30.00728 319.4609 1.678873 27.06408 0.1758907 1.631914 0.08545404         0
#> 2  30.88966 718.2778 1.152569 11.49549 0.1184206 1.075700 0.07788604         0
#> 3  35.54850 424.3582 1.725659 18.88459 0.1563489 1.656205 0.10811280         0
#> 4  30.37972 342.5892 1.641334 21.86955 0.1405720 1.584394 0.08457005         0
#> 5  34.31484 252.9183 2.157704 39.37773 0.1811209 2.116345 0.09374736         0
#> 6  34.75401 266.4354 2.129163 44.00771 0.2058091 2.092670 0.09544486         0
#> 7  32.85575 593.6389 1.348498 14.37304 0.1383538 1.276886 0.08766603         0
#> 8  31.78452 603.9610 1.293335 13.28538 0.1108228 1.218922 0.08067235         0
#> 9  35.69746 407.2364 1.768944 19.11286 0.1488460 1.698609 0.10885973         0
#> 10 28.95069 258.7250 1.799864 75.67389 0.2035750 1.781967 0.06338407         0
#> 11 30.41214 490.1040 1.373735 19.41700 0.1460204 1.319981 0.07298203         0
#> 12 23.59115 396.9700 1.184051 27.24012 0.1865828 1.151148 0.06213821         0
#> 13 31.79540 281.5376 1.894943 46.73551 0.1744817 1.864370 0.06995456         0
#> 14 35.03777 815.8752 1.226661 12.62381 0.1050683 1.152311 0.07607693         0
#> 15 29.49097 540.5210 1.268478 17.28529 0.1505872 1.212631 0.07264176         0
#> 16 27.41652 516.0568 1.206878 19.46286 0.1494031 1.159766 0.06226635         0
#> 17 27.92577 445.0233 1.323774 16.11178 0.1282786 1.261181 0.07480463         0
#> 18 26.76197 277.7094 1.605916 43.17018 0.1878679 1.577854 0.06623132         0
#> 19 30.08199 309.6872 1.709405 57.06269 0.1925364 1.686838 0.06341996         0
#> 20 27.26763 319.9863 1.524340 32.69154 0.2032678 1.489099 0.07855034         0
#>    sigma_sq       phi Tau.case.var((Intercept))
#> 1  187.1795 0.3639078                 132.28148
#> 2  184.7238 0.3951384                 533.55401
#> 3  195.4508 0.3679592                 228.90743
#> 4  166.6547 0.1515021                 175.93458
#> 5  172.8349 0.2795123                  80.08339
#> 6  201.9256 0.4297367                  64.50978
#> 7  210.5367 0.3845998                 383.10222
#> 8  183.0324 0.1491503                 420.92862
#> 9  185.0617 0.2904006                 222.17463
#> 10 224.0342 0.2811519                  34.69081
#> 11 223.5460 0.2559978                 266.55802
#> 12 242.2222 0.4546107                 154.74777
#> 13 200.9425 0.1884228                  80.59509
#> 14 229.2032 0.1276574                 586.67197
#> 15 230.7826 0.3684227                 309.73844
#> 16 237.6001 0.2857327                 278.45670
#> 17 169.8983 0.1936619                 275.12497
#> 18 198.4951 0.3038220                  79.21427
#> 19 242.6606 0.2692564                  67.02657
#> 20 223.7041 0.5251094                  96.28223