Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object
a g_REML object

nsim
number of models to simulate

seed
seed value. See documentation for simulate

parallel
if TRUE, run in parallel using foreach backend.

...
additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment, 
                 random = ~ 1 | case, 
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

simulate(Laski_g, nsim = 20)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#>      p_beta  r_theta delta_AB       nu     kappa     g_AB     V_g_AB cnvg_warn
#> 1  32.04770 625.3738 1.281525 13.60924 0.1231319 1.209579 0.08071561         0
#> 2  33.01146 489.0573 1.492742 17.75251 0.1314544 1.428777 0.08408491         0
#> 3  27.66723 316.2200 1.555862 29.22669 0.1751632 1.515592 0.07399634         0
#> 4  31.45065 277.8468 1.886805 39.16483 0.1864405 1.850441 0.08175158         0
#> 5  30.75727 308.1957 1.752001 23.04754 0.1275755 1.694363 0.08602703         0
#> 6  34.13643 298.5704 1.975580 41.83682 0.1572030 1.939952 0.07269477         0
#> 7  28.93065 577.8444 1.203518 12.71876 0.1094962 1.131126 0.07415207         0
#> 8  37.64901 234.2891 2.459674 50.44776 0.1687496 2.422924 0.08981843         0
#> 9  29.65479 403.7224 1.475888 22.81564 0.1478759 1.426835 0.07214336         0
#> 10 29.47680 543.3786 1.264530 13.60103 0.1097266 1.193494 0.07578783         0
#> 11 31.60813 350.8914 1.687378 21.87889 0.1673794 1.628867 0.09666463         0
#> 12 24.57510 343.0091 1.326913 24.60375 0.1690340 1.286049 0.06638041         0
#> 13 33.06437 357.4548 1.748838 19.88763 0.1422155 1.682047 0.10145454         0
#> 14 26.61686 423.5957 1.293246 19.88272 0.1629334 1.243842 0.07142011         0
#> 15 24.97389 272.0316 1.514177 50.00917 0.1874019 1.491354 0.05882888         0
#> 16 28.28438 266.9165 1.731247 57.80729 0.1932545 1.708688 0.06401477         0
#> 17 35.28129 334.2952 1.929654 27.63886 0.1737299 1.876813 0.10043837         0
#> 18 27.41725 448.1534 1.295122 15.90093 0.1154561 1.233059 0.06988296         0
#> 19 31.21343 464.1401 1.448829 18.35165 0.1442373 1.388800 0.08170438         0
#> 20 33.24348 537.6035 1.433757 14.66306 0.1193013 1.359150 0.08974671         0
#>    sigma_sq        phi Tau.case.var((Intercept))
#> 1  200.3937 0.26506347                 424.98005
#> 2  203.2656 0.16766576                 285.79167
#> 3  190.8644 0.32807589                 125.35556
#> 4  191.8727 0.32028667                  85.97409
#> 5  151.8708 0.01422425                 156.32496
#> 6  200.1763 0.08632181                  98.39409
#> 7  165.4486 0.17232683                 412.39580
#> 8  169.7808 0.13030973                  64.50834
#> 9  203.8888 0.19761785                 199.83357
#> 10 168.9297 0.11776011                 374.44888
#> 11 181.7050 0.38832198                 169.18639
#> 12 188.9415 0.34631162                 154.06764
#> 13 164.3097 0.20930248                 193.14511
#> 14 205.5209 0.40041229                 218.07482
#> 15 203.0719 0.26398838                  68.95971
#> 16 210.2212 0.27123838                  56.69530
#> 17 196.4170 0.33753156                 137.87818
#> 18 165.2789 0.07174141                 282.87449
#> 19 202.9630 0.27006197                 261.17712
#> 20 184.6387 0.16506490                 352.96475