Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object
a g_REML object

nsim
number of models to simulate

seed
seed value. See documentation for simulate

parallel
if TRUE, run in parallel using foreach backend.

...
additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment, 
                 random = ~ 1 | case, 
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

simulate(Laski_g, nsim = 20)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#>      p_beta   r_theta  delta_AB       nu      kappa      g_AB     V_g_AB
#> 1  31.29549  250.9323 1.9756205 83.41781 0.18644714 1.9578045 0.05862315
#> 2  30.71596  322.8337 1.7095217 36.18976 0.18624918 1.6738470 0.07661163
#> 3  28.71743  384.4818 1.4645618 22.32400 0.14955799 1.4148010 0.07303632
#> 4  31.59313  269.2858 1.9252449 36.47648 0.18306372 1.8853864 0.08610044
#> 5  30.10115  524.0217 1.3149477 18.78648 0.14404500 1.2617439 0.06979716
#> 6  37.35455  394.5869 1.8804952 14.72331 0.11471214 1.7830488 0.14203348
#> 7  33.30155  468.1912 1.5390502 21.97489 0.15635906 1.4859182 0.08132621
#> 8  33.62741  329.0811 1.8537108 28.09782 0.17645660 1.8037865 0.09493934
#> 9  28.03025  588.0647 1.1558844 16.88256 0.13375751 1.1037629 0.06040096
#> 10 28.99882  732.7482 1.0712799 11.10783 0.08966689 0.9972817 0.06511362
#> 11 26.35761  264.8158 1.6196982 38.37981 0.19442746 1.5878393 0.07332584
#> 12 28.97608  506.1086 1.2880056 14.36481 0.12694587 1.2195665 0.07861211
#> 13 31.14463  315.6928 1.7528744 29.70748 0.15259892 1.7082455 0.07705868
#> 14 25.56171  370.4555 1.3280728 23.53616 0.17127735 1.2852983 0.06901978
#> 15 27.72283  426.5206 1.3423557 17.07323 0.15541412 1.2825118 0.08080537
#> 16 29.52443  398.8490 1.4783502 22.78400 0.15168019 1.4291462 0.07355014
#> 17 26.81285  612.3880 1.0835019 15.46815 0.13721928 1.0301034 0.05995821
#> 18 30.34074  426.9419 1.4683912 16.40663 0.12265342 1.4002277 0.08527770
#> 19 31.14129  501.3555 1.3907968 15.72855 0.12997979 1.3234070 0.08293294
#> 20 29.32453 1034.7420 0.9116225 10.35257 0.09393032 0.8439450 0.05370839
#>    cnvg_warn sigma_sq        phi Tau.case.var((Intercept))
#> 1          0 215.4683 0.15394307                  35.46395
#> 2          0 217.0685 0.34257238                 105.76514
#> 3          0 192.3986 0.22141256                 192.08319
#> 4          0 180.0856 0.31602249                  89.20021
#> 5          0 232.7717 0.25532269                 291.24999
#> 6          0 134.9211 0.11355735                 259.66576
#> 7          0 235.6317 0.28767403                 232.55946
#> 8          0 196.1428 0.35354865                 132.93824
#> 9          0 236.0567 0.21872529                 352.00791
#> 10         0 164.1763 0.06664429                 568.57188
#> 11         0 185.7793 0.38637264                  79.03659
#> 12         0 173.2128 0.26126350                 332.89579
#> 13         0 183.5241 0.14225077                 132.16873
#> 14         0 201.2499 0.38656963                 169.20555
#> 15         0 183.9390 0.42757694                 242.58155
#> 16         0 202.7439 0.23101551                 196.10511
#> 17         0 230.5931 0.31175942                 381.79487
#> 18         0 163.8390 0.12633405                 263.10293
#> 19         0 188.1041 0.22641856                 313.25136
#> 20         0 204.9268 0.21768323                 829.81523