Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object
a g_REML object

nsim
number of models to simulate

seed
seed value. See documentation for simulate

parallel
if TRUE, run in parallel using foreach backend.

...
additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment, 
                 random = ~ 1 | case, 
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

simulate(Laski_g, nsim = 20)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#>      p_beta  r_theta delta_AB       nu     kappa     g_AB     V_g_AB cnvg_warn
#> 1  30.63641 416.4538 1.501255 26.20855 0.1755942 1.457880 0.07603014         0
#> 2  33.61854 351.0922 1.794189 23.62799 0.1644773 1.736629 0.09856452         0
#> 3  31.40088 330.2974 1.727783 27.76758 0.1568879 1.680692 0.08067706         0
#> 4  32.54026 416.3101 1.594823 21.49193 0.1479587 1.538514 0.08429143         0
#> 5  25.38580 281.8495 1.512106 37.47650 0.1933269 1.481642 0.06916508         0
#> 6  30.07621 369.0370 1.565626 20.31068 0.1442434 1.507093 0.08461005         0
#> 7  28.27283 464.5020 1.311824 16.87900 0.1329047 1.252658 0.07224437         0
#> 8  30.19507 321.3232 1.684477 33.96852 0.1939668 1.647009 0.08112732         0
#> 9  28.93677 454.3190 1.357593 13.59963 0.1298134 1.281322 0.09046792         0
#> 10 32.93945 340.2654 1.785696 19.19932 0.1342314 1.715019 0.10581779         0
#> 11 30.96192 345.7387 1.665152 25.11553 0.1615567 1.614928 0.08396241         0
#> 12 28.24986 331.6777 1.551165 28.52212 0.1590355 1.510016 0.06934667         0
#> 13 31.77969 427.7394 1.536597 17.30149 0.1276307 1.469011 0.08896885         0
#> 14 32.98004 254.3453 2.067946 83.82558 0.2115027 2.049389 0.07078849         0
#> 15 27.28407 470.1320 1.258344 23.04793 0.1580910 1.216947 0.06137493         0
#> 16 35.73502 426.4834 1.730386 23.34325 0.1546070 1.674188 0.09123618         0
#> 17 25.89296 355.0601 1.374139 39.18402 0.1629950 1.347669 0.05158980         0
#> 18 28.32768 362.4405 1.487965 15.96116 0.1266781 1.416934 0.09033771         0
#> 19 32.24650 476.7793 1.476808 15.81181 0.1184955 1.405633 0.08789485         0
#> 20 27.73737 313.0750 1.567621 24.90577 0.1540536 1.519936 0.07547745         0
#>    sigma_sq       phi Tau.case.var((Intercept))
#> 1  240.5248 0.3749424                 175.92900
#> 2  187.3266 0.3250999                 163.76555
#> 3  187.4931 0.1990701                 142.80427
#> 4  203.1269 0.2242106                 213.18322
#> 5  195.6166 0.3855904                  86.23294
#> 6  172.6380 0.2174440                 196.39897
#> 7  186.0936 0.2105518                 278.40840
#> 8  216.8242 0.4249204                 104.49896
#> 9  148.3630 0.3399643                 305.95603
#> 10 150.5022 0.1532887                 189.76315
#> 11 188.7974 0.2745512                 156.94123
#> 12 191.3452 0.2073388                 140.33246
#> 13 173.0445 0.1449645                 254.69488
#> 14 230.9805 0.3074879                  23.36481
#> 15 243.8040 0.2811696                 226.32799
#> 16 220.9950 0.2458635                 205.48846
#> 17 234.4134 0.1461127                 120.64674
#> 18 136.8029 0.1836155                 225.63753
#> 19 175.7975 0.1065053                 300.98185
#> 20 167.6534 0.2151585                 145.42162