Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object

a g_REML object

nsim

number of models to simulate

seed

seed value. See documentation for simulate

parallel

if TRUE, run in parallel using foreach backend.

...

additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment,
                 random = ~ 1 | case,
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)

suppressWarnings(
  Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))
)

if (requireNamespace("plyr", quietly = TRUE)) {
  simulate(Laski_g, nsim = 20)
}
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#>      p_beta  r_theta  delta_AB        nu     kappa      g_AB     V_g_AB
#> 1  34.06266 464.6969 1.5801329  15.55556 0.1267453 1.5027035 0.10210408
#> 2  31.70023 197.1925 2.2574454 116.62324 0.1921925 2.2428967 0.05911478
#> 3  33.52444 348.1012 1.7968376  22.37726 0.1573557 1.7359341 0.10061653
#> 4  30.00657 426.0096 1.4538066  19.31804 0.1533929 1.3966243 0.08168951
#> 5  33.22305 700.1133 1.2556119  13.09064 0.1280099 1.1822737 0.08211176
#> 6  28.04785 235.4172 1.8280182  43.04754 0.1956188 1.7959834 0.07837780
#> 7  27.67472 608.9670 1.1214665  12.51284 0.1203703 1.0528772 0.06963504
#> 8  21.81682 853.9732 0.7465679  12.05574 0.1097963 0.6991397 0.03780029
#> 9  29.81436 334.8790 1.6292269  24.21509 0.1640157 1.5782394 0.08448296
#> 10 25.03373 300.0876 1.4451123  24.54691 0.1883644 1.4005044 0.08046880
#> 11 34.07442 321.0248 1.9017751  31.60956 0.1827559 1.8562920 0.09287071
#> 12 30.87477 338.7641 1.6774712  29.22791 0.1716799 1.6340553 0.07970648
#> 13 33.61859 504.0248 1.4974539  16.10187 0.1381414 1.4266049 0.09372495
#> 14 37.03155 508.0263 1.6429668  18.54606 0.1488947 1.5756177 0.09947451
#> 15 30.03258 496.3168 1.3480721  16.21452 0.1384335 1.2847407 0.07934128
#> 16 39.71249 590.7778 1.6338608  12.70729 0.1132573 1.5354931 0.12705885
#> 17 33.09278 426.1037 1.6031555  19.28899 0.1568846 1.5400026 0.09533055
#> 18 32.32720 471.2900 1.4891004  18.02445 0.1257246 1.4262672 0.08117931
#> 19 32.48401 296.5945 1.8862014  26.13138 0.1433029 1.8315424 0.09150977
#> 20 27.81418 352.7560 1.4809114  19.83913 0.1459881 1.4242124 0.07990377
#>    cnvg_warn sigma_sq       phi Tau.case.var((Intercept))
#> 1          0 171.4396 0.2011271                293.257322
#> 2          0 190.6788 0.1203086                  6.513645
#> 3          0 177.3692 0.2880194                170.731980
#> 4          0 197.2376 0.3266780                228.772036
#> 5          0 218.3500 0.3570884                481.763249
#> 6          0 171.5164 0.3607241                 63.900830
#> 7          0 175.7664 0.3147825                433.200628
#> 8          0 227.7683 0.2288956                626.204942
#> 9          0 180.6517 0.3103823                154.227288
#> 10         0 179.5158 0.5331314                120.571885
#> 11         0 203.8657 0.3605906                117.159154
#> 12         0 202.7353 0.3001125                136.028858
#> 13         0 196.7378 0.2924016                307.287073
#> 14         0 226.3569 0.3077898                281.669407
#> 15         0 194.9776 0.2905575                301.339263
#> 16         0 170.4228 0.2165285                420.354961
#> 17         0 199.1207 0.3603888                226.983024
#> 18         0 196.2476 0.1045203                275.042422
#> 19         0 159.9205 0.1073292                136.673989
#> 20         0 163.1488 0.2447087                189.607240