Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment,
                 random = ~ 1 | case,
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
simulate(Laski_g, nsim = 20)
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#>      p_beta  r_theta delta_AB       nu     kappa     g_AB     V_g_AB cnvg_warn
#> 1  31.11151 339.4954 1.688512 29.25639 0.1799549 1.644853 0.08327529         0
#> 2  31.38655 247.5522 1.994850 36.73484 0.1680406 1.953843 0.08417684         0
#> 3  24.22039 453.0630 1.137894 14.54627 0.1055297 1.078199 0.05920060         0
#> 4  26.94054 258.0259 1.677160 59.59891 0.1891655 1.655966 0.06005296         0
#> 5  32.47658 218.3660 2.197747 62.04025 0.2049165 2.171071 0.08183699         0
#> 6  24.06119 520.0766 1.055075 15.48218 0.1223759 1.003126 0.05382774         0
#> 7  31.06972 666.1624 1.203780 11.83540 0.1129404 1.125852 0.08008485         0
#> 8  30.78519 442.6399 1.463243 15.85675 0.1158323 1.392926 0.08568724         0
#> 9  27.51563 534.8290 1.189795 14.14927 0.1123118 1.125594 0.06677194         0
#> 10 32.57317 535.6466 1.407411 12.34110 0.1010033 1.320111 0.09774705         0
#> 11 29.32553 438.5089 1.400415 17.43169 0.1547766 1.339285 0.08419035         0
#> 12 28.77957 454.8834 1.349380 13.37823 0.1012617 1.272292 0.08399867         0
#> 13 28.61471 673.9634 1.102227 13.24810 0.1151858 1.038628 0.06329128         0
#> 14 30.88625 297.2420 1.791472 42.39488 0.1906878 1.759592 0.07548675         0
#> 15 31.39797 295.5596 1.826329 38.52910 0.1875891 1.790546 0.07999802         0
#> 16 26.08121 440.2688 1.242993 16.64172 0.1240622 1.186120 0.06512260         0
#> 17 29.58636 666.5368 1.145986 12.72333 0.1081044 1.077080 0.06806529         0
#> 18 26.84648 503.2192 1.196765 14.10508 0.1114519 1.131982 0.06738676         0
#> 19 31.01612 631.8119 1.233939 12.83293 0.1095347 1.160390 0.07668039         0
#> 20 32.24763 368.1946 1.680580 25.48102 0.1634141 1.630624 0.08476684         0
#>    sigma_sq        phi Tau.case.var((Intercept))
#> 1  207.7125 0.36662025                 131.78289
#> 2  160.9262 0.20109089                  86.62609
#> 3  150.7119 0.02224737                 302.35110
#> 4  203.2140 0.23724234                  54.81191
#> 5  180.2604 0.33267495                  38.10561
#> 6  189.3821 0.15971672                 330.69449
#> 7  174.6986 0.28755577                 491.46381
#> 8  162.9249 0.07731659                 279.71497
#> 9  174.6345 0.11610305                 360.19448
#> 10 144.7591 0.10032083                 390.88745
#> 11 191.0953 0.40528005                 247.41360
#> 12 136.6505 0.03448837                 318.23288
#> 13 205.4746 0.20047711                 468.48885
#> 14 212.7281 0.32910707                  84.51395
#> 15 203.6080 0.33359966                  91.95167
#> 16 171.4385 0.13203276                 268.83026
#> 17 190.3631 0.15589319                 476.17366
#> 18 163.5005 0.10902514                 339.71870
#> 19 182.9901 0.16462261                 448.82176
#> 20 203.3186 0.28426855                 164.87605