Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object

a g_REML object

nsim

number of models to simulate

seed

seed value. See documentation for simulate

parallel

if TRUE, run in parallel using foreach backend.

...

additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment,
                 random = ~ 1 | case,
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)

suppressWarnings(
  Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))
)

if (requireNamespace("plyr", quietly = TRUE)) {
  simulate(Laski_g, nsim = 20)
}
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#>      p_beta  r_theta delta_AB       nu     kappa     g_AB     V_g_AB cnvg_warn
#> 1  28.26985 273.3788 1.709783 61.84764 0.1920492 1.688965 0.06117199         0
#> 2  31.33989 230.8121 2.062852 58.50859 0.1896190 2.036296 0.07320390         0
#> 3  28.03734 495.8736 1.259074 16.22248 0.1256171 1.199954 0.06821247         0
#> 4  33.63951 319.1997 1.882862 23.84064 0.1441460 1.823001 0.09859079         0
#> 5  29.75931 330.2523 1.637570 18.39550 0.1355612 1.569885 0.09572578         0
#> 6  26.77094 275.9560 1.611549 34.83341 0.1610420 1.576600 0.06460592         0
#> 7  33.14913 550.6877 1.412601 13.42541 0.1256637 1.332190 0.09650551         0
#> 8  27.26962 551.6562 1.161034 15.05750 0.1451114 1.102227 0.06964429         0
#> 9  31.60770 538.1456 1.362520 20.41970 0.1492235 1.311856 0.07049995         0
#> 10 31.47164 526.6077 1.371437 14.16547 0.1204001 1.297521 0.08626829         0
#> 11 28.90829 534.5020 1.250397 12.78573 0.1223165 1.175587 0.08177435         0
#> 12 30.47050 375.0076 1.573474 22.66455 0.1485200 1.520825 0.07953195         0
#> 13 30.69443 602.3544 1.250644 16.85514 0.1270793 1.194156 0.06583557         0
#> 14 26.47680 587.5512 1.092302 13.18391 0.1152397 1.028962 0.06267834         0
#> 15 29.77632 411.7582 1.467405 16.47596 0.1499079 1.399578 0.09256091         0
#> 16 25.90133 275.2387 1.561232 62.07262 0.1733204 1.542292 0.05021130         0
#> 17 33.74982 367.6950 1.760061 21.55273 0.1484864 1.698095 0.09770319         0
#> 18 32.94569 507.6059 1.462296 20.69338 0.1657538 1.408649 0.08230720         0
#> 19 31.64149 314.9872 1.782832 29.96152 0.1402201 1.737829 0.07472029         0
#> 20 32.81232 453.5443 1.540732 20.86539 0.1452659 1.484679 0.08119445         0
#>    sigma_sq        phi Tau.case.var((Intercept))
#> 1  218.7649 0.24873362                  54.61382
#> 2  181.0158 0.24419829                  49.79624
#> 3  189.4507 0.16281222                 306.42286
#> 4  164.1451 0.14790319                 155.05456
#> 5  142.0355 0.18748473                 188.21683
#> 6  173.7670 0.16222972                 102.18907
#> 7  175.1230 0.30541112                 375.56468
#> 8  208.8466 0.41865817                 342.80964
#> 9  255.2443 0.25954187                 282.90128
#> 10 174.9601 0.20249317                 351.64756
#> 11 159.2775 0.31495891                 375.22444
#> 12 188.9030 0.20596337                 186.10452
#> 13 238.3373 0.15443933                 364.01707
#> 14 178.1272 0.20533666                 409.42399
#> 15 169.6781 0.39613297                 242.08004
#> 16 213.5413 0.12091306                  61.69743
#> 17 179.8967 0.22757579                 187.79837
#> 18 254.1108 0.40396536                 253.49506
#> 19 180.7619 0.03789212                 134.22531
#> 20 216.2151 0.21390487                 237.32918