Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object
a g_REML object

nsim
number of models to simulate

seed
seed value. See documentation for simulate

parallel
if TRUE, run in parallel using foreach backend.

...
additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment, 
                 random = ~ 1 | case, 
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

simulate(Laski_g, nsim = 20)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#>      p_beta  r_theta delta_AB       nu     kappa     g_AB     V_g_AB cnvg_warn
#> 1  36.85990 310.5055 2.091797 38.13343 0.1847492 2.050384 0.09337092         0
#> 2  34.13526 276.3663 2.053339 51.23216 0.1732055 2.023133 0.07218637         0
#> 3  28.87958 358.1014 1.526117 21.97086 0.1512299 1.473422 0.07877758         0
#> 4  30.41401 312.3569 1.720871 29.44216 0.1479159 1.676658 0.07417802         0
#> 5  31.61299 428.3919 1.527373 18.72242 0.1448426 1.465360 0.08717978         0
#> 6  34.10010 250.8520 2.153014 32.07217 0.1682620 2.102271 0.10315713         0
#> 7  31.52175 521.6208 1.380171 16.04615 0.1321083 1.314640 0.08114063         0
#> 8  28.75064 439.4755 1.371451 24.07397 0.1623238 1.328276 0.06756816         0
#> 9  34.88418 324.8262 1.935544 31.18261 0.2010188 1.888614 0.10298118         0
#> 10 27.17669 350.0922 1.452464 23.56071 0.1667897 1.405732 0.07506161         0
#> 11 35.07654 410.8301 1.730556 27.12740 0.1612843 1.682265 0.08365865         0
#> 12 34.65989 335.0699 1.893475 23.03307 0.1326538 1.831143 0.09908748         0
#> 13 28.84387 737.7499 1.061937 13.97739 0.1265823 1.003918 0.05999646         0
#> 14 28.69019 724.9572 1.065558 12.99969 0.1189529 1.002877 0.06195455         0
#> 15 34.19887 422.1133 1.664552 18.70558 0.1347415 1.596907 0.09664920         0
#> 16 33.79102 542.3827 1.450936 12.83660 0.1038537 1.364479 0.09991432         0
#> 17 27.06793 488.0826 1.225204 17.18302 0.1503758 1.170937 0.06957118         0
#> 18 32.95535 557.3006 1.395986 19.59459 0.1640578 1.341863 0.07989468         0
#> 19 27.10672 329.4323 1.493461 25.12047 0.1600307 1.448424 0.07224658         0
#> 20 34.53462 392.4077 1.743356 17.22705 0.1321192 1.666339 0.11132086         0
#>    sigma_sq        phi Tau.case.var((Intercept))
#> 1  211.6681 0.31546553                  98.83735
#> 2  202.5701 0.15893814                  73.79622
#> 3  178.1703 0.24296664                 179.93106
#> 4  179.6054 0.10690887                 132.75144
#> 5  190.1945 0.26458718                 238.19731
#> 6  154.8140 0.24231725                  96.03802
#> 7  200.0699 0.23448024                 321.55096
#> 8  235.4142 0.29850031                 204.06133
#> 9  221.9151 0.52691040                 102.91104
#> 10 187.7056 0.34643733                 162.38660
#> 11 232.6117 0.24302478                 178.21840
#> 12 166.1249 0.06030701                 168.94506
#> 13 245.3928 0.27978176                 492.35704
#> 14 218.1497 0.26000865                 506.80751
#> 15 183.4636 0.17110576                 238.64973
#> 16 155.3792 0.09858916                 387.00345
#> 17 207.4039 0.37037918                 280.67875
#> 18 269.4279 0.42060768                 287.87266
#> 19 179.3142 0.26174220                 150.11805
#> 20 159.5327 0.19091990                 232.87500