Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object

a g_REML object

nsim

number of models to simulate

seed

seed value. See documentation for simulate

parallel

if TRUE, run in parallel using foreach backend.

...

additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment,
                 random = ~ 1 | case,
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)

suppressWarnings(
  Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))
)

if (requireNamespace("plyr", quietly = TRUE)) {
  simulate(Laski_g, nsim = 20)
}
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#>      p_beta   r_theta  delta_AB        nu      kappa      g_AB     V_g_AB
#> 1  32.98833  241.6752 2.1219944 44.717994 0.18628296 2.0862047 0.08648948
#> 2  29.83205  268.0953 1.8219583 31.294817 0.16551862 1.7779423 0.08249012
#> 3  22.97301  354.5126 1.2201185 38.299568 0.20168842 1.1960686 0.06114739
#> 4  26.47666  402.3451 1.3199695 19.489029 0.14386303 1.2685128 0.06823573
#> 5  30.08921  613.2445 1.2150495 15.769930 0.14125906 1.1563323 0.07047845
#> 6  34.74035  320.5706 1.9403154 27.524926 0.17750853 1.8869610 0.10286849
#> 7  24.12722  311.3505 1.3673591 38.514581 0.17463315 1.3405583 0.05577233
#> 8  22.20646  510.1588 0.9831658 19.171618 0.17514444 0.9441958 0.05808074
#> 9  28.74093  451.8748 1.3520468 16.607260 0.12454258 1.2900539 0.07438654
#> 10 27.16685  664.1949 1.0541236 12.757700 0.10717698 0.9909151 0.05912577
#> 11 27.50697 1177.5697 0.8015848  9.882372 0.09699604 0.7391715 0.04614240
#> 12 26.47781  472.6385 1.2179165 18.964669 0.13602610 1.1691079 0.06018347
#> 13 29.29551  422.2185 1.4257142 16.243404 0.13390600 1.3588562 0.08498594
#> 14 27.18933  387.1802 1.3817897 26.921691 0.14869603 1.3429342 0.05926633
#> 15 27.46638  408.4543 1.3590321 23.758085 0.17393118 1.3156737 0.07139173
#> 16 36.74216  280.2446 2.1948054 20.653220 0.14775481 2.1141268 0.14451665
#> 17 33.82248  510.3653 1.4971483 13.091607 0.11547411 1.4097090 0.10630762
#> 18 29.04425  306.1430 1.6599613 40.521545 0.18115736 1.6290469 0.06802030
#> 19 29.02265  367.9933 1.5129238 24.281381 0.13973432 1.4657066 0.06895496
#> 20 31.00460  644.9515 1.2208507 14.940218 0.13707683 1.1585209 0.07278745
#>    cnvg_warn sigma_sq       phi Tau.case.var((Intercept))
#> 1          0 174.0101 0.2833285                  67.66513
#> 2          0 162.9441 0.2286156                 105.15124
#> 3          0 255.0505 0.4440167                  99.46204
#> 4          0 183.1099 0.2343804                 219.23513
#> 5          0 237.4535 0.3390120                 375.79100
#> 6          0 190.0529 0.3705530                 130.51777
#> 7          0 208.3763 0.2376162                 102.97424
#> 8          0 262.2287 0.5719880                 247.93003
#> 9          0 175.7974 0.1380195                 276.07742
#> 10         0 189.9907 0.1426993                 474.20416
#> 11         0 214.0270 0.3422971                 963.54273
#> 12         0 207.9230 0.1760824                 264.71551
#> 13         0 164.2850 0.2443699                 257.93353
#> 14         0 213.6217 0.1423644                 173.55848
#> 15         0 224.8619 0.4059104                 183.59242
#> 16         0 133.4313 0.2408916                 146.81326
#> 17         0 153.5227 0.2142237                 356.84261
#> 18         0 211.2341 0.2720031                  94.90891
#> 19         0 189.8660 0.1028968                 178.12733
#> 20         0 235.5862 0.3372207                 409.36530