Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object

a g_REML object

nsim

number of models to simulate

seed

seed value. See documentation for simulate

parallel

if TRUE, run in parallel using foreach backend.

...

additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment,
                 random = ~ 1 | case,
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)

suppressWarnings(
  Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))
)

if (requireNamespace("plyr", quietly = TRUE)) {
  simulate(Laski_g, nsim = 20)
}
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#>      p_beta  r_theta delta_AB        nu     kappa     g_AB     V_g_AB cnvg_warn
#> 1  31.15212 706.1399 1.172309  14.17688 0.1428316 1.109177 0.07321263         0
#> 2  27.81108 527.8263 1.210521  13.81870 0.1300418 1.143610 0.07459027         0
#> 3  31.15107 356.7092 1.649361  20.92789 0.1321889 1.589538 0.08594208         0
#> 4  29.11018 476.3675 1.333748  13.95336 0.1245889 1.260751 0.08459672         0
#> 5  26.28169 312.0249 1.487848  30.84069 0.1633815 1.451370 0.06413993         0
#> 6  31.49443 255.7500 1.969363 103.72851 0.2190006 1.955090 0.06705359         0
#> 7  32.76837 338.4661 1.781137  22.82633 0.1575177 1.721966 0.09782744         0
#> 8  31.22140 392.8393 1.575233  21.33213 0.1487918 1.519194 0.08351380         0
#> 9  28.35329 298.6398 1.640702  38.88876 0.1774537 1.608855 0.06735200         0
#> 10 34.06057 309.5650 1.935869  21.06692 0.1498668 1.866123 0.11608524         0
#> 11 29.64803 466.6687 1.372434  14.92436 0.1229380 1.302289 0.08308744         0
#> 12 34.78874 408.5953 1.721044  26.42123 0.1364408 1.671724 0.07706405         0
#> 13 32.22333 439.7389 1.536643  17.83815 0.1463797 1.471117 0.09195922         0
#> 14 34.25900 373.6796 1.772250  15.44239 0.1250393 1.684760 0.12454308         0
#> 15 32.38154 442.7592 1.538912  15.74483 0.1294857 1.464424 0.09738380         0
#> 16 33.90614 433.6337 1.628234  24.35111 0.1608841 1.577565 0.08303832         0
#> 17 30.33232 277.3303 1.821407  47.66994 0.1762866 1.792599 0.06688338         0
#> 18 33.90840 682.7621 1.297695  11.83987 0.1012671 1.213719 0.08822708         0
#> 19 29.71120 479.9999 1.356125  15.84546 0.1417883 1.290907 0.08254497         0
#> 20 32.37837 435.9138 1.550796  20.41394 0.1427563 1.493114 0.08264645         0
#>    sigma_sq        phi Tau.case.var((Intercept))
#> 1  252.5679 0.45136512                453.571979
#> 2  175.3029 0.32768543                352.523419
#> 3  166.6218 0.09489329                190.087427
#> 4  157.3431 0.25965022                319.024321
#> 5  187.7401 0.21645272                124.284784
#> 6  254.7179 0.29635522                  1.032147
#> 7  174.3906 0.28049324                164.075531
#> 8  191.1479 0.23494418                201.691394
#> 9  201.6594 0.25588364                 96.980344
#> 10 149.8453 0.25025927                159.719674
#> 11 164.3045 0.19055773                302.364209
#> 12 219.5251 0.04493036                189.070169
#> 13 189.5280 0.30696750                250.210869
#> 14 136.4632 0.18865928                237.216396
#> 15 166.0903 0.22074718                276.668889
#> 16 232.8397 0.28149341                200.794051
#> 17 199.7901 0.19680612                 77.540113
#> 18 173.8646 0.14458951                508.897508
#> 19 186.8773 0.34091042                293.122624
#> 20 203.9911 0.20166942                231.922676