Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object

a g_REML object

nsim

number of models to simulate

seed

seed value. See documentation for simulate

parallel

if TRUE, run in parallel using foreach backend.

...

additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment,
                 random = ~ 1 | case,
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)

suppressWarnings(
  Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))
)

if (requireNamespace("plyr", quietly = TRUE)) {
  simulate(Laski_g, nsim = 20)
}
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#>      p_beta  r_theta  delta_AB       nu      kappa      g_AB     V_g_AB
#> 1  35.29193 739.7718 1.2975574 11.65852 0.11100262 1.2122555 0.09172282
#> 2  28.97745 668.7707 1.1205249 14.07674 0.13961780 1.0597446 0.06814948
#> 3  28.70985 528.3628 1.2490066 14.33173 0.13955040 1.1824839 0.07857016
#> 4  30.78326 372.1596 1.5956953 22.33141 0.16201827 1.5414972 0.08636992
#> 5  33.95248 379.6146 1.7426094 27.27516 0.15046760 1.6942486 0.08068916
#> 6  31.91082 442.2092 1.5174840 18.49459 0.14725668 1.4551031 0.08791176
#> 7  29.85702 257.3504 1.8611617 68.56025 0.18439967 1.8407274 0.05985406
#> 8  25.95690 660.8806 1.0096978 12.92392 0.12088488 0.9499473 0.05790125
#> 9  29.62911 364.9819 1.5508970 21.93309 0.12915618 1.4972528 0.07434566
#> 10 32.19232 576.4360 1.3408392 12.68895 0.10703947 1.2599940 0.08864721
#> 11 30.21416 702.1956 1.1402010 11.64628 0.09967446 1.0651633 0.07133933
#> 12 29.02016 655.6483 1.1333507 13.41135 0.12384203 1.0687667 0.06758576
#> 13 32.01654 578.7129 1.3308921 15.75678 0.13633546 1.2665223 0.07905953
#> 14 27.89991 422.2722 1.3577086 16.85141 0.14847806 1.2963715 0.08072106
#> 15 28.83555 895.8801 0.9633925 11.13347 0.10523038 0.8970033 0.05737183
#> 16 35.33873 674.1125 1.3610833 11.82274 0.11388001 1.2728749 0.09896960
#> 17 28.57303 380.9409 1.4639542 18.27429 0.14419941 1.4030383 0.08323614
#> 18 29.81744 363.7756 1.5633406 21.37628 0.15528590 1.5078407 0.08449501
#> 19 30.83243 547.3568 1.3178699 20.97963 0.16320872 1.2701893 0.07065162
#> 20 28.91732 242.0504 1.8586819 51.65068 0.19439278 1.8315615 0.07220416
#>    cnvg_warn sigma_sq        phi Tau.case.var((Intercept))
#> 1          0 188.7702 0.28146813                 551.00156
#> 2          0 234.1817 0.41947504                 434.58897
#> 3          0 187.8035 0.40095411                 340.55927
#> 4          0 191.6534 0.32977102                 180.50619
#> 5          0 211.3704 0.15260236                 168.24423
#> 6          0 195.8380 0.29398574                 246.37112
#> 7          0 208.8085 0.17732153                  48.54184
#> 8          0 198.5678 0.28763787                 462.31275
#> 9          0 174.8737 0.04775251                 190.10820
#> 10         0 163.6897 0.14599453                 412.74636
#> 11         0 173.7647 0.14243223                 528.43085
#> 12         0 207.1684 0.28601377                 448.47993
#> 13         0 220.8194 0.28931378                 357.89359
#> 14         0 175.9067 0.36477748                 246.36555
#> 15         0 209.2841 0.26624534                 686.59592
#> 16         0 177.0407 0.30037398                 497.07187
#> 17         0 166.0764 0.27209977                 214.86449
#> 18         0 179.7912 0.29128173                 183.98443
#> 19         0 273.6233 0.37227161                 273.73342
#> 20         0 185.2406 0.30471286                  56.80983