Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object

a g_REML object

nsim

number of models to simulate

seed

seed value. See documentation for simulate

parallel

if TRUE, run in parallel using foreach backend.

...

additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment,
                 random = ~ 1 | case,
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)

suppressWarnings(
  Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))
)

if (requireNamespace("plyr", quietly = TRUE)) {
  simulate(Laski_g, nsim = 20)
}
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#>      p_beta  r_theta delta_AB       nu      kappa      g_AB     V_g_AB
#> 1  28.51455 532.4669 1.235720 18.65521 0.15493528 1.1853657 0.06781707
#> 2  29.74228 560.9366 1.255791 15.45635 0.13664806 1.1938535 0.07371450
#> 3  31.07813 304.8851 1.779864 33.33924 0.15521299 1.7395214 0.07331651
#> 4  31.27333 685.1320 1.194777 12.44734 0.11949610 1.1213118 0.07712398
#> 5  32.76380 419.7512 1.599183 17.93349 0.13692179 1.5313581 0.09456457
#> 6  31.57941 533.8465 1.366771 16.75091 0.15051771 1.3046484 0.08250755
#> 7  37.48200 390.3524 1.897118 17.16207 0.14404348 1.8129861 0.13235140
#> 8  34.59667 660.9454 1.345710 13.55222 0.11540026 1.2698368 0.08578034
#> 9  29.23828 471.0409 1.347170 23.45467 0.14813424 1.3036284 0.06272994
#> 10 31.16641 384.7938 1.588813 22.61570 0.15884292 1.5355345 0.08403078
#> 11 35.18040 525.2818 1.534987 17.04622 0.13950214 1.4664452 0.09325122
#> 12 34.93846 503.8954 1.556444 15.06374 0.13005143 1.4776436 0.10336884
#> 13 28.84919 343.4138 1.556771 40.03121 0.18351381 1.5274212 0.06509130
#> 14 30.51949 499.1804 1.365993 15.72540 0.12939284 1.2997918 0.08047652
#> 15 27.69628 474.6089 1.271316 19.54033 0.13982044 1.2218876 0.06353713
#> 16 27.42728 550.9977 1.168444 13.38517 0.12774329 1.1017273 0.07197168
#> 17 29.88420 449.2655 1.409906 26.65570 0.16792362 1.3698602 0.06738873
#> 18 29.96579 791.2146 1.065316 10.19504 0.09225924 0.9849761 0.07064546
#> 19 31.56148 454.8286 1.479904 16.78913 0.14301659 1.4127944 0.09024264
#> 20 31.10114 443.7944 1.476337 19.84481 0.14676875 1.4198294 0.07976363
#>    cnvg_warn sigma_sq       phi Tau.case.var((Intercept))
#> 1          0 242.3117 0.3617990                  290.1553
#> 2          0 210.7430 0.3064438                  350.1936
#> 3          0 186.9723 0.1289685                  117.9128
#> 4          0 195.9343 0.3100679                  489.1978
#> 5          0 177.6686 0.2140960                  242.0826
#> 6          0 223.0257 0.3900703                  310.8208
#> 7          0 162.6190 0.3084696                  227.7334
#> 8          0 206.9691 0.1836043                  453.9763
#> 9          0 241.6911 0.1885663                  229.3498
#> 10         0 197.8919 0.2961187                  186.9020
#> 11         0 215.2189 0.2685826                  310.0629
#> 12         0 181.8711 0.2574811                  322.0243
#> 13         0 237.2168 0.2925646                  106.1970
#> 14         0 187.0135 0.2206371                  312.1669
#> 15         0 214.6362 0.1961295                  259.9728
#> 16         0 175.7953 0.3317201                  375.2024
#> 17         0 256.0305 0.3038678                  193.2350
#> 18         0 151.0841 0.2181004                  640.1305
#> 19         0 185.8784 0.3126970                  268.9501
#> 20         0 205.6387 0.2516412                  238.1557