Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object
a g_REML object

nsim
number of models to simulate

seed
seed value. See documentation for simulate

parallel
if TRUE, run in parallel using foreach backend.

...
additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment,
                 random = ~ 1 | case,
                 correlation = corAR1(0, ~ time | case),
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))
simulate(Laski_g, nsim = 20)

#>      p_beta  r_theta delta_AB       nu     kappa     g_AB     V_g_AB cnvg_warn
#> 1  29.62406 483.2743 1.347559 14.44969 0.1127522 1.276383 0.08048654         0
#> 2  31.97675 292.2949 1.870354 36.94495 0.1754553 1.832126 0.07975869         0
#> 3  33.28597 328.2686 1.837158 28.51422 0.1564674 1.788408 0.08608637         0
#> 4  33.94081 426.6243 1.643234 24.15506 0.1572972 1.591679 0.08341013         0
#> 5  28.19241 615.6216 1.136254 12.84542 0.1156927 1.068595 0.06832614         0
#> 6  31.13966 468.1536 1.439195 25.58394 0.1955737 1.396588 0.08105400         0
#> 7  26.73176 619.0728 1.074377 15.30860 0.1203733 1.020867 0.05522596         0
#> 8  32.55998 227.0071 2.161048 58.29463 0.2015878 2.133125 0.08167979         0
#> 9  31.85814 382.0806 1.629832 24.78660 0.1651074 1.580014 0.08350958         0
#> 10 31.49839 348.7527 1.686668 34.27216 0.1803855 1.649486 0.07567767         0
#> 11 27.05622 328.5405 1.492700 26.63768 0.1558317 1.450274 0.06809189         0
#> 12 33.66873 435.6089 1.613163 22.83513 0.1569543 1.559594 0.08460830         0
#> 13 30.46689 370.7049 1.582392 41.59520 0.1911186 1.553687 0.06775609         0
#> 14 29.20605 604.1107 1.188268 15.61961 0.1432601 1.130284 0.06939716         0
#> 15 27.27485 364.0289 1.429534 17.53947 0.1391823 1.367522 0.08154394         0
#> 16 33.72233 366.5670 1.761330 19.44069 0.1325097 1.692495 0.10186219         0
#> 17 28.59614 306.6662 1.632956 27.35333 0.1550694 1.587769 0.07495193         0
#> 18 26.49537 355.6687 1.404905 24.73153 0.1357459 1.361866 0.06027797         0
#> 19 29.90296 275.3239 1.802155 30.68798 0.1531217 1.757750 0.07838962         0
#> 20 31.97828 652.2058 1.252169 14.68177 0.1346376 1.187095 0.07594012         0
#>    sigma_sq        phi Tau.case.var((Intercept))
#> 1  161.5095 0.10567920                 321.76484
#> 2  193.0713 0.25522632                  99.22355
#> 3  188.4990 0.18670999                 139.76961
#> 4  226.3512 0.25454816                 200.27306
#> 5  181.0759 0.23419377                 434.54572
#> 6  301.2658 0.59926756                 166.88785
#> 7  222.2189 0.14655496                 396.85386
#> 8  182.7706 0.32502435                  44.23647
#> 9  209.0380 0.30973160                 173.04260
#> 10 226.6761 0.31491970                 122.07664
#> 11 182.5123 0.20498574                 146.02823
#> 12 224.2010 0.27546444                 211.40792
#> 13 264.0482 0.33755699                 106.65670
#> 14 233.6499 0.36728237                 370.46084
#> 15 152.2929 0.24827742                 211.73604
#> 16 163.0771 0.13124423                 203.48998
#> 17 172.3013 0.18941882                 134.36496
#> 18 184.3424 0.06151188                 171.32638
#> 19 162.5082 0.13648998                 112.81579
#> 20 232.7996 0.32544220                 419.40619