Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object
a g_REML object

nsim
number of models to simulate

seed
seed value. See documentation for simulate

parallel
if TRUE, run in parallel using foreach backend.

...
additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment, 
                 random = ~ 1 | case, 
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

simulate(Laski_g, nsim = 20)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#>      p_beta  r_theta delta_AB       nu     kappa     g_AB     V_g_AB cnvg_warn
#> 1  25.74852 470.7882 1.186696 17.93110 0.1412019 1.136358 0.06200158         0
#> 2  31.68796 304.6059 1.815621 32.28756 0.1879391 1.773117 0.08843449         0
#> 3  27.90450 291.6288 1.634026 32.29937 0.1622233 1.595788 0.06928203         0
#> 4  29.24883 342.2602 1.580994 30.02662 0.1857560 1.541173 0.07805016         0
#> 5  29.40773 701.8533 1.110039 12.65346 0.1103349 1.042919 0.06544830         0
#> 6  32.08318 638.7136 1.269476 12.39733 0.1244900 1.191096 0.08672833         0
#> 7  29.98216 357.3213 1.586111 27.27817 0.1710371 1.542098 0.07754925         0
#> 8  31.05253 502.3719 1.385429 14.97460 0.1297614 1.314862 0.08591235         0
#> 9  31.57100 468.7141 1.458258 17.71830 0.1393329 1.395647 0.08339295         0
#> 10 31.40786 248.2789 1.993280 28.38437 0.1431650 1.940144 0.09319921         0
#> 11 35.43673 714.7403 1.325499 11.00271 0.1044108 1.233046 0.09912131         0
#> 12 37.07041 471.2791 1.707608 16.86313 0.1465357 1.630518 0.11375779         0
#> 13 34.57596 712.7128 1.295141 11.40656 0.1153171 1.208075 0.09438893         0
#> 14 30.06245 440.0861 1.433031 20.76105 0.1559759 1.380631 0.07675240         0
#> 15 33.79554 380.3397 1.732900 21.90620 0.1616839 1.672886 0.09836281         0
#> 16 30.35676 564.7963 1.277349 16.19596 0.1289884 1.217270 0.07071255         0
#> 17 31.24903 361.4828 1.643587 22.88220 0.1704298 1.589121 0.09125429         0
#> 18 33.26982 334.2440 1.819779 27.68896 0.1862963 1.770038 0.09722767         0
#> 19 33.13846 394.7264 1.667955 25.68057 0.1667210 1.618763 0.08455955         0
#> 20 36.71217 396.1263 1.844562 20.68624 0.1298884 1.776867 0.10341893         0
#>    sigma_sq        phi Tau.case.var((Intercept))
#> 1  201.0963 0.25466054                  269.6918
#> 2  198.0038 0.39457924                  106.6021
#> 3  178.5047 0.19314833                  113.1241
#> 4  215.3607 0.40449950                  126.8995
#> 5  200.0001 0.18681644                  501.8532
#> 6  185.0259 0.37504720                  453.6877
#> 7  207.3121 0.32050882                  150.0093
#> 8  180.1988 0.25886514                  322.1731
#> 9  197.6395 0.24380507                  271.0747
#> 10 139.4252 0.07897410                  108.8537
#> 11 163.3376 0.27138687                  551.4028
#> 12 195.2133 0.34471451                  276.0658
#> 13 178.9543 0.36466365                  533.7585
#> 14 214.2439 0.31186436                  225.8423
#> 15 193.7034 0.33612989                  186.6363
#> 16 216.9052 0.19743973                  347.8911
#> 17 193.2679 0.39298709                  168.2149
#> 18 204.8348 0.44418819                  129.4092
#> 19 220.4892 0.30867552                  174.2372
#> 20 183.0260 0.07877587                  213.1003