Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object

a g_REML object

nsim

number of models to simulate

seed

seed value. See documentation for simulate

parallel

if TRUE, run in parallel using foreach backend.

...

additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment,
                 random = ~ 1 | case,
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)

suppressWarnings(
  Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))
)

if (requireNamespace("plyr", quietly = TRUE)) {
  simulate(Laski_g, nsim = 20)
}
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#>      p_beta  r_theta delta_AB       nu     kappa     g_AB     V_g_AB cnvg_warn
#> 1  28.79960 575.7479 1.200246 12.02809 0.1077480 1.123817 0.07731207         0
#> 2  32.46999 287.2121 1.915935 39.97358 0.1877736 1.879761 0.08270094         0
#> 3  33.25333 285.6600 1.967481 50.49550 0.1817379 1.938113 0.07239326         0
#> 4  34.68942 516.4225 1.526492 17.28495 0.1329823 1.459285 0.08953749         0
#> 5  27.08640 569.9881 1.134536 12.79877 0.1290072 1.066729 0.07179348         0
#> 6  33.64786 551.3527 1.432989 14.91395 0.1317182 1.359697 0.09154423         0
#> 7  28.08713 331.5740 1.542472 36.48114 0.1842087 1.510542 0.06786638         0
#> 8  31.38700 613.3367 1.267361 14.24600 0.1207972 1.199447 0.07560331         0
#> 9  29.45827 583.7467 1.219256 12.37267 0.1001913 1.143824 0.07567604         0
#> 10 32.03840 204.4441 2.240699 44.43469 0.1682290 2.202665 0.08629782         0
#> 11 31.78663 442.9423 1.510327 15.62288 0.1244394 1.436642 0.09375880         0
#> 12 27.93918 262.6116 1.724077 22.64812 0.1409357 1.666347 0.08875444         0
#> 13 28.57557 467.2263 1.322000 18.79628 0.1501543 1.268539 0.07214255         0
#> 14 32.90248 376.5254 1.695631 20.42649 0.1343363 1.632601 0.09226444         0
#> 15 33.63142 561.9963 1.418660 13.90300 0.1324467 1.340729 0.09599669         0
#> 16 29.50331 364.6535 1.545007 27.57103 0.1629262 1.502595 0.07184134         0
#> 17 36.89623 266.6506 2.259491 38.41947 0.1784715 2.215094 0.10033145         0
#> 18 32.80599 296.3183 1.905785 25.27298 0.1708252 1.848664 0.10437450         0
#> 19 29.23873 510.2109 1.294445 14.67156 0.1347455 1.227127 0.07993590         0
#> 20 32.92518 284.0996 1.953407 25.40811 0.1704002 1.895173 0.10756237         0
#>    sigma_sq        phi Tau.case.var((Intercept))
#> 1  152.2792 0.20674318                 423.46870
#> 2  200.3170 0.32427547                  86.89514
#> 3  211.4964 0.22227285                  74.16360
#> 4  210.8636 0.19731610                 305.55894
#> 5  174.2296 0.39373434                 395.75846
#> 6  198.0113 0.28225876                 353.34139
#> 7  222.3782 0.32462418                 109.19575
#> 8  205.1617 0.20235510                 408.17508
#> 9  158.0782 0.08817141                 425.66845
#> 10 142.1135 0.15470852                  62.33056
#> 11 163.2915 0.17484149                 279.65081
#> 12 130.5004 0.14025204                 132.11117
#> 13 210.7249 0.31192121                 256.50136
#> 14 173.7004 0.12516850                 202.82495
#> 15 189.2535 0.34892042                 372.74280
#> 16 208.7056 0.25052255                 155.94787
#> 17 179.6806 0.26674694                  86.97008
#> 18 166.0853 0.34992122                 130.23298
#> 19 182.0603 0.32716656                 328.15057
#> 20 159.4335 0.34402137                 124.66611