Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object

a g_REML object

nsim

number of models to simulate

seed

seed value. See documentation for simulate

parallel

if TRUE, run in parallel using foreach backend.

...

additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment,
                 random = ~ 1 | case,
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)

suppressWarnings(
  Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))
)

if (requireNamespace("plyr", quietly = TRUE)) {
  simulate(Laski_g, nsim = 20)
}
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#>      p_beta  r_theta  delta_AB        nu      kappa      g_AB     V_g_AB
#> 1  29.61511 447.1344 1.4005361 18.260178 0.15734093 1.3422134 0.08213262
#> 2  30.23602 286.5174 1.7862785 24.796988 0.16885839 1.7317012 0.09595546
#> 3  30.89889 269.3135 1.8828414 34.794273 0.16122619 1.8419626 0.07869728
#> 4  26.27656 762.8109 0.9513932 11.630068 0.10822976 0.8886919 0.05478910
#> 5  32.19970 349.8077 1.7216195 19.836230 0.14316682 1.6556949 0.09945752
#> 6  35.48601 932.3725 1.1621505 11.047453 0.10108878 1.0814266 0.07783576
#> 7  32.11575 226.3171 2.1348110 26.692013 0.17156398 2.0742594 0.11843393
#> 8  30.90956 318.0683 1.7331356 27.138608 0.16169674 1.6847935 0.08394023
#> 9  29.82202 463.6667 1.3849502 24.739525 0.16902237 1.3425356 0.06946357
#> 10 31.30010 275.3158 1.8863841 34.587201 0.17628375 1.8451813 0.08438081
#> 11 29.54313 321.9525 1.6464955 37.061683 0.20835708 1.6129499 0.08152427
#> 12 29.03696 893.1565 0.9715996  9.713473 0.09033418 0.8945983 0.06280761
#> 13 29.64541 394.9714 1.4916763 23.678438 0.14283329 1.4439242 0.06976135
#> 14 33.24619 250.8880 2.0989493 53.455308 0.19918449 2.0693618 0.08197159
#> 15 30.68877 212.8544 2.1034791 56.531197 0.18240024 2.0754482 0.07333923
#> 16 27.53837 564.5282 1.1590318 17.062900 0.13849594 1.1073290 0.06148289
#> 17 29.64273 275.5084 1.7858736 35.586590 0.17377666 1.7479694 0.07663108
#> 18 30.16860 267.4281 1.8448097 35.376384 0.16356248 1.8054203 0.07652033
#> 19 32.59790 284.8897 1.9313066 25.891798 0.16195949 1.8748175 0.10145490
#> 20 34.79667 631.9171 1.3842280 14.100638 0.11538941 1.3092733 0.08673297
#>    cnvg_warn sigma_sq       phi Tau.case.var((Intercept))
#> 1          0 202.4676 0.3992451                 244.66684
#> 2          0 158.3246 0.3413804                 128.19276
#> 3          0 169.5590 0.1639639                  99.75455
#> 4          0 192.3856 0.2500070                 570.42528
#> 5          0 160.8260 0.2192015                 188.98168
#> 6          0 212.4653 0.2227644                 719.90724
#> 7          0 130.2413 0.3334144                  96.07578
#> 8          0 180.2742 0.2462334                 137.79415
#> 9          0 256.0529 0.3437942                 207.61379
#> 10         0 177.8570 0.2807674                  97.45877
#> 11         0 237.5315 0.5144607                  84.42106
#> 12         0 152.1140 0.2748079                 741.04250
#> 13         0 201.9191 0.1392103                 193.05230
#> 14         0 196.1543 0.3292990                  54.73370
#> 15         0 163.0940 0.2023777                  49.76040
#> 16         0 230.9225 0.2582264                 333.60574
#> 17         0 178.9194 0.2533134                  96.58898
#> 18         0 170.1002 0.1773487                  97.32788
#> 19         0 158.0179 0.2659534                 126.87179
#> 20         0 206.7666 0.1520938                 425.15045