Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object
a g_REML object

nsim
number of models to simulate

seed
seed value. See documentation for simulate

parallel
if TRUE, run in parallel using foreach backend.

...
additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment, 
                 random = ~ 1 | case, 
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

simulate(Laski_g, nsim = 20)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#>      p_beta  r_theta delta_AB       nu     kappa      g_AB     V_g_AB cnvg_warn
#> 1  33.31134 356.3809 1.764554 19.09362 0.1466348 1.6943225 0.10785209         0
#> 2  35.14503 291.1628 2.059662 51.39102 0.1666569 2.0294565 0.07006277         0
#> 3  27.59518 540.7160 1.186721 14.68517 0.1356472 1.1250631 0.07040030         0
#> 4  33.63325 309.3747 1.912170 33.01653 0.1990797 1.8684018 0.09721312         0
#> 5  28.80649 261.3130 1.782009 24.36596 0.1221365 1.7265885 0.08296855         0
#> 6  28.29368 387.1741 1.437925 30.44940 0.1706301 1.4022144 0.06462877         0
#> 7  33.84704 506.1391 1.504478 15.07033 0.1389380 1.4283418 0.10018203         0
#> 8  31.01414 460.3957 1.445420 17.21292 0.1426952 1.3815125 0.08522348         0
#> 9  30.22565 299.8953 1.745384 26.54548 0.1312141 1.6956017 0.07699813         0
#> 10 26.97123 407.2675 1.336475 23.37320 0.1470715 1.2931268 0.06191804         0
#> 11 25.89679 323.3381 1.440183 29.81491 0.1373893 1.4036482 0.05509691         0
#> 12 33.24579 219.5058 2.243953 50.37870 0.1908253 2.2103800 0.08767204         0
#> 13 33.03683 381.9263 1.690474 18.64390 0.1369676 1.6215464 0.10000197         0
#> 14 32.90201 435.8854 1.575928 21.15847 0.1656345 1.5193981 0.08953279         0
#> 15 32.11602 308.7762 1.827679 37.37884 0.1796515 1.7907596 0.07852363         0
#> 16 26.46993 416.2294 1.297438 23.96926 0.1599327 1.2564128 0.06268291         0
#> 17 27.99202 688.9280 1.066467 12.19043 0.1002833 0.9994801 0.06121732         0
#> 18 27.28538 258.7587 1.696221 45.51938 0.1751020 1.6681193 0.06325509         0
#> 19 28.92137 459.9025 1.348609 19.93889 0.1690490 1.2972368 0.07722155         0
#> 20 32.89794 628.2643 1.312495 14.09610 0.1346207 1.2414008 0.08441331         0
#>    sigma_sq          phi Tau.case.var((Intercept))
#> 1  160.9893  0.270524182                 195.39157
#> 2  211.5979  0.111279013                  79.56487
#> 3  193.6474  0.336061915                 347.06866
#> 4  211.6530  0.480365263                  97.72168
#> 5  132.2040 -0.055394615                 129.10906
#> 6  235.1067  0.277635832                 152.06743
#> 7  187.3798  0.350011613                 318.75930
#> 8  191.5377  0.293476412                 268.85798
#> 9  160.5179 -0.002134569                 139.37738
#> 10 208.1756  0.180837118                 199.09185
#> 11 184.5226  0.015497138                 138.81548
#> 12 165.4313  0.286019010                  54.07445
#> 13 166.2962  0.193605307                 215.63013
#> 14 220.6237  0.390229085                 215.26174
#> 15 206.6160  0.283256244                 102.16024
#> 16 221.2331  0.280049014                 194.99621
#> 17 182.7255  0.103296621                 506.20255
#> 18 183.3272  0.199007368                  75.43148
#> 19 229.1978  0.460574283                 230.70474
#> 20 215.9735  0.360549668                 412.29080