Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object
a g_REML object

nsim
number of models to simulate

seed
seed value. See documentation for simulate

parallel
if TRUE, run in parallel using foreach backend.

...
additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment, 
                 random = ~ 1 | case, 
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

simulate(Laski_g, nsim = 20)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#>      p_beta  r_theta  delta_AB       nu     kappa      g_AB     V_g_AB
#> 1  33.03674 366.6640 1.7252935 23.61154 0.1513882 1.6699046 0.08904231
#> 2  29.94542 700.1357 1.1317206 12.11664 0.1009971 1.0601931 0.06814047
#> 3  30.82650 283.0321 1.8323415 37.52188 0.1856975 1.7954703 0.08081611
#> 4  35.53988 320.2186 1.9860611 27.65096 0.1626716 1.9317000 0.10074320
#> 5  32.82852 289.5366 1.9292993 30.55380 0.1523307 1.8815504 0.08639457
#> 6  25.69632 510.4083 1.1373971 21.14883 0.1389501 1.0965791 0.05180266
#> 7  29.95497 367.8563 1.5618159 24.68181 0.1538374 1.5138718 0.07551597
#> 8  29.68506 447.2288 1.4036958 16.81862 0.1277126 1.3401557 0.07890755
#> 9  30.42877 238.6324 1.9697887 48.20395 0.1726883 1.9389812 0.07115501
#> 10 30.48802 517.1869 1.3406191 16.77506 0.1386767 1.2797743 0.07663754
#> 11 30.10816 503.3365 1.3420075 17.93093 0.1488447 1.2850815 0.07584246
#> 12 33.09342 516.3686 1.4563365 13.28508 0.1145895 1.3725432 0.09973710
#> 13 25.39102 305.3239 1.4531143 30.43142 0.1748996 1.4170049 0.06689567
#> 14 28.10634 490.0349 1.2696699 17.11115 0.1350552 1.2131937 0.06868917
#> 15 26.23005 706.1415 0.9870822 12.51380 0.1243223 0.9267166 0.05838188
#> 16 31.22653 215.4980 2.1271692 76.42493 0.1938861 2.1062256 0.06779689
#> 17 30.71907 545.4680 1.3152956 13.94592 0.1217739 1.2432689 0.08203236
#> 18 30.17568 583.6093 1.2490965 14.38979 0.1451026 1.1828422 0.07996745
#> 19 27.94863 282.6838 1.6623024 29.76622 0.1682836 1.6200637 0.07671428
#> 20 31.66617 348.9535 1.6951641 30.48911 0.1626315 1.6531201 0.07549093
#>    cnvg_warn sigma_sq       phi Tau.case.var((Intercept))
#> 1          0 189.9555 0.2136945                 176.70855
#> 2          0 184.3466 0.1178589                 515.78916
#> 3          0 192.3523 0.3273340                  90.67975
#> 4          0 183.4208 0.2474265                 136.79785
#> 5          0 170.3667 0.1314542                 119.16987
#> 6          0 242.5835 0.1514794                 267.82474
#> 7          0 195.9964 0.2168349                 171.85994
#> 8          0 176.8468 0.1619262                 270.38205
#> 9          0 171.4939 0.1685501                  67.13850
#> 10         0 208.8708 0.2706204                 308.31615
#> 11         0 219.1914 0.3274680                 284.14508
#> 12         0 157.7341 0.1913970                 358.63449
#> 13         0 187.1876 0.3115542                 118.13628
#> 14         0 199.3587 0.2232424                 290.67617
#> 15         0 206.6487 0.3620986                 499.49283
#> 16         0 183.1710 0.2182221                  32.32702
#> 17         0 178.8242 0.2296785                 366.64376
#> 18         0 213.6405 0.4638751                 369.96886
#> 19         0 169.1890 0.2667510                 113.49486
#> 20         0 208.6551 0.2141792                 140.29839