Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object

a g_REML object

nsim

number of models to simulate

seed

seed value. See documentation for simulate

parallel

if TRUE, run in parallel using foreach backend.

...

additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment,
                 random = ~ 1 | case,
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)

suppressWarnings(
  Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))
)

if (requireNamespace("plyr", quietly = TRUE)) {
  simulate(Laski_g, nsim = 20)
}
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#>      p_beta  r_theta delta_AB       nu     kappa     g_AB     V_g_AB cnvg_warn
#> 1  28.14904 377.1051 1.449547 18.47251 0.1522849 1.389887 0.08380269         0
#> 2  25.06756 515.1975 1.104397 14.89460 0.1379956 1.047837 0.06352157         0
#> 3  32.79152 405.6960 1.628025 18.72920 0.1223147 1.561950 0.08987432         0
#> 4  27.89553 284.3730 1.654209 29.47893 0.1619256 1.611763 0.07459386         0
#> 5  30.89362 681.7694 1.183178 13.41984 0.1341909 1.115798 0.07497050         0
#> 6  32.73231 462.8167 1.521501 22.63219 0.1618021 1.470517 0.08013681         0
#> 7  27.09576 517.6580 1.190913 14.44307 0.1340913 1.127982 0.07127186         0
#> 8  36.46110 301.3757 2.100272 31.76980 0.1610472 2.050297 0.09784086         0
#> 9  28.49732 400.0125 1.424844 19.49900 0.1532973 1.369328 0.07884569         0
#> 10 27.36780 361.5147 1.439385 29.55148 0.1620403 1.402543 0.06290747         0
#> 11 30.75415 394.1179 1.549140 23.63682 0.1532016 1.499460 0.07683919         0
#> 12 30.35055 475.7008 1.391553 13.38808 0.1126696 1.312115 0.09113909         0
#> 13 29.58740 572.1222 1.236979 12.71462 0.1028384 1.162550 0.07616566         0
#> 14 29.56033 357.4963 1.563412 19.51796 0.1354908 1.502557 0.08463420         0
#> 15 26.94178 568.6744 1.129781 14.87722 0.1313474 1.071853 0.06375091         0
#> 16 25.55451 336.0913 1.393924 26.92314 0.1643116 1.354729 0.06489898         0
#> 17 31.01355 473.5663 1.425152 27.23349 0.1760540 1.385540 0.07020042         0
#> 18 29.07212 332.2555 1.594926 18.80633 0.1446440 1.530464 0.09270819         0
#> 19 26.13156 283.5947 1.551730 25.32255 0.1437632 1.505313 0.07044499         0
#> 20 37.91587 330.9711 2.084135 21.83933 0.1593824 2.011734 0.12988388         0
#>    sigma_sq        phi Tau.case.var((Intercept))
#> 1  169.1706 0.34231726                  207.9345
#> 2  188.2423 0.34954047                  326.9552
#> 3  173.0135 0.05246482                  232.6825
#> 4  167.3670 0.21950209                  117.0060
#> 5  223.3722 0.40433914                  458.3972
#> 6  239.8820 0.32155506                  222.9347
#> 7  181.7070 0.33325199                  335.9510
#> 8  182.8454 0.18891930                  118.5303
#> 9  186.2283 0.32059845                  213.7842
#> 10 213.0425 0.21960583                  148.4722
#> 11 205.0143 0.22870240                  189.1036
#> 12 146.0338 0.16341200                  329.6670
#> 13 161.5610 0.09503780                  410.5612
#> 14 160.2814 0.15693446                  197.2148
#> 15 203.5690 0.28027528                  365.1054
#> 16 190.8306 0.27047638                  145.2607
#> 17 278.3282 0.36271574                  195.2381
#> 18 147.8946 0.26030457                  184.3609
#> 19 150.5466 0.12226336                  133.0481
#> 20 167.2500 0.31719768                  163.7211