Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object

a g_REML object

nsim

number of models to simulate

seed

seed value. See documentation for simulate

parallel

if TRUE, run in parallel using foreach backend.

...

additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment,
                 random = ~ 1 | case,
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
simulate(Laski_g, nsim = 20)
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#>      p_beta  r_theta  delta_AB       nu     kappa      g_AB     V_g_AB
#> 1  29.27735 406.8405 1.4515087 35.20815 0.1828738 1.4203678 0.06466046
#> 2  28.46779 497.3780 1.2764695 12.83805 0.1149796 1.2004171 0.08242667
#> 3  35.98914 346.4347 1.9335739 33.56030 0.1863289 1.8900384 0.09252024
#> 4  28.33041 457.4630 1.3245701 19.02720 0.1488579 1.2716640 0.07129576
#> 5  36.11315 471.4602 1.6631941 21.47939 0.1518707 1.6044361 0.09095131
#> 6  32.60691 566.9572 1.3694133 12.97737 0.1192331 1.2887163 0.09289055
#> 7  31.55111 297.2502 1.8300107 27.61893 0.1708204 1.7798623 0.09245498
#> 8  32.37322 262.9307 1.9964811 33.49584 0.1644728 1.9514420 0.08863743
#> 9  32.71495 482.9440 1.4886680 14.58391 0.1160691 1.4107757 0.09527988
#> 10 30.58187 712.7901 1.1454685 12.87733 0.1248330 1.0774334 0.07136253
#> 11 31.19823 431.1203 1.5025563 23.85208 0.1657876 1.4548098 0.07734587
#> 12 21.10715 489.5865 0.9539263 14.96938 0.1221903 0.9053207 0.04796523
#> 13 30.41520 430.3361 1.4661785 20.29737 0.1430696 1.4113267 0.07652512
#> 14 31.10771 213.8554 2.1271981 73.62848 0.1942453 2.1054560 0.06909973
#> 15 30.03276 260.3470 1.8613107 25.56517 0.1671644 1.8061666 0.09881895
#> 16 28.57733 444.3771 1.3556443 17.74492 0.1291158 1.2975284 0.07187150
#> 17 30.36712 525.5436 1.3246443 12.43536 0.1118913 1.2431134 0.08960106
#> 18 35.15465 430.8990 1.6935384 20.35549 0.1415308 1.6303639 0.09439377
#> 19 30.45527 689.7625 1.1596119 13.00761 0.1179299 1.0914402 0.07035280
#> 20 33.87915 250.2766 2.1415212 66.00648 0.2136329 2.1170956 0.08121237
#>    cnvg_warn sigma_sq       phi Tau.case.var((Intercept))
#> 1          0 268.6249 0.3254667                 138.21555
#> 2          0 145.9344 0.2265913                 351.44361
#> 3          0 227.0696 0.3679319                 119.36505
#> 4          0 207.2839 0.2929791                 250.17915
#> 5          0 231.8405 0.2588834                 239.61970
#> 6          0 170.3877 0.2648184                 396.56950
#> 7          0 173.2991 0.3139752                 123.95113
#> 8          0 163.9575 0.1998636                  98.97321
#> 9          0 163.9369 0.1346744                 319.00703
#> 10         0 215.9238 0.3369552                 496.86629
#> 11         0 231.9051 0.3321707                 199.21519
#> 12         0 172.6120 0.1806609                 316.97460
#> 13         0 200.7678 0.2071975                 229.56834
#> 14         0 180.1097 0.2278056                  33.74575
#> 15         0 145.2807 0.3142100                 115.06633
#> 16         0 183.8630 0.1454597                 260.51414
#> 17         0 146.9491 0.2217433                 378.59450
#> 18         0 200.8111 0.1919883                 230.08784
#> 19         0 207.1525 0.2478456                 482.60995
#> 20         0 215.2862 0.3745339                  34.99034