Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object

a g_REML object

nsim

number of models to simulate

seed

seed value. See documentation for simulate

parallel

if TRUE, run in parallel using foreach backend.

...

additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment,
                 random = ~ 1 | case,
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)

suppressWarnings(
  Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))
)

if (requireNamespace("plyr", quietly = TRUE)) {
  simulate(Laski_g, nsim = 20)
}
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#>      p_beta  r_theta  delta_AB       nu      kappa      g_AB     V_g_AB
#> 1  29.17350 489.3668 1.3187769 15.48333 0.12277026 1.2538480 0.07545095
#> 2  22.18898 416.1365 1.0877257 17.65289 0.13472169 1.0408488 0.05412651
#> 3  27.55949 801.2943 0.9735879 12.83521 0.13282067 0.9155681 0.05838486
#> 4  27.41467 680.8337 1.0506609 12.99903 0.11542621 0.9888526 0.05978283
#> 5  28.45314 474.6082 1.3060582 14.31473 0.11802951 1.2364128 0.07832425
#> 6  38.17834 658.9559 1.4872667 12.56179 0.10248647 1.3966665 0.10634683
#> 7  34.41479 416.5158 1.6862787 17.46148 0.13101614 1.6127981 0.10376159
#> 8  32.31150 525.3572 1.4097098 14.71746 0.13294015 1.3366298 0.09054311
#> 9  30.85485 311.3566 1.7486161 25.46616 0.15466380 1.6966073 0.08671012
#> 10 28.67886 630.2068 1.1424058 11.39436 0.09928957 1.0655235 0.07299884
#> 11 29.27004 471.3633 1.3481723 16.52313 0.12776215 1.2860374 0.07521124
#> 12 26.19314 524.7747 1.1434077 20.42167 0.15782084 1.1008948 0.05912598
#> 13 26.08785 577.0852 1.0859712 15.80351 0.12513793 1.0336050 0.05591719
#> 14 29.15197 457.1816 1.3634007 14.60509 0.12397597 1.2921681 0.08404816
#> 15 34.19812 353.5955 1.8186489 31.33662 0.17911577 1.7747720 0.08698324
#> 16 30.67989 305.8805 1.7541953 31.27420 0.16821030 1.7117882 0.07945133
#> 17 28.90717 829.4487 1.0037158 11.11293 0.10857502 0.9344171 0.06213357
#> 18 33.48719 453.5584 1.5723965 20.84286 0.13689041 1.5151292 0.08131005
#> 19 26.08561 367.5380 1.3606608 22.15187 0.15898844 1.3140668 0.06951824
#> 20 29.23083 271.0030 1.7756363 48.46261 0.19650056 1.7480143 0.07217287
#>    cnvg_warn sigma_sq       phi Tau.case.var((Intercept))
#> 1          0 178.3460 0.1637068                 311.02072
#> 2          0 173.2967 0.2020044                 242.83980
#> 3          0 250.0357 0.4388432                 551.25857
#> 4          0 203.0400 0.2201147                 477.79362
#> 5          0 158.6293 0.1691533                 315.97890
#> 6          0 182.9841 0.1016904                 475.97185
#> 7          0 170.7786 0.1725977                 245.73719
#> 8          0 187.0056 0.3054404                 338.35157
#> 9          0 168.8172 0.2118038                 142.53934
#> 10         0 150.4957 0.1620273                 479.71101
#> 11         0 183.6765 0.1729446                 287.68680
#> 12         0 254.3060 0.3371770                 270.46872
#> 13         0 215.2571 0.1744612                 361.82807
#> 14         0 157.9431 0.2171356                 299.23849
#> 15         0 221.6093 0.3347311                 131.98619
#> 16         0 186.8375 0.2498820                 119.04297
#> 17         0 195.3043 0.3122313                 634.14434
#> 18         0 212.8630 0.1393722                 240.69543
#> 19         0 186.9833 0.3069623                 180.55470
#> 20         0 204.7786 0.3350986                  66.22444