Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object
a g_REML object

nsim
number of models to simulate

seed
seed value. See documentation for simulate

parallel
if TRUE, run in parallel using foreach backend.

...
additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment, 
                 random = ~ 1 | case, 
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

simulate(Laski_g, nsim = 20)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#>      p_beta  r_theta delta_AB       nu     kappa     g_AB     V_g_AB cnvg_warn
#> 1  34.31624 305.1859 1.964343 38.20897 0.1755604 1.925531 0.08296537         0
#> 2  27.72482 285.3928 1.641146 34.01700 0.1650651 1.604695 0.06834703         0
#> 3  31.71139 300.9184 1.828062 27.53965 0.1414880 1.777821 0.08316725         0
#> 4  37.77496 713.2056 1.414480 13.40137 0.1258837 1.333814 0.09693372         0
#> 5  32.13841 200.0810 2.272068 53.76232 0.1821277 2.240224 0.08231507         0
#> 6  33.21071 541.6133 1.427031 13.92625 0.1158671 1.348774 0.09247857         0
#> 7  30.49310 238.1717 1.975861 97.55085 0.2254570 1.960631 0.07129133         0
#> 8  33.83629 438.8216 1.615246 18.69782 0.1535924 1.549578 0.09773432         0
#> 9  31.56114 444.7284 1.496598 20.31673 0.1490755 1.440662 0.08059632         0
#> 10 33.17938 378.9674 1.704383 20.78081 0.1497330 1.642121 0.09617773         0
#> 11 29.99898 312.8755 1.695980 41.19793 0.1879786 1.664917 0.07147896         0
#> 12 31.91984 562.6569 1.345671 17.08352 0.1492845 1.285716 0.07911400         0
#> 13 32.77555 573.1911 1.368990 12.98680 0.1233815 1.288378 0.09383384         0
#> 14 28.69523 474.2773 1.317630 15.01408 0.1170172 1.250696 0.07593513         0
#> 15 27.66087 453.5450 1.298840 15.67884 0.1290621 1.235703 0.07451868         0
#> 16 29.28708 429.4226 1.413298 18.17530 0.1501073 1.354165 0.08116336         0
#> 17 30.06102 274.6145 1.814020 47.14896 0.1843276 1.785010 0.06992877         0
#> 18 31.07136 344.6925 1.673572 32.01322 0.1603480 1.634055 0.07116524         0
#> 19 29.20332 695.5098 1.107339 12.37524 0.1222046 1.038845 0.06939362         0
#> 20 31.06134 369.6235 1.615625 19.04386 0.1451736 1.551150 0.09375565         0
#>    sigma_sq       phi Tau.case.var((Intercept))
#> 1  204.0472 0.2466652               101.1386870
#> 2  179.2378 0.1999479               106.1550257
#> 3  166.1588 0.0745015               134.7596038
#> 4  226.5174 0.3095555               486.6882821
#> 5  151.0186 0.2111773                49.0623767
#> 6  175.0068 0.1668476               366.6065046
#> 7  237.7212 0.3455290                 0.4505582
#> 8  199.1610 0.3475140               239.6606359
#> 9  210.2096 0.2607078               234.5188696
#> 10 181.8838 0.2555749               197.0836613
#> 11 220.4337 0.3172960                92.4417832
#> 12 237.2793 0.3633866               325.3775496
#> 13 174.4867 0.3114847               398.7043684
#> 14 166.0568 0.1245823               308.2205112
#> 15 169.3490 0.2192874               284.1959673
#> 16 189.3985 0.3313003               240.0241133
#> 17 200.1423 0.2563905                74.4721985
#> 18 209.5436 0.1811316               135.1488130
#> 19 199.3175 0.3490405               496.1923242
#> 20 166.1250 0.2584457               203.4984719