Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object
a g_REML object

nsim
number of models to simulate

seed
seed value. See documentation for simulate

parallel
if TRUE, run in parallel using foreach backend.

...
additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment,
                 random = ~ 1 | case,
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
simulate(Laski_g, nsim = 20)
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#>      p_beta  r_theta  delta_AB       nu     kappa      g_AB     V_g_AB
#> 1  25.79932 704.4801 0.9720171 11.85725 0.1129229 0.9092104 0.05679465
#> 2  28.16101 484.1418 1.2798583 16.79958 0.1323535 1.2218572 0.06975641
#> 3  26.65312 566.7878 1.1195351 13.01556 0.1115443 1.0537604 0.06499399
#> 4  32.63687 609.7644 1.3216835 14.25515 0.1354944 1.2509051 0.08477088
#> 5  33.14303 337.4641 1.8041739 23.75201 0.1404959 1.7465989 0.09147730
#> 6  26.91422 698.3372 1.0184723 12.20968 0.1194581 0.9546032 0.06110707
#> 7  22.09085 376.8268 1.1379983 22.31110 0.1630081 1.0993103 0.05750947
#> 8  31.79831 551.7083 1.3537838 12.29362 0.1004294 1.2694787 0.09146175
#> 9  29.80722 442.7699 1.4165515 16.30853 0.1417531 1.3503926 0.08609367
#> 10 32.23186 375.6541 1.6629961 23.65387 0.1574512 1.6097037 0.08619319
#> 11 24.09402 305.2535 1.3790464 24.29540 0.1510418 1.3360325 0.06400755
#> 12 25.74565 265.3976 1.5803576 49.51820 0.1604651 1.5563001 0.05170474
#> 13 30.67785 438.1833 1.4655378 18.72011 0.1425524 1.4060280 0.08130863
#> 14 35.79889 545.1202 1.5332867 14.51751 0.1356042 1.4526864 0.10576916
#> 15 35.65885 317.7374 2.0004752 38.66051 0.1856420 1.9614141 0.08792979
#> 16 28.99692 353.4741 1.5423147 24.28660 0.1522117 1.4941907 0.07459089
#> 17 33.65077 732.0870 1.2436944 13.20917 0.1220359 1.1717167 0.07869304
#> 18 31.83339 557.7221 1.3479505 14.64061 0.1146036 1.2776989 0.08001235
#> 19 37.88869 627.6994 1.5122852 12.81850 0.1272699 1.4220426 0.11338534
#> 20 35.84457 601.4911 1.4615334 13.59532 0.1210231 1.3793960 0.09977945
#>    cnvg_warn sigma_sq        phi Tau.case.var((Intercept))
#> 1          0 185.2130 0.28519318                 519.26715
#> 2          0 193.0060 0.20801977                 291.13584
#> 3          0 167.7997 0.17496971                 398.98807
#> 4          0 212.4485 0.36008881                 397.31595
#> 5          0 172.1949 0.11776115                 165.26921
#> 6          0 194.9410 0.33107347                 503.39612
#> 7          0 194.4895 0.33895488                 182.33727
#> 8          0 148.1331 0.09703333                 403.57516
#> 9          0 176.3304 0.31982727                 266.43956
#> 10         0 197.2128 0.26454633                 178.44129
#> 11         0 160.4698 0.19940656                 144.78369
#> 12         0 189.1038 0.07364870                  76.29380
#> 13         0 193.5315 0.24338574                 244.65177
#> 14         0 193.1836 0.34528108                 351.93660
#> 15         0 218.0609 0.31811137                  99.67643
#> 16         0 186.1849 0.20944036                 167.28929
#> 17         0 226.2952 0.27981945                 505.79189
#> 18         0 189.5702 0.11631426                 368.15189
#> 19         0 190.8509 0.37080552                 436.84843
#> 20         0 191.4917 0.24277839                 409.99938