Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object
a g_REML object

nsim
number of models to simulate

seed
seed value. See documentation for simulate

parallel
if TRUE, run in parallel using foreach backend.

...
additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment, 
                 random = ~ 1 | case, 
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

simulate(Laski_g, nsim = 20)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#>      p_beta  r_theta delta_AB       nu     kappa     g_AB     V_g_AB cnvg_warn
#> 1  30.58490 302.1013 1.759668 27.56708 0.1457984 1.711356 0.07976215         0
#> 2  32.82892 741.7396 1.205399 13.37127 0.1293166 1.136500 0.07598986         0
#> 3  26.62225 263.0423 1.641467 48.96774 0.1965640 1.616197 0.06707658         0
#> 4  31.14838 292.1212 1.822443 27.95194 0.1756737 1.773102 0.09287766         0
#> 5  34.68658 448.9268 1.637094 20.85163 0.1518832 1.577496 0.09093909         0
#> 6  30.89218 340.3374 1.674533 22.73410 0.1603609 1.618675 0.09061900         0
#> 7  30.11217 257.4056 1.876865 61.77001 0.2018804 1.853984 0.07003019         0
#> 8  31.00306 379.6138 1.591232 19.51913 0.1443201 1.529298 0.08953282         0
#> 9  32.21463 716.0925 1.203840 11.82924 0.1139115 1.125866 0.08036184         0
#> 10 29.18175 459.0415 1.362026 22.77154 0.1731276 1.316668 0.07315563         0
#> 11 25.83793 332.2754 1.417453 25.81164 0.1701604 1.375863 0.06992144         0
#> 12 33.59671 402.1436 1.675352 16.22953 0.1153239 1.596719 0.10556475         0
#> 13 30.31377 295.5972 1.763152 31.22431 0.1569023 1.720460 0.07631518         0
#> 14 25.43615 439.9589 1.212677 17.08899 0.1324716 1.158665 0.06366482         0
#> 15 29.61382 599.0519 1.209935 14.17111 0.1244312 1.144750 0.07150013         0
#> 16 30.53456 436.3260 1.461794 14.93616 0.1194752 1.387142 0.09120840         0
#> 17 33.08340 443.2119 1.571464 17.06020 0.1305216 1.501352 0.09419348         0
#> 18 29.38693 332.1570 1.612436 36.85854 0.1993814 1.579402 0.07648361         0
#> 19 30.61140 358.4502 1.616847 28.15647 0.1614219 1.573393 0.07453724         0
#> 20 31.66863 513.4835 1.397546 14.45615 0.1336150 1.323764 0.09088508         0
#>    sigma_sq        phi Tau.case.var((Intercept))
#> 1  167.8897 0.11033447                 134.21159
#> 2  237.6468 0.35072814                 504.09287
#> 3  199.3781 0.33289968                  63.66420
#> 4  173.3617 0.34913389                 118.75953
#> 5  216.9416 0.27308008                 231.98525
#> 6  176.1690 0.30692065                 164.16844
#> 7  210.5709 0.31379173                  46.83468
#> 8  173.1072 0.23776574                 206.50660
#> 9  188.2278 0.30011037                 527.86466
#> 10 247.2691 0.42053150                 211.77236
#> 11 187.6830 0.33537072                 144.59240
#> 12 151.0561 0.05827886                 251.08748
#> 13 176.8203 0.16144016                 118.77690
#> 14 177.8533 0.19900794                 262.10565
#> 15 200.9763 0.24530037                 398.07554
#> 16 152.6824 0.15389826                 283.64364
#> 17 178.2442 0.18106569                 264.96764
#> 18 234.3297 0.43910379                  97.82732
#> 19 206.5050 0.23096780                 151.94524
#> 20 180.1266 0.32733676                 333.35692