Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object
a g_REML object

nsim
number of models to simulate

seed
seed value. See documentation for simulate

parallel
if TRUE, run in parallel using foreach backend.

...
additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment, 
                 random = ~ 1 | case, 
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

simulate(Laski_g, nsim = 20)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#>      p_beta  r_theta delta_AB       nu     kappa     g_AB     V_g_AB cnvg_warn
#> 1  30.79921 320.6130 1.720082 30.98958 0.1569890 1.678114 0.07425323         0
#> 2  25.08705 250.0706 1.586420 56.17271 0.1888442 1.565144 0.05875705         0
#> 3  30.87708 350.9852 1.648132 42.03935 0.2034553 1.618552 0.07492874         0
#> 4  25.67296 562.6665 1.082307 14.01833 0.1279407 1.023350 0.06188935         0
#> 5  34.43701 367.1403 1.797254 23.73343 0.1603164 1.739854 0.09709136         0
#> 6  27.71285 438.0112 1.324154 25.11810 0.1583304 1.284219 0.06184264         0
#> 7  29.21310 249.6524 1.848884 50.13432 0.1813682 1.821087 0.06795056         0
#> 8  32.31279 306.9034 1.844479 25.25883 0.1642788 1.789164 0.09745432         0
#> 9  26.42190 515.5787 1.163634 19.29586 0.1477211 1.117812 0.05932592         0
#> 10 30.61304 379.8121 1.570804 21.45998 0.1481385 1.515259 0.08259223         0
#> 11 27.46793 480.9131 1.252543 15.44429 0.1204311 1.190717 0.06915155         0
#> 12 34.03124 306.6587 1.943346 29.06433 0.1470639 1.892763 0.08910233         0
#> 13 30.63447 545.2984 1.311877 15.66565 0.1235116 1.248052 0.07427217         0
#> 14 32.27267 443.0896 1.533166 16.04787 0.1328447 1.460379 0.09608742         0
#> 15 32.94404 644.5570 1.297616 11.85856 0.1058859 1.213780 0.08908968         0
#> 16 26.14078 333.7289 1.430941 20.16600 0.1510328 1.377054 0.07666383         0
#> 17 23.28703 268.3131 1.421651 34.58389 0.1685907 1.390596 0.05886855         0
#> 18 32.51971 586.0006 1.343376 15.08898 0.1217965 1.275479 0.07921160         0
#> 19 28.17083 555.1757 1.195596 15.81200 0.1451911 1.137975 0.06992598         0
#> 20 32.77626 487.6270 1.484279 15.31541 0.1329049 1.410387 0.09499123         0
#>    sigma_sq       phi Tau.case.var((Intercept))
#> 1  191.1912 0.1644185                 129.42178
#> 2  193.5972 0.2477250                  56.47343
#> 3  260.5736 0.4263560                  90.41159
#> 4  188.4161 0.2920269                 374.25048
#> 5  194.3270 0.2875101                 172.81325
#> 6  237.5567 0.2476026                 200.45447
#> 7  184.3131 0.2213003                  65.33931
#> 8  169.1010 0.2949941                 137.80238
#> 9  235.1181 0.2748829                 280.46059
#> 10 185.2215 0.2264651                 194.59060
#> 11 174.0874 0.1413455                 306.82566
#> 12 175.0710 0.1039137                 131.58763
#> 13 201.1645 0.1636258                 344.13394
#> 14 170.2479 0.2417536                 272.84168
#> 15 166.2737 0.1993741                 478.28325
#> 16 157.7275 0.2820821                 176.00137
#> 17 170.6712 0.2222012                  97.64197
#> 18 208.0324 0.1710457                 377.96814
#> 19 218.2326 0.3781033                 336.94314
#> 20 179.9378 0.2747302                 307.68917