Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object
a g_REML object

nsim
number of models to simulate

seed
seed value. See documentation for simulate

parallel
if TRUE, run in parallel using foreach backend.

...
additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment,
                 random = ~ 1 | case,
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

simulate(Laski_g, nsim = 20)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#>      p_beta   r_theta  delta_AB       nu     kappa      g_AB     V_g_AB
#> 1  32.06342  499.1909 1.4350815 13.08395 0.1227223 1.3512171 0.10067016
#> 2  32.76844  313.1512 1.8517342 40.61490 0.1700584 1.8173281 0.07246686
#> 3  33.27896  683.6928 1.2727388 12.27912 0.1057275 1.1933852 0.08330913
#> 4  28.38680  523.6331 1.2405174 17.55664 0.1507559 1.1867584 0.06976436
#> 5  31.15640 1071.6905 0.9517275 10.08120 0.1037846 0.8791224 0.06124628
#> 6  24.84618  625.0896 0.9937760 12.85934 0.1139211 0.9346665 0.05508531
#> 7  30.99916  342.0645 1.6760844 19.84345 0.1400915 1.6119271 0.09444179
#> 8  34.16682  538.9350 1.4717577 13.53063 0.1153530 1.3886428 0.10001265
#> 9  24.10237  496.9627 1.0811795 14.33202 0.1281797 1.0235965 0.06078723
#> 10 29.68467  513.2211 1.3103281 14.21506 0.1268383 1.2399563 0.08148008
#> 11 28.04323  316.5969 1.5760672 26.32568 0.1503994 1.5307357 0.07196644
#> 12 32.35773  270.7119 1.9666377 22.11256 0.1389144 1.8991718 0.11104293
#> 13 34.58575  389.4532 1.7525468 21.54425 0.1420994 1.6908208 0.09519014
#> 14 25.12950  594.9728 1.0302326 13.98990 0.1279233 0.9739968 0.05776631
#> 15 34.33932  418.5738 1.6784393 19.45159 0.1548358 1.6128807 0.10071880
#> 16 30.64332  249.7269 1.9391131 43.73529 0.1975792 1.9056688 0.08339820
#> 17 31.01312  449.9715 1.4620189 17.16801 0.1448969 1.3972055 0.08754484
#> 18 29.23564  377.4590 1.5047961 19.24341 0.1579696 1.4453756 0.08751316
#> 19 29.95435  515.0191 1.3199220 17.39992 0.1375118 1.2621991 0.07244376
#> 20 27.11360  326.7989 1.4998469 33.09298 0.1878027 1.4655965 0.07078818
#>    cnvg_warn sigma_sq       phi Tau.case.var((Intercept))
#> 1          0 152.9482 0.2966342                 346.24269
#> 2          0 211.7558 0.1899011                 101.39547
#> 3          0 185.2622 0.1615967                 498.43063
#> 4          0 226.0732 0.3596161                 297.55991
#> 5          0 209.3652 0.4073001                 862.32526
#> 6          0 183.3109 0.2129588                 441.77869
#> 7          0 156.3689 0.1910855                 185.69557
#> 8          0 168.4389 0.1844031                 370.49608
#> 9          0 170.2407 0.2762257                 326.72193
#> 10         0 173.7719 0.2685467                 339.44920
#> 11         0 173.2298 0.1642803                 143.36707
#> 12         0 132.2101 0.1322091                 138.50175
#> 13         0 188.2324 0.1714440                 201.22075
#> 14         0 198.8158 0.2935727                 396.15693
#> 15         0 195.4076 0.3362595                 223.16614
#> 16         0 183.8213 0.3705194                  65.90551
#> 17         0 187.9410 0.3165878                 262.03046
#> 18         0 176.7476 0.3721608                 200.71143
#> 19         0 213.3743 0.2369857                 301.64479
#> 20         0 214.1254 0.3845322                 112.67345