Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object

a g_REML object

nsim

number of models to simulate

seed

seed value. See documentation for simulate

parallel

if TRUE, run in parallel using foreach backend.

...

additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment,
                 random = ~ 1 | case,
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
simulate(Laski_g, nsim = 20)
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#>      p_beta  r_theta delta_AB       nu     kappa     g_AB     V_g_AB cnvg_warn
#> 1  28.27176 388.3351 1.434662 20.32704 0.1538720 1.381068 0.07737981         0
#> 2  31.09495 421.1438 1.515216 20.59749 0.1475037 1.459366 0.08070672         0
#> 3  28.71693 388.1189 1.457658 25.93747 0.1684012 1.415098 0.07141647         0
#> 4  30.90402 474.7590 1.418333 19.98611 0.1179473 1.364435 0.06707998         0
#> 5  30.39853 369.2360 1.581979 36.66759 0.1941754 1.549399 0.07324111         0
#> 6  29.40896 562.8737 1.239579 14.82261 0.1378286 1.175782 0.07512340         0
#> 7  25.58218 345.6082 1.376086 25.87941 0.1518107 1.335817 0.06146188         0
#> 8  33.97187 386.4741 1.728062 32.43252 0.1897226 1.687790 0.08395365         0
#> 9  26.69221 581.5814 1.106826 13.68101 0.1310208 1.045020 0.06604630         0
#> 10 28.80411 237.4950 1.869077 55.08401 0.1778912 1.843513 0.06417826         0
#> 11 27.08734 545.9545 1.159279 13.09339 0.1239282 1.091582 0.07143699         0
#> 12 32.46826 419.9644 1.584356 17.41984 0.1369383 1.515149 0.09550992         0
#> 13 29.94533 559.0364 1.266511 15.92190 0.1320563 1.205901 0.07161730         0
#> 14 29.78405 530.0384 1.293690 12.12194 0.1024788 1.211962 0.08603261         0
#> 15 29.84488 321.8170 1.663663 19.33064 0.1548750 1.598270 0.09989574         0
#> 16 27.85914 326.4134 1.541998 26.33465 0.1628711 1.497662 0.07384803         0
#> 17 31.80434 848.5757 1.091795 10.87708 0.1107927 1.014742 0.07318130         0
#> 18 27.44150 270.1545 1.669559 25.63884 0.1396150 1.620239 0.07628203         0
#> 19 31.63073 390.1873 1.601300 26.61936 0.1670376 1.555756 0.07835140         0
#> 20 29.37316 283.9106 1.743250 58.47923 0.2024989 1.720797 0.06775662         0
#>    sigma_sq         phi Tau.case.var((Intercept))
#> 1  185.7097  0.30368186                 202.62539
#> 2  200.0778  0.23994306                 221.06599
#> 3  218.7095  0.31865545                 169.40931
#> 4  210.9758 -0.01927234                 263.78315
#> 5  255.2477  0.39940046                 113.98830
#> 6  204.6860  0.35177150                 358.18771
#> 7  187.9439  0.18212948                 157.66437
#> 8  253.1068  0.40745976                 133.36732
#> 9  191.8428  0.34794994                 389.73867
#> 10 179.2330  0.17661000                  58.26191
#> 11 168.0203  0.30968038                 377.93419
#> 12 173.9278  0.23081861                 246.03661
#> 13 212.9756  0.23910252                 346.06077
#> 14 140.0720  0.13536893                 389.96644
#> 15 149.6817  0.34016585                 172.13534
#> 16 182.8706  0.26698346                 143.54281
#> 17 194.8705  0.37389099                 653.70515
#> 18 143.4907  0.08243447                 126.66374
#> 19 221.7580  0.29710803                 168.42931
#> 20 229.3215  0.33035516                  54.58909