Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object
a g_REML object

nsim
number of models to simulate

seed
seed value. See documentation for simulate

parallel
if TRUE, run in parallel using foreach backend.

...
additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment,
                 random = ~ 1 | case,
                 correlation = corAR1(0, ~ time | case),
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))
simulate(Laski_g, nsim = 20)

#>      p_beta  r_theta delta_AB       nu     kappa     g_AB     V_g_AB cnvg_warn
#> 1  34.64830 397.8976 1.736986 22.29512 0.1441470 1.677891 0.09187879         0
#> 2  28.59858 305.7104 1.635646 19.90005 0.1212658 1.573217 0.08562821         0
#> 3  31.05097 250.2404 1.962892 39.87944 0.1729929 1.925744 0.07974856         0
#> 4  35.88797 472.2405 1.651457 16.86249 0.1564157 1.576899 0.11093589         0
#> 5  32.77743 322.4907 1.825225 27.68001 0.1404934 1.775319 0.08235595         0
#> 6  31.92891 415.9212 1.565592 25.57457 0.1647607 1.519226 0.07742903         0
#> 7  35.06650 287.4073 2.068443 38.65565 0.1993287 2.028050 0.09694082         0
#> 8  33.64066 512.5679 1.485897 18.38239 0.1550912 1.424437 0.08806376         0
#> 9  30.83359 373.8215 1.594747 22.03854 0.1338554 1.539853 0.07859032         0
#> 10 29.86657 334.1426 1.633877 35.84903 0.1822870 1.599455 0.07192033         0
#> 11 35.48750 333.7329 1.942567 28.57221 0.1676315 1.891125 0.09684591         0
#> 12 35.43776 276.0572 2.132881 72.32253 0.2203873 2.110686 0.08075971         0
#> 13 31.13391 463.7733 1.445709 23.18986 0.1621961 1.398443 0.07386352         0
#> 14 25.45381 397.6579 1.276433 27.06402 0.1868824 1.240731 0.06678285         0
#> 15 37.45659 366.4548 1.956672 19.10035 0.1361083 1.878822 0.12442668         0
#> 16 31.39257 529.0995 1.364766 15.10478 0.1265789 1.295860 0.08241913         0
#> 17 32.03247 550.2407 1.365570 13.28677 0.1246615 1.287010 0.09185964         0
#> 18 31.61283 473.0832 1.453432 17.45732 0.1423614 1.390082 0.08486408         0
#> 19 30.46664 322.2136 1.697276 37.16536 0.1843736 1.662793 0.07420200         0
#> 20 31.60494 464.2444 1.466836 16.36927 0.1380890 1.398587 0.08946888         0
#>    sigma_sq        phi Tau.case.var((Intercept))
#> 1  196.9814 0.17483859                 200.91623
#> 2  135.9981 0.01457741                 169.71229
#> 3  169.0080 0.21626043                  81.23238
#> 4  203.0441 0.44861410                 269.19636
#> 5  178.2833 0.06446216                 144.20734
#> 6  230.7911 0.29401070                 185.13005
#> 7  205.4255 0.42199820                  81.98179
#> 8  231.2349 0.37252607                 281.33306
#> 9  180.8081 0.08842440                 193.01337
#> 10 221.6993 0.31542375                 112.44334
#> 11 196.0007 0.27574492                 137.73218
#> 12 248.3991 0.39455982                  27.65812
#> 13 243.6785 0.31381576                 220.09479
#> 14 242.4411 0.46060023                 155.21672
#> 15 162.0410 0.17326182                 204.41385
#> 16 189.9103 0.21971230                 339.18923
#> 17 172.5004 0.30388794                 377.74021
#> 18 198.7109 0.28153816                 274.37231
#> 19 217.5591 0.32025541                 104.65454
#> 20 183.6172 0.28074739                 280.62717