Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object

a g_REML object

nsim

number of models to simulate

seed

seed value. See documentation for simulate

parallel

if TRUE, run in parallel using foreach backend.

...

additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment,
                 random = ~ 1 | case,
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
simulate(Laski_g, nsim = 20)
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#>      p_beta  r_theta delta_AB       nu     kappa     g_AB     V_g_AB cnvg_warn
#> 1  31.20960 346.6490 1.676267 24.06289 0.1614585 1.623472 0.08737014         0
#> 2  30.73762 372.9108 1.591724 18.47851 0.1556418 1.526234 0.09716862         0
#> 3  29.33290 396.9194 1.472326 29.00327 0.1894850 1.433922 0.07515968         0
#> 4  31.47106 432.3793 1.513488 15.77782 0.1225975 1.440386 0.09278724         0
#> 5  25.85849 386.6792 1.315006 18.31830 0.1365217 1.260422 0.06895114         0
#> 6  31.02402 581.8265 1.286179 16.31907 0.1294607 1.226148 0.07113973         0
#> 7  26.62724 310.1384 1.511988 32.26081 0.1697755 1.476563 0.06576171         0
#> 8  36.92043 246.3930 2.352082 37.14340 0.1685889 2.304267 0.10515686         0
#> 9  29.60256 330.0795 1.629371 22.47822 0.1419697 1.574395 0.08223242         0
#> 10 29.05446 503.6643 1.294620 18.13672 0.1434710 1.240335 0.06994562         0
#> 11 29.69102 317.8785 1.665308 39.15501 0.1889311 1.633205 0.07242770         0
#> 12 30.40042 476.1833 1.393133 18.51864 0.1399626 1.335939 0.07537141         0
#> 13 32.23559 486.8148 1.461012 21.12629 0.1547485 1.408524 0.07741404         0
#> 14 36.15440 308.1831 2.059476 21.63770 0.1267019 1.987256 0.11884556         0
#> 15 33.38044 450.5872 1.572544 17.37334 0.1340165 1.503666 0.09377570         0
#> 16 32.55909 427.9991 1.573805 21.86849 0.1440830 1.519205 0.08041571         0
#> 17 29.17299 365.5587 1.525817 24.56981 0.1718195 1.478762 0.07940023         0
#> 18 29.88837 440.5746 1.423942 19.10496 0.1377052 1.367302 0.07530692         0
#> 19 31.05420 364.7839 1.625933 20.93440 0.1498421 1.566978 0.08911501         0
#> 20 29.27798 493.0806 1.318506 16.15005 0.1345514 1.256313 0.07590787         0
#>    sigma_sq        phi Tau.case.var((Intercept))
#> 1  185.2662 0.29134798                 161.38278
#> 2  169.0859 0.37461034                 203.82490
#> 3  250.3623 0.45161778                 146.55711
#> 4  160.2696 0.14975473                 272.10969
#> 5  166.1040 0.19874288                 220.57520
#> 6  225.1068 0.19737900                 356.71968
#> 7  192.4308 0.25223033                 117.70759
#> 8  160.9806 0.20222641                  85.41239
#> 9  163.5840 0.15238507                 166.49555
#> 10 218.0418 0.26955567                 285.62254
#> 11 220.8951 0.33884075                  96.98341
#> 12 207.5860 0.22511117                 268.59732
#> 13 238.6694 0.29220964                 248.14541
#> 14 145.9441 0.03062501                 162.23900
#> 15 185.0971 0.20433330                 265.49002
#> 16 209.4606 0.18248670                 218.53851
#> 17 202.8813 0.37102588                 162.67737
#> 18 195.4238 0.18792189                 245.15084
#> 19 175.8977 0.25302576                 188.88616
#> 20 191.2252 0.25455825                 301.85546