Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object

a g_REML object

nsim

number of models to simulate

seed

seed value. See documentation for simulate

parallel

if TRUE, run in parallel using foreach backend.

...

additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment,
                 random = ~ 1 | case,
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)

suppressWarnings(
  Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))
)

if (requireNamespace("plyr", quietly = TRUE)) {
  simulate(Laski_g, nsim = 20)
}
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#>      p_beta  r_theta delta_AB       nu     kappa     g_AB     V_g_AB cnvg_warn
#> 1  32.10218 424.2548 1.558551 20.37790 0.1440533 1.500477 0.08376435         0
#> 2  31.13532 472.2738 1.432704 25.26019 0.1637199 1.389740 0.06954725         0
#> 3  30.58581 477.9691 1.399008 16.18249 0.1466149 1.333151 0.08647460         0
#> 4  28.49241 394.7378 1.434085 17.38361 0.1302715 1.371310 0.08005105         0
#> 5  29.08281 533.5009 1.259125 17.75322 0.1480668 1.205172 0.06976499         0
#> 6  33.97583 572.7578 1.419661 15.59200 0.1324586 1.350260 0.08699566         0
#> 7  32.10103 561.4332 1.354784 14.26499 0.1196558 1.282283 0.08383740         0
#> 8  31.38813 464.8646 1.455802 16.48475 0.1337895 1.388548 0.08669575         0
#> 9  31.79887 473.8012 1.460877 17.66437 0.1614481 1.397960 0.09069748         0
#> 10 32.05490 461.9849 1.491354 15.09561 0.1355911 1.416010 0.09769263         0
#> 11 27.57230 534.7361 1.192349 14.42246 0.1204591 1.129250 0.06786729         0
#> 12 33.50383 376.7937 1.726007 35.51332 0.1948356 1.689298 0.08157182         0
#> 13 35.27855 384.9180 1.798153 21.86069 0.1549962 1.735748 0.10185018         0
#> 14 31.82593 562.6434 1.341728 13.64786 0.1128286 1.266619 0.08419615         0
#> 15 23.18658 281.3354 1.382370 44.95790 0.2037944 1.359180 0.06370520         0
#> 16 34.09794 335.6145 1.861263 24.65490 0.1526229 1.804064 0.09678456         0
#> 17 34.23656 261.1164 2.118717 61.51054 0.2008152 2.092778 0.07769354         0
#> 18 31.19494 417.1251 1.527394 21.32028 0.1517664 1.473026 0.08082866         0
#> 19 28.98183 366.5214 1.513827 25.40084 0.1618387 1.468684 0.07355645         0
#> 20 30.56175 376.7993 1.574429 22.57300 0.1420321 1.521532 0.07791335         0
#>    sigma_sq       phi Tau.case.var((Intercept))
#> 1  198.8111 0.2141589                 225.44367
#> 2  259.8858 0.2903033                 212.38801
#> 3  192.1364 0.3751992                 285.83271
#> 4  161.0602 0.1679193                 233.67766
#> 5  230.2355 0.3261445                 303.26541
#> 6  214.4790 0.2574289                 358.27883
#> 7  187.5988 0.1891511                 373.83434
#> 8  183.0472 0.2337678                 281.81746
#> 9  214.3279 0.4675824                 259.47331
#> 10 169.5353 0.3133696                 292.44952
#> 11 181.0225 0.1894210                 353.71361
#> 12 258.5324 0.4157209                 118.26129
#> 13 192.5450 0.2781572                 192.37293
#> 14 176.6629 0.1494853                 385.98044
#> 15 212.8439 0.4081097                  68.49157
#> 16 178.3104 0.2070480                 157.30410
#> 17 212.8041 0.3077414                  48.31235
#> 18 204.1855 0.2614343                 212.93957
#> 19 201.3766 0.2724144                 165.14480
#> 20 187.2159 0.1512095                 189.58340