Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object
a g_REML object

nsim
number of models to simulate

seed
seed value. See documentation for simulate

parallel
if TRUE, run in parallel using foreach backend.

...
additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment, 
                 random = ~ 1 | case, 
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

simulate(Laski_g, nsim = 20)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#>      p_beta  r_theta  delta_AB       nu      kappa      g_AB     V_g_AB
#> 1  29.34002 347.7029 1.5734620 25.97070 0.15133993 1.5275808 0.07279311
#> 2  26.09932 341.0025 1.4133521 34.45903 0.18162061 1.3823657 0.06326572
#> 3  29.09283 656.6538 1.1353185 13.22687 0.12725583 1.0697026 0.06945600
#> 4  25.80083 340.0286 1.3991874 27.03892 0.16243700 1.3600148 0.06438707
#> 5  32.48531 368.5794 1.6920823 17.45801 0.13400221 1.6183340 0.10519068
#> 6  31.97330 273.3450 1.9338902 36.89538 0.19566202 1.8943103 0.09079155
#> 7  32.40244 301.6228 1.8657165 38.29661 0.20294278 1.8289383 0.08830084
#> 8  31.12006 457.1190 1.4555452 19.50792 0.13859029 1.3988590 0.07678095
#> 9  32.98747 386.1705 1.6786472 25.23666 0.15990659 1.6282609 0.08405923
#> 10 28.49858 403.8677 1.4180894 20.14349 0.14968892 1.3646264 0.07535900
#> 11 31.11914 522.8086 1.3609940 16.48400 0.13150324 1.2981169 0.07748169
#> 12 25.17836 443.3150 1.1958349 19.07425 0.14829834 1.1481902 0.06205559
#> 13 35.28780 330.2308 1.9418507 38.05472 0.16144250 1.9033267 0.07717525
#> 14 32.00639 426.4021 1.5499836 13.70814 0.12517317 1.4636055 0.11052908
#> 15 28.41896 724.5112 1.0558096 15.15980 0.14257184 1.0026998 0.06032845
#> 16 22.17220 353.9440 1.1785323 28.24688 0.19033161 1.1469610 0.06233088
#> 17 31.51712 501.3333 1.4076129 16.70540 0.14652471 1.3434571 0.08504695
#> 18 35.04418 431.6621 1.6867236 23.33924 0.15090759 1.6319344 0.08676375
#> 19 30.84451 572.8443 1.2887230 14.91165 0.13505758 1.2227999 0.07842538
#> 20 28.13363 820.1176 0.9823987  9.90701 0.08751532 0.9061019 0.06225340
#>    cnvg_warn sigma_sq       phi Tau.case.var((Intercept))
#> 1          0 189.2639 0.1769396                 158.43894
#> 2          0 222.7481 0.3226664                 118.25445
#> 3          0 206.5375 0.3378034                 450.11627
#> 4          0 192.6836 0.2536025                 147.34508
#> 5          0 151.9748 0.2014098                 216.60459
#> 6          0 190.2667 0.4089797                  83.07839
#> 7          0 218.1275 0.4542551                  83.49534
#> 8          0 206.0185 0.1856799                 251.10054
#> 9          0 210.6211 0.2588906                 175.54939
#> 10         0 190.1481 0.2705120                 213.71956
#> 11         0 204.8454 0.2112964                 317.96317
#> 12         0 200.9211 0.2864397                 242.39392
#> 13         0 215.1988 0.1417556                 115.03198
#> 14         0 138.4558 0.2811555                 287.94627
#> 15         0 272.9638 0.3843173                 451.54736
#> 16         0 222.2704 0.4719456                 131.67362
#> 17         0 206.2377 0.3511106                 295.09565
#> 18         0 222.0200 0.2143189                 209.64212
#> 19         0 207.6188 0.3173735                 365.22548
#> 20         0 144.9719 0.1968561                 675.14574