Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object

a g_REML object

nsim

number of models to simulate

seed

seed value. See documentation for simulate

parallel

if TRUE, run in parallel using foreach backend.

...

additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment,
                 random = ~ 1 | case,
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
simulate(Laski_g, nsim = 20)
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#>      p_beta  r_theta delta_AB       nu     kappa      g_AB     V_g_AB cnvg_warn
#> 1  28.43456 622.2975 1.139850 15.79368 0.1518839 1.0848505 0.06766583         0
#> 2  29.98378 453.5843 1.407853 17.34863 0.1374396 1.3461003 0.07989805         0
#> 3  29.21906 587.2765 1.205716 15.51327 0.1383478 1.1464699 0.06975357         0
#> 4  27.33197 305.0051 1.565011 33.65516 0.1759340 1.5298739 0.06884054         0
#> 5  32.56226 477.0307 1.490876 12.46785 0.1116838 1.3993572 0.10966958         0
#> 6  31.72335 455.2377 1.486825 16.05964 0.1306844 1.4162913 0.09080790         0
#> 7  27.63340 330.0273 1.521106 22.61015 0.1559001 1.4700851 0.07824275         0
#> 8  30.19317 204.0075 2.113906 79.68142 0.1957333 2.0939463 0.06691554         0
#> 9  28.98608 411.2954 1.429264 19.82727 0.1413870 1.3745096 0.07460437         0
#> 10 32.47785 495.6775 1.458773 14.57606 0.1193512 1.3824031 0.09289881         0
#> 11 32.32056 308.5035 1.840131 44.95243 0.1867681 1.8092585 0.07372198         0
#> 12 30.36592 537.3000 1.310020 14.65333 0.1397673 1.2418060 0.08293291         0
#> 13 27.66378 234.2525 1.807463 84.59737 0.1878143 1.7913913 0.05500189         0
#> 14 25.12745 261.1768 1.554824 45.26883 0.1931352 1.5289208 0.06497714         0
#> 15 26.46227 449.9145 1.247562 20.83184 0.1660365 1.2021007 0.06740553         0
#> 16 30.84847 886.4966 1.036084 11.95117 0.1112076 0.9696751 0.06186505         0
#> 17 33.56389 277.9135 2.013342 53.88367 0.1777545 1.9851877 0.07015147         0
#> 18 31.56461 714.4332 1.180918 12.40074 0.1178853 1.1080261 0.07554171         0
#> 19 35.06524 298.9051 2.028197 24.89380 0.1644057 1.9664718 0.11340850         0
#> 20 33.90565 907.5064 1.125505 11.26880 0.1080028 1.0488966 0.07382278         0
#>    sigma_sq       phi Tau.case.var((Intercept))
#> 1  251.3723 0.4535469                 370.92513
#> 2  187.4700 0.2380345                 266.11431
#> 3  222.4033 0.3212265                 364.87320
#> 4  194.8923 0.2864955                 110.11277
#> 5  133.7728 0.2167293                 343.25785
#> 6  174.1873 0.2197337                 281.05035
#> 7  168.5497 0.2707429                 161.47758
#> 8  175.9944 0.2217976                  28.01305
#> 9  188.3786 0.2032629                 222.91675
#> 10 169.2355 0.1698846                 326.44203
#> 11 222.7242 0.2854605                  85.77933
#> 12 194.7866 0.3829227                 342.51335
#> 13 202.3846 0.1601280                  31.86793
#> 14 191.8343 0.3290195                  69.34252
#> 15 226.2324 0.4028161                 223.68211
#> 16 234.4725 0.2553696                 652.02411
#> 17 208.3311 0.1802812                  69.58234
#> 18 202.3466 0.2950070                 512.08662
#> 19 163.5931 0.3020222                 135.31199
#> 20 217.9906 0.2861158                 689.51579