Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object
a g_REML object

nsim
number of models to simulate

seed
seed value. See documentation for simulate

parallel
if TRUE, run in parallel using foreach backend.

...
additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment, 
                 random = ~ 1 | case, 
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

simulate(Laski_g, nsim = 20)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#>      p_beta  r_theta delta_AB       nu     kappa      g_AB     V_g_AB cnvg_warn
#> 1  28.99373 504.6527 1.290648 15.75280 0.1385455 1.2282085 0.07614444         0
#> 2  27.60662 548.0192 1.179276 14.90861 0.1260012 1.1189388 0.06630759         0
#> 3  33.66977 455.9254 1.576860 13.96277 0.1174613 1.4906163 0.10993782         0
#> 4  31.92111 323.0888 1.775894 33.38733 0.1601587 1.7357003 0.07461115         0
#> 5  34.23411 287.5068 2.018994 47.72476 0.1875784 1.9870982 0.07909743         0
#> 6  29.59814 297.2271 1.716802 39.11984 0.1906157 1.6836760 0.07538804         0
#> 7  27.54799 428.4052 1.330952 25.61091 0.1518091 1.2915920 0.05940637         0
#> 8  29.37146 558.9860 1.242296 11.62749 0.1009148 1.1604043 0.08312008         0
#> 9  29.53104 190.9093 2.137299 85.71001 0.1977326 2.1185416 0.06625957         0
#> 10 32.75742 354.0926 1.740809 19.58893 0.1352704 1.6732975 0.10002535         0
#> 11 33.45710 520.3407 1.466711 15.27324 0.1274505 1.3934886 0.09193616         0
#> 12 27.91903 660.3298 1.086475 13.21311 0.1032809 1.0236157 0.05930851         0
#> 13 30.23088 412.8492 1.487836 20.71574 0.1528192 1.4333117 0.07991579         0
#> 14 32.06482 387.1009 1.629734 23.47573 0.1489125 1.5771069 0.08157375         0
#> 15 31.30939 539.9600 1.347391 18.08063 0.1410080 1.2907168 0.07345016         0
#> 16 29.76952 533.1018 1.289338 15.55841 0.1195026 1.2261699 0.07169080         0
#> 17 35.51898 158.1187 2.824677 64.12788 0.1823391 2.7915120 0.09660289         0
#> 18 27.15395 716.5405 1.014408 12.26032 0.1259583 0.9510619 0.06220499         0
#> 19 31.44214 389.1906 1.593789 19.70054 0.1615928 1.5323340 0.09451970         0
#> 20 31.28254 539.0107 1.347421 15.73861 0.1288719 1.2821753 0.07857686         0
#>    sigma_sq        phi Tau.case.var((Intercept))
#> 1  193.6800 0.31196211                 310.97269
#> 2  194.0307 0.22312746                 353.98848
#> 3  148.2291 0.18209841                 307.69636
#> 4  199.7606 0.16732846                 123.32817
#> 5  211.7447 0.27635322                  75.76206
#> 6  207.4221 0.35166780                  89.80504
#> 7  231.7493 0.18591694                 196.65592
#> 8  138.3608 0.15964891                 420.62515
#> 9  168.6614 0.21997966                  22.24794
#> 10 159.0885 0.15315790                 195.00406
#> 11 189.0677 0.22085123                 331.27300
#> 12 196.0184 0.06770823                 464.31140
#> 13 199.1657 0.28464197                 213.68352
#> 14 198.9980 0.19487384                 188.10290
#> 15 231.9275 0.24810227                 308.03254
#> 16 193.9878 0.12702518                 339.11393
#> 17 125.5253 0.17645472                  32.59346
#> 18 206.2502 0.40740090                 510.29033
#> 19 186.9828 0.39296301                 202.20773
#> 20 201.8575 0.21482573                 337.15315