Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object
a g_REML object

nsim
number of models to simulate

seed
seed value. See documentation for simulate

parallel
if TRUE, run in parallel using foreach backend.

...
additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment, 
                 random = ~ 1 | case, 
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

simulate(Laski_g, nsim = 20)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#>      p_beta  r_theta delta_AB       nu     kappa     g_AB     V_g_AB cnvg_warn
#> 1  29.98047 634.1751 1.190512 12.31757 0.1093895 1.116522 0.07497162         0
#> 2  33.40401 472.4660 1.536786 15.11155 0.1251129 1.459231 0.09961982         0
#> 3  28.79291 255.8154 1.800206 41.71948 0.1559366 1.767648 0.06432033         0
#> 4  28.33697 228.4193 1.874938 42.04190 0.1670383 1.841291 0.07097512         0
#> 5  33.75239 353.4800 1.795238 18.53462 0.1417545 1.721601 0.11225590         0
#> 6  33.08653 281.0450 1.973619 27.03243 0.1643230 1.918351 0.10211324         0
#> 7  26.98056 201.9297 1.898676 78.18572 0.1940768 1.880405 0.06123287         0
#> 8  31.05310 474.7028 1.425260 16.49395 0.1406577 1.359454 0.08577808         0
#> 9  28.30822 374.4677 1.462869 31.79648 0.1850199 1.428090 0.06945315         0
#> 10 28.94795 294.7922 1.686008 31.03967 0.1757774 1.644939 0.07860601         0
#> 11 34.55072 639.9258 1.365816 12.36609 0.1196463 1.281271 0.09681483         0
#> 12 28.13340 191.2797 2.034172 92.18266 0.1913290 2.017577 0.05947211         0
#> 13 29.01896 356.3793 1.537183 24.97282 0.1539065 1.490551 0.07332243         0
#> 14 29.94718 461.4079 1.394163 13.85271 0.1022358 1.317294 0.08623632         0
#> 15 32.14543 322.0849 1.791159 37.60687 0.1825866 1.755198 0.07751399         0
#> 16 33.04767 356.2509 1.750906 28.75750 0.1617350 1.704842 0.08169845         0
#> 17 32.79256 269.7976 1.996440 33.35816 0.1507525 1.951215 0.08450174         0
#> 18 31.58782 626.4115 1.262088 13.56762 0.1223739 1.191012 0.07877949         0
#> 19 29.93393 394.0974 1.507863 32.38089 0.1836341 1.472667 0.07040200         0
#> 20 26.98662 244.2355 1.726808 40.78407 0.1884244 1.694857 0.07334722         0
#>    sigma_sq        phi Tau.case.var((Intercept))
#> 1  174.0278 0.20185952                 460.14728
#> 2  169.1143 0.20428721                 303.35174
#> 3  171.0653 0.07730517                  84.75015
#> 4  155.5516 0.15903297                  72.86770
#> 5  154.7761 0.24135259                 198.70387
#> 6  159.8766 0.26906377                 121.16833
#> 7  172.7492 0.21501373                  29.18047
#> 8  190.1512 0.30116349                 284.55154
#> 9  239.9064 0.37660948                 134.56129
#> 10 182.5393 0.31166910                 112.25298
#> 11 181.6158 0.31900469                 458.31005
#> 12 170.3097 0.16614141                  20.97004
#> 13 191.0528 0.21289574                 165.32643
#> 14 144.6830 0.01945332                 316.72490
#> 15 217.4123 0.30339696                 104.67265
#> 16 207.3168 0.22625435                 148.93406
#> 17 164.4632 0.09341418                 105.33443
#> 18 199.6626 0.25939832                 426.74892
#> 19 253.1006 0.35905300                 140.99680
#> 20 171.7139 0.32337568                  72.52158