Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment,
                 random = ~ 1 | case,
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
simulate(Laski_g, nsim = 20)
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#>      p_beta  r_theta delta_AB       nu      kappa     g_AB     V_g_AB cnvg_warn
#> 1  28.22088 370.6537 1.465839 26.26857 0.15518884 1.423586 0.06695977         0
#> 2  30.41561 650.8299 1.192237 11.91651 0.10499076 1.115593 0.07649581         0
#> 3  32.85879 276.1982 1.977157 45.07342 0.19694304 1.944074 0.08348571         0
#> 4  31.27839 310.6739 1.774566 20.70260 0.12415998 1.709492 0.09544785         0
#> 5  28.47833 342.3794 1.539078 20.45522 0.13459191 1.481949 0.07926055         0
#> 6  33.51193 390.8499 1.695097 16.72813 0.13613131 1.617945 0.11015918         0
#> 7  31.01619 657.6871 1.209425 12.14649 0.11177674 1.133178 0.07852840         0
#> 8  31.52511 531.4661 1.367473 15.96357 0.13362967 1.302205 0.08075056         0
#> 9  35.69452 412.5258 1.757422 13.53485 0.09928849 1.658206 0.13304128         0
#> 10 31.35967 470.8411 1.445221 17.76999 0.13441883 1.383354 0.08068270         0
#> 11 29.35681 621.8965 1.177199 11.47684 0.09325866 1.098557 0.07511376         0
#> 12 28.32179 395.4251 1.424258 28.51124 0.16509298 1.386461 0.06452448         0
#> 13 37.39201 343.8391 2.016514 33.88160 0.16906189 1.971544 0.09068989         0
#> 14 25.18104 541.5795 1.082038 11.66167 0.08936004 1.010924 0.06315281         0
#> 15 28.21939 341.9269 1.526093 22.60111 0.15484494 1.474884 0.07827981         0
#> 16 31.64348 219.4160 2.136240 67.71557 0.22167761 2.112492 0.08366185         0
#> 17 30.38439 384.1490 1.550246 22.58503 0.15053949 1.498189 0.07868777         0
#> 18 33.48455 364.8247 1.753082 14.11742 0.12002758 1.658269 0.13171949         0
#> 19 32.26378 268.1422 1.970302 53.49658 0.18901732 1.942549 0.07297692         0
#> 20 33.35423 503.8031 1.486006 17.11706 0.13367143 1.419930 0.08670761         0
#>    sigma_sq         phi Tau.case.var((Intercept))
#> 1  204.2775 0.204715988                 166.37624
#> 2  168.7302 0.183278867                 482.09974
#> 3  204.6916 0.357467790                  71.50660
#> 4  142.5041 0.025082901                 168.16981
#> 5  158.1581 0.126871004                 184.22129
#> 6  156.5587 0.247512615                 234.29123
#> 7  178.1607 0.244584927                 479.52630
#> 8  203.6654 0.253101377                 327.80071
#> 9  125.3039 0.002683427                 287.22197
#> 10 196.9448 0.195421491                 273.89632
#> 11 148.3598 0.077066618                 473.53670
#> 12 230.7684 0.256131562                 164.65675
#> 13 217.1823 0.231810416                 126.65687
#> 14 131.6588 0.010065410                 409.92073
#> 15 174.1828 0.261800643                 167.74406
#> 16 195.3816 0.419193375                  24.03441
#> 17 193.8764 0.224916255                 190.27263
#> 18 120.6786 0.201145440                 244.14603
#> 19 204.8059 0.259695062                  63.33628
#> 20 204.4237 0.209659674                 299.37934