Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object

a g_REML object

nsim

number of models to simulate

seed

seed value. See documentation for simulate

parallel

if TRUE, run in parallel using foreach backend.

...

additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment,
                 random = ~ 1 | case,
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)

suppressWarnings(
  Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))
)

if (requireNamespace("plyr", quietly = TRUE)) {
  simulate(Laski_g, nsim = 20)
}
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#>      p_beta  r_theta delta_AB        nu     kappa      g_AB     V_g_AB
#> 1  33.85333 555.3854 1.436495  16.53655 0.1220498 1.3703440 0.08156693
#> 2  33.97908 353.7624 1.806574  29.86002 0.1616344 1.7608150 0.08296692
#> 3  32.91567 703.6042 1.240905  12.71620 0.1313585 1.1662487 0.08357139
#> 4  33.87047 382.1008 1.732736  21.35373 0.1667118 1.6711564 0.10200222
#> 5  36.10070 217.1712 2.449709  66.23214 0.2022514 2.4218636 0.08714631
#> 6  31.41823 356.3503 1.664344  22.99406 0.1311363 1.6094613 0.08035681
#> 7  25.50312 627.7881 1.017857  13.13652 0.1100129 0.9586175 0.05519174
#> 8  27.17611 526.3125 1.184583  18.69336 0.1330397 1.1364118 0.05774036
#> 9  25.48768 375.5526 1.315209  22.41032 0.1440335 1.2706972 0.06151002
#> 10 28.29462 401.4419 1.412188  22.98412 0.1657710 1.3656000 0.07333822
#> 11 26.49959 203.8446 1.856049  96.21681 0.2270405 1.8415441 0.06989013
#> 12 30.99961 494.0164 1.394715  16.91318 0.1345812 1.3319401 0.07961220
#> 13 30.49615 412.5585 1.501420  20.68815 0.1426085 1.4463239 0.07792359
#> 14 32.46087 294.6992 1.890909  34.37872 0.1628766 1.8493555 0.08035036
#> 15 29.06882 433.1915 1.396649  14.12596 0.1062167 1.3211601 0.08578284
#> 16 25.98285 214.6695 1.773379 122.13314 0.1802813 1.7624667 0.04560559
#> 17 31.44205 402.8532 1.566525  26.50042 0.1516116 1.5217681 0.07141643
#> 18 35.58330 439.5086 1.697315  16.77736 0.1331548 1.6202923 0.10928138
#> 19 27.58535 347.1739 1.480488  19.46798 0.1311024 1.4227109 0.07680208
#> 20 35.89428 297.6155 2.080643  21.16481 0.1529476 2.0060315 0.13095290
#>    cnvg_warn sigma_sq        phi Tau.case.var((Intercept))
#> 1          0 214.3606 0.11564801              3.410249e+02
#> 2          0 209.2616 0.21298561              1.445009e+02
#> 3          0 215.9714 0.43137183              4.876328e+02
#> 4          0 195.0283 0.39526291              1.870724e+02
#> 5          0 181.0767 0.30088926              3.609451e+01
#> 6          0 176.1456 0.04733162              1.802046e+02
#> 7          0 187.3741 0.14954335              4.404140e+02
#> 8          0 227.9505 0.15549508              2.983619e+02
#> 9          0 186.4420 0.17169211              1.891106e+02
#> 10         0 212.0051 0.34910840              1.894368e+02
#> 11         0 203.8446 0.35710603              8.811185e-06
#> 12         0 198.9398 0.22563753              2.950765e+02
#> 13         0 194.6441 0.19412597              2.179144e+02
#> 14         0 185.1634 0.18006493              1.095357e+02
#> 15         0 139.7339 0.05019501              2.934577e+02
#> 16         0 208.3999 0.03391522              6.269594e+00
#> 17         0 221.5940 0.17200906              1.812592e+02
#> 18         0 175.3176 0.21664225              2.641910e+02
#> 19         0 154.2486 0.11747599              1.929253e+02
#> 20         0 145.4477 0.27534631              1.521678e+02