Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object
a g_REML object

nsim
number of models to simulate

seed
seed value. See documentation for simulate

parallel
if TRUE, run in parallel using foreach backend.

...
additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment,
                 random = ~ 1 | case,
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

simulate(Laski_g, nsim = 20)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#>      p_beta  r_theta delta_AB       nu     kappa     g_AB     V_g_AB cnvg_warn
#> 1  30.96050 332.2573 1.698520 53.57074 0.1819431 1.674629 0.06081010         0
#> 2  29.21704 379.2068 1.500369 22.52398 0.1551874 1.449849 0.07678114         0
#> 3  27.06017 340.3341 1.466823 28.87461 0.1522551 1.428391 0.06208865         0
#> 4  27.02749 241.8604 1.737894 74.42424 0.2023875 1.720321 0.06178100         0
#> 5  35.29824 352.8560 1.879119 19.59056 0.1323451 1.806250 0.11262642         0
#> 6  27.08556 527.6068 1.179187 17.76896 0.1640077 1.128705 0.06906716         0
#> 7  29.24580 560.1636 1.235680 14.41000 0.1261306 1.170231 0.07328128         0
#> 8  30.96806 365.2369 1.620417 17.43268 0.1434894 1.549688 0.10085033         0
#> 9  30.39838 341.8306 1.644164 17.93927 0.1493564 1.574453 0.10250180         0
#> 10 26.58342 496.5448 1.192976 15.52704 0.1212921 1.134409 0.06406153         0
#> 11 25.32750 287.8328 1.492871 22.69635 0.1219663 1.442990 0.06641384         0
#> 12 32.76937 243.7725 2.098823 62.67897 0.1878009 2.073608 0.07120810         0
#> 13 33.28980 672.4634 1.283740 13.55266 0.1245033 1.211363 0.08159237         0
#> 14 30.03674 764.1962 1.086552 10.55740 0.1037894 1.007491 0.07304851         0
#> 15 28.87231 404.8334 1.434972 19.19104 0.1468076 1.378152 0.07857028         0
#> 16 27.64078 442.5593 1.313907 23.50351 0.1627458 1.271529 0.06533166         0
#> 17 37.46040 608.0317 1.519180 13.43694 0.1163532 1.432778 0.10659156         0
#> 18 35.74472 311.7318 2.024516 47.70482 0.1702799 1.992520 0.07309598         0
#> 19 29.19983 360.0011 1.538964 27.53041 0.1611086 1.496654 0.07095982         0
#> 20 28.94401 208.9584 2.002298 95.02847 0.1813398 1.986453 0.05435470         0
#>    sigma_sq         phi Tau.case.var((Intercept))
#> 1  250.4386  0.21066952                  81.81865
#> 2  192.9557  0.26627229                 186.25118
#> 3  195.2046  0.14845981                 145.12955
#> 4  207.9319  0.27718968                  33.92855
#> 5  157.7699  0.12607072                 195.08611
#> 6  242.8105  0.49283514                 284.79634
#> 7  191.9214  0.25013794                 368.24218
#> 8  153.6883  0.29326285                 211.54860
#> 9  149.1235  0.33211512                 192.70705
#> 10 180.9312  0.14671618                 315.61358
#> 11 139.6229 -0.03174538                 148.20987
#> 12 194.1658  0.21769503                  49.60668
#> 13 215.3123  0.28387506                 457.15109
#> 14 162.2820  0.32382001                 601.91420
#> 15 183.5661  0.26938920                 221.26727
#> 16 234.4582  0.31254548                 208.10109
#> 17 188.9453  0.20130012                 419.08637
#> 18 222.4541  0.15351448                  89.27769
#> 19 205.1559  0.23631488                 154.84517
#> 20 184.6757  0.09677323                  24.28278