Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment,
                 random = ~ 1 | case,
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
simulate(Laski_g, nsim = 20)
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#>      p_beta  r_theta delta_AB        nu      kappa      g_AB     V_g_AB
#> 1  30.30498 398.7780 1.517569  23.97737 0.17239440 1.4696001 0.08034328
#> 2  29.98013 392.2800 1.513685  19.21369 0.12043178 1.4538198 0.07758856
#> 3  32.29833 260.0807 2.002746  70.81833 0.21281970 1.9814612 0.07430201
#> 4  32.15030 613.4414 1.298071  13.62236 0.12104898 1.2252676 0.08180054
#> 5  25.76821 462.8305 1.197769  17.11741 0.13552930 1.1445113 0.06332027
#> 6  31.65845 464.6389 1.468696  15.73532 0.13153871 1.3975627 0.09085419
#> 7  36.20605 425.8820 1.754432  17.88561 0.14923703 1.6798202 0.11377323
#> 8  31.06335 349.0843 1.662582  28.72812 0.16836608 1.6187966 0.07856856
#> 9  27.53104 672.7548 1.061437  10.62920 0.09903556 0.9847376 0.06876274
#> 10 34.30397 512.6195 1.515119  13.99507 0.11609861 1.4324469 0.10204887
#> 11 33.26508 348.0769 1.782998  22.87687 0.15435800 1.7238982 0.09680055
#> 12 31.09369 245.7173 1.983601  36.00991 0.17220566 1.9419987 0.08613405
#> 13 33.59384 402.3645 1.674749  23.01772 0.14761299 1.6195808 0.08577913
#> 14 34.22915 402.6586 1.705798  21.14878 0.14476099 1.6445819 0.09344776
#> 15 34.27237 523.5364 1.497858  15.50217 0.11825540 1.4242034 0.09156534
#> 16 31.12969 368.8807 1.620809  17.11455 0.14118015 1.5487285 0.10179536
#> 17 23.88293 323.4886 1.327878  41.94864 0.17979514 1.3039941 0.05421177
#> 18 32.17978 376.0905 1.659346  18.47772 0.14965707 1.5910703 0.10155279
#> 19 26.02630 240.0151 1.679938 100.71181 0.22294922 1.6673959 0.06409449
#> 20 30.36596 541.4672 1.304971  16.95390 0.12324494 1.2463786 0.06886916
#>    cnvg_warn sigma_sq        phi Tau.case.var((Intercept))
#> 1          0 219.2846 0.38754435              1.794934e+02
#> 2          0 170.0227 0.02226647              2.222573e+02
#> 3          0 227.0425 0.35294932              3.303824e+01
#> 4          0 195.7356 0.24136079              4.177058e+02
#> 5          0 188.5361 0.22760631              2.742944e+02
#> 6          0 175.0091 0.24183525              2.896298e+02
#> 7          0 185.3492 0.33282546              2.405328e+02
#> 8          0 205.8312 0.27970743              1.432531e+02
#> 9          0 142.3275 0.24845315              5.304273e+02
#> 10         0 166.6092 0.16553063              3.460103e+02
#> 11         0 178.3121 0.25231858              1.697649e+02
#> 12         0 159.7746 0.23806140              8.594263e+01
#> 13         0 204.1158 0.19170959              1.982487e+02
#> 14         0 193.3654 0.20324208              2.092932e+02
#> 15         0 189.4475 0.11647074              3.340890e+02
#> 16         0 152.2225 0.28236793              2.166583e+02
#> 17         0 225.1350 0.25298792              9.835367e+01
#> 18         0 167.5002 0.31713815              2.085903e+02
#> 19         0 240.0151 0.32404337              1.008934e-06
#> 20         0 213.8555 0.11327647              3.276118e+02