Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object
a g_REML object

nsim
number of models to simulate

seed
seed value. See documentation for simulate

parallel
if TRUE, run in parallel using foreach backend.

...
additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment, 
                 random = ~ 1 | case, 
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

simulate(Laski_g, nsim = 20)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#>      p_beta  r_theta delta_AB       nu     kappa     g_AB     V_g_AB cnvg_warn
#> 1  29.73960 465.4497 1.378475 16.40808 0.1127478 1.314491 0.07458253         0
#> 2  33.82208 376.8930 1.742173 16.55218 0.1326013 1.662022 0.11539627         0
#> 3  27.86701 505.5646 1.239373 18.30302 0.1402464 1.187884 0.06449795         0
#> 4  40.70614 371.2137 2.112749 22.98859 0.1437075 2.043063 0.12228375         0
#> 5  27.95739 407.1597 1.385525 23.08872 0.1476172 1.340025 0.06561238         0
#> 6  27.77054 608.0478 1.126200 13.70594 0.1159765 1.063428 0.06376680         0
#> 7  28.60064 365.4677 1.496068 22.00099 0.1525070 1.444482 0.07694110         0
#> 8  28.72124 302.1898 1.652203 37.78205 0.1633949 1.619187 0.06408333         0
#> 9  31.86060 569.0222 1.335640 14.74809 0.1349802 1.266546 0.08358209         0
#> 10 34.44099 401.6735 1.718459 32.58227 0.1657889 1.678596 0.07455465         0
#> 11 34.68210 487.9706 1.570032 15.97341 0.1451596 1.495142 0.10383231         0
#> 12 31.07971 617.7472 1.250465 12.04483 0.1143391 1.170952 0.08430454         0
#> 13 29.56494 332.0120 1.622558 29.68044 0.1544933 1.581209 0.07005731         0
#> 14 24.02989 261.5100 1.485962 52.90810 0.1816205 1.464797 0.05453810         0
#> 15 35.84875 245.4052 2.288401 56.57736 0.1874844 2.257931 0.08249634         0
#> 16 36.13586 241.6649 2.324511 44.22792 0.1588640 2.284869 0.08789012         0
#> 17 32.31596 247.3058 2.054944 79.95004 0.1771622 2.035606 0.05829412         0
#> 18 28.40896 448.6791 1.341181 15.02371 0.1203695 1.273095 0.07894340         0
#> 19 33.44054 415.4072 1.640727 27.27023 0.1668761 1.595185 0.07947273         0
#> 20 32.53512 425.2891 1.577649 23.86046 0.1550501 1.527534 0.07884338         0
#>    sigma_sq        phi Tau.case.var((Intercept))
#> 1  175.8882 0.02542096                 289.56151
#> 2  148.5212 0.21953133                 228.37184
#> 3  218.7000 0.23413419                 286.86469
#> 4  186.9067 0.15842926                 184.30694
#> 5  206.9065 0.19048621                 200.25323
#> 6  193.0815 0.18071557                 414.96632
#> 7  182.4690 0.25348955                 182.99867
#> 8  196.9870 0.15845431                 105.20284
#> 9  204.1514 0.32541277                 364.87082
#> 10 248.3041 0.21827350                 153.36932
#> 11 193.2813 0.36987710                 294.68936
#> 12 166.6528 0.28450127                 451.09445
#> 13 193.5038 0.15782316                 138.50823
#> 14 196.2797 0.21113139                  65.23036
#> 15 189.8587 0.23693524                  55.54649
#> 16 165.5372 0.08617744                  76.12772
#> 17 206.9787 0.10052278                  40.32707
#> 18 158.1756 0.15920180                 290.50348
#> 19 238.5952 0.28661020                 176.81208
#> 20 223.1633 0.24055097                 202.12585