Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object
a g_REML object

nsim
number of models to simulate

seed
seed value. See documentation for simulate

parallel
if TRUE, run in parallel using foreach backend.

...
additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment,
                 random = ~ 1 | case,
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

simulate(Laski_g, nsim = 20)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#>      p_beta  r_theta delta_AB        nu      kappa     g_AB     V_g_AB
#> 1  30.92944 557.0854 1.310422 13.276587 0.11212078 1.234975 0.08282948
#> 2  28.62167 689.9635 1.089637 12.194267 0.11308832 1.021217 0.06629443
#> 3  30.95467 488.7896 1.400120 18.220750 0.13088916 1.341686 0.07436601
#> 4  35.52971 952.3771 1.151297  9.320722 0.08043397 1.056103 0.08651888
#> 5  35.91496 496.4253 1.611939 20.467286 0.15599206 1.552141 0.09148701
#> 6  31.13323 713.7734 1.165317 12.663086 0.11008927 1.094908 0.07072074
#> 7  35.71236 333.4523 1.955698 22.113827 0.14700561 1.888611 0.11239696
#> 8  29.56791 434.6703 1.418210 20.913341 0.14104773 1.366735 0.07074119
#> 9  31.28109 415.4652 1.534669 21.802620 0.15651311 1.481264 0.08152382
#> 10 33.26630 487.0990 1.507287 14.801815 0.13361319 1.429601 0.10055476
#> 11 33.48720 432.3742 1.610456 24.384783 0.18105267 1.560410 0.08878227
#> 12 34.43308 523.6587 1.504706 15.586139 0.12994426 1.431120 0.09483084
#> 13 31.04415 633.6041 1.233306 12.060080 0.11053441 1.154985 0.08138568
#> 14 29.27556 350.1960 1.564407 36.521910 0.18748812 1.532059 0.07002091
#> 15 30.33243 359.9561 1.598757 22.143542 0.13514665 1.543989 0.07894712
#> 16 34.92634 496.9447 1.566748 13.399847 0.10818645 1.477388 0.11085160
#> 17 30.95807 327.5489 1.710550 41.224848 0.17717450 1.679241 0.06807433
#> 18 31.55387 396.3924 1.584857 15.337637 0.12341155 1.506074 0.10306706
#> 19 34.22866 362.9157 1.796748 22.777400 0.15329029 1.736929 0.09792294
#> 20 31.55758 722.1349 1.174342 12.762076 0.11244693 1.103949 0.07166942
#>    cnvg_warn sigma_sq        phi Tau.case.var((Intercept))
#> 1          0 169.1431 0.16429700                  387.9423
#> 2          0 188.5666 0.25602743                  501.3969
#> 3          0 206.9817 0.14837225                  281.8078
#> 4          0 140.8886 0.20369476                  811.4884
#> 5          0 239.6806 0.31933481                  256.7447
#> 6          0 203.4838 0.18326009                  510.2896
#> 7          0 165.1456 0.20327912                  168.3066
#> 8          0 205.8893 0.17522669                  228.7811
#> 9          0 208.2564 0.29270785                  207.2088
#> 10         0 174.6341 0.30800141                  312.4649
#> 11         0 247.5459 0.45923262                  184.8283
#> 12         0 194.8821 0.23222385                  328.7766
#> 13         0 169.4031 0.23736023                  464.2010
#> 14         0 236.9768 0.34913911                  113.2192
#> 15         0 174.9338 0.09808539                  185.0223
#> 16         0 151.3765 0.11237528                  345.5683
#> 17         0 225.4865 0.23839288                  102.0624
#> 18         0 143.3760 0.17657771                  253.0164
#> 19         0 185.0548 0.24500774                  177.8608
#> 20         0 209.0457 0.20315475                  513.0892