Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object
a g_REML object

nsim
number of models to simulate

seed
seed value. See documentation for simulate

parallel
if TRUE, run in parallel using foreach backend.

...
additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment, 
                 random = ~ 1 | case, 
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

simulate(Laski_g, nsim = 20)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#>      p_beta  r_theta delta_AB       nu     kappa      g_AB     V_g_AB cnvg_warn
#> 1  28.38452 453.8145 1.332424 16.31937 0.1260985 1.2702359 0.07418548         0
#> 2  35.44996 529.3623 1.540775 17.08022 0.1487210 1.4721141 0.09639371         0
#> 3  35.50201 195.6132 2.538364 77.09454 0.1691949 2.5135895 0.07109360         0
#> 4  29.96822 855.0730 1.024848 11.06697 0.1031124 0.9537893 0.06326636         0
#> 5  32.46859 363.9398 1.701957 17.06324 0.1361251 1.6260368 0.10896741         0
#> 6  32.50657 632.9156 1.292106 13.41896 0.1300531 1.2185183 0.08464757         0
#> 7  34.38200 673.1872 1.325144 14.10932 0.1397432 1.2534338 0.08707913         0
#> 8  29.86936 497.0352 1.339776 18.09410 0.1535430 1.2834647 0.07662118         0
#> 9  26.27081 526.4282 1.144996 12.58333 0.1104319 1.0753676 0.06918596         0
#> 10 27.30931 418.3011 1.335261 21.70516 0.1464432 1.2885848 0.06489416         0
#> 11 27.95445 509.2549 1.238749 18.06480 0.1441927 1.1865979 0.06623711         0
#> 12 32.93457 384.9825 1.678539 23.54566 0.1597333 1.6244990 0.08837620         0
#> 13 22.44282 288.4173 1.321499 31.91338 0.1749600 1.2901971 0.05929000         0
#> 14 23.50513 273.9387 1.420156 43.27678 0.1703750 1.3954010 0.05317796         0
#> 15 32.18947 621.8455 1.290840 13.25269 0.1208970 1.2163843 0.08303466         0
#> 16 30.27637 348.8911 1.620910 19.44930 0.1366530 1.5575910 0.09014914         0
#> 17 28.63774 402.9695 1.426601 24.01861 0.1452277 1.3815860 0.06562795         0
#> 18 26.78756 279.9093 1.601122 66.41878 0.1900520 1.5829743 0.05596254         0
#> 19 26.60024 426.6427 1.287815 23.21555 0.1716003 1.2457580 0.06734397         0
#> 20 30.36100 303.7475 1.742046 30.21675 0.1574171 1.6984470 0.07699775         0
#>    sigma_sq        phi Tau.case.var((Intercept))
#> 1  174.4326 0.16396594                 279.38195
#> 2  222.8032 0.35785265                 306.55908
#> 3  160.4592 0.05255775                  35.15391
#> 4  196.5480 0.24670752                 658.52495
#> 5  148.0772 0.23526844                 215.86260
#> 6  204.0506 0.35567568                 428.86496
#> 7  236.3205 0.41861721                 436.86667
#> 8  220.8693 0.36744340                 276.16589
#> 9  148.9682 0.19320386                 377.45998
#> 10 204.6934 0.20648723                 213.60769
#> 11 220.2242 0.27866196                 289.03079
#> 12 202.6635 0.28594909                 182.31898
#> 13 180.1915 0.29596678                 108.22574
#> 14 189.3898 0.17665978                  84.54891
#> 15 192.3387 0.26399876                 429.50678
#> 16 156.3671 0.16935267                 192.52399
#> 17 208.4356 0.15433720                 194.53386
#> 18 227.4597 0.22095937                  52.44961
#> 19 230.5623 0.39631049                 196.08043
#> 20 179.2871 0.17563102                 124.46042