Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object

a g_REML object

nsim

number of models to simulate

seed

seed value. See documentation for simulate

parallel

if TRUE, run in parallel using foreach backend.

...

additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment,
                 random = ~ 1 | case,
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)

suppressWarnings(
  Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))
)

if (requireNamespace("plyr", quietly = TRUE)) {
  simulate(Laski_g, nsim = 20)
}
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#>      p_beta  r_theta delta_AB       nu     kappa     g_AB     V_g_AB cnvg_warn
#> 1  25.63613 491.7819 1.156022 21.29590 0.1495908 1.114826 0.05576692         0
#> 2  32.07280 321.0184 1.790078 50.73717 0.1846111 1.763486 0.06657864         0
#> 3  30.73439 564.4941 1.293585 13.58646 0.1178507 1.220838 0.08073456         0
#> 4  36.25099 563.1698 1.527567 14.61947 0.1324170 1.447837 0.10358420         0
#> 5  30.81901 443.4697 1.463480 16.64125 0.1240338 1.396517 0.08411723         0
#> 6  28.62197 289.8988 1.681033 42.19439 0.1901363 1.650975 0.07082088         0
#> 7  26.99450 291.7766 1.580338 51.90949 0.2079355 1.557394 0.06815124         0
#> 8  34.85694 232.3948 2.286525 49.25307 0.1863526 2.251530 0.08915519         0
#> 9  34.63955 398.1873 1.735915 30.75936 0.1864015 1.693242 0.08583075         0
#> 10 34.62815 412.9247 1.704095 24.17073 0.1678245 1.650666 0.09125211         0
#> 11 24.57562 346.3637 1.320500 17.89355 0.1345979 1.264367 0.07010513         0
#> 12 31.05980 505.7563 1.381109 17.61073 0.1375838 1.321444 0.07673948         0
#> 13 33.73197 419.9435 1.646063 16.17676 0.1181478 1.568549 0.10337654         0
#> 14 24.29265 465.8748 1.125487 17.32906 0.1372687 1.076063 0.05810557         0
#> 15 31.27091 349.2248 1.673354 40.46854 0.1848405 1.642150 0.06999736         0
#> 16 32.58546 397.3744 1.634647 20.11988 0.1520647 1.572946 0.09338334         0
#> 17 32.51958 392.4937 1.641454 26.94431 0.1766324 1.595336 0.08358575         0
#> 18 28.12608 360.9100 1.480504 30.00903 0.1777458 1.443192 0.06982610         0
#> 19 34.60190 544.9995 1.482183 15.05058 0.1232374 1.407076 0.09365751         0
#> 20 35.79450 503.9803 1.594445 15.77909 0.1170562 1.517439 0.09988202         0
#>    sigma_sq        phi Tau.case.var((Intercept))
#> 1  239.4460 0.24277385                 252.33589
#> 2  239.3150 0.24124105                  81.70345
#> 3  178.2350 0.20849093                 386.25903
#> 4  198.9297 0.30525821                 364.24003
#> 5  172.6724 0.13176587                 270.79739
#> 6  206.8812 0.32640739                  83.01763
#> 7  232.2702 0.39714406                  59.50640
#> 8  172.3274 0.26024367                  60.06736
#> 9  253.2439 0.40036319                 144.94340
#> 10 224.7938 0.34379487                 188.13082
#> 11 145.6863 0.19329682                 200.67735
#> 12 211.4855 0.23066271                 294.27072
#> 13 157.9780 0.08913119                 261.96550
#> 14 192.3116 0.23706068                 273.56320
#> 15 242.8234 0.29929317                 106.40142
#> 16 188.0108 0.29259432                 209.36355
#> 17 230.0566 0.37200078                 162.43714
#> 18 221.6402 0.33920398                 139.26979
#> 19 193.5660 0.18773975                 351.43343
#> 20 185.0300 0.09296579                 318.95024