Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object
a g_REML object

nsim
number of models to simulate

seed
seed value. See documentation for simulate

parallel
if TRUE, run in parallel using foreach backend.

...
additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment, 
                 random = ~ 1 | case, 
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

simulate(Laski_g, nsim = 20)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#>      p_beta  r_theta delta_AB       nu     kappa     g_AB     V_g_AB cnvg_warn
#> 1  29.64184 237.8247 1.922103 34.51625 0.1817100 1.880033 0.08848983         0
#> 2  27.06072 305.3574 1.548585 22.80836 0.1513485 1.497099 0.07824719         0
#> 3  31.50813 405.0150 1.565622 18.90415 0.1575330 1.502676 0.09375204         0
#> 4  30.34565 309.7127 1.724317 31.02764 0.1524201 1.682298 0.07299430         0
#> 5  32.14138 557.3834 1.361405 14.42426 0.1299194 1.289370 0.08634530         0
#> 6  30.88482 563.9926 1.300494 13.23418 0.1155206 1.225374 0.08282782         0
#> 7  36.03946 403.8380 1.793390 19.22635 0.1358580 1.722510 0.10689248         0
#> 8  31.85836 422.5563 1.549819 19.11844 0.1530501 1.488216 0.09014950         0
#> 9  29.27177 351.6363 1.560997 29.41388 0.1864270 1.520853 0.07814023         0
#> 10 28.85288 468.0660 1.333631 16.07596 0.1212421 1.270429 0.07398908         0
#> 11 24.60233 343.4611 1.327509 25.26146 0.1667661 1.287702 0.06461050         0
#> 12 32.79980 353.5153 1.744485 28.67390 0.1916683 1.698454 0.09224124         0
#> 13 28.55363 247.4842 1.815046 58.27971 0.1871873 1.791588 0.06405547         0
#> 14 35.22700 533.7728 1.524746 16.06497 0.1339852 1.452437 0.09546250         0
#> 15 35.29273 624.9193 1.411800 12.61654 0.1118547 1.326178 0.09861432         0
#> 16 31.40599 377.1787 1.617108 32.64093 0.1977855 1.579664 0.08096640         0
#> 17 33.00519 398.4051 1.653559 28.41456 0.1701929 1.609526 0.07922858         0
#> 18 24.68953 451.3896 1.162083 14.22175 0.1252399 1.099702 0.06721919         0
#> 19 30.34658 454.1879 1.423941 23.63525 0.1556581 1.378273 0.06942666         0
#> 20 27.84667 631.8564 1.107807 12.21593 0.1101122 1.038372 0.06726667         0
#>    sigma_sq       phi Tau.case.var((Intercept))
#> 1  155.4774 0.3228247                  82.34734
#> 2  155.2097 0.2277044                 150.14770
#> 3  187.3842 0.3787970                 217.63083
#> 4  183.4955 0.1275489                 126.21717
#> 5  193.0380 0.2895700                 364.34541
#> 6  171.8659 0.2051399                 392.12668
#> 7  179.3082 0.1676756                 224.52976
#> 8  194.1868 0.3293573                 228.36949
#> 9  220.1791 0.4186283                 131.45727
#> 10 176.0112 0.1243502                 292.05489
#> 11 190.4010 0.3157895                 153.06012
#> 12 224.4421 0.4772201                 129.07314
#> 13 192.9600 0.2286146                  54.52420
#> 14 205.8185 0.2524211                 327.95431
#> 15 178.0341 0.2071821                 446.88522
#> 16 255.6680 0.4737012                 121.51074
#> 17 234.7800 0.2983279                 163.62511
#> 18 152.2909 0.2510832                 299.09864
#> 19 237.4342 0.2496249                 216.75376
#> 20 171.7560 0.2188233                 460.10041