Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object
a g_REML object

nsim
number of models to simulate

seed
seed value. See documentation for simulate

parallel
if TRUE, run in parallel using foreach backend.

...
additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment,
                 random = ~ 1 | case,
                 correlation = corAR1(0, ~ time | case),
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

simulate(Laski_g, nsim = 20)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#>      p_beta  r_theta delta_AB       nu     kappa     g_AB     V_g_AB cnvg_warn
#> 1  33.09668 360.6672 1.742734 27.09007 0.1833800 1.694036 0.09231848         0
#> 2  30.07987 305.0574 1.722206 36.35824 0.1691343 1.686435 0.07085423         0
#> 3  29.74114 475.5613 1.363811 14.73627 0.1346741 1.293203 0.08631004         0
#> 4  32.27697 285.5483 1.910086 29.30308 0.1565813 1.860778 0.08922067         0
#> 5  34.75316 390.6387 1.758356 16.33533 0.1363395 1.676370 0.11966144         0
#> 6  30.69401 398.9044 1.536807 25.15604 0.1693709 1.490529 0.07803139         0
#> 7  34.95153 324.2832 1.940904 20.35446 0.1566268 1.868498 0.12216061         0
#> 8  34.35678 515.0565 1.513857 15.25175 0.1314171 1.438173 0.09802784         0
#> 9  32.39385 359.8471 1.707669 19.75837 0.1585908 1.642017 0.10330275         0
#> 10 29.21714 573.3080 1.220236 14.19290 0.1253198 1.154598 0.07258361         0
#> 11 29.07224 604.5505 1.182394 14.78955 0.1387834 1.121402 0.07053055         0
#> 12 35.52761 665.9977 1.376669 13.38630 0.1133583 1.298070 0.08965609         0
#> 13 28.43948 603.2020 1.157951 15.37263 0.1280221 1.100523 0.06348172         0
#> 14 36.21569 619.8110 1.454679 13.73802 0.1255815 1.373792 0.09921437         0
#> 15 29.30293 343.5216 1.581008 30.76259 0.1726068 1.542147 0.07218033         0
#> 16 27.44369 320.9478 1.531882 26.87527 0.1669844 1.488731 0.07364611         0
#> 17 34.00568 498.1188 1.523649 19.35887 0.1647352 1.463848 0.09091615         0
#> 18 27.14089 517.5269 1.193047 16.60937 0.1171383 1.138351 0.05961523         0
#> 19 29.99795 495.0924 1.348182 16.18221 0.1311471 1.284717 0.07744395         0
#> 20 33.54182 569.9531 1.404970 14.44457 0.1170033 1.330735 0.08738388         0
#>    sigma_sq       phi Tau.case.var((Intercept))
#> 1  216.7759 0.4280989                  143.8913
#> 2  197.9346 0.2121809                  107.1228
#> 3  170.3389 0.3228065                  305.2224
#> 4  166.0386 0.1786982                  119.5096
#> 5  153.5810 0.2647866                  237.0576
#> 6  222.2137 0.3395430                  176.6907
#> 7  156.3290 0.3280123                  167.9542
#> 8  188.6467 0.2624906                  326.4099
#> 9  171.4682 0.3625734                  188.3789
#> 10 193.0640 0.2535673                  380.2440
#> 11 220.1269 0.3640074                  384.4236
#> 12 204.7149 0.1712079                  461.2828
#> 13 220.7625 0.2221803                  382.4395
#> 14 201.9466 0.2837303                  417.8644
#> 15 210.3903 0.2897567                  133.1313
#> 16 183.1779 0.2930002                  137.7699
#> 17 239.9665 0.4354392                  258.1523
#> 18 198.9414 0.0635924                  318.5855
#> 19 190.8589 0.2194359                  304.2335
#> 20 191.9052 0.1515353                  378.0479