Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object
a g_REML object

nsim
number of models to simulate

seed
seed value. See documentation for simulate

parallel
if TRUE, run in parallel using foreach backend.

...
additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment, 
                 random = ~ 1 | case, 
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

simulate(Laski_g, nsim = 20)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#>      p_beta  r_theta  delta_AB       nu      kappa      g_AB     V_g_AB
#> 1  29.90072 386.7044 1.5205198 21.81755 0.15858997 1.4676445 0.08111981
#> 2  29.51949 385.4103 1.5036517 24.57483 0.14624557 1.4572901 0.06965326
#> 3  31.03280 339.7493 1.6836105 22.18842 0.14968257 1.6260536 0.08961153
#> 4  28.75888 523.3710 1.2570920 13.47806 0.12918515 1.1858178 0.08052696
#> 5  21.91354 582.5920 0.9078836 12.36617 0.08591655 0.8516849 0.04388654
#> 6  30.97033 285.2106 1.8338467 41.30278 0.17171568 1.8003439 0.07148263
#> 7  30.63892 390.2283 1.5510081 20.46079 0.15752955 1.4934519 0.08707214
#> 8  28.65116 295.7593 1.6659924 61.10808 0.21317488 1.6454612 0.06887652
#> 9  34.88275 342.0388 1.8861359 41.30971 0.20089154 1.8516836 0.08489287
#> 10 30.64709 378.5407 1.5751897 21.16597 0.16684655 1.5187069 0.08986596
#> 11 36.40636 516.4508 1.6020012 15.27174 0.13080200 1.5220170 0.10733134
#> 12 31.51547 544.0064 1.3512068 12.97366 0.11550009 1.2715595 0.08993653
#> 13 29.55231 392.9615 1.4907898 13.69119 0.12039464 1.4076058 0.10236614
#> 14 29.35549 616.0740 1.1826954 14.86303 0.13122012 1.1219946 0.06816000
#> 15 29.42067 214.5957 2.0083621 84.04096 0.19897706 1.9903856 0.06408753
#> 16 27.67957 341.0913 1.4987322 31.81700 0.17048770 1.4631238 0.06589380
#> 17 32.81518 341.5996 1.7754814 29.31651 0.16874901 1.7296689 0.08449001
#> 18 30.86476 457.6845 1.4427126 18.83688 0.14832942 1.3844977 0.08078719
#> 19 33.77986 430.2019 1.6286272 16.20599 0.11601488 1.5520747 0.10081709
#> 20 25.62551 298.3064 1.4836834 25.76471 0.15673645 1.4400708 0.06938313
#>    cnvg_warn sigma_sq         phi Tau.case.var((Intercept))
#> 1          0 194.9103  0.31067535                 191.79407
#> 2          0 202.1648  0.15435256                 183.24550
#> 3          0 169.4641  0.22514014                 170.28517
#> 4          0 168.9928  0.34143788                 354.37826
#> 5          0 154.0848 -0.08851463                 428.50726
#> 6          0 194.5278  0.19791717                  90.68283
#> 7          0 189.1597  0.33349684                 201.06860
#> 8          0 249.3434  0.39042417                  46.41593
#> 9          0 250.4132  0.41233904                  91.62565
#> 10         0 192.4442  0.40146444                 186.09652
#> 11         0 189.0969  0.25523873                 327.35384
#> 12         0 161.8813  0.22274303                 382.12501
#> 13         0 125.8797  0.22992922                 267.08185
#> 14         0 220.2682  0.27967642                 395.80582
#> 15         0 189.0752  0.23152929                  25.52050
#> 16         0 210.7772  0.26213906                 130.31404
#> 17         0 203.3418  0.27566630                 138.25781
#> 18         0 205.7463  0.29364955                 251.93815
#> 19         0 161.5557  0.06622993                 268.64618
#> 20         0 163.3211  0.22464173                 134.98524