Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object
a g_REML object

nsim
number of models to simulate

seed
seed value. See documentation for simulate

parallel
if TRUE, run in parallel using foreach backend.

...
additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment, 
                 random = ~ 1 | case, 
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

simulate(Laski_g, nsim = 20)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#>      p_beta  r_theta delta_AB       nu      kappa      g_AB     V_g_AB
#> 1  28.01423 293.4532 1.635345 27.21194 0.15570692 1.5898542 0.07557670
#> 2  27.17840 329.6072 1.497013 16.83593 0.13406247 1.4293192 0.08907141
#> 3  29.76954 534.6649 1.287453 13.97197 0.11828419 1.2170847 0.07824056
#> 4  28.01660 354.0552 1.488949 35.42878 0.18263915 1.4572055 0.06596798
#> 5  30.44657 282.0825 1.812802 40.68097 0.17836295 1.7791743 0.07353256
#> 6  36.09626 646.8560 1.419248 12.62869 0.09913509 1.3332591 0.09661341
#> 7  29.03518 513.8779 1.280839 18.07295 0.13447828 1.2269405 0.06651578
#> 8  29.76061 639.2559 1.177076 15.33385 0.13766762 1.1185492 0.06782237
#> 9  27.06018 380.5819 1.387096 16.88346 0.14090414 1.3245523 0.08086456
#> 10 32.86358 422.0774 1.599627 24.12751 0.15442980 1.5493827 0.07951924
#> 11 34.74684 378.0708 1.787017 22.27639 0.16768449 1.7261692 0.10359327
#> 12 33.79363 998.5907 1.069402 10.28038 0.10113397 0.9894399 0.07236100
#> 13 30.17321 354.1188 1.603419 33.30053 0.16550221 1.5670336 0.06751486
#> 14 33.41284 599.6998 1.364415 12.95224 0.10807995 1.2838535 0.08983245
#> 15 28.00543 294.9190 1.630763 24.19822 0.14224603 1.5796917 0.07781254
#> 16 30.81295 565.5068 1.295729 16.54833 0.13094571 1.2361039 0.07152710
#> 17 26.35545 478.2593 1.205144 17.64726 0.13720299 1.1531895 0.06288971
#> 18 27.47398 331.0766 1.509932 19.26198 0.14660467 1.4503669 0.08430804
#> 19 29.18200 456.9419 1.365163 27.99093 0.16779193 1.3282549 0.06311440
#> 20 28.55798 485.4258 1.296182 18.56171 0.15111776 1.2430939 0.07118749
#>    cnvg_warn sigma_sq        phi Tau.case.var((Intercept))
#> 1          0 164.6567 0.19643080                 128.79652
#> 2          0 132.1017 0.22337107                 197.50555
#> 3          0 174.2572 0.19048685                 360.40767
#> 4          0 234.1471 0.32172995                 119.90818
#> 5          0 193.7737 0.25053093                  88.30881
#> 6          0 179.8182 0.05704114                 467.03782
#> 7          0 217.7099 0.18672704                 296.16807
#> 8          0 239.2390 0.32298011                 400.01692
#> 9          0 155.3283 0.28828778                 225.25361
#> 10         0 222.5002 0.23079158                 199.57712
#> 11         0 197.7323 0.38170303                 180.33850
#> 12         0 199.8615 0.33048010                 798.72923
#> 13         0 220.7091 0.20953522                 133.40969
#> 14         0 175.2012 0.13970202                 424.49864
#> 15         0 152.4473 0.12588713                 142.47170
#> 16         0 222.0354 0.20340306                 343.47139
#> 17         0 200.1366 0.22584310                 278.12268
#> 18         0 150.4166 0.26556116                 180.66004
#> 19         0 266.0582 0.28441980                 190.88365
#> 20         0 217.7155 0.32830748                 267.71039