Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object
a g_REML object

nsim
number of models to simulate

seed
seed value. See documentation for simulate

parallel
if TRUE, run in parallel using foreach backend.

...
additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment, 
                 random = ~ 1 | case, 
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

simulate(Laski_g, nsim = 20)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#>      p_beta  r_theta delta_AB       nu      kappa     g_AB     V_g_AB cnvg_warn
#> 1  33.40131 394.5553 1.681549 21.38934 0.14771865 1.621890 0.09147425         0
#> 2  34.50327 860.0829 1.176495 10.18081 0.09003241 1.087643 0.08385916         0
#> 3  23.66136 421.0991 1.153049 24.72272 0.17733400 1.117712 0.06007406         0
#> 4  29.08797 296.8992 1.688142 40.27152 0.18292665 1.656506 0.07009497         0
#> 5  31.78421 674.7455 1.223606 12.72459 0.12518226 1.150040 0.08005114         0
#> 6  31.71725 458.1327 1.481835 21.53950 0.16610690 1.429632 0.08156452         0
#> 7  36.61680 303.4301 2.102089 24.42269 0.12538202 2.036868 0.11007063         0
#> 8  32.76951 344.0542 1.766674 28.37190 0.16027292 1.719558 0.08300978         0
#> 9  33.35734 563.3355 1.405426 14.15021 0.10731926 1.329595 0.08682200         0
#> 10 29.68182 333.9998 1.624117 22.46699 0.15563967 1.569291 0.08603763         0
#> 11 28.61134 462.1922 1.330843 23.14037 0.16571590 1.287238 0.06797532         0
#> 12 29.53462 419.8084 1.441471 15.91131 0.11808306 1.372441 0.08385610         0
#> 13 31.95455 439.8724 1.523594 16.90794 0.12365064 1.454997 0.08848765         0
#> 14 30.09732 639.4035 1.190256 16.31925 0.13867618 1.134703 0.06597587         0
#> 15 28.72449 185.1729 2.110880 45.57687 0.18401130 2.075953 0.08411583         0
#> 16 31.86655 315.9247 1.792847 32.73397 0.17056895 1.751453 0.08006760         0
#> 17 30.18974 267.4594 1.845995 30.89743 0.17708753 1.800820 0.08872419         0
#> 18 35.57367 471.9974 1.637415 18.16343 0.13479848 1.568860 0.09654419         0
#> 19 27.61573 427.0547 1.336333 16.81223 0.13107775 1.275819 0.07402043         0
#> 20 34.32930 308.2764 1.955216 45.53655 0.19789819 1.922836 0.08243668         0
#>    sigma_sq         phi Tau.case.var((Intercept))
#> 1  191.8728  0.22426210                 202.68253
#> 2  162.7152  0.18868643                 697.36767
#> 3  238.7233  0.41712319                 182.37582
#> 4  205.2218  0.28662068                  91.67742
#> 5  201.9095  0.35382383                 472.83601
#> 6  234.3832  0.38490122                 223.74950
#> 7  154.2407 -0.02650357                 149.18937
#> 8  198.4626  0.21909280                 145.59168
#> 9  182.3829  0.06121298                 380.95256
#> 10 169.8866  0.27131710                 164.11320
#> 11 244.8760  0.34542501                 217.31616
#> 12 155.5737  0.09841962                 264.23471
#> 13 173.4563  0.11882706                 266.41605
#> 14 252.6539  0.28862354                 386.74957
#> 15 133.4685  0.26239390                  51.70439
#> 16 197.4687  0.25383635                 118.45601
#> 17 165.8526  0.32359442                 101.60683
#> 18 200.8294  0.18706470                 271.16801
#> 19 169.9117  0.19511052                 257.14306
#> 20 229.7535  0.36152780                  78.52291