Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object
a g_REML object

nsim
number of models to simulate

seed
seed value. See documentation for simulate

parallel
if TRUE, run in parallel using foreach backend.

...
additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment, 
                 random = ~ 1 | case, 
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

simulate(Laski_g, nsim = 20)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#>      p_beta   r_theta  delta_AB       nu      kappa     g_AB     V_g_AB
#> 1  37.94953  525.6433 1.6552394 14.95950 0.12899168 1.570843 0.11505111
#> 2  30.84279  506.6815 1.3702070 14.30833 0.13854437 1.297108 0.09026416
#> 3  30.06705  449.4158 1.4182953 15.13862 0.11253375 1.346850 0.08401817
#> 4  30.76927  630.9396 1.2249640 12.83981 0.12032455 1.151991 0.07831216
#> 5  32.14412  394.4462 1.6184812 17.37977 0.12176637 1.547619 0.09496536
#> 6  34.13899  234.7327 2.2282499 53.20744 0.17714257 2.196693 0.07914323
#> 7  37.55815  411.5404 1.8513902 17.01467 0.14020588 1.768565 0.12692054
#> 8  23.02224  430.3733 1.1097495 21.80774 0.16537086 1.071141 0.05753359
#> 9  31.75674  450.9510 1.4954478 24.27271 0.16647024 1.448759 0.07622123
#> 10 32.28208  316.1854 1.8154764 37.56330 0.18833205 1.778985 0.08092406
#> 11 29.61005  210.5404 2.0406634 92.64646 0.18290848 2.024099 0.05633193
#> 12 30.18425  359.9971 1.5908560 28.00892 0.18072492 1.547874 0.07999658
#> 13 31.95210  456.9833 1.4946835 19.83891 0.13644143 1.437457 0.07821459
#> 14 31.71123  812.0269 1.1128274 11.75929 0.11463911 1.040310 0.07121004
#> 15 31.04996  376.5790 1.6000478 23.15750 0.16568315 1.547662 0.08567295
#> 16 28.34819  351.2304 1.5126181 26.60056 0.14877096 1.469565 0.06712542
#> 17 32.89228  847.9450 1.1295623 10.44795 0.09134519 1.046490 0.07623300
#> 18 35.38591  344.1638 1.9074262 18.90019 0.12855194 1.830721 0.11825549
#> 19 27.13603  310.9797 1.5387930 33.59491 0.16429439 1.504182 0.06364229
#> 20 30.27288 1214.5828 0.8686407 10.72569 0.10110509 0.806451 0.04952717
#>    cnvg_warn sigma_sq        phi Tau.case.var((Intercept))
#> 1          0 188.0153 0.25163583                 337.62799
#> 2          0 179.1227 0.39096923                 327.55874
#> 3          0 157.4381 0.07177874                 291.97763
#> 4          0 187.8146 0.28764399                 443.12492
#> 5          0 158.6123 0.08502268                 235.83391
#> 6          0 175.1265 0.17867389                  59.60627
#> 7          0 168.6658 0.27657212                 242.87454
#> 8          0 221.0524 0.37159998                 209.32091
#> 9          0 245.0890 0.33029330                 205.86198
#> 10         0 216.4461 0.34681719                  99.73925
#> 11         0 185.1322 0.11166814                  25.40813
#> 12         0 217.0815 0.39047056                 142.91553
#> 13         0 207.5020 0.15787729                 249.48127
#> 14         0 212.2062 0.31626548                 599.82068
#> 15         0 199.5721 0.34479078                 177.00687
#> 16         0 192.6664 0.14710660                 158.56395
#> 17         0 169.8688 0.16826691                 678.07619
#> 18         0 149.1324 0.10773688                 195.03136
#> 19         0 194.0870 0.19762347                 116.89270
#> 20         0 262.8805 0.26325397                 951.70235