Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object
a g_REML object

nsim
number of models to simulate

seed
seed value. See documentation for simulate

parallel
if TRUE, run in parallel using foreach backend.

...
additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment, 
                 random = ~ 1 | case, 
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

simulate(Laski_g, nsim = 20)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#>      p_beta  r_theta delta_AB       nu     kappa      g_AB     V_g_AB cnvg_warn
#> 1  30.72874 376.6762 1.583291 25.73571 0.1637997 1.5366972 0.07790198         0
#> 2  33.20912 496.6116 1.490215 15.88291 0.1388929 1.4187210 0.09431670         0
#> 3  34.97007 309.6846 1.987178 23.28548 0.1429563 1.9224784 0.10920192         0
#> 4  31.88252 348.7846 1.707159 21.31361 0.1404968 1.6463732 0.09172545         0
#> 5  29.58539 447.3199 1.398840 17.32951 0.1426887 1.3374140 0.08071752         0
#> 6  29.38546 353.1060 1.563796 21.01606 0.1587985 1.5073167 0.08674927         0
#> 7  34.09495 390.4754 1.725414 19.00108 0.1419223 1.6564012 0.10311579         0
#> 8  35.35027 237.0457 2.296025 39.11545 0.1794723 2.2517180 0.10162379         0
#> 9  33.59881 470.1372 1.549571 18.10573 0.1558719 1.4844840 0.09497728         0
#> 10 28.59988 629.5176 1.139883 12.57190 0.1018864 1.0705013 0.06683650         0
#> 11 31.33278 396.7419 1.573059 25.26721 0.1541693 1.5258993 0.07509473         0
#> 12 27.16263 436.1327 1.300656 15.31245 0.1260679 1.2358929 0.07537353         0
#> 13 35.42814 440.4720 1.688065 14.23542 0.1330295 1.5975387 0.12567959         0
#> 14 24.85901 361.9868 1.306585 26.06048 0.1717479 1.2686178 0.06406424         0
#> 15 25.51140 408.7360 1.261865 30.67439 0.1942649 1.2307583 0.06515201         0
#> 16 29.42876 379.3013 1.511054 23.57443 0.1804263 1.4624655 0.08371125         0
#> 17 25.08602 821.1900 0.875407 12.32045 0.1176136 0.8210133 0.04825952         0
#> 18 29.48683 841.6169 1.016414 10.60842 0.0977981 0.9428211 0.06377641         0
#> 19 30.85995 448.3728 1.457389 19.04945 0.1649758 1.3992471 0.08664084         0
#> 20 26.73172 556.3698 1.133300 15.44814 0.1529020 1.0773742 0.06854413         0
#>    sigma_sq        phi Tau.case.var((Intercept))
#> 1  209.1750 0.28353547                  167.5012
#> 2  192.0764 0.30957795                  304.5352
#> 3  156.8714 0.14680036                  152.8132
#> 4  167.0059 0.16189878                  181.7788
#> 5  187.0275 0.28920444                  260.2924
#> 6  174.4056 0.33107827                  178.7003
#> 7  173.9996 0.22964994                  216.4758
#> 8  161.0548 0.26919950                   75.9909
#> 9  210.6592 0.39014697                  259.4780
#> 10 174.8015 0.09384385                  454.7161
#> 11 214.0675 0.21063597                  182.6744
#> 12 158.3710 0.20491926                  277.7617
#> 13 152.0191 0.33416296                  288.4529
#> 14 206.2316 0.34473150                  155.7553
#> 15 267.8926 0.46947723                  140.8433
#> 16 213.8658 0.47223768                  165.4355
#> 17 230.4855 0.29878572                  590.7045
#> 18 176.7525 0.23454601                  664.8644
#> 19 214.4625 0.44906815                  233.9103
#> 20 223.0185 0.48798975                  333.3513