Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object
a g_REML object

nsim
number of models to simulate

seed
seed value. See documentation for simulate

parallel
if TRUE, run in parallel using foreach backend.

...
additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment,
                 random = ~ 1 | case,
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

simulate(Laski_g, nsim = 20)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#>      p_beta  r_theta delta_AB       nu     kappa     g_AB     V_g_AB cnvg_warn
#> 1  36.40438 335.3609 1.987913 30.44640 0.1707078 1.938538 0.09655340         0
#> 2  26.59966 306.9522 1.518241 23.46347 0.1656359 1.469188 0.07920703         0
#> 3  30.55284 386.7914 1.553507 18.71439 0.1452993 1.490405 0.08961010         0
#> 4  37.00843 414.3875 1.818014 19.33741 0.1539006 1.746579 0.11415436         0
#> 5  29.92424 445.0569 1.418454 20.30098 0.1558896 1.365397 0.07690134         0
#> 6  28.16565 408.3666 1.393781 15.14440 0.1256461 1.323598 0.08480668         0
#> 7  27.89240 533.0803 1.208063 15.88110 0.1310562 1.150099 0.06665268         0
#> 8  31.32274 425.4963 1.518490 15.41918 0.1298191 1.443412 0.09714035         0
#> 9  29.62819 478.0887 1.355036 16.27100 0.1339758 1.291602 0.07847437         0
#> 10 32.33102 757.7621 1.174500 12.32393 0.1273438 1.101543 0.07773255         0
#> 11 27.53236 400.2189 1.376241 14.47475 0.1154314 1.303679 0.08388456         0
#> 12 30.10593 367.0380 1.571436 28.90525 0.1651044 1.530306 0.07187734         0
#> 13 32.04217 484.1819 1.456189 15.33316 0.1297004 1.383781 0.09115133         0
#> 14 39.03214 534.0163 1.689060 14.68521 0.1238462 1.601303 0.11979028         0
#> 15 27.66444 240.7311 1.783018 40.35446 0.1701480 1.749674 0.06962025         0
#> 16 33.10401 486.3442 1.501097 18.72415 0.1392488 1.440156 0.08329960         0
#> 17 31.83646 574.2904 1.328492 14.03361 0.1282607 1.256206 0.08460343         0
#> 18 28.06252 409.1330 1.387377 20.88784 0.1398153 1.336958 0.06828422         0
#> 19 32.55732 336.0792 1.775938 25.76485 0.1740254 1.723735 0.09439125         0
#> 20 25.75662 306.1011 1.472165 37.92632 0.1888455 1.442859 0.06543190         0
#>    sigma_sq       phi Tau.case.var((Intercept))
#> 1  203.6701 0.2782815                 131.69083
#> 2  163.7153 0.3382955                 143.23695
#> 3  171.8567 0.2690691                 214.93466
#> 4  192.2468 0.3308495                 222.14074
#> 5  213.7929 0.3226770                 231.26407
#> 6  146.6375 0.2082570                 261.72910
#> 7  202.1823 0.2307889                 330.89803
#> 8  156.7871 0.2383949                 268.70915
#> 9  186.3164 0.2439585                 291.77229
#> 10 220.7955 0.4188189                 536.96663
#> 11 134.6455 0.1332310                 265.57335
#> 12 215.5003 0.2514690                 151.53766
#> 13 177.4633 0.2411105                 306.71865
#> 14 185.4329 0.2117134                 348.58340
#> 15 162.4457 0.1921879                  78.28538
#> 16 213.3579 0.2125952                 272.98626
#> 17 192.6927 0.2945701                 381.59771
#> 18 193.2261 0.1647349                 215.90692
#> 19 191.7705 0.3689109                 144.30872
#> 20 210.4912 0.3477483                  95.60993