Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object

a g_REML object

nsim

number of models to simulate

seed

seed value. See documentation for simulate

parallel

if TRUE, run in parallel using foreach backend.

...

additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment,
                 random = ~ 1 | case,
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)

suppressWarnings(
  Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))
)

if (requireNamespace("plyr", quietly = TRUE)) {
  simulate(Laski_g, nsim = 5)
}
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#>     p_beta  r_theta  delta_AB       nu     kappa      g_AB     V_g_AB cnvg_warn
#> 1 27.73134 438.8325 1.3237970 27.35938 0.1770851 1.2871733 0.06511730         0
#> 2 33.14829 454.4643 1.5549314 21.77113 0.1578645 1.5007428 0.08354738         0
#> 3 34.99606 721.8803 1.3025265 13.36043 0.1276930 1.2280137 0.08542337         0
#> 4 24.89200 620.1118 0.9995965 12.73591 0.1152099 0.9395531 0.05633259         0
#> 5 33.96979 333.4580 1.8602543 23.06119 0.1459844 1.7990917 0.09997186         0
#>   sigma_sq       phi Tau.case.var((Intercept))
#> 1 259.1956 0.3694737                  179.6369
#> 2 228.3766 0.3052927                  226.0877
#> 3 229.8276 0.3329390                  492.0526
#> 4 180.2598 0.2368291                  439.8520
#> 5 168.8511 0.1768822                  164.6070