Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object
a g_REML object

nsim
number of models to simulate

seed
seed value. See documentation for simulate

parallel
if TRUE, run in parallel using foreach backend.

...
additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment, 
                 random = ~ 1 | case, 
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

simulate(Laski_g, nsim = 20)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#>      p_beta  r_theta delta_AB       nu     kappa     g_AB     V_g_AB cnvg_warn
#> 1  34.33046 241.8863 2.207364 65.84831 0.1810110 2.182127 0.07052763         0
#> 2  30.27050 393.6738 1.525638 22.19431 0.1601011 1.473495 0.08098569         0
#> 3  27.08786 349.1930 1.449579 24.73511 0.1434190 1.405177 0.06513846         0
#> 4  32.06468 598.9294 1.310205 14.12309 0.1288417 1.239373 0.08246143         0
#> 5  22.35604 300.0518 1.290615 47.08383 0.1881299 1.269947 0.05382038         0
#> 6  30.10659 273.9650 1.818922 24.43999 0.1576076 1.762527 0.09572674         0
#> 7  34.33189 485.2466 1.558534 19.05069 0.1422205 1.496361 0.08785404         0
#> 8  26.76693 390.2477 1.354966 27.85712 0.1733661 1.318156 0.06471264         0
#> 9  29.09272 353.8486 1.546591 25.96155 0.1585061 1.501478 0.07340521         0
#> 10 30.83019 614.1660 1.244037 15.79831 0.1350461 1.184029 0.07099991         0
#> 11 28.77515 173.1311 2.186905 93.32284 0.2033213 2.169283 0.06743605         0
#> 12 35.31090 224.7594 2.355319 58.18969 0.1761013 2.324831 0.07969931         0
#> 13 28.78749 628.2860 1.148485 12.24933 0.0961477 1.076700 0.06814234         0
#> 14 31.32814 404.9743 1.556757 17.23298 0.1321788 1.488008 0.09241032         0
#> 15 28.66190 488.3457 1.297004 18.55498 0.1457924 1.243863 0.06963659         0
#> 16 30.76248 430.7040 1.482285 20.03994 0.1514066 1.426110 0.08104951         0
#> 17 33.07881 386.4049 1.682785 18.93143 0.1418852 1.615226 0.09939170         0
#> 18 28.45865 396.4439 1.429300 22.81417 0.1599397 1.381792 0.07285994         0
#> 19 30.56612 320.3257 1.707829 25.59048 0.1510954 1.657283 0.08242193         0
#> 20 32.75800 710.5317 1.228926 12.70919 0.1143073 1.154949 0.07796155         0
#>    sigma_sq        phi Tau.case.var((Intercept))
#> 1  193.0211 0.16234151                  48.86524
#> 2  201.0535 0.31582730                 192.62026
#> 3  182.9855 0.12782527                 166.20751
#> 4  202.5733 0.29546807                 396.35612
#> 5  220.3550 0.28353655                  79.69679
#> 6  146.3287 0.25240642                 127.63626
#> 7  216.7502 0.23103791                 268.49647
#> 8  229.8352 0.33149981                 160.41248
#> 9  195.1260 0.23642080                 158.72254
#> 10 234.0931 0.27443542                 380.07289
#> 11 158.6136 0.23649652                  14.51754
#> 12 171.8683 0.15293774                  52.89112
#> 13 166.5216 0.04827382                 461.76440
#> 14 164.7055 0.19129750                 240.26880
#> 15 215.9710 0.27842134                 272.37474
#> 16 202.9570 0.28868163                 227.74701
#> 17 171.7404 0.23123288                 214.66448
#> 18 205.3795 0.30166733                 191.06443
#> 19 172.9988 0.18025147                 147.32685
#> 20 205.5358 0.22846910                 504.99583