Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object
a g_REML object

nsim
number of models to simulate

seed
seed value. See documentation for simulate

parallel
if TRUE, run in parallel using foreach backend.

...
additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment, 
                 random = ~ 1 | case, 
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

simulate(Laski_g, nsim = 20)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#>      p_beta  r_theta delta_AB       nu     kappa     g_AB     V_g_AB cnvg_warn
#> 1  32.72464 251.6672 2.062821 70.88927 0.2072594 2.040919 0.07366435         0
#> 2  28.38828 274.7385 1.712692 50.40786 0.1961763 1.687083 0.06850737         0
#> 3  28.99364 425.6753 1.405282 18.19551 0.1438616 1.346551 0.07854610         0
#> 4  34.16777 452.4916 1.606243 15.32867 0.1248090 1.526350 0.10585452         0
#> 5  30.72398 364.8483 1.608501 20.93221 0.1597317 1.550172 0.09085723         0
#> 6  35.63573 594.0703 1.462065 13.33012 0.1184608 1.378232 0.10103831         0
#> 7  30.45075 395.2216 1.531714 26.44116 0.1665013 1.487853 0.07425157         0
#> 8  27.41951 509.1274 1.215196 14.22515 0.1256717 1.149981 0.07207430         0
#> 9  31.21707 320.7304 1.743099 26.91280 0.1558330 1.694067 0.08320942         0
#> 10 30.24102 301.1983 1.742489 31.18932 0.1804460 1.700250 0.08325914         0
#> 11 32.75290 290.7377 1.920875 33.15858 0.1597261 1.877098 0.08312691         0
#> 12 31.01974 438.6907 1.481013 16.06524 0.1272811 1.410780 0.08930029         0
#> 13 32.33770 605.7977 1.313849 17.55890 0.1380781 1.256919 0.07160956         0
#> 14 34.07175 334.9136 1.861779 28.29810 0.1625976 1.811995 0.09022961         0
#> 15 27.08146 241.5567 1.742459 47.13056 0.1676642 1.714583 0.06125985         0
#> 16 34.00173 392.6490 1.715927 18.79273 0.1332317 1.646523 0.10070936         0
#> 17 35.00849 501.9168 1.562635 14.14186 0.1070095 1.478271 0.10448601         0
#> 18 33.61716 220.7008 2.262867 60.89543 0.1857719 2.234882 0.07747458         0
#> 19 28.08449 240.2979 1.811722 60.51374 0.1702859 1.789174 0.05677275         0
#> 20 28.36745 329.4470 1.562887 24.48766 0.1450813 1.514526 0.07333859         0
#>    sigma_sq       phi Tau.case.var((Intercept))
#> 1  216.2967 0.3180025                  35.37057
#> 2  209.7372 0.3229438                  65.00132
#> 3  184.8704 0.2713841                 240.80486
#> 4  164.0405 0.1913035                 288.45105
#> 5  180.2708 0.3415930                 184.57752
#> 6  183.8300 0.2315940                 410.24037
#> 7  223.6493 0.2952945                 171.57229
#> 8  171.9970 0.2555016                 337.13041
#> 9  179.0561 0.2013434                 141.67435
#> 10 189.1037 0.3469423                 112.09466
#> 11 179.1338 0.1659313                 111.60387
#> 12 166.7273 0.1855276                 271.96344
#> 13 253.0286 0.2370855                 352.76913
#> 14 193.8256 0.2386965                 141.08800
#> 15 171.0393 0.1372094                  70.51736
#> 16 170.7660 0.1546126                 221.88298
#> 17 162.3109 0.0582007                 339.60595
#> 18 173.7411 0.2099524                  46.95962
#> 19 184.3039 0.1045341                  55.99403
#> 20 172.1577 0.1457587                 157.28933