Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object
a g_REML object

nsim
number of models to simulate

seed
seed value. See documentation for simulate

parallel
if TRUE, run in parallel using foreach backend.

...
additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment,
                 random = ~ 1 | case,
                 correlation = corAR1(0, ~ time | case),
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))
simulate(Laski_g, nsim = 20)

#>      p_beta  r_theta delta_AB       nu      kappa      g_AB     V_g_AB
#> 1  30.54812 249.9648 1.932169 58.75086 0.19118356 1.9073977 0.06913722
#> 2  36.13837 758.4711 1.312197 11.56882 0.09988473 1.2252490 0.09173191
#> 3  33.96533 518.9435 1.490994 14.67739 0.13409753 1.4134854 0.09966479
#> 4  32.55051 346.4830 1.748706 31.74193 0.17636588 1.7070595 0.08122041
#> 5  29.62375 286.3869 1.750505 34.66827 0.20314510 1.7123606 0.08728429
#> 6  28.91328 501.0664 1.291665 21.63539 0.14878877 1.2463651 0.06298682
#> 7  30.35065 490.6353 1.370215 17.10379 0.11918631 1.3092395 0.07274961
#> 8  36.35639 492.1708 1.638788 19.64384 0.15009836 1.5754128 0.09492741
#> 9  32.76123 555.9484 1.389450 15.72428 0.13307773 1.3221070 0.08367088
#> 10 26.37008 431.9256 1.268840 22.19778 0.17980158 1.2254815 0.07097744
#> 11 31.66365 321.4467 1.766064 31.26661 0.15727455 1.7233600 0.07653129
#> 12 27.22606 696.5398 1.031601 11.59008 0.11051444 0.9633739 0.06296662
#> 13 27.23415 623.9693 1.090265 13.86572 0.13914357 1.0302097 0.06622269
#> 14 31.27589 512.5288 1.381499 19.30787 0.14880413 1.3271319 0.07471686
#> 15 32.92467 374.9493 1.700337 27.03083 0.16423328 1.6527194 0.08286686
#> 16 32.36672 828.0918 1.124759 11.89322 0.11719825 1.0523072 0.07233712
#> 17 37.83962 379.1135 1.943400 20.73622 0.13154037 1.8722523 0.11309255
#> 18 29.79431 205.3496 2.079153 52.92900 0.18816789 2.0495514 0.07729539
#> 19 32.12868 267.9425 1.962782 43.11387 0.17728953 1.9284388 0.07744636
#> 20 28.96534 704.2444 1.091483 13.56552 0.12107976 1.0300047 0.06254259
#>    cnvg_warn sigma_sq        phi Tau.case.var((Intercept))
#> 1          0 196.8853 0.25383174                  53.07953
#> 2          0 185.8280 0.15236817                 572.64313
#> 3          0 184.8819 0.31992454                 334.06160
#> 4          0 216.7303 0.30872654                 129.75266
#> 5          0 202.4030 0.49590568                  83.98395
#> 6          0 245.8525 0.22850738                 255.21399
#> 7          0 193.9526 0.06800873                 296.68265
#> 8          0 228.3243 0.28718982                 263.84651
#> 9          0 210.0331 0.25783099                 345.91532
#> 10         0 237.7962 0.50411669                 194.12947
#> 11         0 192.5088 0.16398471                 128.93792
#> 12         0 175.9565 0.28250661                 520.58334
#> 13         0 215.2530 0.42953494                 408.71638
#> 14         0 234.4258 0.28450853                 278.10299
#> 15         0 213.2485 0.26835404                 161.70079
#> 16         0 221.6980 0.33433561                 606.39383
#> 17         0 175.8655 0.09292968                 203.24798
#> 18         0 156.0929 0.25626649                  49.25666
#> 19         0 187.2517 0.22790236                  80.69078
#> 20         0 223.6948 0.24529110                 480.54969