Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object

a g_REML object

nsim

number of models to simulate

seed

seed value. See documentation for simulate

parallel

if TRUE, run in parallel using foreach backend.

...

additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment,
                 random = ~ 1 | case,
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)

suppressWarnings(
  Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))
)

if (requireNamespace("plyr", quietly = TRUE)) {
  simulate(Laski_g, nsim = 20)
}
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#>      p_beta  r_theta delta_AB       nu      kappa      g_AB     V_g_AB
#> 1  36.92705 550.9625 1.573199 13.38590 0.10042810 1.4833763 0.11008665
#> 2  33.30126 288.4143 1.960885 26.21524 0.14364379 1.9042456 0.09705181
#> 3  36.53942 368.7204 1.902887 20.10238 0.14671468 1.8309982 0.11654644
#> 4  34.66260 502.7761 1.545873 16.74755 0.13680736 1.4755955 0.09494202
#> 5  34.97715 424.5739 1.697492 19.23508 0.13822830 1.6304334 0.09837550
#> 6  31.38555 362.4098 1.648655 20.54999 0.15022596 1.5877441 0.09243908
#> 7  29.80919 822.6586 1.039299 11.67808 0.09790171 0.9710916 0.06052017
#> 8  31.54587 263.1399 1.944684 50.31393 0.19397279 1.9155514 0.07628039
#> 9  25.01192 378.3885 1.285814 24.13620 0.15574356 1.2454409 0.06041601
#> 10 28.39393 477.9840 1.298730 20.45662 0.14953215 1.2505254 0.06615724
#> 11 25.88852 391.8812 1.307766 24.49744 0.16911676 1.2673153 0.06551028
#> 12 30.16855 248.1894 1.914974 45.92998 0.18652703 1.8835324 0.07590753
#> 13 29.29550 466.2458 1.356730 17.28465 0.14637271 1.2969963 0.07843572
#> 14 26.50480 237.1355 1.721180 66.59904 0.22518992 1.7017239 0.07364750
#> 15 29.43893 559.0309 1.245100 12.14881 0.11739173 1.1666190 0.08378112
#> 16 32.13685 309.4862 1.826765 26.76214 0.15381570 1.7750874 0.08869339
#> 17 26.09684 521.8200 1.142425 12.63118 0.11073761 1.0732215 0.06876923
#> 18 37.13839 251.5264 2.341700 48.19936 0.18966895 2.3050725 0.09432819
#> 19 27.83898 608.0455 1.128977 14.58591 0.12422424 1.0699136 0.06278945
#> 20 32.02212 330.7741 1.760696 29.89081 0.16246808 1.7161450 0.08035802
#>    cnvg_warn sigma_sq        phi Tau.case.var((Intercept))
#> 1          0 165.4104 0.02436889                 385.55212
#> 2          0 155.8409 0.10911319                 132.57339
#> 3          0 172.2402 0.24470641                 196.48020
#> 4          0 201.8913 0.25338977                 300.88481
#> 5          0 189.3957 0.18933823                 235.17814
#> 6          0 172.9499 0.26539859                 189.45991
#> 7          0 203.6717 0.11735644                 618.98695
#> 8          0 199.6924 0.30813282                  63.44749
#> 9          0 200.0333 0.24174583                 178.35523
#> 10         0 227.1171 0.26142086                 250.86690
#> 11         0 215.4751 0.34894096                 176.40602
#> 12         0 180.2712 0.27834599                  67.91823
#> 13         0 196.5079 0.32666112                 269.73787
#> 14         0 213.4821 0.44584849                  23.65339
#> 15         0 153.9950 0.31170601                 405.03591
#> 16         0 171.6994 0.18675117                 137.78680
#> 17         0 148.4963 0.19314157                 373.32364
#> 18         0 186.6479 0.28860934                  64.87850
#> 19         0 209.9063 0.22072811                 398.13925
#> 20         0 196.0513 0.21927068                 134.72276