Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object

a g_REML object

nsim

number of models to simulate

seed

seed value. See documentation for simulate

parallel

if TRUE, run in parallel using foreach backend.

...

additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment,
                 random = ~ 1 | case,
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)

suppressWarnings(
  Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))
)

if (requireNamespace("plyr", quietly = TRUE)) {
  simulate(Laski_g, nsim = 20)
}
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#>      p_beta  r_theta delta_AB       nu      kappa      g_AB     V_g_AB
#> 1  34.85252 420.7645 1.699083 16.80114 0.13098111 1.6220906 0.10873876
#> 2  32.25392 463.5099 1.498142 17.71610 0.13700200 1.4338116 0.08624937
#> 3  32.52576 808.9017 1.143614 13.14401 0.12822512 1.0770943 0.07085578
#> 4  33.67280 495.2329 1.513124 15.14662 0.11836157 1.4369425 0.09516262
#> 5  29.82821 558.7099 1.261926 16.51690 0.15373818 1.2037440 0.07558003
#> 6  32.46635 485.7356 1.473104 16.37889 0.13106004 1.4046045 0.08806231
#> 7  33.21060 216.0646 2.259357 92.38732 0.22172945 2.2409661 0.07733106
#> 8  30.94189 208.4988 2.142866 42.79505 0.18138208 2.1050909 0.08810652
#> 9  32.69513 397.2787 1.640346 16.23382 0.10700202 1.5633767 0.09982734
#> 10 31.19764 319.9424 1.744158 40.77681 0.20199930 1.7118799 0.07947811
#> 11 28.58694 469.1624 1.319794 17.42764 0.11639208 1.2621695 0.06680295
#> 12 33.00268 661.7039 1.282972 15.55595 0.13958383 1.2201062 0.07653987
#> 13 29.81749 562.6446 1.257054 15.36012 0.13928704 1.1946594 0.07495037
#> 14 34.37282 201.3250 2.422514 80.00763 0.19579111 2.3997336 0.07566748
#> 15 33.61133 426.2814 1.627937 34.20109 0.16639415 1.5919747 0.06791849
#> 16 31.83305 352.0496 1.696588 22.39086 0.14727129 1.6391174 0.08926374
#> 17 24.60749 581.9608 1.020048 16.20706 0.14105176 0.9721044 0.05468900
#> 18 34.44836 738.7048 1.267457 10.53185 0.09274338 1.1750032 0.09320189
#> 19 30.92457 329.9811 1.702390 32.36778 0.16797907 1.6626365 0.07474562
#> 20 31.48500 321.0135 1.757284 28.66828 0.15825285 1.7109072 0.08114493
#>    cnvg_warn sigma_sq         phi Tau.case.var((Intercept))
#> 1          0 167.2869  0.19456060                253.477654
#> 2          0 194.4494  0.22170958                269.060478
#> 3          0 253.5701  0.35539679                555.331654
#> 4          0 175.3131  0.13275092                319.919811
#> 5          0 234.2704  0.43570913                324.439404
#> 6          0 189.1323  0.21091542                296.603300
#> 7          0 208.4624  0.34200732                  7.602217
#> 8          0 146.5231  0.25935087                 61.975616
#> 9          0 147.5483 -0.02888760                249.730444
#> 10         0 234.3432  0.42504149                 85.599230
#> 11         0 187.5970  0.03001007                281.565454
#> 12         0 252.0872  0.33194834                409.616709
#> 13         0 211.9085  0.33856758                350.736048
#> 14         0 173.8906  0.22136451                 27.434412
#> 15         0 268.9160  0.20860464                157.365433
#> 16         0 175.7033  0.20026813                176.346260
#> 17         0 230.1690  0.31712476                351.791723
#> 18         0 150.8925  0.17611992                587.812240
#> 19         0 204.2871  0.23727516                125.694005
#> 20         0 185.3621  0.19932463                135.651479