Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object

a g_REML object

nsim

number of models to simulate

seed

seed value. See documentation for simulate

parallel

if TRUE, run in parallel using foreach backend.

...

additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment,
                 random = ~ 1 | case,
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
simulate(Laski_g, nsim = 20)
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#>      p_beta   r_theta  delta_AB        nu      kappa      g_AB     V_g_AB
#> 1  29.01191  405.8034 1.4401854 21.466376 0.13744451 1.3892748 0.06986713
#> 2  33.79731  416.2257 1.6566003 21.203697 0.15449760 1.5973052 0.09216342
#> 3  31.97042  330.5560 1.7584333 40.254930 0.16271133 1.7254668 0.06610902
#> 4  30.68799  277.4690 1.8423039 38.605237 0.18162131 1.8062794 0.07845088
#> 5  29.80824  454.3634 1.3984104 20.756673 0.15632729 1.3472657 0.07440532
#> 6  31.16885  331.6808 1.7114357 38.145690 0.19459397 1.6775644 0.07771763
#> 7  31.39414  807.0174 1.1051140 11.060881 0.11177834 1.0284472 0.07374097
#> 8  27.71286  485.5607 1.2576495 14.764514 0.12474176 1.1926637 0.07341934
#> 9  30.34901  579.9652 1.2602120 13.117388 0.10568786 1.1867581 0.07698337
#> 10 33.64573  632.3078 1.3380294 13.965543 0.12621131 1.2648626 0.08539546
#> 11 29.13999  311.2903 1.6516065 33.755158 0.17191327 1.6146359 0.07154379
#> 12 27.57681  554.9629 1.1706095 15.692275 0.13110091 1.1137554 0.06428197
#> 13 34.97101  302.2782 2.0114295 53.997413 0.21771745 1.9833617 0.08595366
#> 14 28.71291  500.3786 1.2835945 13.708442 0.12826697 1.2120635 0.08175785
#> 15 29.16659  469.1043 1.3466381 13.646151 0.13825934 1.2712449 0.09139342
#> 16 31.34805  419.6918 1.5301899 17.885523 0.15616238 1.4651143 0.09423244
#> 17 31.74577 1021.5876 0.9932261  9.196237 0.07314221 0.9099599 0.06588406
#> 18 26.44735  304.3027 1.5161049 24.933768 0.14653801 1.4700391 0.06980065
#> 19 33.19017  278.7289 1.9880094 28.472657 0.17604833 1.9351792 0.10328300
#> 20 28.35154  349.4364 1.5166752 32.211048 0.20113689 1.4810848 0.07787441
#>    cnvg_warn sigma_sq       phi Tau.case.var((Intercept))
#> 1          0 194.1842 0.1315555                 211.61917
#> 2          0 204.3711 0.2882049                 211.85459
#> 3          0 220.2674 0.1372726                 110.28856
#> 4          0 188.5552 0.2887218                  88.91377
#> 5          0 221.3688 0.3151361                 232.99461
#> 6          0 232.3784 0.3896129                  99.30238
#> 7          0 190.9515 0.3616690                 616.06588
#> 8          0 169.8347 0.2171133                 315.72603
#> 9          0 171.3492 0.1014174                 408.61593
#> 10         0 209.9278 0.2764782                 422.38000
#> 11         0 197.4410 0.2547755                 113.84936
#> 12         0 208.3205 0.2392839                 346.64241
#> 13         0 252.0648 0.4551534                  50.21337
#> 14         0 163.9369 0.3152621                 336.44171
#> 15         0 158.9698 0.4360440                 310.13445
#> 16         0 186.8346 0.4014043                 232.85725
#> 17         0 142.1264 0.1125292                 879.46116
#> 18         0 160.9268 0.1514465                 143.37585
#> 19         0 166.8440 0.3450568                 111.88491
#> 20         0 240.3754 0.5112691                 109.06100