Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object
a g_REML object

nsim
number of models to simulate

seed
seed value. See documentation for simulate

parallel
if TRUE, run in parallel using foreach backend.

...
additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment, 
                 random = ~ 1 | case, 
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

simulate(Laski_g, nsim = 20)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#>      p_beta  r_theta delta_AB       nu      kappa     g_AB     V_g_AB cnvg_warn
#> 1  34.56977 433.1318 1.661064 16.10526 0.13125328 1.582491 0.10882317         0
#> 2  27.27350 487.9476 1.234679 21.62984 0.16725166 1.191367 0.06550794         0
#> 3  32.77212 344.7276 1.765089 21.35910 0.14484405 1.702376 0.09775926         0
#> 4  32.11105 665.6411 1.244613 11.60815 0.09737836 1.162429 0.08278573         0
#> 5  33.84568 516.8923 1.488686 14.11826 0.11738191 1.408178 0.09851242         0
#> 6  35.97372 330.8006 1.977891 21.34787 0.13733181 1.907579 0.11512498         0
#> 7  31.34387 396.4511 1.574192 19.02131 0.13589501 1.511296 0.08750791         0
#> 8  32.97996 387.0408 1.676377 20.51039 0.15380584 1.614321 0.09605239         0
#> 9  30.31144 267.7734 1.852349 37.46489 0.16371167 1.815018 0.07410510         0
#> 10 32.35843 351.7031 1.725438 22.08226 0.14563016 1.666164 0.09209184         0
#> 11 30.09251 265.5700 1.846583 43.10591 0.17829877 1.814267 0.07257467         0
#> 12 33.46798 380.7238 1.715238 22.24116 0.14975982 1.656741 0.09198960         0
#> 13 35.93503 357.0379 1.901782 23.82430 0.14610534 1.841278 0.10079144         0
#> 14 29.37384 642.7491 1.158617 12.99637 0.10935607 1.090444 0.06827004         0
#> 15 31.19298 320.1912 1.743220 29.14050 0.16616885 1.697966 0.08194899         0
#> 16 31.73187 535.0971 1.371764 17.40525 0.12782006 1.311793 0.07400566         0
#> 17 32.67304 261.8604 2.019085 32.13318 0.16523081 1.971589 0.09303589         0
#> 18 33.53036 440.7879 1.597068 15.65627 0.13165416 1.519320 0.10465646         0
#> 19 28.27031 433.8574 1.357241 17.39385 0.15016246 1.297865 0.07925892         0
#> 20 31.67459 556.1604 1.343108 13.03076 0.12291618 1.264292 0.09051858         0
#>    sigma_sq       phi Tau.case.var((Intercept))
#> 1  166.3362 0.2235634                 266.79565
#> 2  251.1318 0.3932824                 236.81576
#> 3  166.5710 0.1995280                 178.15660
#> 4  163.0747 0.1172072                 502.56641
#> 5  170.0350 0.1726561                 346.85735
#> 6  157.7227 0.1329131                 173.07792
#> 7  174.7300 0.1733622                 221.72105
#> 8  186.0555 0.2984907                 200.98532
#> 9  174.0942 0.1630590                  93.67922
#> 10 173.5971 0.1918758                 178.10601
#> 11 185.9357 0.2352496                  79.63431
#> 12 190.1894 0.2247775                 190.53446
#> 13 184.1276 0.1649972                 172.91032
#> 14 189.0678 0.1513422                 453.68134
#> 15 189.0701 0.2571840                 131.12108
#> 16 217.5815 0.1434873                 317.51565
#> 17 160.8037 0.2181593                 101.05661
#> 18 165.3405 0.2464407                 275.44744
#> 19 185.7785 0.3599275                 248.07894
#> 20 169.7134 0.3028106                 386.44697