Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object
a g_REML object

nsim
number of models to simulate

seed
seed value. See documentation for simulate

parallel
if TRUE, run in parallel using foreach backend.

...
additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment, 
                 random = ~ 1 | case, 
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

simulate(Laski_g, nsim = 20)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#>      p_beta  r_theta  delta_AB        nu      kappa      g_AB     V_g_AB
#> 1  29.59943 252.9343 1.8611419 38.902154 0.16139292 1.8250286 0.07197963
#> 2  31.75112 321.1381 1.7717936 25.328729 0.17255514 1.7188066 0.09471723
#> 3  23.87132 615.2533 0.9623864 11.420825 0.10471116 0.8977726 0.05587697
#> 4  30.66164 190.7832 2.2198583 94.601470 0.17788067 2.2022127 0.05810406
#> 5  27.64738 592.7588 1.1355728 13.100017 0.13041367 1.0692943 0.07089224
#> 6  31.04538 483.2300 1.4122773 22.307897 0.15984735 1.3642579 0.07282140
#> 7  35.94807 264.5865 2.2099971 72.491933 0.20911747 2.1870534 0.07814716
#> 8  30.68305 395.0691 1.5436966 19.512358 0.14720303 1.4835911 0.08641353
#> 9  28.31450 202.0986 1.9917152 91.870369 0.18143455 1.9754111 0.05490242
#> 10 36.23923 374.8809 1.8716832 21.111543 0.13689342 1.8043937 0.10600527
#> 11 31.84145 294.0821 1.8567718 26.957577 0.16513424 1.8046300 0.09400984
#> 12 31.28678 274.4254 1.8886381 36.761082 0.16626423 1.8498422 0.07778202
#> 13 28.39364 456.9951 1.3282056 18.408984 0.15208686 1.2733482 0.07431915
#> 14 32.43393 980.4211 1.0358412  9.990774 0.08831690 0.9560857 0.06787688
#> 15 29.23589 317.0488 1.6419246 62.028930 0.19896192 1.6219916 0.06196546
#> 16 34.48125 410.4826 1.7019064 20.645180 0.13425905 1.6393215 0.09195809
#> 17 30.10237 606.7208 1.2220985 13.656385 0.12753240 1.1537301 0.07576641
#> 18 36.18469 813.2261 1.2688761 11.028398 0.09411882 1.1805831 0.08942806
#> 19 28.42476 338.2436 1.5455464 34.190018 0.19906565 1.5113932 0.07613194
#> 20 29.77088 237.2433 1.9328345 63.220548 0.17661323 1.9098138 0.06141696
#>    cnvg_warn sigma_sq        phi Tau.case.var((Intercept))
#> 1          0 166.1558 0.13579063                  86.77855
#> 2          0 181.0786 0.36376073                 140.05950
#> 3          0 149.4776 0.22765302                 465.77575
#> 4          0 167.5961 0.07493913                  23.18707
#> 5          0 186.7017 0.38486758                 406.05713
#> 6          0 247.3082 0.31119694                 235.92184
#> 7          0 229.8563 0.32471279                  34.73021
#> 8          0 181.2560 0.26425820                 213.81317
#> 9          0 176.8975 0.10364866                  25.20112
#> 10         0 177.3732 0.13378199                 197.50771
#> 11         0 167.3767 0.27671575                 126.70543
#> 12         0 177.9021 0.18746008                  96.52335
#> 13         0 204.4329 0.34250952                 252.56224
#> 14         0 177.0905 0.19439697                 803.33059
#> 15         0 257.8021 0.29372481                  59.24668
#> 16         0 190.6625 0.11971622                 219.82005
#> 17         0 197.5577 0.31056431                 409.16312
#> 18         0 181.5963 0.13404325                 631.62977
#> 19         0 233.4918 0.46541639                 104.75182
#> 20         0 185.9074 0.14018690                  51.33589