Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object
a g_REML object

nsim
number of models to simulate

seed
seed value. See documentation for simulate

parallel
if TRUE, run in parallel using foreach backend.

...
additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment, 
                 random = ~ 1 | case, 
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

simulate(Laski_g, nsim = 20)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#>      p_beta  r_theta delta_AB       nu     kappa     g_AB     V_g_AB cnvg_warn
#> 1  29.05430 374.6118 1.501135 25.42919 0.1579545 1.456421 0.07145202         0
#> 2  32.72611 329.5575 1.802722 32.17772 0.1917127 1.760375 0.08933178         0
#> 3  32.50157 646.5711 1.278192 16.43941 0.1455477 1.218978 0.07463572         0
#> 4  25.48232 399.2563 1.275302 24.74793 0.1583193 1.236259 0.05974744         0
#> 5  34.17368 495.2279 1.535639 17.34724 0.1476546 1.468276 0.09437174         0
#> 6  27.29940 381.8472 1.397038 34.76767 0.1863356 1.366683 0.06407401         0
#> 7  29.75328 361.9708 1.563861 33.40898 0.1784263 1.528489 0.06996996         0
#> 8  33.08233 269.7112 2.014404 67.76795 0.2213226 1.992028 0.07969559         0
#> 9  32.68500 561.0466 1.379904 15.60871 0.1178868 1.312521 0.07933626         0
#> 10 29.97619 315.2037 1.688421 23.97010 0.1476967 1.635035 0.08415155         0
#> 11 36.43764 388.3546 1.848996 24.79127 0.1593690 1.792490 0.09756238         0
#> 12 33.48844 458.1611 1.564537 15.00711 0.1168295 1.485022 0.10122404         0
#> 13 30.79900 564.5240 1.296270 13.74412 0.1160587 1.224223 0.07973903         0
#> 14 27.29102 429.4517 1.316930 15.51215 0.1196481 1.252214 0.07437812         0
#> 15 30.72063 517.3855 1.350588 18.88035 0.1542243 1.296218 0.07531382         0
#> 16 31.45034 231.7466 2.065945 75.89982 0.1924450 2.045463 0.06573666         0
#> 17 30.12686 653.7479 1.178280 15.07712 0.1459218 1.118680 0.07124891         0
#> 18 25.60882 613.7508 1.033698 14.06047 0.1205213 0.977561 0.05589709         0
#> 19 32.73463 366.2915 1.710385 19.57726 0.1470720 1.644013 0.10069033         0
#> 20 26.14485 223.2414 1.749842 75.23295 0.1874647 1.732339 0.05597706         0
#>    sigma_sq       phi Tau.case.var((Intercept))
#> 1  204.2622 0.2397101                 170.34966
#> 2  216.8261 0.4270130                 112.73139
#> 3  262.0422 0.3525705                 384.52886
#> 4  214.9273 0.2533652                 184.32904
#> 5  210.0395 0.3368926                 285.18843
#> 6  253.3408 0.3561017                 128.50643
#> 7  231.9851 0.3079772                 129.98569
#> 8  239.8774 0.4167587                  29.83381
#> 9  204.1987 0.1083081                 356.84787
#> 10 163.5369 0.1762334                 151.66679
#> 11 209.7075 0.2614589                 178.64710
#> 12 160.2869 0.1229241                 297.87419
#> 13 179.8494 0.1793718                 384.67458
#> 14 155.8775 0.1304364                 273.57417
#> 15 236.7036 0.3477395                 280.68183
#> 16 196.0699 0.2103477                  35.67668
#> 17 248.6219 0.4267905                 405.12602
#> 18 202.3765 0.2096235                 411.37439
#> 19 168.3645 0.2613381                 197.92705
#> 20 186.7217 0.1798441                  36.51967