Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object

a g_REML object

nsim

number of models to simulate

seed

seed value. See documentation for simulate

parallel

if TRUE, run in parallel using foreach backend.

...

additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment,
                 random = ~ 1 | case,
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)
Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
simulate(Laski_g, nsim = 20)
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#>      p_beta  r_theta  delta_AB         nu      kappa      g_AB     V_g_AB
#> 1  33.85292 579.5243 1.4062430  15.951415 0.12284056 1.3390718 0.08151503
#> 2  29.82900 361.4386 1.5689946  32.262123 0.18421948 1.5322352 0.07375945
#> 3  28.72927 341.5420 1.5545424  32.728490 0.18658347 1.5186446 0.07335660
#> 4  27.02060 384.1542 1.3786127  19.185752 0.15251433 1.3240091 0.07600288
#> 5  35.95834 603.8118 1.4633520  14.221777 0.12612107 1.3847998 0.09724738
#> 6  28.83085 487.0466 1.3063882  22.264456 0.15323700 1.2618815 0.06405233
#> 7  29.91239 281.8988 1.7815768  33.036977 0.17511594 1.7408233 0.08055679
#> 8  26.49156 461.5370 1.2331171  17.280271 0.13919656 1.1788116 0.06653989
#> 9  36.81976 270.0554 2.2405493  32.010057 0.17321380 2.1876397 0.11120870
#> 10 27.61815 584.6113 1.1422495  12.285724 0.11140210 1.0710708 0.07064573
#> 11 24.90773 610.9296 1.0077169  15.929658 0.14139089 0.9595151 0.05460221
#> 12 28.58154 396.2326 1.4358547  18.853184 0.15288982 1.3779673 0.08159807
#> 13 30.73331 968.0663 0.9877721  10.391363 0.09812908 0.9147219 0.06206152
#> 14 27.02075 219.7497 1.8227758 104.906356 0.16945580 1.8097132 0.04482555
#> 15 31.57588 452.3940 1.4845577  15.738379 0.12072733 1.4126704 0.08959267
#> 16 24.40630 361.4496 1.2837432  19.889148 0.12038578 1.2347183 0.05836098
#> 17 31.20523 457.9758 1.4581632  15.683272 0.12665570 1.3873019 0.08876162
#> 18 23.33044 408.0808 1.1549147  23.075565 0.16761570 1.1169667 0.05887372
#> 19 28.57936 841.5787 0.9851564   9.722892 0.08750884 0.9071583 0.06372990
#> 20 31.36576 294.2899 1.8283870  28.367239 0.17819445 1.7796166 0.09324223
#>    cnvg_warn sigma_sq          phi Tau.case.var((Intercept))
#> 1          0 217.0337  0.145330377                 362.49054
#> 2          0 232.2105  0.364990871                 129.22811
#> 3          0 222.0977  0.378738100                 119.44423
#> 4          0 176.6693  0.322353075                 207.48489
#> 5          0 204.1661  0.260493611                 399.64567
#> 6          0 245.1953  0.254515058                 241.85131
#> 7          0 178.5439  0.286024623                 103.35497
#> 8          0 190.9398  0.257280550                 270.59719
#> 9          0 168.2579  0.281393105                 101.79748
#> 10         0 160.6635  0.228233169                 423.94775
#> 11         0 238.5705  0.332916482                 372.35912
#> 12         0 180.4072  0.335927292                 215.82536
#> 13         0 195.6576  0.270274474                 772.40873
#> 14         0 197.5099 -0.001173557                  22.23978
#> 15         0 166.7399  0.132277556                 285.65413
#> 16         0 160.5540  0.006370438                 200.89558
#> 17         0 170.1787  0.194782074                 287.79706
#> 18         0 217.1059  0.363514380                 190.97483
#> 19         0 142.2892  0.230208632                 699.28942
#> 20         0 176.9649  0.364149262                 117.32505