Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object

a g_REML object

nsim

number of models to simulate

seed

seed value. See documentation for simulate

parallel

if TRUE, run in parallel using foreach backend.

...

additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment,
                 random = ~ 1 | case,
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)

suppressWarnings(
  Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))
)

if (requireNamespace("plyr", quietly = TRUE)) {
  simulate(Laski_g, nsim = 20)
}
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#>      p_beta  r_theta  delta_AB       nu      kappa      g_AB     V_g_AB
#> 1  29.50929 470.7152 1.3601274 17.19748 0.14932411 1.2999358 0.07992869
#> 2  31.79526 451.1653 1.4969061 17.44335 0.14255064 1.4316088 0.08885950
#> 3  27.76994 799.0042 0.9824272 12.34611 0.11789580 0.9215134 0.05698335
#> 4  27.86171 619.1825 1.1196918 13.09292 0.11649541 1.0543041 0.06585226
#> 5  33.18606 217.4049 2.2507173 60.81823 0.18999070 2.2228473 0.07867122
#> 6  35.05599 417.8909 1.7148686 22.04607 0.15637684 1.6558601 0.09468093
#> 7  35.21309 486.4214 1.5966065 15.53530 0.11060441 1.5182662 0.10004103
#> 8  32.48700 605.4325 1.3203126 14.39902 0.12449794 1.2503265 0.08094357
#> 9  32.14776 300.3849 1.8548620 28.01166 0.16346957 1.8047516 0.09072066
#> 10 30.08704 340.5537 1.6303716 18.82617 0.16890780 1.5645465 0.10366114
#> 11 35.47632 340.8550 1.9215591 23.64128 0.17385218 1.8599477 0.11218294
#> 12 32.74666 660.3898 1.2742857 11.86662 0.09155288 1.1920144 0.08330113
#> 13 29.08448 558.2916 1.2309225 17.26836 0.16645848 1.1766757 0.07502080
#> 14 35.27650 437.8444 1.6858758 21.09471 0.14390133 1.6252174 0.09171294
#> 15 29.47432 313.4883 1.6646891 17.57922 0.14151833 1.5926422 0.10392936
#> 16 30.93976 431.3880 1.4896452 17.86297 0.13287935 1.4262128 0.08376319
#> 17 29.45219 315.1905 1.6589412 39.57794 0.19592200 1.6273045 0.07447855
#> 18 32.13800 400.8588 1.6051781 18.51603 0.14231708 1.5392698 0.09414232
#> 19 30.70447 530.2030 1.3334617 12.29088 0.11319025 1.2504033 0.09193411
#> 20 32.86158 346.2082 1.7661181 23.00686 0.15035207 1.7079120 0.09376356
#>    cnvg_warn sigma_sq        phi Tau.case.var((Intercept))
#> 1          0 199.4532 0.35922595                 271.26197
#> 2          0 189.4820 0.28384595                 261.68336
#> 3          0 225.0478 0.29988558                 573.95640
#> 4          0 186.7340 0.22563784                 432.44856
#> 5          0 172.4611 0.23853025                  44.94381
#> 6          0 210.7037 0.28633030                 207.18721
#> 7          0 174.2286 0.03501728                 312.19270
#> 8          0 206.4633 0.23344302                 398.96925
#> 9          0 173.3522 0.24927776                 127.03267
#> 10         0 165.4800 0.50270727                 175.07372
#> 11         0 187.1870 0.40767243                 153.66805
#> 12         0 165.6267 0.02097394                 494.76308
#> 13         0 259.3728 0.55599243                 298.91885
#> 14         0 209.5935 0.19676957                 228.25093
#> 15         0 132.0811 0.26926321                 181.40722
#> 16         0 180.6219 0.17788714                 250.76613
#> 17         0 224.2193 0.38803896                  90.97119
#> 18         0 175.6143 0.24714649                 225.24444
#> 19         0 146.4754 0.24891914                 383.72756
#> 20         0 176.4794 0.21548558                 169.72883