Simulate data from a fitted g_REML object

Simulates data from the linear mixed effects model used to estimate the specified standardized mean difference effect size. Suitable for parametric bootstrapping.

# S3 method for g_REML
simulate(object, nsim = 1, seed = NULL, parallel = FALSE, ...)

Arguments

object

a g_REML object

nsim

number of models to simulate

seed

seed value. See documentation for simulate

parallel

if TRUE, run in parallel using foreach backend.

...

additional optional arguments

Value

A matrix with one row per simulation, with columns corresponding to the output of g_REML.

Examples

data(Laski)
Laski_RML <- lme(fixed = outcome ~ treatment,
                 random = ~ 1 | case,
                 correlation = corAR1(0, ~ time | case), 
                 data = Laski)

suppressWarnings(
  Laski_g <- g_REML(Laski_RML, p_const = c(0,1), r_const = c(1,0,1))
)

if (requireNamespace("plyr", quietly = TRUE)) {
  simulate(Laski_g, nsim = 20)
}
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#>      p_beta  r_theta delta_AB       nu     kappa     g_AB     V_g_AB cnvg_warn
#> 1  31.70271 482.9631 1.442578 14.79481 0.1199761 1.368192 0.09010801         0
#> 2  29.63610 272.0306 1.796852 66.33302 0.1669889 1.776458 0.05277112         0
#> 3  36.22226 446.7081 1.713814 16.21485 0.1294223 1.633302 0.11349554         0
#> 4  30.87111 399.5002 1.544521 20.31054 0.1497153 1.486776 0.08457066         0
#> 5  32.27181 438.9542 1.540330 22.47460 0.1565690 1.488349 0.08016241         0
#> 6  23.72748 294.0443 1.383711 54.39268 0.2103873 1.364543 0.06258242         0
#> 7  33.15235 334.4539 1.812785 32.20541 0.1538300 1.770238 0.07655698         0
#> 8  30.84933 351.1124 1.646352 26.97231 0.1725109 1.600145 0.08237014         0
#> 9  28.39293 511.8894 1.254938 13.78565 0.1291492 1.185402 0.07876492         0
#> 10 23.53506 513.3855 1.038708 14.76676 0.1278605 0.985044 0.05603982         0
#> 11 30.85774 437.7552 1.474852 30.68734 0.1821648 1.438510 0.07029164         0
#> 12 23.72327 218.3936 1.605294 44.75564 0.1604181 1.578242 0.05541594         0
#> 13 27.33903 609.3973 1.107472 14.61847 0.1202279 1.049665 0.05990021         0
#> 14 32.80540 355.1680 1.740718 36.25674 0.1759530 1.704460 0.07427005         0
#> 15 25.24302 408.3805 1.249134 22.37859 0.1716008 1.206797 0.06654061         0
#> 16 34.18396 402.6573 1.703549 18.49556 0.1297503 1.633523 0.09996934         0
#> 17 30.22349 339.9839 1.639138 31.26420 0.1637138 1.599499 0.07151919         0
#> 18 29.80228 399.4335 1.491171 19.49424 0.1406065 1.433056 0.08023001         0
#> 19 36.04202 400.8000 1.800302 17.45041 0.1391912 1.721802 0.11813755         0
#> 20 32.16713 680.6913 1.232927 12.28906 0.1082661 1.156119 0.07941220         0
#>    sigma_sq        phi Tau.case.var((Intercept))
#> 1  167.5857 0.16579471                 315.37741
#> 2  213.3361 0.06403166                  58.69447
#> 3  171.8778 0.20101816                 274.83029
#> 4  189.0673 0.26661059                 210.43288
#> 5  223.7833 0.27922158                 215.17095
#> 6  238.7621 0.40165659                  55.28216
#> 7  201.7625 0.12788080                 132.69135
#> 8  203.4918 0.33709414                 147.62065
#> 9  169.1290 0.31995686                 342.76040
#> 10 180.8973 0.24954456                 332.48817
#> 11 274.5051 0.36660187                 163.25019
#> 12 150.5031 0.09518984                  67.89057
#> 13 209.0436 0.17703248                 400.35373
#> 14 233.2595 0.26432329                 121.90850
#> 15 216.9956 0.41556854                 191.38495
#> 16 172.0777 0.12985535                 230.57957
#> 17 205.8606 0.21477737                 134.12332
#> 18 180.6259 0.20449188                 218.80762
#> 19 167.0412 0.25136716                 233.75884
#> 20 185.7138 0.19090202                 494.97745