Simulate Model MB2 from Pustejovsky, Hedges, & Shadish (2014)

Simulates data from a linear mixed effects model, then calculates REML effect size estimator as described in Pustejovsky, Hedges, & Shadish (2014).

simulate_MB2(
  iterations,
  beta,
  rho,
  phi,
  tau1_ratio,
  tau_corr,
  design,
  m,
  n,
  MB = TRUE
)

Arguments

iterations
number of independent iterations of the simulation

beta
vector of fixed effect parameters

rho
intra-class correlation parameter

phi
autocorrelation parameter

tau1_ratio
ratio of treatment effect variance to intercept variance

tau_corr
correlation between case-specific treatment effects and intercepts

design
design matrix. If not specified, it will be calculated based on m, n, and MB.

m
number of cases. Not used if design is specified.

n
number of measurement occasions. Not used if design is specified.

MB
If true, a multiple baseline design will be used; otherwise, an AB design will be used. Not used if design is specified.

Value

A matrix reporting the mean and variance of the effect size estimates and various associated statistics.

References

Pustejovsky, J. E., Hedges, L. V., & Shadish, W. R. (2014). Design-comparable effect sizes in multiple baseline designs: A general modeling framework. Journal of Educational and Behavioral Statistics, 39(4), 211-227. doi:10.3102/1076998614547577

Examples

set.seed(8)
simulate_MB2(iterations = 5, beta = c(0,1,0,0), rho = 0.4, phi = 0.5, 
             tau1_ratio = 0.5, tau_corr = -0.4, design = design_matrix(m=3, n=8))

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: nlminb problem, convergence error code = 1
#>   message = iteration limit reached without convergence (10)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: nlminb problem, convergence error code = 1
#>   message = iteration limit reached without convergence (10)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#>                                      mean        var
#> p_beta                          0.9388914 0.16476996
#> r_theta                         1.6971957 1.32455481
#> delta_AB                        0.7970308 0.13807232
#> nu                              3.4958949 0.73276819
#> kappa                           0.4161736 0.01666201
#> g_AB                            0.6111397 0.10416382
#> V_g_AB                          0.5168009 0.10671954
#> cnvg_warn                       0.4000000 0.30000000
#> sigma_sq                        0.4672461 0.01517582
#> phi                             0.1993965 0.04850646
#> Tau.id.var(constant)            1.2299495 1.35869787
#> Tau.id.cov(treatment,constant) -0.8827947 1.18307730
#> Tau.id.var(treatment)           0.8139623 1.06938258
#> RML_coverage1                   1.0000000 0.00000000
#> RML_coverage2                   1.0000000 0.00000000

             
set.seed(8)
simulate_MB2(iterations = 5, beta = c(0,1,0,0), rho = 0.4, phi = 0.5, 
             tau1_ratio = 0.5, tau_corr = -0.4, m = 3, n = 8, MB = FALSE)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: nlminb problem, convergence error code = 1
#>   message = iteration limit reached without convergence (10)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: nlminb problem, convergence error code = 1
#>   message = iteration limit reached without convergence (10)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#>                                      mean        var
#> p_beta                          0.9388914 0.16476996
#> r_theta                         1.6971957 1.32455481
#> delta_AB                        0.7970308 0.13807232
#> nu                              3.4958949 0.73276819
#> kappa                           0.4161736 0.01666201
#> g_AB                            0.6111397 0.10416382
#> V_g_AB                          0.5168009 0.10671954
#> cnvg_warn                       0.4000000 0.30000000
#> sigma_sq                        0.4672461 0.01517582
#> phi                             0.1993965 0.04850646
#> Tau.id.var(constant)            1.2299495 1.35869787
#> Tau.id.cov(treatment,constant) -0.8827947 1.18307730
#> Tau.id.var(treatment)           0.8139623 1.06938258
#> RML_coverage1                   1.0000000 0.00000000
#> RML_coverage2                   1.0000000 0.00000000