Simulate Model MB4 from Pustejovsky, Hedges, & Shadish (2014)

Simulates data from a linear mixed effects model, then calculates REML effect size estimator as described in Pustejovsky, Hedges, & Shadish (2014).

simulate_MB4(
  iterations,
  beta,
  rho,
  phi,
  tau2_ratio,
  tau_corr,
  p_const,
  r_const,
  design,
  m,
  n,
  MB = TRUE
)

Arguments

iterations
number of independent iterations of the simulation

beta
vector of fixed effect parameters

rho
intra-class correlation parameter

phi
autocorrelation parameter

tau2_ratio
ratio of trend variance to intercept variance

tau_corr
correlation between case-specific trends and intercepts

p_const
vector of constants for calculating numerator of effect size

r_const
vector of constants for calculating denominator of effect size

design
design matrix. If not specified, it will be calculated based on m, n, and MB.

m
number of cases. Not used if design is specified.

n
number of measurement occasions. Not used if design is specified.

MB
If true, a multiple baseline design will be used; otherwise, an AB design will be used. Not used if design is specified.

Value

A matrix reporting the mean and variance of the effect size estimates and various associated statistics.

References

Pustejovsky, J. E., Hedges, L. V., & Shadish, W. R. (2014). Design-comparable effect sizes in multiple baseline designs: A general modeling framework. Journal of Educational and Behavioral Statistics, 39(4), 211-227. doi:10.3102/1076998614547577

Examples

simulate_MB4(iterations = 10, beta = c(0,1,0,0), rho = 0.8, phi = 0.5, 
             tau2_ratio = 0.5, tau_corr = 0, 
             p_const = c(0,1,0,7), r_const = c(1,0,1,0,0), 
             design = design_matrix(3, 16, treat_times=c(5,9,13), center = 12))

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: nlminb problem, convergence error code = 1
#>   message = iteration limit reached without convergence (10)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: nlminb problem, convergence error code = 1
#>   message = iteration limit reached without convergence (10)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: nlminb problem, convergence error code = 1
#>   message = iteration limit reached without convergence (10)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: nlminb problem, convergence error code = 1
#>   message = iteration limit reached without convergence (10)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#>                                  mean        var
#> p_beta                     1.05876376 0.47212223
#> r_theta                    0.93656835 0.47406203
#> delta_AB                   1.32493353 1.52193443
#> nu                         3.83709611 6.89153502
#> kappa                      0.55666166 0.03771016
#> g_AB                       1.01480354 1.26207011
#> V_g_AB                     1.60213726 0.61635782
#> cnvg_warn                  0.40000000 0.26666667
#> sigma_sq                   0.21559938 0.01261079
#> phi                        0.41649662 0.07881349
#> Tau.id.var(constant)       0.72096897 0.56776384
#> Tau.id.cov(trend,constant) 0.09849735 0.04566565
#> Tau.id.var(trend)          0.34142756 0.04725949
#> RML_coverage1              0.80000000 0.17777778
#> RML_coverage2              0.90000000 0.10000000

simulate_MB4(iterations = 10, beta = c(0,1,0,0), rho = 0.8, phi = 0.5, 
             tau2_ratio = 0.5, tau_corr = 0, m = 6, n = 8)

#> Warning: nlminb problem, convergence error code = 1
#>   message = iteration limit reached without convergence (10)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: nlminb problem, convergence error code = 1
#>   message = iteration limit reached without convergence (10)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: nlminb problem, convergence error code = 1
#>   message = iteration limit reached without convergence (10)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: nlminb problem, convergence error code = 1
#>   message = iteration limit reached without convergence (10)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: nlminb problem, convergence error code = 1
#>   message = iteration limit reached without convergence (10)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: nlminb problem, convergence error code = 1
#>   message = iteration limit reached without convergence (10)

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.

#>                                 mean         var
#> p_beta                     0.8414394  0.05702994
#> r_theta                    0.7704342  0.17994483
#> delta_AB                   1.0731913  0.21340143
#> nu                         8.7611127 13.56242662
#> kappa                      0.3627880  0.01432465
#> g_AB                       0.9758855  0.20415253
#> V_g_AB                     0.2355153  0.01438466
#> cnvg_warn                  0.6000000  0.26666667
#> sigma_sq                   0.3716201  0.06728972
#> phi                        0.5963314  0.05035935
#> Tau.id.var(constant)       0.3988140  0.22482836
#> Tau.id.cov(trend,constant) 0.1117323  0.05807754
#> Tau.id.var(trend)          0.4664484  0.04103693
#> RML_coverage1              1.0000000  0.00000000
#> RML_coverage2              1.0000000  0.00000000