Simulate Model MB4 from Pustejovsky, Hedges, & Shadish (2014)

Simulates data from a linear mixed effects model, then calculates REML effect size estimator as described in Pustejovsky, Hedges, & Shadish (2014).

simulate_MB4(
  iterations,
  beta,
  rho,
  phi,
  tau2_ratio,
  tau_corr,
  p_const,
  r_const,
  design,
  m,
  n,
  MB = TRUE
)

Arguments

iterations

number of independent iterations of the simulation

beta

vector of fixed effect parameters

rho

intra-class correlation parameter

phi

autocorrelation parameter

tau2_ratio

ratio of trend variance to intercept variance

tau_corr

correlation between case-specific trends and intercepts

p_const

vector of constants for calculating numerator of effect size

r_const

vector of constants for calculating denominator of effect size

design

design matrix. If not specified, it will be calculated based on m, n, and MB.

m

number of cases. Not used if design is specified.

n

number of measurement occasions. Not used if design is specified.

MB

If true, a multiple baseline design will be used; otherwise, an AB design will be used. Not used if design is specified.

Value

A matrix reporting the mean and variance of the effect size estimates and various associated statistics.

References

Pustejovsky, J. E., Hedges, L. V., & Shadish, W. R. (2014). Design-comparable effect sizes in multiple baseline designs: A general modeling framework. Journal of Educational and Behavioral Statistics, 39(4), 211-227. doi:10.3102/1076998614547577

Examples

simulate_MB4(iterations = 5, beta = c(0,1,0,0), rho = 0.8, phi = 0.5, 
             tau2_ratio = 0.5, tau_corr = 0, 
             p_const = c(0,1,0,7), r_const = c(1,0,1,0,0), 
             design = design_matrix(3, 16, treat_times=c(5,9,13), center = 12))
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: nlminb problem, convergence error code = 1
#>   message = iteration limit reached without convergence (10)
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: nlminb problem, convergence error code = 1
#>   message = iteration limit reached without convergence (10)
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: nlminb problem, convergence error code = 1
#>   message = iteration limit reached without convergence (10)
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: nlminb problem, convergence error code = 1
#>   message = iteration limit reached without convergence (10)
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#>                                  mean         var
#> p_beta                     0.90793548  0.60999262
#> r_theta                    0.72189553  0.29739446
#> delta_AB                   1.26284750  1.23256900
#> nu                         4.79829762 13.68077631
#> kappa                      0.56651109  0.04464539
#> g_AB                       1.00517231  0.88262381
#> V_g_AB                     1.26344315  0.47463295
#> cnvg_warn                  0.00000000  0.00000000
#> sigma_sq                   0.26579500  0.04140776
#> phi                        0.54831551  0.04809307
#> Tau.id.var(constant)       0.45610053  0.30820167
#> Tau.id.cov(trend,constant) 0.05594046  0.15775347
#> Tau.id.var(trend)          0.52751809  0.07826889
#> RML_coverage1              0.80000000  0.20000000
#> RML_coverage2              1.00000000  0.00000000
             
simulate_MB4(iterations = 5, beta = c(0,1,0,0), rho = 0.8, phi = 0.5, 
             tau2_ratio = 0.5, tau_corr = 0, m = 6, n = 8)
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#> Warning: nlminb problem, convergence error code = 1
#>   message = iteration limit reached without convergence (10)
#> Warning: 'g_REML()' is deprecated and may be removed in a later version of the package. Please use 'g_mlm()' instead.
#>                                 mean          var
#> p_beta                     1.0192984 0.0535019165
#> r_theta                    0.6543302 0.0259036261
#> delta_AB                   1.2629154 0.0525448248
#> nu                         7.6337552 0.5653216274
#> kappa                      0.3737259 0.0019086704
#> g_AB                       1.1326461 0.0399042768
#> V_g_AB                     0.2814867 0.0008022736
#> cnvg_warn                  0.0000000 0.0000000000
#> sigma_sq                   0.3761682 0.0491868935
#> phi                        0.6218321 0.0364734240
#> Tau.id.var(constant)       0.2781620 0.0691607705
#> Tau.id.cov(trend,constant) 0.1026425 0.0539387043
#> Tau.id.var(trend)          0.4106622 0.0415238206
#> RML_coverage1              1.0000000 0.0000000000
#> RML_coverage2              1.0000000 0.0000000000